Retrieve batter boxscore data for a single game played

This function allows a user to retrieve a boxscore of batter statistics for any game played in the PITCHf/x era (2008-current). The function takes a boxscore.xml url as it's only argument and returns boxscore data for both the home and away batters.

batter_boxscore(x)

Arguments

x
A boxscore.xml url for a given game from the MLBAM GameDay app data.

Examples

# \donttest{
url_base <- "http://gd2.mlb.com/components/game/mlb/"
url <- paste0(url_base, "year_2016/month_05/day_21/gid_2016_05_21_milmlb_nynmlb_1/boxscore.xml")
batter_boxscore(url)# }
#>        id           name name_display_first_last pos  bo ab po r a bb sac t sf
#> 1  542340         Villar         Jonathan Villar  SS 100  5  1 0 0  0   0 0  0
#> 2  571697        Gennett         Scooter Gennett  2B 200  4  2 0 3  1   0 0  0
#> 3  518960         Lucroy         Jonathan Lucroy   C 300  5  8 0 1  0   0 0  0
#> 4  474892         Carter            Chris Carter  1B 400  4 10 0 2  0   0 0  0
#> 5  543590    Nieuwenhuis        Kirk Nieuwenhuis  CF 500  4  0 0 0  0   0 0  0
#> 6  431094        Hill, A              Aaron Hill  3B 600  1  1 2 4  3   0 0  0
#> 7  502100        Presley            Alex Presley  LF 700  3  0 1 0  1   0 0  0
#> 8  570717      Flores, R            Ramon Flores  RF 800  3  2 1 0  0   0 0  1
#> 9  605200         Davies             Zach Davies   P 900  2  1 0 1  0   1 0  0
#> 10 430641          Boyer            Blaine Boyer   P 901  0  0 0 0  0   0 0  0
#> 11 448614         Torres           Carlos Torres   P 902  0  0 0 0  0   0 0  0
#> 12 592804      Thornburg         Tyler Thornburg   P 903  0  0 0 0  0   0 0  0
#> 13 595023          Walsh             Colin Walsh  PH 904  1  0 0 0  0   0 0  0
#> 14 518468         Blazek          Michael Blazek   P 905  0  0 0 0  0   0 0  0
#> 15 434158     Granderson       Curtis Granderson  RF 100  3  1 1 0  2   0 0  0
#> 16 431151         Wright            David Wright  3B 200  4  2 0 0  1   0 0  0
#> 17 624424       Conforto        Michael Conforto  LF 300  3  3 1 0  1   0 0  0
#> 18 493316       Cespedes         Yoenis Cespedes  CF 400  4  3 2 0  0   0 0  0
#> 19 435522         Walker             Neil Walker  2B 500  4  3 0 0  0   0 0  0
#> 20 452678     Cabrera, A        Asdrubal Cabrera  SS 600  4  0 0 1  0   0 0  0
#> 21 493472       Campbell           Eric Campbell  1B 700  4  2 1 1  0   0 0  0
#> 22 608700       Plawecki          Kevin Plawecki   C 800  3 13 0 0  1   0 0  0
#> 23 594798         deGrom            Jacob deGrom   P 900  1  0 0 0  0   0 0  0
#> 24 457477         De Aza        Alejandro De Aza  PH 901  0  0 0 0  1   0 0  0
#> 25 570663         Robles           Hansel Robles   P 902  0  0 0 0  0   0 0  0
#> 26 501571        Lagares            Juan Lagares  PH 903  1  0 0 0  0   0 0  0
#> 27 592665        Reed, A            Addison Reed   P 904  0  0 0 0  0   0 0  0
#> 28 544727        Familia          Jeurys Familia   P 905  0  0 0 0  0   0 0  0
#> 29 608703 Reynolds, Matt           Matt Reynolds  PH 906  0  0 0 0  0   1 0  0
#>    h e d hbp so hr rbi lob  fldg sb cs s_hr s_rbi s_h s_bb s_r s_so  avg  obp
#> 1  3 0 2   0  0  0   1   0  .964  0  0    1    13  44   23  18   43 .293 .391
#> 2  0 0 0   0  2  0   0   3  .978  0  0    4     5  22   12  18   28 .232 .318
#> 3  0 0 0   0  0  0   0   3  .989  0  0    6    16  42   14  23   29 .288 .354
#> 4  1 0 0   0  3  0   0   1  .992  0  0   13    31  37   15  25   56 .247 .306
#> 5  0 0 0   0  3  0   0   2 1.000  0  0    2    14  22   15  12   30 .256 .366
#> 6  1 0 0   0  0  0   0   0  .992  1  0    4    17  31   13  13   21 .252 .321
#> 7  1 0 0   0  1  0   0   2  .977  0  0    3    10  18    9   9   14 .243 .321
#> 8  1 0 0   0  2  1   3   3 1.000  0  0    1     9  16   12   7   24 .195 .292
#> 9  0 0 0   0  1  0   0   1 1.000  0  0    0     0   1    0   0    4 .091 .091
#> 10 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    0 .000 .000
#> 11 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    0 .000 .000
#> 12 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    0 .000 .000
#> 13 0 0 0   0  1  0   0   0  .960  0  0    0     2   4   14   4   22 .087 .311
#> 14 0 0 0   0  0  0   0   0  .667  0  0    0     0   0    0   0    0 .000 .000
#> 15 1 0 0   0  0  1   1   1 1.000  0  1    8    13  31   18  22   43 .197 .292
#> 16 1 0 0   0  2  0   1   0  .959  0  0    4     9  27   26  15   50 .223 .361
#> 17 1 0 1   0  0  0   0   1  .972  0  0    7    23  39   14  20   35 .283 .355
#> 18 2 0 0   0  1  1   2   1  .985  0  0   14    35  41   16  26   33 .297 .378
#> 19 1 0 0   0  2  0   0   2  .993  0  0   10    20  37   11  17   36 .255 .308
#> 20 1 0 0   0  1  0   1   1  .987  0  0    2    13  43    7  13   30 .279 .335
#> 21 2 0 0   0  0  0   0   1 1.000  0  0    0     1   9    7   6   14 .231 .375
#> 22 0 0 0   0  0  0   0   2 1.000  0  0    1     7  16   10   5   18 .203 .300
#> 23 0 0 0   0  0  0   0   0 1.000  0  0    0     2   1    1   1    3 .100 .182
#> 24 0 0 0   0  0  0   0   0 1.000  0  0    1     1   9    4   3   13 .205 .271
#> 25 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    0 .000 .000
#> 26 0 0 0   0  1  0   0   0  .973  0  0    0     2  16    3   3    8 .267 .302
#> 27 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    0 .000 .000
#> 28 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    0 .000 .000
#> 29 0 0 0   0  0  0   0   0 1.000  0  0    0     0   0    0   0    1 .000 .000
#>     slg   ops go ao gidp note team
#> 1  .400  .791  0  0    0 <NA>  158
#> 2  .400  .718  1  1    0 <NA>  158
#> 3  .479  .834  0  1    0 <NA>  158
#> 4  .587  .892  0  0    0 <NA>  158
#> 5  .419  .785  1  0    0 <NA>  158
#> 6  .390  .712  0  0    0 <NA>  158
#> 7  .378  .700  0  0    0 <NA>  158
#> 8  .268  .560  0  1    0 <NA>  158
#> 9  .091  .182  1  0    0 <NA>  158
#> 10 .000  .000  0  0    0 <NA>  158
#> 11 .000  .000  0  0    0 <NA>  158
#> 12 .000  .000  0  0    0 <NA>  158
#> 13 .109  .420  0  0    0   a-  158
#> 14 .000  .000  0  0    0 <NA>  158
#> 15 .420  .712  1  0    1 <NA>  121
#> 16 .388  .749  0  1    0 <NA>  121
#> 17 .536  .891  1  1    0 <NA>  121
#> 18 .659 1.038  1  0    0 <NA>  121
#> 19 .476  .784  1  0    0 <NA>  121
#> 20 .377  .712  1  0    0 <NA>  121
#> 21 .256  .631  2  0    0 <NA>  121
#> 22 .291  .591  3  0    0 <NA>  121
#> 23 .100  .282  1  0    0 <NA>  121
#> 24 .295  .566  0  0    0   a-  121
#> 25 .000  .000  0  0    0 <NA>  121
#> 26 .350  .652  0  0    0   b-  121
#> 27 .000  .000  0  0    0 <NA>  121
#> 28 .000  .000  0  0    0 <NA>  121
#> 29 .000  .000  1  0    0   c-  121