Directly download the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.

Directly download the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.

get_chadwick_lu()

Examples

get_chadwick_lu()
#> Rows: 409913 Columns: 40
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (10): key_person, key_uuid, key_retro, key_bbref, key_bbref_minors, name...
#> dbl (16): key_mlbam, key_fangraphs, birth_year, birth_month, birth_day, deat...
#> lgl (14): key_npb, key_sr_nfl, key_sr_nba, key_sr_nhl, key_findagrave, name_...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 409,913 × 40
#>    key_person key_uuid            key_mlbam key_retro key_bbref key_bbref_minors
#>    <chr>      <chr>                   <dbl> <chr>     <chr>     <chr>           
#>  1 663ecca1   663ecca1-6b4e-4a11…    439524 NA        NA        thoen-001eve    
#>  2 3e8b4a3e   3e8b4a3e-0866-4402…        NA NA        NA        ahara-000sco    
#>  3 8e7b24b0   8e7b24b0-6dec-40b2…        NA NA        NA        ahease001---    
#>  4 aaf20538   aaf20538-b53e-4d63…        NA NA        NA        aaberg001alf    
#>  5 34c63185   34c63185-85a0-4369…        NA NA        NA        aadlan001kir    
#>  6 a1f1855b   a1f1855b-c13c-4887…        NA NA        NA        aaker-000zac    
#>  7 2f06a161   2f06a161-2cac-4bba…        NA NA        NA        aakhus001zac    
#>  8 754d5b99   754d5b99-f754-45b9…    444058 NA        NA        aalber001bra    
#>  9 735c68ec   735c68ec-17d6-419f…        NA NA        NA        aamold000ryl    
#> 10 cbcc7093   cbcc7093-038c-4ddd…        NA NA        NA        aander000bri    
#> # … with 409,903 more rows, and 34 more variables: key_fangraphs <dbl>,
#> #   key_npb <lgl>, key_sr_nfl <lgl>, key_sr_nba <lgl>, key_sr_nhl <lgl>,
#> #   key_findagrave <lgl>, name_last <chr>, name_first <chr>, name_given <chr>,
#> #   name_suffix <lgl>, name_matrilineal <chr>, name_nick <chr>,
#> #   birth_year <dbl>, birth_month <dbl>, birth_day <dbl>, death_year <dbl>,
#> #   death_month <dbl>, death_day <dbl>, pro_played_first <dbl>,
#> #   pro_played_last <dbl>, mlb_played_first <dbl>, mlb_played_last <dbl>, …