Load cleaned NCAA baseball play-by-play data from the baseballr data repo

helper that loads multiple seasons from the data repo either into memory or writes it into a db using some forwarded arguments in the dots

Usage

load_ncaa_baseball_pbp(
  seasons = most_recent_ncaa_baseball_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)

Arguments

seasons

A vector of 4-digit years associated with given NCAA college baseball seasons. (Min: 2022)

...

Additional arguments passed to an underlying function that writes the season data into a database.

dbConnection

A DBIConnection object, as returned by

tablename

The name of the schedule data table within the database

Value

Returns a tibble

Examples

# \donttest{
  load_ncaa_baseball_pbp(seasons = 2022)
#> ── NCAA Play-by-Play Information from baseballr data repository ────────
#> ℹ Data updated: 2023-03-03 03:26:33 UTC
#> # A tibble: 2,342,771 × 12
#>    game_d…¹ locat…² atten…³ inning innin…⁴ score batting field…⁵ descr…⁶
#>    <chr>    <chr>   <lgl>   <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#>  1 06/26/2… Omaha,… NA      1      top     0-0   Oklaho… Ole Mi… Spiker…
#>  2 06/26/2… Omaha,… NA      1      top     0-0   Oklaho… Ole Mi… Graham…
#>  3 06/26/2… Omaha,… NA      1      top     0-0   Oklaho… Ole Mi… Robert…
#>  4 06/26/2… Omaha,… NA      1      top     0-0   Oklaho… Ole Mi… H: 1   
#>  5 06/26/2… Omaha,… NA      1      bot     0-0   Ole Mi… Oklaho… Bench …
#>  6 06/26/2… Omaha,… NA      1      bot     0-0   Ole Mi… Oklaho… Gonzal…
#>  7 06/26/2… Omaha,… NA      1      bot     0-0   Ole Mi… Oklaho… Elko s…
#>  8 06/26/2… Omaha,… NA      2      top     0-0   Oklaho… Ole Mi… Tredaw…
#>  9 06/26/2… Omaha,… NA      2      top     0-0   Oklaho… Ole Mi… Crooks…
#> 10 06/26/2… Omaha,… NA      2      top     0-0   Oklaho… Ole Mi… Clark …
#> # … with 2,342,761 more rows, 3 more variables: year <int>,
#> #   game_pbp_url <chr>, game_pbp_id <int>, and abbreviated variable
#> #   names ¹​game_date, ²​location, ³​attendance, ⁴​inning_top_bot,
#> #   ⁵​fielding, ⁶​description
# }