Retrieve latest draft information by year

Retrieve latest draft information by year

Usage

mlb_draft_latest(year)

Arguments

year

The year for which to return data

Value

Returns a data frame with the latest draft information for the year requested:

col_name	types

bis_player_id	integer

pick_round	character

pick_number	integer

round_pick_number	integer

rank	integer

pick_value	character

signing_bonus	character

home_city	character

home_state	character

home_country	character

scouting_report	character

school_name	character

school_school_class	character

school_country	character

school_state	character

blurb	character

headshot_link	character

person_id	integer

person_full_name	character

person_link	character

person_first_name	character

person_last_name	character

person_primary_number	character

person_birth_date	character

person_current_age	integer

person_birth_city	character

person_birth_state_province	character

person_birth_country	character

person_height	character

person_weight	integer

person_active	logical

person_primary_position_code	character

person_primary_position_name	character

person_primary_position_type	character

person_primary_position_abbreviation	character

person_use_name	character

person_middle_name	character

person_boxscore_name	character

person_gender	character

person_is_player	logical

person_is_verified	logical

person_draft_year	integer

person_bat_side_code	character

person_bat_side_description	character

person_pitch_hand_code	character

person_pitch_hand_description	character

person_name_first_last	character

person_name_slug	character

person_first_last_name	character

person_last_first_name	character

person_last_init_name	character

person_init_last_name	character

person_full_fml_name	character

person_full_lfm_name	character

person_strike_zone_top	numeric

person_strike_zone_bottom	numeric

team_id	integer

team_name	character

team_link	character

team_season	integer

team_venue_id	integer

team_venue_name	character

team_venue_link	character

team_spring_venue_id	integer

team_spring_venue_link	character

team_team_code	character

team_file_code	character

team_abbreviation	character

team_team_name	character

team_location_name	character

team_first_year_of_play	character

team_league_id	integer

team_league_name	character

team_league_link	character

team_division_id	integer

team_division_name	character

team_division_link	character

team_sport_id	integer

team_sport_link	character

team_sport_name	character

team_short_name	character

team_franchise_name	character

team_club_name	character

team_spring_league_id	integer

team_spring_league_name	character

team_spring_league_link	character

team_spring_league_abbreviation	character

team_all_star_status	character

team_active	logical

draft_type_code	character

draft_type_description	character

is_drafted	logical

is_pass	logical

year	character

Examples

# \donttest{
 try(mlb_draft_latest(year = 2020))
#>   bis_player_id pick_round pick_number round_pick_number rank
#> 1        790203       CB-B          66                 6   51
#>   pick_value signing_bonus            home_city home_state home_country
#> 1    1003300       1196500 North Richland Hills      Texas          USA
#>             scouting_report school_name school_school_class
#> 1 https://atmlb.com/3fh9KkE  Texas Tech              4YR JR
#>   school_country school_state
#> 1            USA           TX
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          blurb
#> 1 Beeter has had as many elbow operations as quality starts (two) in college, yet he pitched so well this spring that some scouts think he belongs in the first round. He blew out his elbow during fall practice before what would have been his first season at Texas Tech, having Tommy John surgery in December 2017 and an arthroscopic procedure seven months later. He led the Red Raiders' 2019 College World Series semifinalists with eight saves as a redshirt freshman, then really opened eyes with nearly twice as many strikeouts (33) as baserunners (18) in 21 innings in 2020. Beeter added rather than lost velocity as he worked longer stints this year, sitting at 93-96 mph and peaking at 98 with his fastball. He has a pair of power breaking balls, a wipeout curveball in the low 80s and a slider in the mid-80s that some evaluators believe is even better. He even flashes at least a solid changeup, though he doesn't turn to his fourth option very often. Employing an overhand delivery that creates deception and steep downhill plane, Beeter tunnels his pitches extremely well. Scouts can't figure out how he went from averaging 8.7 walks per nine innings as a reliever to 1.7 as a starter without any significant mechanical changes. His lack of track record is worrisome but it's hard to argue with the quality stuff and strikes he unleashed this spring.
#>                                                                                                                                            headshot_link
#> 1 https://img.mlbstatic.com/mlb-photos/image/upload/d_people:generic:headshot:silo:current.png/w_120,q_auto:best/v1/people/690925/headshot/draft/current
#>   person_id person_full_name           person_link person_first_name
#> 1    690925   Clayton Beeter /api/v1/people/690925           Clayton
#>   person_last_name person_primary_number person_birth_date
#> 1           Beeter                    93        1998-10-09
#>   person_current_age person_birth_city person_birth_state_province
#> 1                 23        Fort Worth                          TX
#>   person_birth_country person_height person_weight person_active
#> 1                  USA         6' 2"           220          TRUE
#>   person_primary_position_code person_primary_position_name
#> 1                            1                      Pitcher
#>   person_primary_position_type person_primary_position_abbreviation
#> 1                      Pitcher                                    P
#>   person_use_name person_middle_name person_boxscore_name person_gender
#> 1         Clayton             Howard               Beeter             M
#>   person_is_player person_is_verified person_draft_year
#> 1             TRUE               TRUE              2020
#>   person_bat_side_code person_bat_side_description
#> 1                    R                       Right
#>   person_pitch_hand_code person_pitch_hand_description
#> 1                      R                         Right
#>   person_name_first_last      person_name_slug person_first_last_name
#> 1         Clayton Beeter clayton-beeter-690925         Clayton Beeter
#>   person_last_first_name person_last_init_name person_init_last_name
#> 1        Beeter, Clayton             Beeter, C              C Beeter
#>    person_full_fml_name   person_full_lfm_name person_strike_zone_top
#> 1 Clayton Howard Beeter Beeter, Clayton Howard                  3.467
#>   person_strike_zone_bottom team_spring_league_id
#> 1                     1.589                   114
#>   team_spring_league_name team_spring_league_link
#> 1           Cactus League      /api/v1/league/114
#>   team_spring_league_abbreviation team_all_star_status team_id
#> 1                              CL                    N     119
#>             team_name         team_link team_season team_venue_id
#> 1 Los Angeles Dodgers /api/v1/teams/119        2022            22
#>   team_venue_name   team_venue_link team_spring_venue_id
#> 1  Dodger Stadium /api/v1/venues/22                 3809
#>   team_spring_venue_link team_team_code team_file_code
#> 1    /api/v1/venues/3809            lan             la
#>   team_abbreviation team_team_name team_location_name
#> 1               LAD        Dodgers        Los Angeles
#>   team_first_year_of_play team_league_id team_league_name
#> 1                    1884            104  National League
#>     team_league_link team_division_id   team_division_name
#> 1 /api/v1/league/104              203 National League West
#>      team_division_link team_sport_id  team_sport_link
#> 1 /api/v1/divisions/203             1 /api/v1/sports/1
#>         team_sport_name team_short_name team_franchise_name
#> 1 Major League Baseball      LA Dodgers         Los Angeles
#>   team_club_name team_active draft_type_code draft_type_description
#> 1        Dodgers        TRUE              JR     June Amateur Draft
#>   is_drafted is_pass year
#> 1       TRUE   FALSE 2020
# }