Get Play-By-Play Data for NCAA Baseball Games

Get Play-By-Play Data for NCAA Baseball Games

Usage

ncaa_baseball_pbp(game_info_url)

Arguments

game_info_url

The url for the game's play-by-play data. This can be found using the get_ncaa_schedule_info function.

Value

A data frame with play-by-play data for an individual game.

col_name	types

date	character

location	character

attendance	logical

inning	character

inning_top_bot	character

score	character

batting	character

fielding	character

description	character

Examples

# \donttest{
  x <- ncaa_schedule_info(736, 2021)$game_info_url[2]
  ncaa_baseball_pbp(game_info_url = x)
#>          date         location attendance inning inning_top_bot score
#> 1  02/22/2021 Nashville, Tenn.         NA      1            top   0-0
#> 2  02/22/2021 Nashville, Tenn.         NA      1            top   0-0
#> 3  02/22/2021 Nashville, Tenn.         NA      1            top   0-0
#> 4  02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 5  02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 6  02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 7  02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 8  02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 9  02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 10 02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 11 02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 12 02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 13 02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 14 02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 15 02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 16 02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 17 02/22/2021 Nashville, Tenn.         NA      3            top   0-0
#> 18 02/22/2021 Nashville, Tenn.         NA      3            top   0-0
#> 19 02/22/2021 Nashville, Tenn.         NA      3            top   0-0
#> 20 02/22/2021 Nashville, Tenn.         NA      3            bot   0-0
#> 21 02/22/2021 Nashville, Tenn.         NA      3            bot   0-0
#> 22 02/22/2021 Nashville, Tenn.         NA      3            bot   0-0
#> 23 02/22/2021 Nashville, Tenn.         NA      3            bot   0-0
#> 24 02/22/2021 Nashville, Tenn.         NA      3            bot   0-0
#> 25 02/22/2021 Nashville, Tenn.         NA      3            bot   0-0
#> 26 02/22/2021 Nashville, Tenn.         NA      4            top   0-0
#> 27 02/22/2021 Nashville, Tenn.         NA      4            top   0-0
#> 28 02/22/2021 Nashville, Tenn.         NA      4            top   0-0
#> 29 02/22/2021 Nashville, Tenn.         NA      4            top   0-0
#> 30 02/22/2021 Nashville, Tenn.         NA      4            bot   0-0
#> 31 02/22/2021 Nashville, Tenn.         NA      4            bot   0-0
#> 32 02/22/2021 Nashville, Tenn.         NA      4            bot   0-0
#> 33 02/22/2021 Nashville, Tenn.         NA      4            bot   0-0
#> 34 02/22/2021 Nashville, Tenn.         NA      4            bot   0-0
#> 35 02/22/2021 Nashville, Tenn.         NA      5            top   0-0
#> 36 02/22/2021 Nashville, Tenn.         NA      5            top   0-0
#> 37 02/22/2021 Nashville, Tenn.         NA      5            top   0-0
#> 38 02/22/2021 Nashville, Tenn.         NA      5            top   0-0
#> 39 02/22/2021 Nashville, Tenn.         NA      5            top   0-1
#> 40 02/22/2021 Nashville, Tenn.         NA      5            bot   0-0
#> 41 02/22/2021 Nashville, Tenn.         NA      5            bot   0-0
#> 42 02/22/2021 Nashville, Tenn.         NA      5            bot   0-0
#> 43 02/22/2021 Nashville, Tenn.         NA      5            bot   0-1
#> 44 02/22/2021 Nashville, Tenn.         NA      5            bot   0-1
#> 45 02/22/2021 Nashville, Tenn.         NA      5            bot   0-1
#> 46 02/22/2021 Nashville, Tenn.         NA      5            top   0-1
#> 47 02/22/2021 Nashville, Tenn.         NA      6            top   0-1
#> 48 02/22/2021 Nashville, Tenn.         NA      6            top   0-1
#> 49 02/22/2021 Nashville, Tenn.         NA      6            top   0-1
#> 50 02/22/2021 Nashville, Tenn.         NA      6            top   0-1
#> 51 02/22/2021 Nashville, Tenn.         NA      6            top   0-1
#> 52 02/22/2021 Nashville, Tenn.         NA      6            top   0-1
#> 53 02/22/2021 Nashville, Tenn.         NA      6            bot   0-1
#> 54 02/22/2021 Nashville, Tenn.         NA      6            bot   0-1
#> 55 02/22/2021 Nashville, Tenn.         NA      6            bot   0-1
#> 56 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 57 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 58 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 59 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 60 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 61 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 62 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 63 02/22/2021 Nashville, Tenn.         NA      7            bot   0-1
#> 64 02/22/2021 Nashville, Tenn.         NA      7            bot   0-1
#> 65 02/22/2021 Nashville, Tenn.         NA      7            bot   0-1
#> 66 02/22/2021 Nashville, Tenn.         NA      7            bot   0-1
#> 67 02/22/2021 Nashville, Tenn.         NA      7            top   0-1
#> 68 02/22/2021 Nashville, Tenn.         NA      8            top   0-1
#> 69 02/22/2021 Nashville, Tenn.         NA      8            top   0-1
#> 70 02/22/2021 Nashville, Tenn.         NA      8            top   0-1
#> 71 02/22/2021 Nashville, Tenn.         NA      8            top   0-1
#> 72 02/22/2021 Nashville, Tenn.         NA      8            top   0-1
#> 73 02/22/2021 Nashville, Tenn.         NA      8            bot   0-1
#> 74 02/22/2021 Nashville, Tenn.         NA      8            bot   0-1
#> 75 02/22/2021 Nashville, Tenn.         NA      8            bot   0-1
#> 76 02/22/2021 Nashville, Tenn.         NA      8            bot   0-1
#> 77 02/22/2021 Nashville, Tenn.         NA      8            bot   0-1
#> 78 02/22/2021 Nashville, Tenn.         NA      9            top   0-1
#> 79 02/22/2021 Nashville, Tenn.         NA      9            top   0-1
#> 80 02/22/2021 Nashville, Tenn.         NA      9            top   0-1
#> 81 02/22/2021 Nashville, Tenn.         NA      9            top   0-1
#> 82 02/22/2021 Nashville, Tenn.         NA      9            top   0-1
#> 83 02/22/2021 Nashville, Tenn.         NA      9            top   0-1
#>       batting   fielding
#> 1  Wright St. Vanderbilt
#> 2  Wright St. Vanderbilt
#> 3  Wright St. Vanderbilt
#> 4  Vanderbilt Wright St.
#> 5  Vanderbilt Wright St.
#> 6  Vanderbilt Wright St.
#> 7  Vanderbilt Wright St.
#> 8  Vanderbilt Wright St.
#> 9  Wright St. Vanderbilt
#> 10 Wright St. Vanderbilt
#> 11 Wright St. Vanderbilt
#> 12 Vanderbilt Wright St.
#> 13 Vanderbilt Wright St.
#> 14 Vanderbilt Wright St.
#> 15 Vanderbilt Wright St.
#> 16 Vanderbilt Wright St.
#> 17 Wright St. Vanderbilt
#> 18 Wright St. Vanderbilt
#> 19 Wright St. Vanderbilt
#> 20 Vanderbilt Wright St.
#> 21 Vanderbilt Wright St.
#> 22 Vanderbilt Wright St.
#> 23 Vanderbilt Wright St.
#> 24 Vanderbilt Wright St.
#> 25 Vanderbilt Wright St.
#> 26 Wright St. Vanderbilt
#> 27 Wright St. Vanderbilt
#> 28 Wright St. Vanderbilt
#> 29 Wright St. Vanderbilt
#> 30 Vanderbilt Wright St.
#> 31 Vanderbilt Wright St.
#> 32 Vanderbilt Wright St.
#> 33 Vanderbilt Wright St.
#> 34 Vanderbilt Wright St.
#> 35 Wright St. Vanderbilt
#> 36 Wright St. Vanderbilt
#> 37 Wright St. Vanderbilt
#> 38 Wright St. Vanderbilt
#> 39 Wright St. Vanderbilt
#> 40 Vanderbilt Wright St.
#> 41 Vanderbilt Wright St.
#> 42 Vanderbilt Wright St.
#> 43 Vanderbilt Wright St.
#> 44 Vanderbilt Wright St.
#> 45 Vanderbilt Wright St.
#> 46 Wright St. Vanderbilt
#> 47 Wright St. Vanderbilt
#> 48 Wright St. Vanderbilt
#> 49 Wright St. Vanderbilt
#> 50 Wright St. Vanderbilt
#> 51 Wright St. Vanderbilt
#> 52 Wright St. Vanderbilt
#> 53 Vanderbilt Wright St.
#> 54 Vanderbilt Wright St.
#> 55 Vanderbilt Wright St.
#> 56 Wright St. Vanderbilt
#> 57 Wright St. Vanderbilt
#> 58 Wright St. Vanderbilt
#> 59 Wright St. Vanderbilt
#> 60 Wright St. Vanderbilt
#> 61 Wright St. Vanderbilt
#> 62 Wright St. Vanderbilt
#> 63 Vanderbilt Wright St.
#> 64 Vanderbilt Wright St.
#> 65 Vanderbilt Wright St.
#> 66 Vanderbilt Wright St.
#> 67 Wright St. Vanderbilt
#> 68 Wright St. Vanderbilt
#> 69 Wright St. Vanderbilt
#> 70 Wright St. Vanderbilt
#> 71 Wright St. Vanderbilt
#> 72 Wright St. Vanderbilt
#> 73 Vanderbilt Wright St.
#> 74 Vanderbilt Wright St.
#> 75 Vanderbilt Wright St.
#> 76 Vanderbilt Wright St.
#> 77 Vanderbilt Wright St.
#> 78 Wright St. Vanderbilt
#> 79 Wright St. Vanderbilt
#> 80 Wright St. Vanderbilt
#> 81 Wright St. Vanderbilt
#> 82 Wright St. Vanderbilt
#> 83 Wright St. Vanderbilt
#>                                                                                                          description
#> 1                                                                           Dues, D. struck out looking (2-2 BKBSK).
#> 2                                                                        Hamilton, Q. struck out looking (1-2 SBKK).
#> 3                                                                          Black, T. struck out looking (2-2 BBSKK).
#> 4                                                                               Bulger grounded out to 3b (2-1 FBB).
#> 5                                                                           Young tripled to center field (2-1 SBB).
#> 6                                                                            Thomas struck out swinging (1-2 SBKFS).
#> 7                                                                             Noland struck out swinging (1-2 BFSS).
#> 8                                                                                                        H: 1 LOB: 1
#> 9                                                                         Harris, Z. struck out looking (2-2 BKBSK).
#> 10                                                                               Anglin, G. popped up to 1b (0-1 K).
#> 11                                                                         Alders, A. struck out looking (1-2 SKBK).
#> 12                                                                                  Kolwyck flied out to rf (0-1 F).
#> 13                                                                        Romero Jr. struck out swinging (1-2 KBKS).
#> 14                                                                                       Keegan walked (3-2 KBBBFB).
#> 15                                                                       Bradfield Jr struck out swinging (0-2 KKS).
#> 16                                                                                                            LOB: 1
#> 17                                                                       Piotto, K. struck out swinging (2-2 FBBSS).
#> 18                                                                            Greenwell, J grounded out to 2b (0-0).
#> 19                                                                          Luikart, J. flied out to 2b (1-2 BKFFF).
#> 20                                                                             Gonzalez singled to left field (0-0).
#> 21                                                                                  Bulger flied out to 2b (1-1 KB).
#> 22                                                                        Gonzalez out at first p to 1b, picked off.
#> 23                                                                          Young doubled to left center (2-2 BKSB).
#> 24                                                                                  Thomas flied out to cf (1-1 BK).
#> 25                                                                                                       H: 2 LOB: 1
#> 26                                                                           Dues, D. struck out swinging (0-2 KKS).
#> 27                                                            Hamilton, Q. grounded out to 1b unassisted (2-2 BFSB).
#> 28                                                                                Black, T. hit by pitch (2-2 KBKB).
#> 29                                                                 Black, T. out at second c to ss, caught stealing.
#> 30                                                                                   Noland popped up to ss (1-0 B).
#> 31                                                                         Kolwyck singled to center field (0-2 KK).
#> 32                                                         Romero Jr. walked (3-1 BBBSB) Kolwyck advanced to second.
#> 33     Keegan reached on a fielders choice (2-1 BBK) Romero Jr. out on the play Kolwyck out on double play 3b to 2b.
#> 34                                                                                                       H: 1 LOB: 1
#> 35                                                                      Harris, Z. struck out swinging (3-2 KBFBBS).
#> 36                                                                             Anglin, G. flied out to cf (1-2 BKS).
#> 37                                                                  Alders, A. singled through the right side (0-0).
#> 38                                                                        Piotto, K. grounded out to ss (2-2 KBKFB).
#> 39                                                                                                       H: 1 LOB: 1
#> 40                                                                   Bradfield Jr singled to first base, bunt (0-0).
#> 41     Gonzalez singled through the left side, advanced to second on the throw (0-0) Bradfield Jr advanced to third.
#> 42                                                                                   Cline, A. to p for Brehmer, B..
#> 43                         Bulger out at first 1b to p, RBI (2-0 BB) Gonzalez advanced to third Bradfield Jr scored.
#> 44                                                                              Young struck out swinging (0-2 KKS).
#> 45                                                                            Thomas struck out swinging (1-2 KBFS).
#> 46                                                                                                  R: 1 H: 2 LOB: 1
#> 47                                                                                          Schultz to p for Leiter.
#> 48                                                                  Greenwell, J doubled down the lf line (1-2 KSB).
#> 49                                              Luikart, J. grounded out to 2b (0-0) Greenwell, J advanced to third.
#> 50                                                                         Dues, D. struck out looking (3-2 FFBBBK).
#> 51                                                                      Hamilton, Q. struck out swinging (1-2 KKBS).
#> 52                                                                                                       H: 1 LOB: 1
#> 53                                                                            Noland struck out swinging (1-2 FKBS).
#> 54                                                                                 Kolwyck flied out to cf (1-1 BF).
#> 55                                                                       Romero Jr. struck out looking (3-2 BBBKFK).
#> 56                                                                                       Maldonado to p for Schultz.
#> 57                                                                                  Black, T. walked (3-2 KBFBBFFB).
#> 58                                                      Harris, Z. walked (3-2 BBBKFB) Black, T. advanced to second.
#> 59                                                                       Anglin, G. struck out looking (3-2 SBFBBK).
#> 60 Alders, A. reached on a fielders choice (3-1 BKBB) Harris, Z. out at second 2b to ss Black, T. advanced to third.
#> 61                                                                            Piotto, K. grounded out to 2b (1-0 B).
#> 62                                                                                                            LOB: 2
#> 63                                                                            Keegan struck out swinging (1-2 BSKS).
#> 64                                                                         Bradfield Jr flied out to lf (3-2 BBKBK).
#> 65                                                                     Gonzalez singled through the left side (0-0).
#> 66                                                                                 Bulger fouled out to 1b (0-2 KS).
#> 67                                                                                                       H: 1 LOB: 1
#> 68                                                                                        Murphy to p for Maldonado.
#> 69                                                                                          Hewett to lf for Bulger.
#> 70                                                                     Greenwell, J struck out swinging (2-2 KBSBS).
#> 71                                                                      Luikart, J. struck out swinging (2-2 KBBKS).
#> 72                                                                           Dues, D. struck out looking (1-2 BFSK).
#> 73                                                                                  Young popped up to ss (2-1 BBF).
#> 74                                                                            Thomas singled to third base (0-2 KS).
#> 75                                   Noland reached on a fielders choice (1-0 B) Thomas out at second ss unassisted.
#> 76                                                                             Kolwyck struck out looking (0-2 KSK).
#> 77                                                                                                       H: 1 LOB: 1
#> 78                                                                                            Smith to p for Murphy.
#> 79                                                                             Hamilton, Q. flied out to cf (0-1 K).
#> 80                                                                                    Black, T. walked (3-2 BFFBBB).
#> 81                                                                     Harris, Z. struck out swinging (3-2 BBBKKFS).
#> 82                                                                         Anglin, G. struck out swinging (0-2 KSS).
#> 83                                                                                                            LOB: 1
# }