Retrieve pitcher boxscore data for a single game played

This function allows a user to retrieve a boxscore of pitcher statistics for any game played in the PITCHf/x era (2008-current). The function takes a boxscore.xml url as it's only argument and returns boxscore data for both the home and away batters.

pitcher_boxscore(x)

Arguments

x
A boxscore.xml url for a given game from the MLBAM GameDay app data.

Examples

# \donttest{
# pitchers
url_base <- "http://gd2.mlb.com/components/game/mlb/"
url <- paste0(url_base,
  "year_2016/month_05/day_21/gid_2016_05_21_milmlb_nynmlb_1/boxscore.xml")
pitcher_boxscore(url)# }
#>       id      name name_display_first_last pos out bf er r h so hr bb  np  s w
#> 1 605200    Davies             Zach Davies   P  16 22  4 4 5  5  2  2  86 51 1
#> 2 430641     Boyer            Blaine Boyer   P   2  4  0 0 2  0  0  0   8  7 1
#> 3 448614    Torres           Carlos Torres   P   3  4  0 0 0  1  0  2  21 11 0
#> 4 592804 Thornburg         Tyler Thornburg   P   3  3  0 0 0  1  0  0  14  8 2
#> 5 518468    Blazek          Michael Blazek   P   1  5  1 1 2  0  0  2  23 10 1
#> 6 594798    deGrom            Jacob deGrom   P  15 23  4 4 5  7  1  3 100 65 3
#> 7 570663    Robles           Hansel Robles   P   6  7  0 0 0  3  0  1  35 22 0
#> 8 592665   Reed, A            Addison Reed   P   3  5  0 0 1  2  0  1  28 18 0
#> 9 544727   Familia          Jeurys Familia   P   3  4  0 0 1  1  0  0  18 13 1
#>   l sv bs hld s_ip s_h s_r s_er s_bb s_so  era     note team
#> 1 3  0  0   0 36.0  41  24   23   15   25 5.75     <NA>   18
#> 2 0  0  1   0 21.1  22   4    4    5    7 1.69     <NA>   18
#> 3 1  0  0   2 22.1  22   9    9   14   21 3.63     <NA>   18
#> 4 1  1  2   8 18.2  13   8    8    7   29 3.86     <NA>   18
#> 5 1  0  1   8 21.1  23   6    6   14   18 2.53 (L, 1-1)   18
#> 6 1  0  0   0 41.0  40  17   14   11   30 3.07     <NA>  121
#> 7 2  0  0   5 19.2  14   4    3    5   23 1.37     <NA>  121
#> 8 0  1  1   9 20.0  16   5    5    5   27 2.25     <NA>  121
#> 9 0 14  0   0 20.1  18   5    5    4   17 2.21 (W, 1-0)  121