Look up Baseball Player IDs

This function allows you to query the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.

playerid_lookup(last_name = NULL, first_name = NULL)

Arguments

last_name
A text string used to return results for players with that string in their last name.

first_name
A text string used to return results for players with that string in their first name.

Examples

# \donttest{
playerid_lookup("Garcia", "Karim")
#> Be patient, this may take a few seconds...
#> Data courtesy of the Chadwick Bureau Register (https://github.com/chadwickbureau/register)
#> Rows: 409913 Columns: 40
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (10): key_person, key_uuid, key_retro, key_bbref, key_bbref_minors, name...
#> dbl (16): key_mlbam, key_fangraphs, birth_year, birth_month, birth_day, deat...
#> lgl (14): key_npb, key_sr_nfl, key_sr_nba, key_sr_nhl, key_findagrave, name_...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> [1] "Be patient, this may take a few seconds..."
#> [1] "Data courtesy of the Chadwick Bureau Register (https://github.com/chadwickbureau/register)"
#> # A tibble: 1 × 11
#>   first_name last_name given_name    name_suffix nick_name birth_year
#>   <chr>      <chr>     <chr>         <chr>       <chr>          <dbl>
#> 1 Karim      Garcia    Gustavo Karim NA          NA              1975
#> # … with 5 more variables: mlb_played_first <dbl>, mlbam_id <dbl>,
#> #   retrosheet_id <chr>, bbref_id <chr>, fangraphs_id <dbl>
# }