Scrape Team Results

This function allows you to scrape schedule and results for a major league team from Baseball-Reference.com

team_results_bref(Tm, year)

Arguments

Tm
The abbreviation used by Baseball-Reference.com for the team whose results you want to scrape.

year
Season for which you want to scrape the park factors.

Examples

team_results_bref("NYM", 2015)
#> # A tibble: 162 × 22
#>       Gm Date      Tm    H_A   Opp   Result     R    RA Inn   Record  Rank GB   
#>    <dbl> <chr>     <chr> <chr> <chr> <chr>  <dbl> <dbl> <chr> <chr>  <dbl> <chr>
#>  1     1 Monday, … NYM   A     WSN   W          3     1 ""    1-0        1 Tied 
#>  2     2 Wednesda… NYM   A     WSN   L          1     2 ""    1-1        2 1.5  
#>  3     3 Thursday… NYM   A     WSN   W          6     3 ""    2-1        2 1.0  
#>  4     4 Friday, … NYM   A     ATL   L          3     5 ""    2-2        2 2.0  
#>  5     5 Saturday… NYM   A     ATL   L          3     5 ""    2-3        3 3.0  
#>  6     6 Sunday, … NYM   A     ATL   W          4     3 ""    3-3        2 2.0  
#>  7     7 Monday, … NYM   H     PHI   W          2     0 ""    4-3        2 2.0  
#>  8     8 Tuesday,… NYM   H     PHI   W          6     5 ""    5-3        2 1.0  
#>  9     9 Wednesda… NYM   H     PHI   W          6     1 ""    6-3        1 Tied 
#> 10    10 Thursday… NYM   H     MIA   W          7     5 ""    7-3        1 up 0…
#> # … with 152 more rows, and 10 more variables: Win <chr>, Loss <chr>,
#> #   Save <chr>, Time <chr>, D/N <chr>, Attendance <dbl>, cLI <dbl>,
#> #   Streak <dbl>, Orig_Scheduled <chr>, Year <dbl>
team_results_bref(Tm="TBR", year=2008)
#> # A tibble: 162 × 22
#>       Gm Date      Tm    H_A   Opp   Result     R    RA Inn   Record  Rank GB   
#>    <dbl> <chr>     <chr> <chr> <chr> <chr>  <dbl> <dbl> <chr> <chr>  <dbl> <chr>
#>  1     1 Monday, … TBR   A     BAL   W          6     2 ""    1-0        1 up 0…
#>  2     2 Wednesda… TBR   A     BAL   L          6     9 ""    1-1        2 1.0  
#>  3     3 Friday, … TBR   A     NYY   W         13     4 ""    2-1        1 Tied 
#>  4     4 Saturday… TBR   A     NYY   W          6     3 ""    3-1        1 Tied 
#>  5     5 Sunday, … TBR   A     NYY   L          0     2 ""    3-2        3 1.0  
#>  6     6 Monday, … TBR   A     NYY   L          1     6 ""    3-3        4 2.0  
#>  7     7 Tuesday,… TBR   H     SEA   L          5     6 ""    3-4        5 3.0  
#>  8     8 Wednesda… TBR   H     SEA   L          1     7 ""    3-5        5 3.5  
#>  9     9 Thursday… TBR   H     SEA   W          7     0 ""    4-5        4 2.0  
#> 10    10 Friday, … TBR   H     BAL   W         10     5 ""    5-5        3 1.0  
#> # … with 152 more rows, and 10 more variables: Win <chr>, Loss <chr>,
#> #   Save <chr>, Time <chr>, D/N <chr>, Attendance <dbl>, cLI <dbl>,
#> #   Streak <dbl>, Orig_Scheduled <chr>, Year <dbl>