Download the Chadwick Bureau's public register of baseball players

Download the Chadwick Bureau's public register of baseball players

Usage

chadwick_player_lu()

get_chadwick_lu()

Value

A data frame of baseball players and the various IDs associated with them in different systems of record. |col_name |types | |:-----------------|:---------| |key_person |character | |key_uuid |character | |key_mlbam |integer | |key_retro |character | |key_bbref |character | |key_bbref_minors |character | |key_fangraphs |integer | |key_npb |integer | |key_sr_nfl |character | |key_sr_nba |character | |key_sr_nhl |character | |key_findagrave |integer | |name_last |character | |name_first |character | |name_given |character | |name_suffix |character | |name_matrilineal |character | |name_nick |character | |birth_year |integer | |birth_month |integer | |birth_day |integer | |death_year |integer | |death_month |integer | |death_day |integer | |pro_played_first |integer | |pro_played_last |integer | |mlb_played_first |integer | |mlb_played_last |integer | |col_played_first |integer | |col_played_last |integer | |pro_managed_first |integer | |pro_managed_last |integer | |mlb_managed_first |integer | |mlb_managed_last |integer | |col_managed_first |integer | |col_managed_last |integer | |pro_umpired_first |integer | |pro_umpired_last |integer | |mlb_umpired_first |integer | |mlb_umpired_last |integer | A data frame of baseball players and the various IDs associated with them in different systems of record.

Examples

# \donttest{
  chadwick_player_lu()
#>         key_person                             key_uuid key_mlbam
#>      1:   663ecca1 663ecca1-6b4e-4a11-9494-1caa6d6b2d13    439524
#>      2:   3e8b4a3e 3e8b4a3e-0866-4402-ad11-1905dd9022dd        NA
#>      3:   8e7b24b0 8e7b24b0-6dec-40b2-b895-8685e02e3d3b        NA
#>      4:   aaf20538 aaf20538-b53e-4d63-a3e0-e9e81fbd590c        NA
#>      5:   34c63185 34c63185-85a0-4369-ac23-51f54392566a        NA
#>     ---                                                          
#> 425297:   933c847e 933c847e-f306-4c0b-b888-e8ec93ff44f7        NA
#> 425298:   3135b811 3135b811-ec4b-40d8-af40-ca0adc33a219        NA
#> 425299:   b0678f4c b0678f4c-cc46-4396-b2d3-980b93f2575c    666557
#> 425300:   27f57a23 27f57a23-2441-4bae-8b26-6e792a62d868        NA
#> 425301:   ff617bd4 ff617bd4-3933-43e3-95cc-bef5ea74ed41        NA
#>         key_retro key_bbref key_bbref_minors key_fangraphs key_npb
#>      1:                         thoen-001eve            NA      NA
#>      2:                         ahara-000sco            NA      NA
#>      3:                         ahease001---            NA      NA
#>      4:                         aaberg001alf            NA      NA
#>      5:                         aadlan001kir            NA      NA
#>     ---                                                           
#> 425297:                         zywici000jos            NA      NA
#> 425298:                         zywick000bob            NA      NA
#> 425299:                         zywick000bra            NA      NA
#> 425300:                         zyzda-000nat            NA      NA
#> 425301:                                                 NA      NA
#>         key_sr_nfl key_sr_nba key_sr_nhl key_findagrave name_last
#>      1:                                              NA   't Hoen
#>      2:                                              NA    A'Hara
#>      3:                                              NA   A'Heasy
#>      4:                                              NA    Aaberg
#>      5:                                              NA   Aadland
#>     ---                                                          
#> 425297:                                              NA  Zywiciel
#> 425298:                                              NA   Zywicki
#> 425299:                                              NA   Zywicki
#> 425300:                                              NA     Zyzda
#> 425301:                                              NA     Zyzyk
#>         name_first              name_given name_suffix name_matrilineal
#>      1: Evert Jean Evert Jean Axel Charles                             
#>      2:      Scott                   Scott                             
#>      3:                                                                
#>      4:         Al          Alfred Bernard                             
#>      5:       Kirk              Kirk Allan                             
#>     ---                                                                
#> 425297:     Joseph                  Joseph                             
#> 425298:      Bobby                   Bobby                             
#> 425299:    Brandon                 Brandon                             
#> 425300:       Nate                    Nate                             
#> 425301:     Stormy                  Stormy                             
#>         name_nick birth_year birth_month birth_day death_year
#>      1:     E. J.       1975          11         8         NA
#>      2:                   NA          NA        NA         NA
#>      3:                   NA          NA        NA         NA
#>      4:                 1911           5         5       1976
#>      5:                 1958           7        14         NA
#>     ---                                                      
#> 425297:                   NA          NA        NA         NA
#> 425298:                   NA          NA        NA         NA
#> 425299:                 1993           1         8         NA
#> 425300:                   NA          NA        NA         NA
#> 425301:                   NA          NA        NA         NA
#>         death_month death_day pro_played_first pro_played_last
#>      1:          NA        NA             1996            2007
#>      2:          NA        NA               NA              NA
#>      3:          NA        NA             1901            1901
#>      4:          11        12             1936            1936
#>      5:          NA        NA             1978            1978
#>     ---                                                       
#> 425297:          NA        NA               NA              NA
#> 425298:          NA        NA               NA              NA
#> 425299:          NA        NA             2015            2015
#> 425300:          NA        NA               NA              NA
#> 425301:          NA        NA               NA              NA
#>         mlb_played_first mlb_played_last col_played_first
#>      1:               NA              NA               NA
#>      2:               NA              NA             2008
#>      3:               NA              NA               NA
#>      4:               NA              NA               NA
#>      5:               NA              NA               NA
#>     ---                                                  
#> 425297:               NA              NA             2018
#> 425298:               NA              NA             2016
#> 425299:               NA              NA               NA
#> 425300:               NA              NA               NA
#> 425301:               NA              NA               NA
#>         col_played_last pro_managed_first pro_managed_last
#>      1:              NA                NA               NA
#>      2:            2011                NA               NA
#>      3:              NA                NA               NA
#>      4:              NA                NA               NA
#>      5:              NA                NA               NA
#>     ---                                                   
#> 425297:            2019                NA               NA
#> 425298:            2019                NA               NA
#> 425299:              NA                NA               NA
#> 425300:              NA                NA               NA
#> 425301:              NA                NA               NA
#>         mlb_managed_first mlb_managed_last col_managed_first
#>      1:                NA               NA                NA
#>      2:                NA               NA                NA
#>      3:                NA               NA                NA
#>      4:                NA               NA                NA
#>      5:                NA               NA                NA
#>     ---                                                     
#> 425297:                NA               NA                NA
#> 425298:                NA               NA                NA
#> 425299:                NA               NA                NA
#> 425300:                NA               NA                NA
#> 425301:                NA               NA                NA
#>         col_managed_last pro_umpired_first pro_umpired_last
#>      1:               NA                NA               NA
#>      2:               NA                NA               NA
#>      3:               NA                NA               NA
#>      4:               NA                NA               NA
#>      5:               NA                NA               NA
#>     ---                                                    
#> 425297:               NA                NA               NA
#> 425298:               NA                NA               NA
#> 425299:               NA                NA               NA
#> 425300:               NA                NA               NA
#> 425301:               NA                NA               NA
#>         mlb_umpired_first mlb_umpired_last
#>      1:                NA               NA
#>      2:                NA               NA
#>      3:                NA               NA
#>      4:                NA               NA
#>      5:                NA               NA
#>     ---                                   
#> 425297:                NA               NA
#> 425298:                NA               NA
#> 425299:                NA               NA
#> 425300:                NA               NA
#> 425301:                NA               NA
# }