Scrape MiLB game logs for batters from FanGraphs

This function allows you to scrape MiLB game logs for individual batters from FanGraphs.

Usage

fg_milb_batter_game_logs(playerid, year)

Arguments

playerid

The batter's minor leauge ID from FanGraphs.

year

The season for which game logs should be returned.

Value

Returns a data frame of Minor League batter game logs.

col_name	types

player_name	character

minor_playerid	character

Date	character

Team	character

Level	character

Opp	character

G	numeric

AB	numeric

PA	numeric

H	numeric

1B	numeric

2B	numeric

3B	numeric

HR	numeric

R	numeric

RBI	numeric

BB	numeric

IBB	numeric

SO	numeric

HBP	numeric

SF	numeric

SH	numeric

GDP	numeric

SB	numeric

CS	numeric

AVG	numeric

BB%	numeric

K%	numeric

BB/K	numeric

OBP	numeric

SLG	numeric

OPS	numeric

ISO	numeric

Spd	numeric

BABIP	numeric

wRC	numeric

wRAA	numeric

wOBA	numeric

wRC+	numeric

wBsR	numeric

gamedate	character

dh	integer

Examples

# \donttest{
  fg_milb_batter_game_logs(playerid = "sa3010868", year=2021)
#>       player_name minor_playerid       Date Team Level  Opp G AB PA H
#> 2   Anthony Volpe      sa3010868 2021-09-19  NYY  (A+)  PIT 1  4  5 1
#> 3   Anthony Volpe      sa3010868 2021-09-18  NYY  (A+)  PIT 1  5  6 1
#> 4   Anthony Volpe      sa3010868 2021-09-17  NYY  (A+)  PIT 1  4  4 1
#> 5   Anthony Volpe      sa3010868 2021-09-16  NYY  (A+)  PIT 1  4  5 1
#> 6   Anthony Volpe      sa3010868 2021-09-15  NYY  (A+)  PIT 1  3  5 1
#> 7   Anthony Volpe      sa3010868 2021-09-14  NYY  (A+)  PIT 1  3  4 0
#> 8   Anthony Volpe      sa3010868 2021-09-12  NYY  (A+)  NYM 1  4  5 2
#> 9   Anthony Volpe      sa3010868 2021-09-11  NYY  (A+)  NYM 1  2  4 1
#> 10  Anthony Volpe      sa3010868 2021-09-09  NYY  (A+)  NYM 1  4  5 0
#> 11  Anthony Volpe      sa3010868 2021-09-08  NYY  (A+)  NYM 1  4  4 2
#> 12  Anthony Volpe      sa3010868 2021-09-08  NYY  (A+)  NYM 1  3  4 2
#> 13  Anthony Volpe      sa3010868 2021-09-07  NYY  (A+)  NYM 1  5  5 1
#> 14  Anthony Volpe      sa3010868 2021-09-05  NYY  (A+) @PIT 1  3  3 1
#> 15  Anthony Volpe      sa3010868 2021-09-04  NYY  (A+) @PIT 1  5  5 0
#> 16  Anthony Volpe      sa3010868 2021-09-03  NYY  (A+) @PIT 1  5  5 0
#> 17  Anthony Volpe      sa3010868 2021-09-02  NYY  (A+) @PIT 1  3  3 1
#> 18  Anthony Volpe      sa3010868 2021-09-02  NYY  (A+) @PIT 1  4  4 2
#> 19  Anthony Volpe      sa3010868 2021-08-31  NYY  (A+) @PIT 1  5  5 1
#> 20  Anthony Volpe      sa3010868 2021-08-29  NYY  (A+) @CHW 1  4  5 1
#> 21  Anthony Volpe      sa3010868 2021-08-28  NYY  (A+) @CHW 1  4  5 1
#> 22  Anthony Volpe      sa3010868 2021-08-26  NYY  (A+) @CHW 1  4  4 1
#> 23  Anthony Volpe      sa3010868 2021-08-25  NYY  (A+) @CHW 1  5  5 1
#> 24  Anthony Volpe      sa3010868 2021-08-24  NYY  (A+) @CHW 1  3  5 1
#> 25  Anthony Volpe      sa3010868 2021-08-21  NYY  (A+)  NYM 1  1  4 0
#> 26  Anthony Volpe      sa3010868 2021-08-20  NYY  (A+)  NYM 1  3  4 1
#> 27  Anthony Volpe      sa3010868 2021-08-19  NYY  (A+)  NYM 1  4  5 3
#> 28  Anthony Volpe      sa3010868 2021-08-18  NYY  (A+)  NYM 1  4  5 1
#> 29  Anthony Volpe      sa3010868 2021-08-17  NYY  (A+)  NYM 1  4  5 0
#> 30  Anthony Volpe      sa3010868 2021-08-15  NYY  (A+)  PHI 1  4  4 1
#> 31  Anthony Volpe      sa3010868 2021-08-14  NYY  (A+)  PHI 1  4  5 0
#> 32  Anthony Volpe      sa3010868 2021-08-13  NYY  (A+)  PHI 1  4  4 2
#> 33  Anthony Volpe      sa3010868 2021-08-11  NYY  (A+)  PHI 1  3  5 0
#> 34  Anthony Volpe      sa3010868 2021-08-10  NYY  (A+)  PHI 1  4  5 2
#> 35  Anthony Volpe      sa3010868 2021-08-08  NYY  (A+) @NYM 1  4  5 2
#> 36  Anthony Volpe      sa3010868 2021-08-07  NYY  (A+) @NYM 1  4  5 1
#> 37  Anthony Volpe      sa3010868 2021-08-06  NYY  (A+) @NYM 1  4  5 2
#> 38  Anthony Volpe      sa3010868 2021-08-04  NYY  (A+) @NYM 1  3  4 0
#> 39  Anthony Volpe      sa3010868 2021-08-03  NYY  (A+) @NYM 1  4  5 0
#> 40  Anthony Volpe      sa3010868 2021-08-01  NYY  (A+)  WSN 1  3  3 1
#> 41  Anthony Volpe      sa3010868 2021-07-31  NYY  (A+)  WSN 1  5  5 0
#> 42  Anthony Volpe      sa3010868 2021-07-30  NYY  (A+)  WSN 1  3  5 2
#> 43  Anthony Volpe      sa3010868 2021-07-29  NYY  (A+)  WSN 1  4  4 2
#> 44  Anthony Volpe      sa3010868 2021-07-28  NYY  (A+)  WSN 1  5  5 3
#> 45  Anthony Volpe      sa3010868 2021-07-27  NYY  (A+)  WSN 1  3  4 0
#> 46  Anthony Volpe      sa3010868 2021-07-25  NYY  (A+) @BAL 1  4  6 1
#> 47  Anthony Volpe      sa3010868 2021-07-24  NYY  (A+) @BAL 1  3  5 0
#> 48  Anthony Volpe      sa3010868 2021-07-23  NYY  (A+) @BAL 1  4  4 2
#> 49  Anthony Volpe      sa3010868 2021-07-21  NYY  (A+) @BAL 1  4  5 2
#> 50  Anthony Volpe      sa3010868 2021-07-20  NYY  (A+) @BAL 1  6  7 3
#> 51  Anthony Volpe      sa3010868 2021-07-18  NYY  (A+) @WSN 1  5  5 0
#> 52  Anthony Volpe      sa3010868 2021-07-17  NYY  (A+) @WSN 1  4  5 1
#> 53  Anthony Volpe      sa3010868 2021-07-16  NYY  (A+) @WSN 1  5  5 4
#> 54  Anthony Volpe      sa3010868 2021-07-15  NYY  (A+) @WSN 1  4  4 0
#> 55  Anthony Volpe      sa3010868 2021-07-14  NYY  (A+) @WSN 1  4  4 1
#> 56  Anthony Volpe      sa3010868 2021-07-13  NYY  (A+) @WSN 1  3  5 1
#> 57  Anthony Volpe      sa3010868 2021-07-11  NYY   (A) @TOR 1  5  5 0
#> 58  Anthony Volpe      sa3010868 2021-07-10  NYY   (A) @TOR 1  3  4 0
#> 59  Anthony Volpe      sa3010868 2021-07-10  NYY   (A) @TOR 1  2  5 2
#> 60  Anthony Volpe      sa3010868 2021-07-09  NYY   (A) @TOR 1  3  5 0
#> 61  Anthony Volpe      sa3010868 2021-07-08  NYY   (A) @TOR 1  1  4 0
#> 62  Anthony Volpe      sa3010868 2021-07-08  NYY   (A) @TOR 1  4  4 2
#> 63  Anthony Volpe      sa3010868 2021-07-04  NYY   (A)  MIA 1  2  3 1
#> 64  Anthony Volpe      sa3010868 2021-07-04  NYY   (A)  MIA 1  3  3 0
#> 65  Anthony Volpe      sa3010868 2021-07-02  NYY   (A)  MIA 1  2  4 0
#> 66  Anthony Volpe      sa3010868 2021-07-01  NYY   (A)  MIA 1  0  4 0
#> 67  Anthony Volpe      sa3010868 2021-07-01  NYY   (A)  MIA 1  3  4 2
#> 68  Anthony Volpe      sa3010868 2021-06-27  NYY   (A) @DET 1  4  6 2
#> 69  Anthony Volpe      sa3010868 2021-06-26  NYY   (A) @DET 1  4  4 3
#> 70  Anthony Volpe      sa3010868 2021-06-26  NYY   (A) @DET 1  1  4 1
#> 71  Anthony Volpe      sa3010868 2021-06-24  NYY   (A) @DET 1  5  5 1
#> 72  Anthony Volpe      sa3010868 2021-06-22  NYY   (A) @DET 1  4  5 0
#> 73  Anthony Volpe      sa3010868 2021-06-20  NYY   (A)  PIT 1  3  5 0
#> 74  Anthony Volpe      sa3010868 2021-06-19  NYY   (A)  PIT 1  5  5 3
#> 75  Anthony Volpe      sa3010868 2021-06-18  NYY   (A)  PIT 1  1  1 1
#> 76  Anthony Volpe      sa3010868 2021-06-18  NYY   (A)  PIT 1  3  4 1
#> 77  Anthony Volpe      sa3010868 2021-06-16  NYY   (A)  PIT 1  3  3 1
#> 78  Anthony Volpe      sa3010868 2021-06-15  NYY   (A)  PIT 1  4  5 1
#> 79  Anthony Volpe      sa3010868 2021-06-12  NYY   (A) @PHI 1  4  5 2
#> 80  Anthony Volpe      sa3010868 2021-06-11  NYY   (A) @PHI 1  5  5 1
#> 81  Anthony Volpe      sa3010868 2021-06-10  NYY   (A) @PHI 1  4  5 2
#> 82  Anthony Volpe      sa3010868 2021-06-09  NYY   (A) @PHI 1  3  6 1
#> 83  Anthony Volpe      sa3010868 2021-06-08  NYY   (A) @PHI 1  3  4 1
#> 84  Anthony Volpe      sa3010868 2021-06-06  NYY   (A)  DET 1  5  5 2
#> 85  Anthony Volpe      sa3010868 2021-06-05  NYY   (A)  DET 1  3  4 1
#> 86  Anthony Volpe      sa3010868 2021-06-05  NYY   (A)  DET 1  2  4 2
#> 87  Anthony Volpe      sa3010868 2021-06-03  NYY   (A)  DET 1  5  6 3
#> 88  Anthony Volpe      sa3010868 2021-06-02  NYY   (A)  DET 1  4  4 2
#> 89  Anthony Volpe      sa3010868 2021-06-01  NYY   (A)  DET 1  4  4 0
#> 90  Anthony Volpe      sa3010868 2021-05-30  NYY   (A) @TOR 1  4  6 0
#> 91  Anthony Volpe      sa3010868 2021-05-28  NYY   (A) @TOR 1  5  5 1
#> 92  Anthony Volpe      sa3010868 2021-05-27  NYY   (A) @TOR 1  6  6 0
#> 93  Anthony Volpe      sa3010868 2021-05-26  NYY   (A) @TOR 1  4  4 0
#> 94  Anthony Volpe      sa3010868 2021-05-26  NYY   (A) @TOR 1  5  5 4
#> 95  Anthony Volpe      sa3010868 2021-05-25  NYY   (A) @TOR 1  4  5 1
#> 96  Anthony Volpe      sa3010868 2021-05-23  NYY   (A)  MIN 1  5  5 2
#> 97  Anthony Volpe      sa3010868 2021-05-22  NYY   (A)  MIN 1  5  6 2
#> 98  Anthony Volpe      sa3010868 2021-05-21  NYY   (A)  MIN 1  3  5 0
#> 99  Anthony Volpe      sa3010868 2021-05-20  NYY   (A)  MIN 1  4  6 1
#> 100 Anthony Volpe      sa3010868 2021-05-19  NYY   (A)  MIN 1  5  6 1
#> 101 Anthony Volpe      sa3010868 2021-05-18  NYY   (A)  MIN 1  4  5 0
#> 102 Anthony Volpe      sa3010868 2021-05-15  NYY   (A) @DET 1  4  5 2
#> 103 Anthony Volpe      sa3010868 2021-05-14  NYY   (A) @DET 1  5  5 1
#> 104 Anthony Volpe      sa3010868 2021-05-12  NYY   (A) @DET 1  3  5 0
#> 105 Anthony Volpe      sa3010868 2021-05-11  NYY   (A) @DET 1  4  5 0
#> 106 Anthony Volpe      sa3010868 2021-05-08  NYY   (A)  TOR 1  5  5 0
#> 107 Anthony Volpe      sa3010868 2021-05-07  NYY   (A)  TOR 1  4  7 1
#> 108 Anthony Volpe      sa3010868 2021-05-06  NYY   (A)  TOR 1  5  5 3
#> 109 Anthony Volpe      sa3010868 2021-05-05  NYY   (A)  TOR 1  5  6 1
#> 110 Anthony Volpe      sa3010868 2021-05-04  NYY   (A)  TOR 1  3  6 2
#>     1B 2B 3B HR R RBI BB IBB SO HBP SF SH GDP SB CS       AVG       BB%
#> 2    1  0  0  0 0   0  1   0  1   0  0  0   0  1  0 0.2500000 0.2000000
#> 3    1  0  0  0 1   0  1   1  1   0  0  0   0  0  0 0.2000000 0.1666667
#> 4    0  0  0  1 1   1  0   0  0   0  0  0   0  0  0 0.2500000 0.0000000
#> 5    1  0  0  0 1   0  1   0  3   0  0  0   0  0  0 0.2500000 0.2000000
#> 6    0  0  0  1 2   1  2   0  1   0  0  0   0  0  0 0.3333333 0.4000000
#> 7    0  0  0  0 0   0  0   0  2   1  0  0   0  0  0 0.0000000 0.0000000
#> 8    1  1  0  0 2   0  0   0  2   1  0  0   0  1  0 0.5000000 0.0000000
#> 9    1  0  0  0 2   1  2   0  1   0  0  0   0  0  0 0.5000000 0.5000000
#> 10   0  0  0  0 0   0  0   0  0   1  0  0   0  1  0 0.0000000 0.0000000
#> 11   0  1  0  1 2   1  0   0  1   0  0  0   0  0  0 0.5000000 0.0000000
#> 12   0  2  0  0 0   0  0   0  1   1  0  0   0  0  1 0.6666667 0.0000000
#> 13   0  1  0  0 2   2  0   0  0   0  0  0   0  1  0 0.2000000 0.0000000
#> 14   0  0  0  1 1   3  0   0  1   0  0  0   0  0  0 0.3333333 0.0000000
#> 15   0  0  0  0 0   0  0   0  0   0  0  0   0  0  0 0.0000000 0.0000000
#> 16   0  0  0  0 0   0  0   0  2   0  0  0   0  0  0 0.0000000 0.0000000
#> 17   1  0  0  0 0   0  0   0  0   0  0  0   1  1  0 0.3333333 0.0000000
#> 18   0  1  0  1 2   2  0   0  1   0  0  0   0  0  0 0.5000000 0.0000000
#> 19   0  0  0  1 1   1  0   0  0   0  0  0   0  0  1 0.2000000 0.0000000
#> 20   1  0  0  0 1   1  0   0  1   1  0  0   0  0  0 0.2500000 0.0000000
#> 21   0  0  0  1 1   2  0   0  1   0  1  1   0  0  0 0.2500000 0.0000000
#> 22   1  0  0  0 1   0  0   0  2   0  0  0   0  0  0 0.2500000 0.0000000
#> 23   1  0  0  0 0   0  0   0  2   0  0  0   0  0  0 0.2000000 0.0000000
#> 24   0  1  0  0 1   0  2   0  1   0  0  0   0  0  0 0.3333333 0.4000000
#> 25   0  0  0  0 1   0  3   0  0   0  0  0   1  0  0 0.0000000 0.7500000
#> 26   1  0  0  0 1   0  1   0  2   0  0  0   0  1  1 0.3333333 0.2500000
#> 27   1  0  1  1 3   2  1   0  0   0  0  0   0  0  0 0.7500000 0.2000000
#> 28   1  0  0  0 1   1  0   0  1   1  0  0   0  0  0 0.2500000 0.0000000
#> 29   0  0  0  0 0   0  0   0  0   1  0  0   0  0  0 0.0000000 0.0000000
#> 30   0  1  0  0 0   1  0   0  0   0  0  0   0  0  0 0.2500000 0.0000000
#> 31   0  0  0  0 0   0  1   0  2   0  0  0   0  1  0 0.0000000 0.2000000
#> 32   1  0  0  1 1   1  0   0  0   0  0  0   1  0  0 0.5000000 0.0000000
#> 33   0  0  0  0 2   0  2   0  3   0  0  0   0  0  0 0.0000000 0.4000000
#> 34   1  0  0  1 2   2  1   0  2   0  0  0   0  0  0 0.5000000 0.2000000
#> 35   2  0  0  0 1   0  1   0  1   0  0  0   0  2  0 0.5000000 0.2000000
#> 36   0  1  0  0 1   0  0   0  1   1  0  0   0  0  0 0.2500000 0.0000000
#> 37   1  1  0  0 2   1  0   0  0   1  0  0   0  1  0 0.5000000 0.0000000
#> 38   0  0  0  0 0   1  0   0  1   0  1  1   0  0  0 0.0000000 0.0000000
#> 39   0  0  0  0 1   0  1   0  1   0  0  0   0  0  0 0.0000000 0.2000000
#> 40   0  0  0  1 1   2  0   0  2   0  0  0   0  0  0 0.3333333 0.0000000
#> 41   0  0  0  0 0   0  0   0  1   0  0  0   0  0  0 0.0000000 0.0000000
#> 42   2  0  0  0 2   1  0   0  0   1  1  1   0  1  0 0.6666667 0.0000000
#> 43   2  0  0  0 1   0  0   0  2   0  0  0   0  0  0 0.5000000 0.0000000
#> 44   1  1  0  1 2   2  0   0  1   0  0  0   0  0  0 0.6000000 0.0000000
#> 45   0  0  0  0 1   0  1   0  1   0  0  0   0  1  0 0.0000000 0.2500000
#> 46   0  0  0  1 3   1  1   0  2   1  0  0   0  0  0 0.2500000 0.1666667
#> 47   0  0  0  0 0   0  1   0  1   1  0  0   0  0  0 0.0000000 0.2000000
#> 48   1  0  0  1 1   1  0   0  1   0  0  0   0  0  1 0.5000000 0.0000000
#> 49   0  1  0  1 2   3  1   0  1   0  0  0   0  0  0 0.5000000 0.2000000
#> 50   1  2  0  0 4   1  1   0  2   0  0  0   0  0  0 0.5000000 0.1428571
#> 51   0  0  0  0 0   0  0   0  3   0  0  0   0  0  0 0.0000000 0.0000000
#> 52   0  1  0  0 1   0  1   0  1   0  0  0   0  0  0 0.2500000 0.2000000
#> 53   2  2  0  0 2   1  0   0  0   0  0  0   0  0  0 0.8000000 0.0000000
#> 54   0  0  0  0 0   0  0   0  1   0  0  0   0  0  0 0.0000000 0.0000000
#> 55   1  0  0  0 0   0  0   0  1   0  0  0   0  0  0 0.2500000 0.0000000
#> 56   1  0  0  0 0   1  1   0  0   0  1  1   0  0  0 0.3333333 0.2000000
#> 57   0  0  0  0 0   0  0   0  0   0  0  0   0  0  0 0.0000000 0.0000000
#> 58   0  0  0  0 0   0  1   0  2   0  0  0   0  1  0 0.0000000 0.2500000
#> 59   0  0  1  1 4   2  3   0  0   0  0  0   0  0  1 1.0000000 0.6000000
#> 60   0  0  0  0 1   0  2   0  1   0  0  0   0  0  0 0.0000000 0.4000000
#> 61   0  0  0  0 1   0  1   0  0   2  0  0   0  0  0 0.0000000 0.2500000
#> 62   2  0  0  0 2   1  0   0  1   0  0  0   0  0  0 0.5000000 0.0000000
#> 63   0  0  0  1 1   1  1   0  0   0  0  0   0  0  0 0.5000000 0.3333333
#> 64   0  0  0  0 0   0  0   0  1   0  0  0   1  0  0 0.0000000 0.0000000
#> 65   0  0  0  0 0   0  2   0  1   0  0  0   0  0  0 0.0000000 0.5000000
#> 66   0  0  0  0 1   0  3   0  0   1  0  0   0  1  0 0.0000000 0.7500000
#> 67   1  0  0  1 1   4  1   0  1   0  0  0   0  0  1 0.6666667 0.2500000
#> 68   1  1  0  0 2   1  2   0  2   0  0  0   0  1  0 0.5000000 0.3333333
#> 69   1  1  0  1 3   1  0   0  1   0  0  0   0  1  0 0.7500000 0.0000000
#> 70   0  1  0  0 2   0  1   0  0   2  0  0   0  1  0 1.0000000 0.2500000
#> 71   0  0  1  0 1   2  0   0  2   0  0  0   0  0  0 0.2000000 0.0000000
#> 72   0  0  0  0 0   1  1   0  1   0  0  0   0  0  0 0.0000000 0.2000000
#> 73   0  0  0  0 1   0  2   0  1   0  0  0   0  0  0 0.0000000 0.4000000
#> 74   0  0  1  2 3   4  0   0  0   0  0  0   0  0  0 0.6000000 0.0000000
#> 75   0  1  0  0 0   0  0   0  0   0  0  0   0  0  1 1.0000000 0.0000000
#> 76   0  1  0  0 0   0  1   0  0   0  0  0   0  0  0 0.3333333 0.2500000
#> 77   0  1  0  0 0   1  0   0  1   0  0  0   0  0  0 0.3333333 0.0000000
#> 78   1  0  0  0 1   0  1   0  1   0  0  0   0  0  0 0.2500000 0.2000000
#> 79   1  0  0  1 2   2  1   0  2   0  0  0   0  1  0 0.5000000 0.2000000
#> 80   1  0  0  0 0   0  0   0  1   0  0  0   0  0  0 0.2000000 0.0000000
#> 81   1  1  0  0 0   2  1   0  0   0  0  0   0  0  0 0.5000000 0.2000000
#> 82   1  0  0  0 1   1  3   0  0   0  0  0   0  1  0 0.3333333 0.5000000
#> 83   1  0  0  0 0   0  1   0  0   0  0  0   0  0  1 0.3333333 0.2500000
#> 84   1  0  1  0 1   0  0   0  2   0  0  0   0  0  0 0.4000000 0.0000000
#> 85   1  0  0  0 0   0  1   0  1   0  0  0   0  0  0 0.3333333 0.2500000
#> 86   1  1  0  0 2   0  2   0  0   0  0  0   0  1  0 1.0000000 0.5000000
#> 87   0  1  0  2 2   6  1   0  1   0  0  0   0  2  0 0.6000000 0.1666667
#> 88   1  0  0  1 1   1  0   0  2   0  0  0   0  0  0 0.5000000 0.0000000
#> 89   0  0  0  0 0   0  0   0  1   0  0  0   0  0  0 0.0000000 0.0000000
#> 90   0  0  0  0 0   0  2   0  1   0  0  0   0  0  0 0.0000000 0.3333333
#> 91   0  1  0  0 1   1  0   0  2   0  0  0   0  0  0 0.2000000 0.0000000
#> 92   0  0  0  0 0   0  0   0  0   0  0  0   1  1  0 0.0000000 0.0000000
#> 93   0  0  0  0 0   0  0   0  1   0  0  0   0  0  0 0.0000000 0.0000000
#> 94   2  1  0  1 2   1  0   0  0   0  0  0   0  0  0 0.8000000 0.0000000
#> 95   1  0  0  0 1   0  1   0  0   0  0  0   0  0  1 0.2500000 0.2000000
#> 96   0  1  0  1 1   1  0   0  1   0  0  0   0  0  0 0.4000000 0.0000000
#> 97   2  0  0  0 0   2  1   0  1   0  0  0   0  1  0 0.4000000 0.1666667
#> 98   0  0  0  0 1   0  2   0  0   0  0  0   1  0  0 0.0000000 0.4000000
#> 99   0  1  0  0 4   3  1   0  0   0  1  1   0  3  0 0.2500000 0.1666667
#> 100  1  0  0  0 2   1  1   0  2   0  0  0   0  0  0 0.2000000 0.1666667
#> 101  0  0  0  0 0   0  1   0  2   0  0  0   0  1  0 0.0000000 0.2000000
#> 102  1  1  0  0 0   1  1   0  1   0  0  0   0  1  0 0.5000000 0.2000000
#> 103  1  0  0  0 1   2  0   0  0   0  0  0   0  0  0 0.2000000 0.0000000
#> 104  0  0  0  0 0   0  2   0  1   0  0  0   0  0  0 0.0000000 0.4000000
#> 105  0  0  0  0 0   0  0   0  1   1  0  0   0  0  0 0.0000000 0.0000000
#> 106  0  0  0  0 1   0  0   0  1   0  0  0   1  0  0 0.0000000 0.0000000
#> 107  0  1  0  0 3   2  3   0  1   0  0  0   0  1  0 0.2500000 0.4285714
#> 108  2  1  0  0 1   2  0   0  2   0  0  0   0  0  0 0.6000000 0.0000000
#> 109  0  1  0  0 2   1  1   0  0   0  0  0   0  1  0 0.2000000 0.1666667
#> 110  0  1  1  0 3   2  3   0  0   0  0  0   0  2  0 0.6666667 0.5000000
#>            K%      BB/K       OBP       SLG       OPS       ISO
#> 2   0.2000000 1.0000000 0.4000000 0.2500000 0.6500000 0.0000000
#> 3   0.1666667 1.0000000 0.3333333 0.2000000 0.5333333 0.0000000
#> 4   0.0000000 0.0000000 0.2500000 1.0000000 1.2500000 0.7500000
#> 5   0.6000000 0.3333333 0.4000000 0.2500000 0.6500000 0.0000000
#> 6   0.2000000 2.0000000 0.6000000 1.3333333 1.9333333 1.0000000
#> 7   0.5000000 0.0000000 0.2500000 0.0000000 0.2500000 0.0000000
#> 8   0.4000000 0.0000000 0.6000000 0.7500000 1.3500000 0.2500000
#> 9   0.2500000 2.0000000 0.7500000 0.5000000 1.2500000 0.0000000
#> 10  0.0000000 0.0000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 11  0.2500000 0.0000000 0.5000000 1.5000000 2.0000000 1.0000000
#> 12  0.2500000 0.0000000 0.7500000 1.3333333 2.0833333 0.6666667
#> 13  0.0000000 0.0000000 0.2000000 0.4000000 0.6000000 0.2000000
#> 14  0.3333333 0.0000000 0.3333333 1.3333333 1.6666667 1.0000000
#> 15  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 16  0.4000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 17  0.0000000 0.0000000 0.3333333 0.3333333 0.6666667 0.0000000
#> 18  0.2500000 0.0000000 0.5000000 1.5000000 2.0000000 1.0000000
#> 19  0.0000000 0.0000000 0.2000000 0.8000000 1.0000000 0.6000000
#> 20  0.2000000 0.0000000 0.4000000 0.2500000 0.6500000 0.0000000
#> 21  0.2000000 0.0000000 0.2000000 1.0000000 1.2000000 0.7500000
#> 22  0.5000000 0.0000000 0.2500000 0.2500000 0.5000000 0.0000000
#> 23  0.4000000 0.0000000 0.2000000 0.2000000 0.4000000 0.0000000
#> 24  0.2000000 2.0000000 0.6000000 0.6666667 1.2666667 0.3333333
#> 25  0.0000000 3.0000000 0.7500000 0.0000000 0.7500000 0.0000000
#> 26  0.5000000 0.5000000 0.5000000 0.3333333 0.8333333 0.0000000
#> 27  0.0000000 1.0000000 0.8000000 2.0000000 2.8000000 1.2500000
#> 28  0.2000000 0.0000000 0.4000000 0.2500000 0.6500000 0.0000000
#> 29  0.0000000 0.0000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 30  0.0000000 0.0000000 0.2500000 0.5000000 0.7500000 0.2500000
#> 31  0.4000000 0.5000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 32  0.0000000 0.0000000 0.5000000 1.2500000 1.7500000 0.7500000
#> 33  0.6000000 0.6666667 0.4000000 0.0000000 0.4000000 0.0000000
#> 34  0.4000000 0.5000000 0.6000000 1.2500000 1.8500000 0.7500000
#> 35  0.2000000 1.0000000 0.6000000 0.5000000 1.1000000 0.0000000
#> 36  0.2000000 0.0000000 0.4000000 0.5000000 0.9000000 0.2500000
#> 37  0.0000000 0.0000000 0.6000000 0.7500000 1.3500000 0.2500000
#> 38  0.2500000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 39  0.2000000 1.0000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 40  0.6666667 0.0000000 0.3333333 1.3333333 1.6666667 1.0000000
#> 41  0.2000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 42  0.0000000 0.0000000 0.6000000 0.6666667 1.2666667 0.0000000
#> 43  0.5000000 0.0000000 0.5000000 0.5000000 1.0000000 0.0000000
#> 44  0.2000000 0.0000000 0.6000000 1.4000000 2.0000000 0.8000000
#> 45  0.2500000 1.0000000 0.2500000 0.0000000 0.2500000 0.0000000
#> 46  0.3333333 0.5000000 0.5000000 1.0000000 1.5000000 0.7500000
#> 47  0.2000000 1.0000000 0.4000000 0.0000000 0.4000000 0.0000000
#> 48  0.2500000 0.0000000 0.5000000 1.2500000 1.7500000 0.7500000
#> 49  0.2000000 1.0000000 0.6000000 1.5000000 2.1000000 1.0000000
#> 50  0.2857143 0.5000000 0.5714286 0.8333333 1.4047619 0.3333333
#> 51  0.6000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 52  0.2000000 1.0000000 0.4000000 0.5000000 0.9000000 0.2500000
#> 53  0.0000000 0.0000000 0.8000000 1.2000000 2.0000000 0.4000000
#> 54  0.2500000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 55  0.2500000 0.0000000 0.2500000 0.2500000 0.5000000 0.0000000
#> 56  0.0000000 1.0000000 0.4000000 0.3333333 0.7333333 0.0000000
#> 57  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 58  0.5000000 0.5000000 0.2500000 0.0000000 0.2500000 0.0000000
#> 59  0.0000000 3.0000000 1.0000000 3.5000000 4.5000000 2.5000000
#> 60  0.2000000 2.0000000 0.4000000 0.0000000 0.4000000 0.0000000
#> 61  0.0000000 1.0000000 0.7500000 0.0000000 0.7500000 0.0000000
#> 62  0.2500000 0.0000000 0.5000000 0.5000000 1.0000000 0.0000000
#> 63  0.0000000 1.0000000 0.6666667 2.0000000 2.6666667 1.5000000
#> 64  0.3333333 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 65  0.2500000 2.0000000 0.5000000 0.0000000 0.5000000 0.0000000
#> 66  0.0000000 3.0000000 1.0000000 0.0000000 1.0000000 0.0000000
#> 67  0.2500000 1.0000000 0.7500000 1.6666667 2.4166667 1.0000000
#> 68  0.3333333 1.0000000 0.6666667 0.7500000 1.4166667 0.2500000
#> 69  0.2500000 0.0000000 0.7500000 1.7500000 2.5000000 1.0000000
#> 70  0.0000000 1.0000000 1.0000000 2.0000000 3.0000000 1.0000000
#> 71  0.4000000 0.0000000 0.2000000 0.6000000 0.8000000 0.4000000
#> 72  0.2000000 1.0000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 73  0.2000000 2.0000000 0.4000000 0.0000000 0.4000000 0.0000000
#> 74  0.0000000 0.0000000 0.6000000 2.2000000 2.8000000 1.6000000
#> 75  0.0000000 0.0000000 1.0000000 2.0000000 3.0000000 1.0000000
#> 76  0.0000000 1.0000000 0.5000000 0.6666667 1.1666667 0.3333333
#> 77  0.3333333 0.0000000 0.3333333 0.6666667 1.0000000 0.3333333
#> 78  0.2000000 1.0000000 0.4000000 0.2500000 0.6500000 0.0000000
#> 79  0.4000000 0.5000000 0.6000000 1.2500000 1.8500000 0.7500000
#> 80  0.2000000 0.0000000 0.2000000 0.2000000 0.4000000 0.0000000
#> 81  0.0000000 1.0000000 0.6000000 0.7500000 1.3500000 0.2500000
#> 82  0.0000000 3.0000000 0.6666667 0.3333333 1.0000000 0.0000000
#> 83  0.0000000 1.0000000 0.5000000 0.3333333 0.8333333 0.0000000
#> 84  0.4000000 0.0000000 0.4000000 0.8000000 1.2000000 0.4000000
#> 85  0.2500000 1.0000000 0.5000000 0.3333333 0.8333333 0.0000000
#> 86  0.0000000 2.0000000 1.0000000 1.5000000 2.5000000 0.5000000
#> 87  0.1666667 1.0000000 0.6666667 2.0000000 2.6666667 1.4000000
#> 88  0.5000000 0.0000000 0.5000000 1.2500000 1.7500000 0.7500000
#> 89  0.2500000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 90  0.1666667 2.0000000 0.3333333 0.0000000 0.3333333 0.0000000
#> 91  0.4000000 0.0000000 0.2000000 0.4000000 0.6000000 0.2000000
#> 92  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 93  0.2500000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 94  0.0000000 0.0000000 0.8000000 1.6000000 2.4000000 0.8000000
#> 95  0.0000000 1.0000000 0.4000000 0.2500000 0.6500000 0.0000000
#> 96  0.2000000 0.0000000 0.4000000 1.2000000 1.6000000 0.8000000
#> 97  0.1666667 1.0000000 0.5000000 0.4000000 0.9000000 0.0000000
#> 98  0.0000000 2.0000000 0.4000000 0.0000000 0.4000000 0.0000000
#> 99  0.0000000 1.0000000 0.3333333 0.5000000 0.8333333 0.2500000
#> 100 0.3333333 0.5000000 0.3333333 0.2000000 0.5333333 0.0000000
#> 101 0.4000000 0.5000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 102 0.2000000 1.0000000 0.6000000 0.7500000 1.3500000 0.2500000
#> 103 0.0000000 0.0000000 0.2000000 0.2000000 0.4000000 0.0000000
#> 104 0.2000000 2.0000000 0.4000000 0.0000000 0.4000000 0.0000000
#> 105 0.2000000 0.0000000 0.2000000 0.0000000 0.2000000 0.0000000
#> 106 0.2000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> 107 0.1428571 3.0000000 0.5714286 0.5000000 1.0714286 0.2500000
#> 108 0.4000000 0.0000000 0.6000000 0.8000000 1.4000000 0.2000000
#> 109 0.0000000 1.0000000 0.3333333 0.4000000 0.7333333 0.2000000
#> 110 0.0000000 3.0000000 0.8333333 1.6666667 2.5000000 1.0000000
#>            Spd     BABIP         wRC          wRAA      wOBA
#> 2    5.0000000 0.3333333  0.66675816 -0.0103144678 0.3241058
#> 3    2.6428571 0.2500000  0.05953623 -0.7529509298 0.1801328
#> 4    0.1428571 0.0000000  1.10189593  0.5602378289 0.4898837
#> 5    2.6428571 1.0000000  0.66675816 -0.0103144678 0.3241058
#> 6    2.6428571 0.0000000  2.19167899  1.5146063587 0.6798530
#> 7    0.1428571 0.0000000  0.06411977 -0.4775383327 0.1872566
#> 8    7.5000000 1.0000000  1.76390290  1.0868302706 0.5800576
#> 9    2.6428571 1.0000000  1.42840903  0.8867509235 0.5850986
#> 10   2.5000000 0.0000000 -0.08038645 -0.7574590837 0.1498053
#> 11   2.6428571 0.5000000  2.17404064  1.6323825368 0.8025331
#> 12   0.0000000 1.0000000  2.20840919  1.6667510829 0.8125554
#> 13   5.0000000 0.2000000  0.34961358 -0.3274590469 0.2501195
#> 14   0.1428571 0.0000000  1.24640216  0.8401585799 0.6531783
#> 15   0.1428571 0.0000000 -0.72253113 -1.3996037547 0.0000000
#> 16   0.1428571 0.0000000 -0.72253113 -1.3996037547 0.0000000
#> 17   5.0000000 0.3333333  0.33862597 -0.0676176063 0.3002214
#> 18   2.6428571 0.5000000  2.17404064  1.6323825368 0.8025331
#> 19   0.0000000 0.0000000  0.95738971  0.2803170780 0.3919070
#> 20   2.6428571 0.3333333  0.69175819  0.0146855628 0.3299381
#> 21   0.1428571 0.0000000  0.95738971  0.2803170780 0.3919070
#> 22   2.6428571 0.5000000  0.19411975 -0.3475383573 0.2251660
#> 23   0.1428571 0.3333333  0.04961352 -0.6274591082 0.1801328
#> 24   1.6011905 0.5000000  1.58390286  0.9068302338 0.5380656
#> 25   1.6011905 0.0000000  1.27340902  0.7317509173 0.5398989
#> 26   5.4918567 1.0000000  0.81126439  0.2696062831 0.4051323
#> 27   5.1428571 0.6666667  3.68882371  3.0117510768 1.0291204
#> 28   2.6428571 0.3333333  0.69175819  0.0146855628 0.3299381
#> 29   0.1428571 0.0000000 -0.08038645 -0.7574590837 0.1498053
#> 30   0.1428571 0.2500000  0.49411981 -0.0475382959 0.3126494
#> 31   5.0000000 0.0000000 -0.10538649 -0.7824591143 0.1439730
#> 32   0.1428571 0.3333333  1.87404058  1.3323824754 0.7150497
#> 33   2.6428571 0.0000000  0.51175816 -0.1653144740 0.2879461
#> 34   2.6428571 1.0000000  2.34667899  1.6696063649 0.7160128
#> 35   6.4583333 0.6666667  1.43890281  0.7618301787 0.5042387
#> 36   2.6428571 0.3333333  0.99175825  0.3146856241 0.3999248
#> 37   7.5000000 0.5000000  1.76390290  1.0868302706 0.5800576
#> 38   0.1428571 0.0000000 -0.57802490 -1.1196830038 0.0000000
#> 39   2.6428571 0.0000000 -0.10538649 -0.7824591143 0.1439730
#> 40   0.1428571 0.0000000  1.24640216  0.8401585799 0.6531783
#> 41   0.1428571 0.0000000 -0.72253113 -1.3996037547 0.0000000
#> 42   7.5000000 0.5000000  1.46390284  0.7868302093 0.5100709
#> 43   2.6428571 1.0000000  0.96626439  0.4246062892 0.4503320
#> 44   2.6428571 0.6666667  2.80167906  2.1246064323 0.8221593
#> 45   7.5000000 0.0000000  0.03911974 -0.5025383634 0.1799663
#> 46   2.6428571 0.0000000  2.07217279  1.2596856384 0.5714044
#> 47   0.1428571 0.0000000  0.53675819 -0.1403144433 0.2937783
#> 48   2.5000000 0.5000000  1.87404058  1.3323824754 0.7150497
#> 49   2.6428571 0.5000000  2.64667906  1.9696064262 0.7859995
#> 50   2.6428571 0.7500000  2.52203513  1.5741334459 0.5888178
#> 51   0.1428571 0.0000000 -0.72253113 -1.3996037547 0.0000000
#> 52   2.6428571 0.3333333  0.96675822  0.2896855935 0.3940925
#> 53   2.6428571 0.8000000  2.96604758  2.2889749539 0.8605047
#> 54   0.1428571 0.0000000 -0.57802490 -1.1196830038 0.0000000
#> 55   0.1428571 0.3333333  0.19411975 -0.3475383573 0.2251660
#> 56   0.1428571 0.2500000  0.66675816 -0.0103144678 0.3241058
#> 57   0.1428571 0.0000000 -0.72233586 -1.3859597412 0.0000000
#> 58   5.0000000 0.0000000  0.03809155 -0.4928075601 0.1907796
#> 59   7.0619652 1.0000000  4.14369717  3.4800732830 1.2057140
#> 60   2.6428571 0.0000000  0.50958461 -0.1540392754 0.3052474
#> 61   2.6428571 0.0000000  1.32001202  0.7891129154 0.5878251
#> 62   2.6428571 0.6666667  0.96405182  0.4331527133 0.4775746
#> 63   0.1428571 0.0000000  1.85975077  1.4615764375 0.9470020
#> 64   0.1428571 0.0000000 -0.43340151 -0.8315758447 0.0000000
#> 65   0.1428571 0.0000000  0.65405178  0.1231526729 0.3815592
#> 66   6.0202986 0.0000000  1.91097225  1.3800731435 0.7708616
#> 67   2.5000000 1.0000000  2.48624385  1.9553447423 0.9490388
#> 68   7.0619652 1.0000000  2.20703795  1.4106892923 0.6347013
#> 69   7.5000000 1.0000000  2.94124388  2.4103447722 1.0899646
#> 70   7.5000000 1.0000000  2.39097229  1.8600731782 0.9195306
#> 71   5.1428571 0.3333333  0.61862442 -0.0449994650 0.3322654
#> 72   0.1428571 0.0000000 -0.10637562 -0.7699995083 0.1526237
#> 73   2.6428571 0.0000000  0.50958461 -0.1540392754 0.3052474
#> 74   5.1428571 0.3333333  3.97300852  3.3093846335 1.1634204
#> 75   0.0000000 1.0000000  0.92649309  0.7937683145 1.3268218
#> 76   0.1428571 0.3333333  1.10905181  0.5781527027 0.5224850
#> 77   0.1428571 0.5000000  0.63755875  0.2393844180 0.4422739
#> 78   2.6428571 0.3333333  0.66458463  0.0009607448 0.3436535
#> 79   7.5000000 1.0000000  2.34177668  1.6781527941 0.7592310
#> 80   0.1428571 0.2500000  0.04862440 -0.6149994881 0.1910298
#> 81   0.1428571 0.5000000  1.73554489  1.0719210076 0.6090179
#> 82   5.2232143 0.3333333  1.75203792  0.9556892624 0.5407507
#> 83   2.5000000 0.3333333  0.80905180  0.2781526931 0.4295669
#> 84   5.1428571 0.6666667  1.38958467  0.7259607881 0.5232953
#> 85   0.1428571 0.5000000  0.80905180  0.2781526931 0.4295669
#> 86   7.0619652 1.0000000  2.49597230  1.9650731887 0.9520520
#> 87   5.0000000 0.5000000  4.17450157  3.3781529048 1.0409525
#> 88   0.1428571 1.0000000  1.87028362  1.3393845094 0.7582592
#> 89   0.1428571 0.0000000 -0.57786869 -1.1087677930 0.0000000
#> 90   0.1428571 0.0000000  0.36511744 -0.4312312236 0.2543728
#> 91   2.6428571 0.3333333  0.34862441 -0.3149994785 0.2653644
#> 92   2.5000000 0.0000000 -0.86680303 -1.6631516895 0.0000000
#> 93   0.1428571 0.0000000 -0.57786869 -1.1087677930 0.0000000
#> 94   1.6011905 0.7500000  3.56773696  2.9041130771 1.0630016
#> 95   5.0000000 0.2500000  0.66458463  0.0009607448 0.3436535
#> 96   0.1428571 0.3333333  2.02581646  1.3621925708 0.6809419
#> 97   4.5619652 0.5000000  1.29107771  0.4947290497 0.4455695
#> 98   2.6428571 0.0000000  0.50958461 -0.1540392754 0.3052474
#> 99   7.5000000 0.2000000  0.82011747  0.0237688062 0.3483234
#> 100  2.6428571 0.3333333  0.52011746 -0.2762312034 0.2863779
#> 101  5.0000000 0.0000000 -0.10637562 -0.7699995083 0.1526237
#> 102  5.0000000 0.6666667  1.73554489  1.0719210076 0.6090179
#> 103  2.6428571 0.2000000  0.04862440 -0.6149994881 0.1910298
#> 104  0.1428571 0.0000000  0.50958461 -0.1540392754 0.3052474
#> 105  0.1428571 0.0000000 -0.08137562 -0.7449995035 0.1588182
#> 106  0.1428571 0.0000000 -0.72233586 -1.3859597412 0.0000000
#> 107  7.0619652 0.3333333  1.90757076  0.9784973238 0.5165967
#> 108  1.6011905 1.0000000  1.89054491  1.2269210278 0.6474241
#> 109  7.5000000 0.2000000  0.82011747  0.0237688062 0.3483234
#> 110 10.0000000 0.6666667  3.39299821  2.5966495483 0.8795840
#>            wRC+         wBsR   gamedate dh
#> 2     98.476608  0.192302983 2021-09-19  1
#> 3      7.327651 -0.003848510 2021-09-18  1
#> 4    203.430157  0.000000000 2021-09-17  1
#> 5     98.476608 -0.007697020 2021-09-16  1
#> 6    323.699245 -0.007697020 2021-09-15  1
#> 7     11.837683 -0.003848510 2021-09-14  1
#> 8    260.519008  0.192302983 2021-09-12  1
#> 9    263.710451 -0.011545530 2021-09-11  1
#> 10   -11.872649  0.196151493 2021-09-09  1
#> 11   401.367694  0.000000000 2021-09-08  2
#> 12   407.712757 -0.473296301 2021-09-08  1
#> 13    51.636053  0.200000003 2021-09-07  1
#> 14   306.811535  0.000000000 2021-09-05  1
#> 15  -100.000000  0.000000000 2021-09-04  1
#> 16  -100.000000  0.000000000 2021-09-03  1
#> 17    83.355403  0.196151493 2021-09-02  2
#> 18   401.367694  0.000000000 2021-09-02  1
#> 19   141.401331 -0.469447792 2021-08-31  1
#> 20   102.168979 -0.007697020 2021-08-29  1
#> 21   141.401331  0.000000000 2021-08-28  1
#> 22    35.838058 -0.003848510 2021-08-26  1
#> 23     7.327651 -0.003848510 2021-08-25  1
#> 24   233.933967 -0.007697020 2021-08-24  1
#> 25   235.094613 -0.011545530 2021-08-21  1
#> 26   149.774255 -0.277144808 2021-08-20  1
#> 27   544.819499 -0.007697020 2021-08-19  1
#> 28   102.168979 -0.007697020 2021-08-18  1
#> 29   -11.872649 -0.003848510 2021-08-17  1
#> 30    91.223560  0.000000000 2021-08-15  1
#> 31   -15.565019  0.196151493 2021-08-14  1
#> 32   345.982192 -0.003848510 2021-08-13  1
#> 33    75.583938 -0.007697020 2021-08-11  1
#> 34   346.591915 -0.007697020 2021-08-10  1
#> 35   212.518236  0.388454476 2021-08-08  1
#> 36   146.477381 -0.003848510 2021-08-07  1
#> 37   260.519008  0.192302983 2021-08-06  1
#> 38  -100.000000  0.000000000 2021-08-04  1
#> 39   -15.565019 -0.003848510 2021-08-03  1
#> 40   306.811535  0.000000000 2021-08-01  1
#> 41  -100.000000  0.000000000 2021-07-31  1
#> 42   216.210607  0.188454473 2021-07-30  1
#> 43   178.390093 -0.007697020 2021-07-29  1
#> 44   413.792988 -0.003848510 2021-07-28  1
#> 45     7.222220  0.196151493 2021-07-27  1
#> 46   255.040683 -0.007697020 2021-07-25  1
#> 47    79.276308 -0.007697020 2021-07-24  1
#> 48   345.982192 -0.473296301 2021-07-23  1
#> 49   390.900317 -0.003848510 2021-07-21  1
#> 50   266.065055 -0.007697020 2021-07-20  1
#> 51  -100.000000  0.000000000 2021-07-18  1
#> 52   142.785010 -0.003848510 2021-07-17  1
#> 53   438.069338 -0.007697020 2021-07-16  1
#> 54  -100.000000  0.000000000 2021-07-15  1
#> 55    35.838058 -0.003848510 2021-07-14  1
#> 56    98.476608 -0.007697020 2021-07-13  1
#> 57  -100.000000  0.000000000 2021-07-11  1
#> 58     7.174913  0.193442446 2021-07-10  2
#> 59   624.404465 -0.492209234 2021-07-10  1
#> 60    76.788166 -0.013115114 2021-07-09  1
#> 61   248.637077 -0.019672670 2021-07-08  2
#> 62   181.588518 -0.013115114 2021-07-08  1
#> 63   467.069478 -0.006557557 2021-07-04  2
#> 64  -100.000000  0.000000000 2021-07-04  1
#> 65   123.197001 -0.013115114 2021-07-02  1
#> 66   359.950172  0.173769776 2021-07-01  2
#> 67   468.308162 -0.485651677 2021-07-01  1
#> 68   277.144681  0.180327333 2021-06-27  1
#> 69   554.011834  0.193442446 2021-06-26  2
#> 70   450.362838  0.180327333 2021-06-26  1
#> 71    93.219131  0.000000000 2021-06-24  1
#> 72   -16.029505 -0.006557557 2021-06-22  1
#> 73    76.788166 -0.013115114 2021-06-20  1
#> 74   598.683774  0.000000000 2021-06-19  1
#> 75   698.055867 -0.472536564 2021-06-18  2
#> 76   208.900673 -0.006557557 2021-06-18  1
#> 77   160.120505  0.000000000 2021-06-16  1
#> 78   100.144772 -0.013115114 2021-06-15  1
#> 79   352.877094  0.186884889 2021-06-12  1
#> 80     7.327102 -0.006557557 2021-06-11  1
#> 81   261.525381 -0.013115114 2021-06-10  1
#> 82   220.008899  0.173769776 2021-06-09  1
#> 83   152.392760 -0.485651677 2021-06-08  1
#> 84   209.393409 -0.006557557 2021-06-06  1
#> 85   152.392760 -0.013115114 2021-06-05  2
#> 86   470.140609  0.180327333 2021-06-05  1
#> 87   524.205260  0.393442449 2021-06-03  1
#> 88   352.286073 -0.006557557 2021-06-02  1
#> 89  -100.000000  0.000000000 2021-06-01  1
#> 90    45.848942 -0.013115114 2021-05-30  1
#> 91    52.533433  0.000000000 2021-05-28  1
#> 92  -100.000000  0.200000003 2021-05-27  1
#> 93  -100.000000  0.000000000 2021-05-26  2
#> 94   537.614309 -0.013115114 2021-05-26  1
#> 95   100.144772 -0.485651677 2021-05-25  1
#> 96   305.265754  0.000000000 2021-05-23  1
#> 97   162.124679  0.180327333 2021-05-22  1
#> 98    76.788166 -0.013115114 2021-05-21  1
#> 99   102.984724  0.593442452 2021-05-20  1
#> 100   65.312781 -0.013115114 2021-05-19  1
#> 101  -16.029505  0.193442446 2021-05-18  1
#> 102  261.525381  0.186884889 2021-05-15  1
#> 103    7.327102 -0.006557557 2021-05-14  1
#> 104   76.788166 -0.013115114 2021-05-12  1
#> 105  -12.262310 -0.006557557 2021-05-11  1
#> 106 -100.000000  0.000000000 2021-05-08  1
#> 107  205.319696  0.180327333 2021-05-07  1
#> 108  284.881988 -0.013115114 2021-05-06  1
#> 109  102.984724  0.193442446 2021-05-05  1
#> 110  426.069431  0.380327336 2021-05-04  1
# }