Scrape MiLB game logs for pitchers from FanGraphs

This function allows you to scrape MiLB game logs for individual batters from FanGraphs.com.

Usage

fg_milb_pitcher_game_logs(playerid, year)

Arguments

playerid

The pitcher's minor league ID from FanGraphs.com.

year

The season for which game logs should be returned.

Value

Returns a tibble of Minor League pitcher game logs.

col_name	types

player_name	character

minor_playerid	character

Date	character

Team	character

Level	character

Opp	character

W	numeric

L	numeric

ERA	numeric

G	numeric

GS	numeric

CG	numeric

ShO	numeric

SV	numeric

IP	numeric

TBF	numeric

H	numeric

R	numeric

ER	numeric

HR	numeric

BB	numeric

IBB	numeric

HBP	numeric

WP	numeric

BK	numeric

SO	numeric

K/9	numeric

BB/9	numeric

K/BB	numeric

HR/9	numeric

K%	numeric

K-BB%	numeric

BB%	numeric

AVG	numeric

WHIP	numeric

BABIP	numeric

LOB%	numeric

FIP	numeric

gamedate	character

dh	integer

UPId	character

MLBAMId	character

MinorMasterId	character

RRId	character

FirstName	character

LastName	character

firstLastName	character

Height	character

Weight	character

BirthDate	character

Bats	character

Throws	character

Position	character

BirthCity	character

College	character

Age	character

Examples

# \donttest{
  fg_milb_pitcher_game_logs(playerid = "sa3005315", year=2021)
#> ── MiLB Pitcher Game Logs data from FanGraphs.com ─── baseballr 1.5.0 ──
#> ℹ Data updated: 2023-03-04 05:03:12 UTC
#> # A tibble: 21 × 62
#>    playe…¹ minor…² Date  Team  Level Opp       W     L   ERA     G    GS
#>    <chr>   <chr>   <chr> <chr> <chr> <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Hunter… sa3005… 2021… CIN   (AAA) PIT       0     1  2.25     1     1
#>  2 Hunter… sa3005… 2021… CIN   (AAA) PIT       0     1 13.5      1     1
#>  3 Hunter… sa3005… 2021… CIN   (AAA) MIL       1     0  1.8      1     1
#>  4 Hunter… sa3005… 2021… CIN   (AAA) @PIT      1     0  0        1     1
#>  5 Hunter… sa3005… 2021… CIN   (AAA) ATL       0     1 13.5      1     1
#>  6 Hunter… sa3005… 2021… CIN   (AAA) @MIN      1     0  0        1     1
#>  7 Hunter… sa3005… 2021… CIN   (AAA) @CHC      0     1  3        1     1
#>  8 Hunter… sa3005… 2021… CIN   (AAA) STL       0     1  5.4      1     1
#>  9 Hunter… sa3005… 2021… CIN   (AAA) @MIL      1     0  0        1     1
#> 10 Hunter… sa3005… 2021… CIN   (AAA) STL       0     0  1.8      1     1
#> # … with 11 more rows, 51 more variables: CG <dbl>, ShO <dbl>,
#> #   SV <dbl>, IP <dbl>, TBF <dbl>, H <dbl>, R <dbl>, ER <dbl>,
#> #   HR <dbl>, BB <dbl>, IBB <dbl>, HBP <dbl>, WP <dbl>, BK <dbl>,
#> #   SO <dbl>, `K/9` <dbl>, `BB/9` <dbl>, `K/BB` <dbl>, `HR/9` <dbl>,
#> #   `K%` <dbl>, `K-BB%` <dbl>, `BB%` <dbl>, AVG <dbl>, WHIP <dbl>,
#> #   BABIP <dbl>, `LOB%` <dbl>, FIP <dbl>, gamedate <chr>, dh <int>,
#> #   UPId <chr>, PlayerId <chr>, MLBAMId <chr>, MinorMasterId <chr>, …
# }