Scrape Park Factors from FanGraphs.com

This function allows you to scrape park factors for a given season from FanGraphs.com.

This function allows you to scrape park factors by handedness from FanGraphs.com for a given single year.

fg_park(yr)

fg_park_hand(yr)

Arguments

yr
Season for which you want to scrape the park factors.

Examples

fg_park(2013)

#> Data courtey of FanGraphs.com. Please consider supporting FanGraphs by purchasing a membership: https://plus.fangraphs.com/product/fangraphs-membership/?switch-subscription=254671&item=85029&_wcsnonce=3e893e9b53&auto-switch=true

#> # A tibble: 30 × 16
#>    season home_team basic_5yr `3yr` `1yr` single double triple    hr    so  UIBB
#>     <int> <chr>         <int> <int> <int>  <int>  <int>  <int> <int> <int> <int>
#>  1   2013 Angels           95    95    98    100     96     91    93   102    97
#>  2   2013 Orioles         103   102   103    101     99     86   108    99   100
#>  3   2013 Red Sox         105   104    98    103    114    105    96   100   100
#>  4   2013 White Sox       102   105   100     99     97     91   108   103   107
#>  5   2013 Indians         100    97    97    100    103     81   101   101    99
#>  6   2013 Tigers          101   103   106    101     98    126    98    95    99
#>  7   2013 Royals          101   102   104    101    103    114    93    96    99
#>  8   2013 Twins           101   103   101    102    101    106    98    98    99
#>  9   2013 Yankees         101   100   104    100     97     84   110   101   101
#> 10   2013 Athletics        97    97    95     99    100    105    93    97   101
#> # … with 20 more rows, and 5 more variables: GB <int>, FB <int>, LD <int>,
#> #   IFFB <int>, FIP <int>

fg_park_hand(2013)

#> Data courtey of FanGraphs.com. Please consider supporting FanGraphs by purchasing a membership: https://plus.fangraphs.com/product/fangraphs-membership/?switch-subscription=254671&item=85029&_wcsnonce=3e893e9b53&auto-switch=true

#> # A tibble: 30 × 10
#>    season home_team single_as_LHH single_as_RHH double_as_LHH double_as_RHH
#>     <int> <chr>             <int>         <int>         <int>         <int>
#>  1   2013 Angels               98           100            95            96
#>  2   2013 Orioles             102           102           101            99
#>  3   2013 Red Sox             104           103           121           110
#>  4   2013 White Sox            97           101            99            96
#>  5   2013 Indians             102            99           104           104
#>  6   2013 Tigers              102           101            93           102
#>  7   2013 Royals              100           101           104           101
#>  8   2013 Twins               101           103            97           104
#>  9   2013 Yankees             100            99            98            95
#> 10   2013 Athletics            99           100            99           101
#> # … with 20 more rows, and 4 more variables: triple_as_LHH <int>,
#> #   triple_as_RHH <int>, hr_as_LHH <int>, hr_as_RHH <int>