Scrape Park Factors from FanGraphs.com

This function allows you to scrape park factors for a given season from FanGraphs.com.

This function allows you to scrape park factors by handedness from FanGraphs.com for a given single year.

fg_park(yr)

fg_park_hand(yr)

Arguments

yr
Season for which you want to scrape the park factors.

Examples

fg_park(2013)

#> Data courtey of FanGraphs.com. Please consider supporting FanGraphs by purchasing a membership: https://plus.fangraphs.com/product/fangraphs-membership/?switch-subscription=254671&item=85029&_wcsnonce=3e893e9b53&auto-switch=true

#>    season    home_team basic_5yr 3yr 1yr single double triple  hr  so UIBB  GB
#> 1    2013       Angels        95  95  98    100     96     91  93 102   97 100
#> 2    2013      Orioles       103 102 103    101     99     86 108  99  100 100
#> 3    2013      Red Sox       105 104  98    103    114    105  96 100  100 101
#> 4    2013    White Sox       102 105 100     99     97     91 108 103  107  98
#> 5    2013      Indians       100  97  97    100    103     81 101 101   99 101
#> 6    2013       Tigers       101 103 106    101     98    126  98  95   99 101
#> 7    2013       Royals       101 102 104    101    103    114  93  96   99 102
#> 8    2013        Twins       101 103 101    102    101    106  98  98   99 101
#> 9    2013      Yankees       101 100 104    100     97     84 110 101  101  99
#> 10   2013    Athletics        97  97  95     99    100    105  93  97  101  99
#> 11   2013     Mariners        96  95 100     98     96     85  99 103   99  97
#> 12   2013         Rays        96  97  97     99     95     98  96 102  100  98
#> 13   2013      Rangers       106 103  99    103    101    110 105  98  102 101
#> 14   2013    Blue Jays       102 103 105     97    108    105 106 102   99  99
#> 15   2013 Diamondbacks       104 104  99     99    105    120 102 100   99 100
#> 16   2013       Braves        99  99  98    100     99     93  96 103  101  99
#> 17   2013         Cubs       100 102 109     99     99    102 102 101  102 100
#> 18   2013         Reds       102 101  99     99     98     92 113 103  101  99
#> 19   2013      Rockies       117 120 113    110    108    128 113  95  102 106
#> 20   2013      Marlins        98 101 101     99    100    109  89  99  100 100
#> 21   2013       Astros       100 100 103     99    100    108 105 103  101 100
#> 22   2013      Dodgers        96  94  94     98     98     78 102 100   96  99
#> 23   2013      Brewers       104 104 105     99    100    102 112 101  102  99
#> 24   2013    Nationals       100 102 101    104    102     84  97  97   98 100
#> 25   2013         Mets        95  94  94     96     95     85 102 102  101  97
#> 26   2013     Phillies       100 100 105     98     99     92 107 103  102  99
#> 27   2013      Pirates        96  94  96    101     99     89  90  96   96 102
#> 28   2013    Cardinals        98  99  95    100     99     95  94  98   99 100
#> 29   2013       Padres        95  94  92     98     96     96  96 101  100 100
#> 30   2013       Giants        92  92  94     99     97    115  84 100  100 101
#>     FB  LD IFFB FIP
#> 1   99  97   97  97
#> 2  101 101   99 104
#> 3   97 105  102  98
#> 4  102  99  104 103
#> 5   97 101   95 100
#> 6  103  98  102 100
#> 7  101 101   96  99
#> 8   99 103  101 100
#> 9  102  99  102 103
#> 10 102 100  108  98
#> 11 102  97  107  98
#> 12 101  98  108  98
#> 13 101 103   96 103
#> 14 101 101   99 101
#> 15  99  99   96 100
#> 16 100 101   97  97
#> 17 100  99   99 101
#> 18 100  99  101 103
#> 19  99 109   94 108
#> 20 100  99   96  97
#> 21  99  98  100 101
#> 22 100  97  106  99
#> 23 101 102   96 104
#> 24 101 101   99  99
#> 25 101  98  108 100
#> 26 101  97  104 102
#> 27  98 101   97  97
#> 28 101 100  101  99
#> 29  98  99   97  98
#> 30  95  98   96  94

fg_park_hand(2013)

#> Data courtey of FanGraphs.com. Please consider supporting FanGraphs by purchasing a membership: https://plus.fangraphs.com/product/fangraphs-membership/?switch-subscription=254671&item=85029&_wcsnonce=3e893e9b53&auto-switch=true

#>    season    home_team single_as_LHH single_as_RHH double_as_LHH double_as_RHH
#> 1    2013       Angels            98           100            95            96
#> 2    2013      Orioles           102           102           101            99
#> 3    2013      Red Sox           104           103           121           110
#> 4    2013    White Sox            97           101            99            96
#> 5    2013      Indians           102            99           104           104
#> 6    2013       Tigers           102           101            93           102
#> 7    2013       Royals           100           101           104           101
#> 8    2013        Twins           101           103            97           104
#> 9    2013      Yankees           100            99            98            95
#> 10   2013    Athletics            99           100            99           101
#> 11   2013     Mariners            98            98            96            96
#> 12   2013         Rays           100            97            96            93
#> 13   2013      Rangers           103           102           100           100
#> 14   2013    Blue Jays            97            96           110           106
#> 15   2013 Diamondbacks            99            99           105           105
#> 16   2013       Braves            99           100           101            97
#> 17   2013         Cubs           101            98           100            98
#> 18   2013         Reds            99           100            96           100
#> 19   2013      Rockies           109           111           106           111
#> 20   2013      Marlins           100            98            95           102
#> 21   2013       Astros            98            99            99           101
#> 22   2013      Dodgers            97            97            98            98
#> 23   2013      Brewers           102            97            99           100
#> 24   2013    Nationals           104           104           102           103
#> 25   2013         Mets            95            96            96            94
#> 26   2013     Phillies            98            99           101           100
#> 27   2013      Pirates            99           102            98            99
#> 28   2013    Cardinals           100           100            96           102
#> 29   2013       Padres            98            98           100            93
#> 30   2013       Giants           100            97            97            96
#>    triple_as_LHH triple_as_RHH hr_as_LHH hr_as_RHH
#> 1             91            90        91        94
#> 2             85            89       115       106
#> 3            124            87        89       104
#> 4            104            82       107       109
#> 5             80            83       106        95
#> 6            124           129        98        99
#> 7            122           107        93        93
#> 8            110           102        94       102
#> 9             83            88       115       105
#> 10           114            96        90        95
#> 11            82            90       100        98
#> 12            95           100       100        92
#> 13           108           112       106       104
#> 14           102           107       103       107
#> 15           118           121       103       101
#> 16            91            95        97        95
#> 17            97           107        97       106
#> 18            94            91       112       114
#> 19           133           125       115       114
#> 20           108           108        87        89
#> 21           106           110       105       105
#> 22            80            75       104        99
#> 23            92           109       114       110
#> 24            85            84        96        99
#> 25            88            82       102       102
#> 26            96            89       107       108
#> 27            84            94        92        88
#> 28            96            95        95        93
#> 29            86           109        96        95
#> 30           120           108        81        86