Scrape Pitcher Game Logs from FanGraphs

This function allows you to scrape game logs by year for a pitcher from FanGraphs.com.

Usage

fg_pitcher_game_logs(playerid, year = 2017)

Arguments

playerid

This is the playerid used by FanGraphs for a given player

year

The season for which game logs should be returned (use the YYYY format)

Value

Returns a tibble of pitcher game logs with the following columns:

col_name	types

PlayerName	character

playerid	integer

Date	character

Opp	character

teamid	integer

season	integer

Team	character

HomeAway	character

Age	integer

W	numeric

L	numeric

ERA	numeric

G	numeric

GS	numeric

CG	numeric

ShO	numeric

SV	numeric

HLD	numeric

BS	numeric

IP	numeric

TBF	numeric

H	numeric

R	numeric

ER	numeric

HR	numeric

BB	numeric

IBB	numeric

HBP	numeric

WP	numeric

BK	numeric

SO	numeric

K/9	numeric

BB/9	numeric

H/9	numeric

K/BB	numeric

IFH%	numeric

BUH%	numeric

GB	numeric

FB	numeric

LD	numeric

IFFB	numeric

IFH	numeric

BU	numeric

BUH	numeric

K%	numeric

BB%	numeric

K-BB%	numeric

SIERA	numeric

HR/9	numeric

AVG	numeric

WHIP	numeric

BABIP	numeric

LOB%	numeric

FIP	numeric

E-F	numeric

xFIP	numeric

ERA-	numeric

FIP-	numeric

xFIP-	numeric

GB/FB	numeric

LD%	numeric

GB%	numeric

FB%	numeric

IFFB%	numeric

HR/FB	numeric

RS	numeric

RS/9	numeric

Balls	numeric

Strikes	numeric

Pitches	numeric

WPA	numeric

-WPA	numeric

+WPA	numeric

RE24	numeric

REW	numeric

pLI	numeric

inLI	numeric

gmLI	numeric

exLI	numeric

Pulls	numeric

Games	numeric

WPA/LI	numeric

Clutch	numeric

SD	numeric

MD	numeric

FB%1	numeric

FBv	numeric

SL%	numeric

SLv	numeric

CT%	numeric

CTv	numeric

CB%	numeric

CBv	numeric

CH%	numeric

CHv	numeric

XX%	numeric

PO%	numeric

wFB	numeric

wSL	numeric

wCT	numeric

wCB	numeric

wCH	numeric

wFB/C	numeric

wSL/C	numeric

wCT/C	numeric

wCB/C	numeric

wCH/C	numeric

O-Swing%	numeric

Z-Swing%	numeric

Swing%	numeric

O-Contact%	numeric

Z-Contact%	numeric

Contact%	numeric

Zone%	numeric

F-Strike%	numeric

SwStr%	numeric

Pull	numeric

Cent	numeric

Oppo	numeric

Soft	numeric

Med	numeric

Hard	numeric

bipCount	numeric

Pull%	numeric

Cent%	numeric

Oppo%	numeric

Soft%	numeric

Med%	numeric

Hard%	numeric

tERA	numeric

GSv2	numeric

Events	numeric

gamedate	character

dh	integer

Examples

# \donttest{
  try(fg_pitcher_game_logs(playerid = 104, year = 2006))
#> ── MLB Pitcher Game Log data from FanGraphs.com ───── baseballr 1.4.0 ──
#> ℹ Data updated: 2022-12-13 21:35:18 UTC
#> # A tibble: 34 × 134
#>    PlayerN…¹ playe…² Date  Opp   teamid season Team  HomeA…³   Age     W
#>    <chr>       <int> <chr> <chr>  <int>  <int> <chr> <chr>   <int> <dbl>
#>  1 Greg Mad…     104 2006… @SFG      22   2006 LAD   A          40     1
#>  2 Greg Mad…     104 2006… @COL      22   2006 LAD   A          40     1
#>  3 Greg Mad…     104 2006… PIT       22   2006 LAD   H          40     0
#>  4 Greg Mad…     104 2006… SDP       22   2006 LAD   H          40     1
#>  5 Greg Mad…     104 2006… @NYM      22   2006 LAD   A          40     0
#>  6 Greg Mad…     104 2006… @MIL      22   2006 LAD   A          40     0
#>  7 Greg Mad…     104 2006… CIN       22   2006 LAD   H          40     1
#>  8 Greg Mad…     104 2006… @ARI      22   2006 LAD   A          40     0
#>  9 Greg Mad…     104 2006… @SFG      22   2006 LAD   A          40     1
#> 10 Greg Mad…     104 2006… SFG       22   2006 LAD   H          40     0
#> # … with 24 more rows, 124 more variables: L <dbl>, ERA <dbl>, G <dbl>,
#> #   GS <dbl>, CG <dbl>, ShO <dbl>, SV <dbl>, HLD <dbl>, BS <dbl>,
#> #   IP <dbl>, TBF <dbl>, H <dbl>, R <dbl>, ER <dbl>, HR <dbl>,
#> #   BB <dbl>, IBB <dbl>, HBP <dbl>, WP <dbl>, BK <dbl>, SO <dbl>,
#> #   `K/9` <dbl>, `BB/9` <dbl>, `H/9` <dbl>, `K/BB` <dbl>, `IFH%` <dbl>,
#> #   `BUH%` <dbl>, GB <dbl>, FB <dbl>, LD <dbl>, IFFB <dbl>, IFH <dbl>,
#> #   BU <dbl>, BUH <dbl>, `K%` <dbl>, `BB%` <dbl>, `K-BB%` <dbl>, …
# }