Scrape Team Pitcher Leaderboards from FanGraphs

Scrape Team Pitcher Leaderboards from FanGraphs

Usage

fg_team_pitcher(
  age = "",
  pos = "all",
  stats = "pit",
  lg = "all",
  qual = "y",
  startseason = "2023",
  endseason = "2023",
  startdate = "",
  enddate = "",
  month = "0",
  hand = "",
  team = "0,ts",
  pageitems = "1000",
  pagenum = "1",
  ind = "0",
  rost = "0",
  players = "",
  type = "8",
  postseason = "",
  sortdir = "default",
  sortstat = "WAR"
)

Arguments

age

(integer) Age of players

pos

(character) Position of players, defaults to "all". To exclude pitchers, use "np".

stats

(character) Statistic to return. Defaults to "bat".

lg

(character) League to return. Defaults to "all". Options are "al", "nl", or "all".

qual

(character) Whether you want only batters/pitchers that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimum number of plate appearaces/innings pitched, use the number desired. Defaults to "y".

startseason

(character) Season for which you want to scrape the data.

endseason

(character) Last season for which you want data.

startdate

(character) Start date for which you want data.

enddate

(character) End date for which you want data.

month

(character) Month for which you want data.

hand

(character) Handedness of batter. Options are "L", "R", or "B". Empty string returns all.

team

(character) Teams for which you want data, comma separated.

pageitems

(character) Number of items per page.

pagenum

(character) Page number.

ind

(character) Whether or not to break the seasons out individual, or roll them up together. 1 = split seasons, 0 = aggregate seasons.

rost

(character) Whether or not to include players on the roster. 1 = include, 0 = exclude.

players

(character) Whether or not to include players on the roster. 1 = include only active roster players, 0 = exclude.

type

(character) Defaults to 8, which is the standard leaderboard. The values for the leaderboards appear to go to from type = 0 to 48+, which correspond to links on the leaderboard page.

postseason

(logical) Whether or not to include postseason data. TRUE = include postseason, FALSE = exclude postseason.

sortdir

(character) Sort direction. Options are "asc" or "desc" or "default".

sortstat

(character) Sort by stat. Default is "WAR".

Value

A data frame of pitcher data.

col_name	types

Season	integer

team_name	character

SeasonMin	integer

SeasonMax	integer

W	integer

L	integer

ERA	numeric

G	integer

GS	integer

CG	integer

ShO	integer

SV	integer

BS	integer

IP	numeric

TBF	integer

H	integer

R	integer

ER	integer

HR	integer

BB	integer

IBB	integer

HBP	integer

WP	integer

BK	integer

SO	integer

GB	integer

FB	integer

LD	integer

IFFB	integer

Pitches	integer

Balls	integer

Strikes	integer

RS	integer

IFH	integer

BU	integer

BUH	integer

K_9	numeric

BB_9	numeric

K_BB	numeric

H_9	numeric

HR_9	numeric

AVG	numeric

WHIP	numeric

BABIP	numeric

LOB_pct	numeric

FIP	numeric

GB_FB	numeric

LD_pct	numeric

GB_pct	numeric

FB_pct	numeric

IFFB_pct	numeric

HR_FB	numeric

IFH_pct	numeric

BUH_pct	numeric

TTO_pct	numeric

CFraming	numeric

Starting	numeric

Start_IP	numeric

Relieving	numeric

Relief_IP	numeric

RAR	numeric

WAR	numeric

Dollars	numeric

RA9-Wins	numeric

LOB-Wins	numeric

BIP-Wins	numeric

BS-Wins	numeric

tERA	numeric

xFIP	numeric

WPA	numeric

WPA_minus	numeric

WPA_plus	numeric

RE24	numeric

REW	numeric

pLI	numeric

inLI	numeric

gmLI	numeric

exLI	numeric

Pulls	integer

Games	integer

WPA_LI	numeric

Clutch	numeric

FBall_pct	numeric

FBv	numeric

SL_pct	numeric

SLv	numeric

CT_pct	numeric

CTv	numeric

CB_pct	numeric

CBv	numeric

CH_pct	numeric

CHv	numeric

SF_pct	numeric

SFv	numeric

XX_pct	numeric

wFB	numeric

wSL	numeric

wCT	numeric

wCB	numeric

wCH	numeric

wSF	numeric

wFB_C	numeric

wSL_C	numeric

wCT_C	numeric

wCB_C	numeric

wCH_C	numeric

wSF_C	numeric

O-Swing_pct	numeric

Z-Swing_pct	numeric

Swing_pct	numeric

O-Contact_pct	numeric

Z-Contact_pct	numeric

Contact_pct	numeric

Zone_pct	numeric

F-Strike_pct	numeric

SwStr_pct	numeric

CStr_pct	numeric

C+SwStr_pct	numeric

HLD	integer

SD	integer

MD	integer

ERA-	numeric

FIP-	numeric

xFIP-	numeric

K_pct	numeric

BB_pct	numeric

K-BB_pct	numeric

SIERA	numeric

kwERA	numeric

RS_9	numeric

E-F	numeric

Pull	integer

Cent	integer

Oppo	integer

Soft	integer

Med	integer

Hard	integer

bipCount	integer

Pull_pct	numeric

Cent_pct	numeric

Oppo_pct	numeric

Soft_pct	numeric

Med_pct	numeric

Hard_pct	numeric

K_9+	numeric

BB_9+	numeric

K_BB+	numeric

H_9+	numeric

HR_9+	numeric

AVG+	numeric

WHIP+	numeric

BABIP+	numeric

LOB_pct+	numeric

K_pct+	numeric

BB_pct+	numeric

LD_pct+	numeric

GB_pct+	numeric

FB_pct+	numeric

HRFB_pct+	numeric

Pull_pct+	numeric

Cent_pct+	numeric

Oppo_pct+	numeric

Soft_pct+	numeric

Med_pct+	numeric

Hard_pct+	numeric

rFTeamV	integer

rBTeamV	integer

rTV	integer

pfx_FA_pct	numeric

pfx_FT_pct	numeric

pfx_FC_pct	numeric

pfx_FS_pct	numeric

pfx_SI_pct	numeric

pfx_SL_pct	numeric

pfx_CU_pct	numeric

pfx_KC_pct	numeric

pfx_CH_pct	numeric

pfx_vFA	numeric

pfx_vFT	numeric

pfx_vFC	numeric

pfx_vFS	numeric

pfx_vSI	numeric

pfx_vSL	numeric

pfx_vCU	numeric

pfx_vKC	numeric

pfx_vCH	numeric

pfx_FA-X	numeric

pfx_FT-X	numeric

pfx_FC-X	numeric

pfx_FS-X	numeric

pfx_SI-X	numeric

pfx_SL-X	numeric

pfx_CU-X	numeric

pfx_KC-X	numeric

pfx_CH-X	numeric

pfx_FA-Z	numeric

pfx_FT-Z	numeric

pfx_FC-Z	numeric

pfx_FS-Z	numeric

pfx_SI-Z	numeric

pfx_SL-Z	numeric

pfx_CU-Z	numeric

pfx_KC-Z	numeric

pfx_CH-Z	numeric

pfx_wFA	numeric

pfx_wFT	numeric

pfx_wFC	numeric

pfx_wFS	numeric

pfx_wSI	numeric

pfx_wSL	numeric

pfx_wCU	numeric

pfx_wKC	numeric

pfx_wCH	numeric

pfx_wFA_C	numeric

pfx_wFT_C	numeric

pfx_wFC_C	numeric

pfx_wFS_C	numeric

pfx_wSI_C	numeric

pfx_wSL_C	numeric

pfx_wCU_C	numeric

pfx_wKC_C	numeric

pfx_wCH_C	numeric

pfx_O-Swing_pct	numeric

pfx_Z-Swing_pct	numeric

pfx_Swing_pct	numeric

pfx_O-Contact_pct	numeric

pfx_Z-Contact_pct	numeric

pfx_Contact_pct	numeric

pfx_Zone_pct	numeric

pfx_Pace	numeric

pi_CH_pct	numeric

pi_CU_pct	numeric

pi_FA_pct	numeric

pi_FC_pct	numeric

pi_FS_pct	numeric

pi_SI_pct	numeric

pi_SL_pct	numeric

pi_XX_pct	numeric

pi_vCH	numeric

pi_vCU	numeric

pi_vFA	numeric

pi_vFC	numeric

pi_vFS	numeric

pi_vSI	numeric

pi_vSL	numeric

pi_vXX	numeric

pi_CH-X	numeric

pi_CU-X	numeric

pi_FA-X	numeric

pi_FC-X	numeric

pi_FS-X	numeric

pi_SI-X	numeric

pi_SL-X	numeric

pi_XX-X	numeric

pi_CH-Z	numeric

pi_CU-Z	numeric

pi_FA-Z	numeric

pi_FC-Z	numeric

pi_FS-Z	numeric

pi_SI-Z	numeric

pi_SL-Z	numeric

pi_XX-Z	numeric

pi_wCH	numeric

pi_wCU	numeric

pi_wFA	numeric

pi_wFC	numeric

pi_wFS	numeric

pi_wSI	numeric

pi_wSL	numeric

pi_wXX	numeric

pi_wCH_C	numeric

pi_wCU_C	numeric

pi_wFA_C	numeric

pi_wFC_C	numeric

pi_wFS_C	numeric

pi_wSI_C	numeric

pi_wSL_C	numeric

pi_wXX_C	numeric

pi_O-Swing_pct	numeric

pi_Z-Swing_pct	numeric

pi_Swing_pct	numeric

pi_O-Contact_pct	numeric

pi_Z-Contact_pct	numeric

pi_Contact_pct	numeric

pi_Zone_pct	numeric

pi_Pace	numeric

Events	integer

EV	numeric

LA	numeric

Barrels	integer

Barrel_pct	numeric

maxEV	numeric

HardHit	integer

HardHit_pct	numeric

Q	numeric

TG	integer

TIP	numeric

team_name_abb	character

teamid	integer

pfx_EP_pct	numeric

pfx_vEP	numeric

pfx_EP-X	numeric

pfx_EP-Z	numeric

pfx_wEP	numeric

pfx_wEP_C	numeric

pi_SB_pct	numeric

pi_vSB	numeric

pi_SB-X	numeric

pi_SB-Z	numeric

pi_wSB	numeric

pi_wSB_C	numeric

pi_CS_pct	numeric

pi_vCS	numeric

pi_CS-X	numeric

pi_CS-Z	numeric

pi_wCS	numeric

pi_wCS_C	numeric

pi_KN_pct	numeric

pi_vKN	numeric

pi_KN-X	numeric

pi_KN-Z	numeric

pi_wKN	numeric

pi_wKN_C	numeric

KN_pct	numeric

KNv	numeric

wKN	numeric

wKN_C	numeric

pfx_KN_pct	numeric

pfx_vKN	numeric

pfx_KN-X	numeric

pfx_KN-Z	numeric

pfx_wKN	numeric

pfx_wKN_C	numeric

pfx_SC_pct	numeric

pfx_vSC	numeric

pfx_SC-X	numeric

pfx_SC-Z	numeric

pfx_wSC	numeric

pfx_wSC_C	numeric

pfx_FO_pct	numeric

pfx_vFO	numeric

pfx_FO-X	numeric

pfx_FO-Z	numeric

pfx_wFO	numeric

pfx_wFO_C	numeric

Examples

# \donttest{
  try(fg_team_pitcher(startseason = 2015, endseason = 2015, qual = 150))
#> ── MLB Team Pitching data from FanGraphs.com ──────── baseballr 1.6.0 ──
#> ℹ Data updated: 2024-04-13 21:55:10 UTC
#> # A tibble: 30 × 346
#>    Season team_name SeasonMin SeasonMax     W     L   ERA     G    GS
#>     <int> <chr>         <int>     <int> <int> <int> <dbl> <int> <int>
#>  1   2015 CHC            2015      2015    97    65  3.36   714   162
#>  2   2015 WSN            2015      2015    83    79  3.62   630   162
#>  3   2015 CLE            2015      2015    81    80  3.68   637   161
#>  4   2015 STL            2015      2015   100    62  2.94   677   162
#>  5   2015 LAD            2015      2015    92    70  3.46   677   162
#>  6   2015 PIT            2015      2015    98    64  3.23   662   162
#>  7   2015 NYM            2015      2015    90    72  3.45   647   162
#>  8   2015 HOU            2015      2015    86    76  3.57   644   162
#>  9   2015 NYY            2015      2015    87    75  4.03   659   162
#> 10   2015 CHW            2015      2015    76    86  3.98   576   162
#> # ℹ 20 more rows
#> # ℹ 337 more variables: QS <int>, CG <int>, ShO <int>, SV <int>,
#> #   BS <int>, IP <dbl>, TBF <int>, H <int>, R <int>, ER <int>,
#> #   HR <int>, BB <int>, IBB <int>, HBP <int>, WP <int>, BK <int>,
#> #   SO <int>, GB <int>, FB <int>, LD <int>, IFFB <int>, Pitches <int>,
#> #   Balls <int>, Strikes <int>, RS <int>, IFH <int>, BU <int>,
#> #   BUH <int>, K_9 <dbl>, BB_9 <dbl>, K_BB <dbl>, H_9 <dbl>, …
# }