Label Statcast data as imputed

Based on a series of heuristics, this function attempts to label Statcast data for which the launch angle and speed have been imputed.

label_statcast_imputed_data(
  statcast_df,
  impute_file = NULL,
  inverse_precision = 10000
)

Arguments

statcast_df
A dataframe containing Statcast batted ball data

impute_file
A CSV file giving the launch angle, launch speed, bb_type, events fields to label as imputed. if NULL then it's read from the extdata folder of the package.

inverse_precision
inverse of how many digits to truncate the launch angle and speed to for comparison. Default is 10000, i.e. keep 4 digits of precision.

Value

A copy of the input dataframe with a new column imputed appended. imputed is 1 if launch angle and launch speed are likely imputed, 0 otherwise.

Examples

# \donttest{
statcast_df <- scrape_statcast_savant("2017-05-01", "2017-05-02")
#> 2017-05-01 is not a date. Attempting to coerce...
#> No playerid specified. Collecting data for all batters/pitchers.
#> https://baseballsavant.mlb.com/statcast_search/csv?all=true&hfPT=&hfAB=&hfBBT=&hfPR=&hfZ=&stadium=&hfBBL=&hfNewZones=&hfGT=R%7CPO%7CS%7C&hfC&hfSea=2017%7C&hfSit=&hfOuts=&opponent=&pitcher_throws=&batter_stands=&hfSA=&player_type=batter&hfInfield=&team=&position=&hfOutfield=&hfRO=&home_road=&game_date_gt=2017-05-01&game_date_lt=2017-05-02&hfFlag=&hfPull=&metric_1=&hfInn=&min_pitches=0&min_results=0&group_by=name&sort_col=pitches&player_event_sort=h_launch_speed&sort_order=desc&min_abs=0&type=details
#> These data are from BaseballSavant and are property of MLB Advanced Media, L.P. All rights reserved.
#> Grabbing data, this may take a minute...
#> URL read and payload acquired successfully.
sc_df <- label_statcast_imputed_data(statcast_df)
mean(sc_df$imputed)
#> [1] 0.003561888
# }