baseballr Data Loader Overview

load_umpire_ids(): Download a data frame of all umpires and their mlbamids for games since 2008.

load_game_info_sup(): Download a data frame of supplemental data about MLB games since 2008.

load_ncaa_baseball_pbp(): Load cleaned NCAA baseball play-by-play data from the baseballr data repo.

load_ncaa_baseball_schedule(): Load cleaned NCAA baseball schedule from the baseballr data repo.

load_ncaa_baseball_season_ids(): Load cleaned NCAA men's college baseball season IDs from the baseballr data repo.

load_ncaa_baseball_teams(): Load cleaned NCAA men's college baseball teams from the baseballr data repo.

Details

baseballr Data

These functions load cleaned MLB and NCAA baseball data from the baseballr data repository.