Acquire time codes for Major and Minor League games

Acquire time codes for Major and Minor League games

Usage

mlb_game_changes(updated_since, sport_id)

Arguments

updated_since

Updated since date time

sport_id

Return division(s) for all divisions in a specific sport.

Value

Returns a tibble that includes time codes from the game_pk requested

col_name	types

date	character

total_items	integer

total_events	integer

total_games	integer

total_games_in_progress	integer

game_pk	integer

link	character

game_type	character

season	character

game_date	character

official_date	character

is_tie	logical

game_number	integer

public_facing	logical

double_header	character

gameday_type	character

tiebreaker	character

calendar_event_id	character

season_display	character

day_night	character

description	character

scheduled_innings	integer

reverse_home_away_status	logical

inning_break_length	integer

games_in_series	integer

series_game_number	integer

series_description	character

record_source	character

if_necessary	character

if_necessary_description	character

status_abstract_game_state	character

status_coded_game_state	character

status_detailed_state	character

status_status_code	character

status_start_time_tbd	logical

status_abstract_game_code	character

teams_away_score	integer

teams_away_is_winner	logical

teams_away_split_squad	logical

teams_away_series_number	integer

teams_away_league_record_wins	integer

teams_away_league_record_losses	integer

teams_away_league_record_pct	character

teams_away_team_id	integer

teams_away_team_name	character

teams_away_team_link	character

teams_home_score	integer

teams_home_is_winner	logical

teams_home_split_squad	logical

teams_home_series_number	integer

teams_home_league_record_wins	integer

teams_home_league_record_losses	integer

teams_home_league_record_pct	character

teams_home_team_id	integer

teams_home_team_name	character

teams_home_team_link	character

venue_id	integer

venue_name	character

venue_link	character

content_link	character

status_reason	character

rescheduled_from	character

rescheduled_from_date	character

resumed_from	character

resumed_from_date	character

events	list

Examples

# \donttest{
  try(mlb_game_changes(updated_since = "2021-08-10T19:08:24.000004Z", sport_id = 1))
#> ── MLB Game Changes data from MLB.com ─────────────── baseballr 1.3.0 ──
#> ℹ Data updated: 2022-10-05 17:36:19 UTC
#> # A tibble: 1,000 × 66
#>    date     total…¹ total…² total…³ total…⁴ game_pk link  game_…⁵ season
#>    <chr>      <int>   <int>   <int>   <int>   <int> <chr> <chr>   <chr> 
#>  1 1999-04…       1       0       1       0     916 /api… R       1999  
#>  2 1999-04…       1       0       1       0     200 /api… R       1999  
#>  3 1999-04…       1       0       1       0    2914 /api… R       1999  
#>  4 1999-04…       1       0       1       0    1404 /api… R       1999  
#>  5 1999-04…       1       0       1       0    1504 /api… R       1999  
#>  6 1999-04…       2       0       2       0     526 /api… R       1999  
#>  7 1999-04…       2       0       2       0     602 /api… R       1999  
#>  8 1999-05…       1       0       1       0     529 /api… R       1999  
#>  9 1999-05…       1       0       1       0     690 /api… R       1999  
#> 10 1999-05…       1       0       1       0    1317 /api… R       1999  
#> # … with 990 more rows, 57 more variables: game_date <chr>,
#> #   official_date <chr>, is_tie <lgl>, game_number <int>,
#> #   public_facing <lgl>, double_header <chr>, gameday_type <chr>,
#> #   tiebreaker <chr>, calendar_event_id <chr>, season_display <chr>,
#> #   day_night <chr>, description <chr>, scheduled_innings <int>,
#> #   reverse_home_away_status <lgl>, inning_break_length <int>,
#> #   games_in_series <int>, series_game_number <int>, …
# }