Retrieve game linescores for major and minor league games

Retrieve game linescores for major and minor league games

Usage

mlb_game_linescore(game_pk)

Arguments

game_pk

The unique game_pk identifier for the game

Value

Returns a data frame with the following columns

col_name	types

game_pk	numeric

home_team_id	character

home_team_name	character

away_team_id	character

away_team_name	character

num	integer

ordinal_num	character

home_runs	integer

home_hits	integer

home_errors	integer

home_left_on_base	integer

away_runs	integer

away_hits	integer

away_errors	integer

away_left_on_base	integer

home_team_link	character

home_team_season	character

home_team_venue_id	character

home_team_venue_name	character

home_team_venue_link	character

home_team_team_code	character

home_team_file_code	character

home_team_abbreviation	character

home_team_team_name	character

home_team_location_name	character

home_team_first_year_of_play	character

home_team_league_id	character

home_team_league_name	character

home_team_league_link	character

home_team_division_id	character

home_team_division_name	character

home_team_division_link	character

home_team_sport_id	character

home_team_sport_link	character

home_team_sport_name	character

home_team_short_name	character

home_team_record_games_played	character

home_team_record_wild_card_games_back	character

home_team_record_league_games_back	character

home_team_record_spring_league_games_back	character

home_team_record_sport_games_back	character

home_team_record_division_games_back	character

home_team_record_conference_games_back	character

home_team_record_league_record_wins	character

home_team_record_league_record_losses	character

home_team_record_league_record_pct	character

home_team_record_division_leader	character

home_team_record_wins	character

home_team_record_losses	character

home_team_record_winning_percentage	character

home_team_franchise_name	character

home_team_club_name	character

home_team_all_star_status	character

home_team_active	character

away_team_link	character

away_team_season	character

away_team_venue_id	character

away_team_venue_name	character

away_team_venue_link	character

away_team_team_code	character

away_team_file_code	character

away_team_abbreviation	character

away_team_team_name	character

away_team_location_name	character

away_team_first_year_of_play	character

away_team_league_id	character

away_team_league_name	character

away_team_league_link	character

away_team_division_id	character

away_team_division_name	character

away_team_division_link	character

away_team_sport_id	character

away_team_sport_link	character

away_team_sport_name	character

away_team_short_name	character

away_team_record_games_played	character

away_team_record_wild_card_games_back	character

away_team_record_league_games_back	character

away_team_record_spring_league_games_back	character

away_team_record_sport_games_back	character

away_team_record_division_games_back	character

away_team_record_conference_games_back	character

away_team_record_league_record_wins	character

away_team_record_league_record_losses	character

away_team_record_league_record_pct	character

away_team_record_division_leader	character

away_team_record_wins	character

away_team_record_losses	character

away_team_record_winning_percentage	character

away_team_franchise_name	character

away_team_club_name	character

away_team_all_star_status	character

away_team_active	character

Examples

# \donttest{
  try(mlb_game_linescore(game_pk = 566001))
#>   game_pk home_team_id home_team_name away_team_id  away_team_name num
#> 1  566001          121  New York Mets          113 Cincinnati Reds   1
#> 2  566001          121  New York Mets          113 Cincinnati Reds   2
#> 3  566001          121  New York Mets          113 Cincinnati Reds   3
#> 4  566001          121  New York Mets          113 Cincinnati Reds   4
#> 5  566001          121  New York Mets          113 Cincinnati Reds   5
#> 6  566001          121  New York Mets          113 Cincinnati Reds   6
#> 7  566001          121  New York Mets          113 Cincinnati Reds   7
#> 8  566001          121  New York Mets          113 Cincinnati Reds   8
#> 9  566001          121  New York Mets          113 Cincinnati Reds   9
#>   ordinal_num home_runs home_hits home_errors home_left_on_base
#> 1         1st         0         0           0                 0
#> 2         2nd         2         3           0                 1
#> 3         3rd         0         0           0                 2
#> 4         4th         2         1           0                 3
#> 5         5th         0         0           0                 0
#> 6         6th         0         1           0                 2
#> 7         7th         0         0           0                 0
#> 8         8th         0         1           0                 2
#> 9         9th         0         0           0                 0
#>   away_runs away_hits away_errors away_left_on_base
#> 1         0         1           0                 1
#> 2         4         3           0                 2
#> 3         0         0           0                 0
#> 4         0         1           0                 1
#> 5         0         1           0                 1
#> 6         0         1           0                 0
#> 7         0         1           0                 2
#> 8         0         0           0                 2
#> 9         1         1           0                 0
#>   home_team_all_star_status    home_team_link home_team_season
#> 1                         N /api/v1/teams/121             2019
#> 2                         N /api/v1/teams/121             2019
#> 3                         N /api/v1/teams/121             2019
#> 4                         N /api/v1/teams/121             2019
#> 5                         N /api/v1/teams/121             2019
#> 6                         N /api/v1/teams/121             2019
#> 7                         N /api/v1/teams/121             2019
#> 8                         N /api/v1/teams/121             2019
#> 9                         N /api/v1/teams/121             2019
#>   home_team_venue_id home_team_venue_name home_team_venue_link
#> 1               3289           Citi Field  /api/v1/venues/3289
#> 2               3289           Citi Field  /api/v1/venues/3289
#> 3               3289           Citi Field  /api/v1/venues/3289
#> 4               3289           Citi Field  /api/v1/venues/3289
#> 5               3289           Citi Field  /api/v1/venues/3289
#> 6               3289           Citi Field  /api/v1/venues/3289
#> 7               3289           Citi Field  /api/v1/venues/3289
#> 8               3289           Citi Field  /api/v1/venues/3289
#> 9               3289           Citi Field  /api/v1/venues/3289
#>   home_team_team_code home_team_file_code home_team_abbreviation
#> 1                 nyn                 nym                    NYM
#> 2                 nyn                 nym                    NYM
#> 3                 nyn                 nym                    NYM
#> 4                 nyn                 nym                    NYM
#> 5                 nyn                 nym                    NYM
#> 6                 nyn                 nym                    NYM
#> 7                 nyn                 nym                    NYM
#> 8                 nyn                 nym                    NYM
#> 9                 nyn                 nym                    NYM
#>   home_team_team_name home_team_location_name
#> 1                Mets                Flushing
#> 2                Mets                Flushing
#> 3                Mets                Flushing
#> 4                Mets                Flushing
#> 5                Mets                Flushing
#> 6                Mets                Flushing
#> 7                Mets                Flushing
#> 8                Mets                Flushing
#> 9                Mets                Flushing
#>   home_team_first_year_of_play home_team_league_id
#> 1                         1962                 104
#> 2                         1962                 104
#> 3                         1962                 104
#> 4                         1962                 104
#> 5                         1962                 104
#> 6                         1962                 104
#> 7                         1962                 104
#> 8                         1962                 104
#> 9                         1962                 104
#>   home_team_league_name home_team_league_link home_team_division_id
#> 1       National League    /api/v1/league/104                   204
#> 2       National League    /api/v1/league/104                   204
#> 3       National League    /api/v1/league/104                   204
#> 4       National League    /api/v1/league/104                   204
#> 5       National League    /api/v1/league/104                   204
#> 6       National League    /api/v1/league/104                   204
#> 7       National League    /api/v1/league/104                   204
#> 8       National League    /api/v1/league/104                   204
#> 9       National League    /api/v1/league/104                   204
#>   home_team_division_name home_team_division_link home_team_sport_id
#> 1    National League East   /api/v1/divisions/204                  1
#> 2    National League East   /api/v1/divisions/204                  1
#> 3    National League East   /api/v1/divisions/204                  1
#> 4    National League East   /api/v1/divisions/204                  1
#> 5    National League East   /api/v1/divisions/204                  1
#> 6    National League East   /api/v1/divisions/204                  1
#> 7    National League East   /api/v1/divisions/204                  1
#> 8    National League East   /api/v1/divisions/204                  1
#> 9    National League East   /api/v1/divisions/204                  1
#>   home_team_sport_link  home_team_sport_name home_team_short_name
#> 1     /api/v1/sports/1 Major League Baseball              NY Mets
#> 2     /api/v1/sports/1 Major League Baseball              NY Mets
#> 3     /api/v1/sports/1 Major League Baseball              NY Mets
#> 4     /api/v1/sports/1 Major League Baseball              NY Mets
#> 5     /api/v1/sports/1 Major League Baseball              NY Mets
#> 6     /api/v1/sports/1 Major League Baseball              NY Mets
#> 7     /api/v1/sports/1 Major League Baseball              NY Mets
#> 8     /api/v1/sports/1 Major League Baseball              NY Mets
#> 9     /api/v1/sports/1 Major League Baseball              NY Mets
#>   home_team_record_games_played home_team_record_wild_card_games_back
#> 1                            28                                     -
#> 2                            28                                     -
#> 3                            28                                     -
#> 4                            28                                     -
#> 5                            28                                     -
#> 6                            28                                     -
#> 7                            28                                     -
#> 8                            28                                     -
#> 9                            28                                     -
#>   home_team_record_league_games_back
#> 1                                  -
#> 2                                  -
#> 3                                  -
#> 4                                  -
#> 5                                  -
#> 6                                  -
#> 7                                  -
#> 8                                  -
#> 9                                  -
#>   home_team_record_spring_league_games_back
#> 1                                         -
#> 2                                         -
#> 3                                         -
#> 4                                         -
#> 5                                         -
#> 6                                         -
#> 7                                         -
#> 8                                         -
#> 9                                         -
#>   home_team_record_sport_games_back
#> 1                                 -
#> 2                                 -
#> 3                                 -
#> 4                                 -
#> 5                                 -
#> 6                                 -
#> 7                                 -
#> 8                                 -
#> 9                                 -
#>   home_team_record_division_games_back
#> 1                                    -
#> 2                                    -
#> 3                                    -
#> 4                                    -
#> 5                                    -
#> 6                                    -
#> 7                                    -
#> 8                                    -
#> 9                                    -
#>   home_team_record_conference_games_back
#> 1                                      -
#> 2                                      -
#> 3                                      -
#> 4                                      -
#> 5                                      -
#> 6                                      -
#> 7                                      -
#> 8                                      -
#> 9                                      -
#>   home_team_record_league_record_wins
#> 1                                  14
#> 2                                  14
#> 3                                  14
#> 4                                  14
#> 5                                  14
#> 6                                  14
#> 7                                  14
#> 8                                  14
#> 9                                  14
#>   home_team_record_league_record_losses
#> 1                                    14
#> 2                                    14
#> 3                                    14
#> 4                                    14
#> 5                                    14
#> 6                                    14
#> 7                                    14
#> 8                                    14
#> 9                                    14
#>   home_team_record_league_record_ties
#> 1                                   0
#> 2                                   0
#> 3                                   0
#> 4                                   0
#> 5                                   0
#> 6                                   0
#> 7                                   0
#> 8                                   0
#> 9                                   0
#>   home_team_record_league_record_pct home_team_record_division_leader
#> 1                               .500                            FALSE
#> 2                               .500                            FALSE
#> 3                               .500                            FALSE
#> 4                               .500                            FALSE
#> 5                               .500                            FALSE
#> 6                               .500                            FALSE
#> 7                               .500                            FALSE
#> 8                               .500                            FALSE
#> 9                               .500                            FALSE
#>   home_team_record_wins home_team_record_losses
#> 1                    14                      14
#> 2                    14                      14
#> 3                    14                      14
#> 4                    14                      14
#> 5                    14                      14
#> 6                    14                      14
#> 7                    14                      14
#> 8                    14                      14
#> 9                    14                      14
#>   home_team_record_winning_percentage home_team_franchise_name
#> 1                                .500                 New York
#> 2                                .500                 New York
#> 3                                .500                 New York
#> 4                                .500                 New York
#> 5                                .500                 New York
#> 6                                .500                 New York
#> 7                                .500                 New York
#> 8                                .500                 New York
#> 9                                .500                 New York
#>   home_team_club_name home_team_active away_team_all_star_status
#> 1                Mets             TRUE                         N
#> 2                Mets             TRUE                         N
#> 3                Mets             TRUE                         N
#> 4                Mets             TRUE                         N
#> 5                Mets             TRUE                         N
#> 6                Mets             TRUE                         N
#> 7                Mets             TRUE                         N
#> 8                Mets             TRUE                         N
#> 9                Mets             TRUE                         N
#>      away_team_link away_team_season away_team_venue_id
#> 1 /api/v1/teams/113             2019               2602
#> 2 /api/v1/teams/113             2019               2602
#> 3 /api/v1/teams/113             2019               2602
#> 4 /api/v1/teams/113             2019               2602
#> 5 /api/v1/teams/113             2019               2602
#> 6 /api/v1/teams/113             2019               2602
#> 7 /api/v1/teams/113             2019               2602
#> 8 /api/v1/teams/113             2019               2602
#> 9 /api/v1/teams/113             2019               2602
#>       away_team_venue_name away_team_venue_link away_team_team_code
#> 1 Great American Ball Park  /api/v1/venues/2602                 cin
#> 2 Great American Ball Park  /api/v1/venues/2602                 cin
#> 3 Great American Ball Park  /api/v1/venues/2602                 cin
#> 4 Great American Ball Park  /api/v1/venues/2602                 cin
#> 5 Great American Ball Park  /api/v1/venues/2602                 cin
#> 6 Great American Ball Park  /api/v1/venues/2602                 cin
#> 7 Great American Ball Park  /api/v1/venues/2602                 cin
#> 8 Great American Ball Park  /api/v1/venues/2602                 cin
#> 9 Great American Ball Park  /api/v1/venues/2602                 cin
#>   away_team_file_code away_team_abbreviation away_team_team_name
#> 1                 cin                    CIN                Reds
#> 2                 cin                    CIN                Reds
#> 3                 cin                    CIN                Reds
#> 4                 cin                    CIN                Reds
#> 5                 cin                    CIN                Reds
#> 6                 cin                    CIN                Reds
#> 7                 cin                    CIN                Reds
#> 8                 cin                    CIN                Reds
#> 9                 cin                    CIN                Reds
#>   away_team_location_name away_team_first_year_of_play
#> 1              Cincinnati                         1882
#> 2              Cincinnati                         1882
#> 3              Cincinnati                         1882
#> 4              Cincinnati                         1882
#> 5              Cincinnati                         1882
#> 6              Cincinnati                         1882
#> 7              Cincinnati                         1882
#> 8              Cincinnati                         1882
#> 9              Cincinnati                         1882
#>   away_team_league_id away_team_league_name away_team_league_link
#> 1                 104       National League    /api/v1/league/104
#> 2                 104       National League    /api/v1/league/104
#> 3                 104       National League    /api/v1/league/104
#> 4                 104       National League    /api/v1/league/104
#> 5                 104       National League    /api/v1/league/104
#> 6                 104       National League    /api/v1/league/104
#> 7                 104       National League    /api/v1/league/104
#> 8                 104       National League    /api/v1/league/104
#> 9                 104       National League    /api/v1/league/104
#>   away_team_division_id away_team_division_name away_team_division_link
#> 1                   205 National League Central   /api/v1/divisions/205
#> 2                   205 National League Central   /api/v1/divisions/205
#> 3                   205 National League Central   /api/v1/divisions/205
#> 4                   205 National League Central   /api/v1/divisions/205
#> 5                   205 National League Central   /api/v1/divisions/205
#> 6                   205 National League Central   /api/v1/divisions/205
#> 7                   205 National League Central   /api/v1/divisions/205
#> 8                   205 National League Central   /api/v1/divisions/205
#> 9                   205 National League Central   /api/v1/divisions/205
#>   away_team_sport_id away_team_sport_link  away_team_sport_name
#> 1                  1     /api/v1/sports/1 Major League Baseball
#> 2                  1     /api/v1/sports/1 Major League Baseball
#> 3                  1     /api/v1/sports/1 Major League Baseball
#> 4                  1     /api/v1/sports/1 Major League Baseball
#> 5                  1     /api/v1/sports/1 Major League Baseball
#> 6                  1     /api/v1/sports/1 Major League Baseball
#> 7                  1     /api/v1/sports/1 Major League Baseball
#> 8                  1     /api/v1/sports/1 Major League Baseball
#> 9                  1     /api/v1/sports/1 Major League Baseball
#>   away_team_short_name away_team_record_games_played
#> 1           Cincinnati                            28
#> 2           Cincinnati                            28
#> 3           Cincinnati                            28
#> 4           Cincinnati                            28
#> 5           Cincinnati                            28
#> 6           Cincinnati                            28
#> 7           Cincinnati                            28
#> 8           Cincinnati                            28
#> 9           Cincinnati                            28
#>   away_team_record_wild_card_games_back
#> 1                                     -
#> 2                                     -
#> 3                                     -
#> 4                                     -
#> 5                                     -
#> 6                                     -
#> 7                                     -
#> 8                                     -
#> 9                                     -
#>   away_team_record_league_games_back
#> 1                                  -
#> 2                                  -
#> 3                                  -
#> 4                                  -
#> 5                                  -
#> 6                                  -
#> 7                                  -
#> 8                                  -
#> 9                                  -
#>   away_team_record_spring_league_games_back
#> 1                                         -
#> 2                                         -
#> 3                                         -
#> 4                                         -
#> 5                                         -
#> 6                                         -
#> 7                                         -
#> 8                                         -
#> 9                                         -
#>   away_team_record_sport_games_back
#> 1                                 -
#> 2                                 -
#> 3                                 -
#> 4                                 -
#> 5                                 -
#> 6                                 -
#> 7                                 -
#> 8                                 -
#> 9                                 -
#>   away_team_record_division_games_back
#> 1                                    -
#> 2                                    -
#> 3                                    -
#> 4                                    -
#> 5                                    -
#> 6                                    -
#> 7                                    -
#> 8                                    -
#> 9                                    -
#>   away_team_record_conference_games_back
#> 1                                      -
#> 2                                      -
#> 3                                      -
#> 4                                      -
#> 5                                      -
#> 6                                      -
#> 7                                      -
#> 8                                      -
#> 9                                      -
#>   away_team_record_league_record_wins
#> 1                                  12
#> 2                                  12
#> 3                                  12
#> 4                                  12
#> 5                                  12
#> 6                                  12
#> 7                                  12
#> 8                                  12
#> 9                                  12
#>   away_team_record_league_record_losses
#> 1                                    16
#> 2                                    16
#> 3                                    16
#> 4                                    16
#> 5                                    16
#> 6                                    16
#> 7                                    16
#> 8                                    16
#> 9                                    16
#>   away_team_record_league_record_ties
#> 1                                   0
#> 2                                   0
#> 3                                   0
#> 4                                   0
#> 5                                   0
#> 6                                   0
#> 7                                   0
#> 8                                   0
#> 9                                   0
#>   away_team_record_league_record_pct away_team_record_division_leader
#> 1                               .429                            FALSE
#> 2                               .429                            FALSE
#> 3                               .429                            FALSE
#> 4                               .429                            FALSE
#> 5                               .429                            FALSE
#> 6                               .429                            FALSE
#> 7                               .429                            FALSE
#> 8                               .429                            FALSE
#> 9                               .429                            FALSE
#>   away_team_record_wins away_team_record_losses
#> 1                    12                      16
#> 2                    12                      16
#> 3                    12                      16
#> 4                    12                      16
#> 5                    12                      16
#> 6                    12                      16
#> 7                    12                      16
#> 8                    12                      16
#> 9                    12                      16
#>   away_team_record_winning_percentage away_team_franchise_name
#> 1                                .429               Cincinnati
#> 2                                .429               Cincinnati
#> 3                                .429               Cincinnati
#> 4                                .429               Cincinnati
#> 5                                .429               Cincinnati
#> 6                                .429               Cincinnati
#> 7                                .429               Cincinnati
#> 8                                .429               Cincinnati
#> 9                                .429               Cincinnati
#>   away_team_club_name away_team_active
#> 1                Reds             TRUE
#> 2                Reds             TRUE
#> 3                Reds             TRUE
#> 4                Reds             TRUE
#> 5                Reds             TRUE
#> 6                Reds             TRUE
#> 7                Reds             TRUE
#> 8                Reds             TRUE
#> 9                Reds             TRUE
# }