Retrieve game pace metrics for major and minor league

Retrieve game pace metrics for major and minor league

Usage

mlb_game_pace(
  season,
  league_ids = NULL,
  sport_ids = NULL,
  team_ids = NULL,
  game_type = NULL,
  venue_ids = NULL,
  org_type = NULL,
  start_date = NULL,
  end_date = NULL
)

Arguments

season

Year for which to return information (Required).

league_ids

The league_id(s) for which to return information.

sport_ids

The sport_id(s) for which to return information.

team_ids

The team_id(s) for which to return information.

game_type

The game_type for which to return information.

venue_ids

Venue directorial information based venue_id.

org_type

pace of game metrics based on team ('T'), league ('L') or sport('S')

start_date

Date of first game for which you want data. Format must be in MM/DD/YYYY format.

end_date

Date of last game for which you want data. Format must be in MM/DD/YYYY format.

Value

Returns a data frame with the following columns

col_name	types

hits_per9inn	numeric

runs_per9inn	numeric

pitches_per9inn	numeric

plate_appearances_per9inn	numeric

hits_per_game	numeric

runs_per_game	numeric

innings_played_per_game	numeric

pitches_per_game	numeric

pitchers_per_game	numeric

plate_appearances_per_game	numeric

total_game_time	character

total_innings_played	integer

total_hits	integer

total_runs	integer

total_plate_appearances	integer

total_pitchers	integer

total_pitches	integer

total_games	integer

total7inn_games	integer

total9inn_games	integer

total_extra_inn_games	integer

time_per_game	character

time_per_pitch	character

time_per_hit	character

time_per_run	character

time_per_plate_appearance	character

time_per9inn	character

time_per77plate_appearances	character

total_extra_inn_time	character

time_per7inn_game	character

time_per7inn_game_without_extra_inn	character

total7inn_games_scheduled	integer

total7inn_games_without_extra_inn	integer

total9inn_games_completed_early	integer

total9inn_games_without_extra_inn	integer

total9inn_games_scheduled	integer

hits_per_run	numeric

pitches_per_pitcher	numeric

season	character

sport_id	integer

sport_code	character

sport_link	character

pr_portal_calculated_fields_total7inn_games	integer

pr_portal_calculated_fields_total9inn_games	integer

pr_portal_calculated_fields_total_extra_inn_games	integer

pr_portal_calculated_fields_time_per7inn_game	character

pr_portal_calculated_fields_time_per9inn_game	character

pr_portal_calculated_fields_time_per_extra_inn_game	character

Examples

# \donttest{
  try(mlb_game_pace(season = 2021, start_date = "09/14/2021", end_date = "09/16/2021"))
#>   hits_per9inn runs_per9inn pitches_per9inn plate_appearances_per9inn
#> 1        17.75        10.16          303.63                     77.36
#> 2        18.30        10.75          308.43                     78.71
#> 3        16.43         9.71          299.52                     76.88
#> 4        15.92         9.25          301.88                     77.14
#> 5        16.80        10.86          312.15                     79.66
#> 6        17.30        12.71          166.23                     81.77
#> 7        18.48        12.13          308.65                     81.40
#>   hits_per_game runs_per_game innings_played_per_game pitches_per_game
#> 1         17.51         10.03                    8.91           299.59
#> 2         17.97         10.55                    8.86           302.87
#> 3         15.49          9.15                    8.50           282.31
#> 4         15.52          9.02                    8.80           294.43
#> 5         16.40         10.60                    8.82           304.77
#> 6         14.40         10.59                    7.52           138.41
#> 7         16.43         10.79                    8.04           274.36
#>   pitchers_per_game plate_appearances_per_game total_game_time
#> 1             10.13                      76.33       130:09:00
#> 2              8.63                      77.29       117:06:00
#> 3              7.08                      72.46       111:57:00
#> 4              7.31                      75.24       127:45:00
#> 5              7.67                      77.78       119:23:00
#> 6              7.56                      68.09       239:05:00
#> 7              8.00                      72.36        39:51:00
#>   total_innings_played total_hits total_runs total_plate_appearances
#> 1                347.5        683        391                    2977
#> 2                336.5        683        401                    2937
#> 3                331.5        604        357                    2826
#> 4                369.5        652        379                    3160
#> 5                353.0        656        424                    3111
#> 6                617.0       1181        868                    5583
#> 7                112.5        230        151                    1013
#>   total_pitchers total_pitches total_games total7inn_games
#> 1            395         11684          39               2
#> 2            328         11509          38               1
#> 3            276         11010          39               6
#> 4            307         12366          42               2
#> 5            307         12191          40               3
#> 6            620         11350          82              53
#> 7            112          3841          14               5
#>   total9inn_games total_extra_inn_games time_per_game time_per_pitch
#> 1              32                     5      03:20:13       00:00:40
#> 2              36                     1      03:04:53       00:00:36
#> 3              32                     1      02:52:13       00:00:36
#> 4              37                     3      03:02:29       00:00:37
#> 5              31                     6      02:59:04       00:00:35
#> 6              28                     4      02:54:56       00:01:15
#> 7               9                    NA      02:50:47       00:00:37
#>   time_per_hit time_per_run time_per_plate_appearance time_per9inn
#> 1     00:11:26     00:19:58                  00:02:37     03:22:55
#> 2     00:10:17     00:17:31                  00:02:23     03:08:17
#> 3     00:11:07     00:18:48                  00:02:22     03:02:43
#> 4     00:11:45     00:20:13                  00:02:25     03:07:07
#> 5     00:10:55     00:16:53                  00:02:18     03:03:24
#> 6     00:12:08     00:16:31                  00:02:34     03:30:05
#> 7     00:10:23     00:15:50                  00:02:21     03:12:08
#>   time_per77plate_appearances total_extra_inn_time time_per7inn_game
#> 1                    03:21:58             20:07:59          02:11:30
#> 2                    03:04:12             03:34:00              <NA>
#> 3                    03:03:01             03:29:00          02:20:20
#> 4                    03:06:46             11:07:00          02:16:59
#> 5                    02:57:17             18:37:59          02:10:59
#> 6                    03:17:50             14:13:59          02:38:26
#> 7                    03:01:44             00:00:00          02:30:36
#>   time_per7inn_game_without_extra_inn total7inn_games_scheduled
#> 1                            04:23:00                         2
#> 2                            00:00:00                        NA
#> 3                            14:01:59                         6
#> 4                            04:33:59                         2
#> 5                            06:33:00                         3
#> 6                           126:45:00                        52
#> 7                            12:33:00                         5
#>   total7inn_games_without_extra_inn total9inn_games_without_extra_inn
#> 1                                 2                                32
#> 2                                NA                                36
#> 3                                 6                                32
#> 4                                 2                                37
#> 5                                 3                                31
#> 6                                48                                28
#> 7                                 5                                 9
#>   total9inn_games_scheduled hits_per_run pitches_per_pitcher season
#> 1                        37        1.747              29.580   2021
#> 2                        38        1.703              35.088   2021
#> 3                        33        1.692              39.891   2021
#> 4                        40        1.720              40.280   2021
#> 5                        37        1.547              39.710   2021
#> 6                        30        1.361              18.306   2021
#> 7                         9        1.523              34.295   2021
#>   total9inn_games_completed_early total7inn_games_completed_early
#> 1                              NA                              NA
#> 2                               1                              NA
#> 3                              NA                              NA
#> 4                              NA                              NA
#> 5                              NA                              NA
#> 6                               1                               1
#> 7                              NA                              NA
#>   sport_id sport_code        sport_link
#> 1        1        mlb  /api/v1/sports/1
#> 2       11        aaa /api/v1/sports/11
#> 3       12        aax /api/v1/sports/12
#> 4       13        afa /api/v1/sports/13
#> 5       14        afx /api/v1/sports/14
#> 6       16        rok /api/v1/sports/16
#> 7       23        ind /api/v1/sports/23
#>   pr_portal_calculated_fields_total7inn_games
#> 1                                           2
#> 2                                           1
#> 3                                           6
#> 4                                           2
#> 5                                           3
#> 6                                          53
#> 7                                           5
#>   pr_portal_calculated_fields_total9inn_games
#> 1                                          32
#> 2                                          36
#> 3                                          32
#> 4                                          37
#> 5                                          31
#> 6                                          28
#> 7                                           9
#>   pr_portal_calculated_fields_total_extra_inn_games
#> 1                                                 5
#> 2                                                 1
#> 3                                                 1
#> 4                                                 3
#> 5                                                 6
#> 6                                                 1
#> 7                                                NA
#>   pr_portal_calculated_fields_time_per7inn_game
#> 1                                      02:11:30
#> 2                                      02:31:59
#> 3                                      02:20:20
#> 4                                      02:16:59
#> 5                                      02:10:59
#> 6                                      02:40:09
#> 7                                      02:30:36
#>   pr_portal_calculated_fields_time_per9inn_game
#> 1                                      03:18:03
#> 2                                      03:05:00
#> 3                                      02:57:03
#> 4                                      03:01:43
#> 5                                      03:02:19
#> 6                                      03:20:27
#> 7                                      03:01:59
#>   pr_portal_calculated_fields_time_per_extra_inn_game
#> 1                                            04:01:35
#> 2                                            03:34:00
#> 3                                            03:29:00
#> 4                                            03:42:20
#> 5                                            03:06:20
#> 6                                            04:03:59
#> 7                                                <NA>
# }