Retrieve game pace metrics for major and minor league

Retrieve game pace metrics for major and minor league

Usage

mlb_game_pace(
  season,
  league_ids = NULL,
  sport_ids = NULL,
  team_ids = NULL,
  game_type = NULL,
  venue_ids = NULL,
  org_type = NULL,
  start_date = NULL,
  end_date = NULL
)

Arguments

season

Year for which to return information (Required).

league_ids

The league_id(s) for which to return information.

sport_ids

The sport_id(s) for which to return information.

team_ids

The team_id(s) for which to return information.

game_type

The game_type for which to return information.

venue_ids

Venue directorial information based venue_id.

org_type

pace of game metrics based on team ('T'), league ('L') or sport('S')

start_date

Date of first game for which you want data. Format must be in MM/DD/YYYY format.

end_date

Date of last game for which you want data. Format must be in MM/DD/YYYY format.

Value

Returns a tibble with the following columns

col_name	types

hits_per9inn	numeric

runs_per9inn	numeric

pitches_per9inn	numeric

plate_appearances_per9inn	numeric

hits_per_game	numeric

runs_per_game	numeric

innings_played_per_game	numeric

pitches_per_game	numeric

pitchers_per_game	numeric

plate_appearances_per_game	numeric

total_game_time	character

total_innings_played	integer

total_hits	integer

total_runs	integer

total_plate_appearances	integer

total_pitchers	integer

total_pitches	integer

total_games	integer

total7inn_games	integer

total9inn_games	integer

total_extra_inn_games	integer

time_per_game	character

time_per_pitch	character

time_per_hit	character

time_per_run	character

time_per_plate_appearance	character

time_per9inn	character

time_per77plate_appearances	character

total_extra_inn_time	character

time_per7inn_game	character

time_per7inn_game_without_extra_inn	character

total7inn_games_scheduled	integer

total7inn_games_without_extra_inn	integer

total9inn_games_completed_early	integer

total9inn_games_without_extra_inn	integer

total9inn_games_scheduled	integer

hits_per_run	numeric

pitches_per_pitcher	numeric

season	character

sport_id	integer

sport_code	character

sport_link	character

pr_portal_calculated_fields_total7inn_games	integer

pr_portal_calculated_fields_total9inn_games	integer

pr_portal_calculated_fields_total_extra_inn_games	integer

pr_portal_calculated_fields_time_per7inn_game	character

pr_portal_calculated_fields_time_per9inn_game	character

pr_portal_calculated_fields_time_per_extra_inn_game	character

Examples

# \donttest{
  try(mlb_game_pace(season = 2021, start_date = "09/14/2021", end_date = "09/16/2021"))
#> ── MLB Game Pace data from MLB.com ────────────────── baseballr 1.6.0 ──
#> ℹ Data updated: 2024-01-09 04:42:01 UTC
#> # A tibble: 7 × 49
#>   hits_per9inn runs_per9inn pitches_per9inn plate_appearances_per9inn
#>          <dbl>        <dbl>           <dbl>                     <dbl>
#> 1         17.8        10.2             304.                      77.4
#> 2         18.3        10.8             308.                      78.7
#> 3         16.4         9.71            300.                      76.9
#> 4         15.9         9.25            302.                      77.1
#> 5         16.8        10.9             312.                      79.7
#> 6         17.3        12.7             166.                      81.8
#> 7         18.5        12.1             309.                      81.4
#> # ℹ 45 more variables: hits_per_game <dbl>, runs_per_game <dbl>,
#> #   innings_played_per_game <dbl>, pitches_per_game <dbl>,
#> #   pitchers_per_game <dbl>, plate_appearances_per_game <dbl>,
#> #   total_game_time <chr>, total_innings_played <dbl>,
#> #   total_hits <int>, total_runs <int>, total_plate_appearances <int>,
#> #   total_pitchers <int>, total_pitches <int>, total_games <int>,
#> #   total7inn_games <int>, total9inn_games <int>, …
# }