Get MLB Game Info by Date and Level

Find game_pk values for professional baseball games (major and minor leagues) via the MLB api http://statsapi.mlb.com/api/

Usage

mlb_game_pks(date, level_ids = c(1))

Arguments

date

The date for which you want to find game_pk values for MLB games

level_ids

A numeric vector with ids for each level where game_pks are desired. See below for a reference of level ids.

Value

Returns a data frame that includes game_pk values and additional information for games scheduled or played |col_name |types | |:------------------------------|:---------| |game_pk |integer | |link |character | |gameType |character | |season |character | |gameDate |character | |officialDate |character | |isTie |logical | |gameNumber |integer | |publicFacing |logical | |doubleHeader |character | |gamedayType |character | |tiebreaker |character | |calendarEventID |character | |seasonDisplay |character | |dayNight |character | |scheduledInnings |integer | |reverseHomeAwayStatus |logical | |inningBreakLength |integer | |gamesInSeries |integer | |seriesGameNumber |integer | |seriesDescription |character | |recordSource |character | |ifNecessary |character | |ifNecessaryDescription |character | |status.abstractGameState |character | |status.codedGameState |character | |status.detailedState |character | |status.statusCode |character | |status.startTimeTBD |logical | |status.abstractGameCode |character | |teams.away.score |integer | |teams.away.isWinner |logical | |teams.away.splitSquad |logical | |teams.away.seriesNumber |integer | |teams.away.leagueRecord.wins |integer | |teams.away.leagueRecord.losses |integer | |teams.away.leagueRecord.pct |character | |teams.away.team.id |integer | |teams.away.team.name |character | |teams.away.team.link |character | |teams.home.score |integer | |teams.home.isWinner |logical | |teams.home.splitSquad |logical | |teams.home.seriesNumber |integer | |teams.home.leagueRecord.wins |integer | |teams.home.leagueRecord.losses |integer | |teams.home.leagueRecord.pct |character | |teams.home.team.id |integer | |teams.home.team.name |character | |teams.home.team.link |character | |venue.id |integer | |venue.name |character | |venue.link |character | |content.link |character |

Details

Level IDs:

The following IDs can be passed to the level_ids argument:

1 = MLB 11 = Triple-A 12 = Doubl-A 13 = Class A Advanced 14 = Class A 15 = Class A Short Season 5442 = Rookie Advanced 16 = Rookie 17 = Winter League

Examples

# \donttest{
  mlb_game_pks("2019-04-29")
#>   game_pk                            link gameType season
#> 1  565909 /api/v1.1/game/565909/feed/live        R   2019
#> 2  566001 /api/v1.1/game/566001/feed/live        R   2019
#> 3  565040 /api/v1.1/game/565040/feed/live        R   2019
#> 4  567173 /api/v1.1/game/567173/feed/live        R   2019
#> 5  566975 /api/v1.1/game/566975/feed/live        R   2019
#> 6  567566 /api/v1.1/game/567566/feed/live        R   2019
#> 7  567275 /api/v1.1/game/567275/feed/live        R   2019
#> 8  565717 /api/v1.1/game/565717/feed/live        R   2019
#> 9  566484 /api/v1.1/game/566484/feed/live        R   2019
#>               gameDate officialDate isTie gameNumber publicFacing
#> 1 2019-04-29T23:05:00Z   2019-04-29 FALSE          1         TRUE
#> 2 2019-04-29T23:10:00Z   2019-04-29 FALSE          1         TRUE
#> 3 2019-04-29T23:10:00Z   2019-04-29 FALSE          1         TRUE
#> 4 2019-04-29T23:20:00Z   2019-04-29 FALSE          1         TRUE
#> 5 2019-04-29T23:40:00Z   2019-04-29 FALSE          1         TRUE
#> 6 2019-04-29T23:40:00Z   2019-04-29 FALSE          1         TRUE
#> 7 2019-04-30T00:10:00Z   2019-04-29 FALSE          1         TRUE
#> 8 2019-04-30T00:15:00Z   2019-04-29 FALSE          1         TRUE
#> 9 2019-04-30T01:45:00Z   2019-04-29 FALSE          1         TRUE
#>   doubleHeader gamedayType tiebreaker      calendarEventID
#> 1            N           P          N 14-565909-2019-04-29
#> 2            N           P          N 14-566001-2019-04-29
#> 3            N           P          N 14-565040-2019-04-29
#> 4            N           P          N 14-567173-2019-04-29
#> 5            N           P          N 14-566975-2019-04-29
#> 6            N           P          N 14-567566-2019-04-29
#> 7            N           P          N 14-567275-2019-04-29
#> 8            N           P          N 14-565717-2019-04-29
#> 9            N           P          N 14-566484-2019-04-29
#>   seasonDisplay dayNight scheduledInnings reverseHomeAwayStatus
#> 1          2019    night                9                 FALSE
#> 2          2019    night                9                 FALSE
#> 3          2019    night                9                 FALSE
#> 4          2019    night                9                 FALSE
#> 5          2019    night                9                 FALSE
#> 6          2019    night                9                 FALSE
#> 7          2019    night                9                 FALSE
#> 8          2019    night                9                 FALSE
#> 9          2019    night                9                 FALSE
#>   inningBreakLength gamesInSeries seriesGameNumber seriesDescription
#> 1               120             4                1    Regular Season
#> 2               120             4                1    Regular Season
#> 3               120             3                1    Regular Season
#> 4               120             4                1    Regular Season
#> 5               120             4                1    Regular Season
#> 6               120             4                1    Regular Season
#> 7               120             3                1    Regular Season
#> 8               120             4                1    Regular Season
#> 9               120             3                1    Regular Season
#>   recordSource ifNecessary ifNecessaryDescription
#> 1            S           N            Normal Game
#> 2            S           N            Normal Game
#> 3            S           N            Normal Game
#> 4            S           N            Normal Game
#> 5            S           N            Normal Game
#> 6            S           N            Normal Game
#> 7            S           N            Normal Game
#> 8            S           N            Normal Game
#> 9            S           N            Normal Game
#>   status.abstractGameState status.codedGameState status.detailedState
#> 1                    Final                     F                Final
#> 2                    Final                     F                Final
#> 3                    Final                     F                Final
#> 4                    Final                     F                Final
#> 5                    Final                     F                Final
#> 6                    Final                     F                Final
#> 7                    Final                     F                Final
#> 8                    Final                     F                Final
#> 9                    Final                     F                Final
#>   status.statusCode status.startTimeTBD status.abstractGameCode
#> 1                 F               FALSE                       F
#> 2                 F               FALSE                       F
#> 3                 F               FALSE                       F
#> 4                 F               FALSE                       F
#> 5                 F               FALSE                       F
#> 6                 F               FALSE                       F
#> 7                 F               FALSE                       F
#> 8                 F               FALSE                       F
#> 9                 F               FALSE                       F
#>   teams.away.score teams.away.isWinner teams.away.splitSquad
#> 1                6                TRUE                 FALSE
#> 2                5                TRUE                 FALSE
#> 3                4               FALSE                 FALSE
#> 4                1               FALSE                 FALSE
#> 5                0               FALSE                 FALSE
#> 6                1               FALSE                 FALSE
#> 7                3               FALSE                 FALSE
#> 8                8                TRUE                 FALSE
#> 9                2               FALSE                 FALSE
#>   teams.away.seriesNumber teams.away.leagueRecord.wins
#> 1                      10                           18
#> 2                      10                           12
#> 3                      11                           14
#> 4                      10                           16
#> 5                      10                           17
#> 6                      10                           13
#> 7                      10                           10
#> 8                      10                           19
#> 9                      10                           19
#>   teams.away.leagueRecord.losses teams.away.leagueRecord.pct
#> 1                             10                        .643
#> 2                             16                        .429
#> 3                             17                        .452
#> 4                             13                        .552
#> 5                             12                        .586
#> 6                             16                        .448
#> 7                             20                        .333
#> 8                              9                        .679
#> 9                             12                        .613
#>   teams.away.team.id teams.away.team.name teams.away.team.link
#> 1                138  St. Louis Cardinals    /api/v1/teams/138
#> 2                113      Cincinnati Reds    /api/v1/teams/113
#> 3                133    Oakland Athletics    /api/v1/teams/133
#> 4                135     San Diego Padres    /api/v1/teams/135
#> 5                117       Houston Astros    /api/v1/teams/117
#> 6                115     Colorado Rockies    /api/v1/teams/115
#> 7                110    Baltimore Orioles    /api/v1/teams/110
#> 8                139       Tampa Bay Rays    /api/v1/teams/139
#> 9                119  Los Angeles Dodgers    /api/v1/teams/119
#>   teams.home.score teams.home.isWinner teams.home.splitSquad
#> 1                3               FALSE                 FALSE
#> 2                4               FALSE                 FALSE
#> 3                9                TRUE                 FALSE
#> 4                3                TRUE                 FALSE
#> 5                1                TRUE                 FALSE
#> 6                5                TRUE                 FALSE
#> 7                5                TRUE                 FALSE
#> 8                5               FALSE                 FALSE
#> 9                3                TRUE                 FALSE
#>   teams.home.seriesNumber teams.home.leagueRecord.wins
#> 1                      10                           12
#> 2                      10                           14
#> 3                      10                           12
#> 4                      10                           14
#> 5                      10                           17
#> 6                      10                           16
#> 7                      10                           12
#> 8                      10                            9
#> 9                      10                           12
#>   teams.home.leagueRecord.losses teams.home.leagueRecord.pct
#> 1                             15                        .444
#> 2                             14                        .500
#> 3                             17                        .414
#> 4                             14                        .500
#> 5                              9                        .654
#> 6                             14                        .533
#> 7                             14                        .462
#> 8                             20                        .310
#> 9                             17                        .414
#>   teams.home.team.id teams.home.team.name teams.home.team.link venue.id
#> 1                120 Washington Nationals    /api/v1/teams/120     3309
#> 2                121        New York Mets    /api/v1/teams/121     3289
#> 3                111       Boston Red Sox    /api/v1/teams/111        3
#> 4                144       Atlanta Braves    /api/v1/teams/144     4705
#> 5                142      Minnesota Twins    /api/v1/teams/142     3312
#> 6                158    Milwaukee Brewers    /api/v1/teams/158       32
#> 7                145    Chicago White Sox    /api/v1/teams/145        4
#> 8                118   Kansas City Royals    /api/v1/teams/118        7
#> 9                137 San Francisco Giants    /api/v1/teams/137     2395
#>              venue.name          venue.link                content.link
#> 1        Nationals Park /api/v1/venues/3309 /api/v1/game/565909/content
#> 2            Citi Field /api/v1/venues/3289 /api/v1/game/566001/content
#> 3           Fenway Park    /api/v1/venues/3 /api/v1/game/565040/content
#> 4         SunTrust Park /api/v1/venues/4705 /api/v1/game/567173/content
#> 5          Target Field /api/v1/venues/3312 /api/v1/game/566975/content
#> 6           Miller Park   /api/v1/venues/32 /api/v1/game/567566/content
#> 7 Guaranteed Rate Field    /api/v1/venues/4 /api/v1/game/567275/content
#> 8      Kauffman Stadium    /api/v1/venues/7 /api/v1/game/565717/content
#> 9           Oracle Park /api/v1/venues/2395 /api/v1/game/566484/content
# }