Acquire time codes for Major and Minor League games

Acquire time codes for Major and Minor League games

Usage

mlb_game_timecodes(game_pk)

Arguments

game_pk

The game_pk for the game requested

Value

Returns a data frame that includes time codes from the game_pk requested

col_name	types

timecodes (MMDDYYYY_HHMMSS)	numeric

Examples

# \donttest{
  try(mlb_game_timecodes(game_pk = 632970))
#>           timecodes
#> 1   20210808_213535
#> 2   20210808_224317
#> 3   20210808_230851
#> 4   20210808_230927
#> 5   20210808_230954
#> 6   20210808_231025
#> 7   20210808_231026
#> 8   20210808_231038
#> 9   20210808_231047
#> 10  20210808_231115
#> 11  20210808_231133
#> 12  20210808_231158
#> 13  20210808_231235
#> 14  20210808_231311
#> 15  20210808_231315
#> 16  20210808_231338
#> 17  20210808_231404
#> 18  20210808_231427
#> 19  20210808_231500
#> 20  20210808_231527
#> 21  20210808_231552
#> 22  20210808_231603
#> 23  20210808_231617
#> 24  20210808_231632
#> 25  20210808_231650
#> 26  20210808_231708
#> 27  20210808_231724
#> 28  20210808_231729
#> 29  20210808_231802
#> 30  20210808_231806
#> 31  20210808_231824
#> 32  20210808_231848
#> 33  20210808_231914
#> 34  20210808_231937
#> 35  20210808_231959
#> 36  20210808_232017
#> 37  20210808_232110
#> 38  20210808_232116
#> 39  20210808_232148
#> 40  20210808_232155
#> 41  20210808_232210
#> 42  20210808_232226
#> 43  20210808_232248
#> 44  20210808_232252
#> 45  20210808_232536
#> 46  20210808_232546
#> 47  20210808_232602
#> 48  20210808_232610
#> 49  20210808_232644
#> 50  20210808_232648
#> 51  20210808_232708
#> 52  20210808_232734
#> 53  20210808_232738
#> 54  20210808_232806
#> 55  20210808_232810
#> 56  20210808_232818
#> 57  20210808_232843
#> 58  20210808_232907
#> 59  20210808_232929
#> 60  20210808_232954
#> 61  20210808_233005
#> 62  20210808_233222
#> 63  20210808_233230
#> 64  20210808_233245
#> 65  20210808_233300
#> 66  20210808_233315
#> 67  20210808_233337
#> 68  20210808_233340
#> 69  20210808_233408
#> 70  20210808_233413
#> 71  20210808_233427
#> 72  20210808_233501
#> 73  20210808_233511
#> 74  20210808_233545
#> 75  20210808_233547
#> 76  20210808_233606
#> 77  20210808_233633
#> 78  20210808_233709
#> 79  20210808_233739
#> 80  20210808_233819
#> 81  20210808_233844
#> 82  20210808_233848
#> 83  20210808_233911
#> 84  20210808_233927
#> 85  20210808_234000
#> 86  20210808_234028
#> 87  20210808_234100
#> 88  20210808_234108
#> 89  20210808_234144
#> 90  20210808_234146
#> 91  20210808_234207
#> 92  20210808_234230
#> 93  20210808_234242
#> 94  20210808_234319
#> 95  20210808_234331
#> 96  20210808_234350
#> 97  20210808_234410
#> 98  20210808_234435
#> 99  20210808_234511
#> 100 20210808_234542
#> 101 20210808_234612
#> 102 20210808_234617
#> 103 20210808_234844
#> 104 20210808_234850
#> 105 20210808_234905
#> 106 20210808_234922
#> 107 20210808_234939
#> 108 20210808_234957
#> 109 20210808_235018
#> 110 20210808_235023
#> 111 20210808_235050
#> 112 20210808_235057
#> 113 20210808_235117
#> 114 20210808_235139
#> 115 20210808_235212
#> 116 20210808_235218
#> 117 20210808_235254
#> 118 20210808_235257
#> 119 20210808_235311
#> 120 20210808_235332
#> 121 20210808_235343
#> 122 20210808_235417
#> 123 20210808_235432
#> 124 20210808_235434
#> 125 20210808_235453
#> 126 20210808_235532
#> 127 20210808_235543
#> 128 20210808_235604
#> 129 20210808_235636
#> 130 20210808_235700
#> 131 20210808_235724
#> 132 20210808_235729
#> 133 20210808_235750
#> 134 20210808_235751
#> 135 20210808_235824
#> 136 20210808_235845
#> 137 20210808_235913
#> 138 20210808_235921
#> 139 20210808_235946
#> 140 20210808_235957
#> 141 20210809_000036
#> 142 20210809_000043
#> 143 20210809_000115
#> 144 20210809_000126
#> 145 20210809_000150
#> 146 20210809_000253
#> 147 20210809_000336
#> 148 20210809_000406
#> 149 20210809_000423
#> 150 20210809_000440
#> 151 20210809_000505
#> 152 20210809_000509
#> 153 20210809_000634
#> 154 20210809_000751
#> 155 20210809_000806
#> 156 20210809_000824
#> 157 20210809_000841
#> 158 20210809_000902
#> 159 20210809_000921
#> 160 20210809_000925
#> 161 20210809_000952
#> 162 20210809_000956
#> 163 20210809_001015
#> 164 20210809_001032
#> 165 20210809_001102
#> 166 20210809_001124
#> 167 20210809_001145
#> 168 20210809_001152
#> 169 20210809_001238
#> 170 20210809_001308
#> 171 20210809_001315
#> 172 20210809_001334
#> 173 20210809_001355
#> 174 20210809_001417
#> 175 20210809_001447
#> 176 20210809_001451
#> 177 20210809_001742
#> 178 20210809_001757
#> 179 20210809_001815
#> 180 20210809_001838
#> 181 20210809_001847
#> 182 20210809_001920
#> 183 20210809_001938
#> 184 20210809_001957
#> 185 20210809_002019
#> 186 20210809_002020
#> 187 20210809_002032
#> 188 20210809_002110
#> 189 20210809_002111
#> 190 20210809_002133
#> 191 20210809_002159
#> 192 20210809_002220
#> 193 20210809_002230
#> 194 20210809_002248
#> 195 20210809_002254
#> 196 20210809_002333
#> 197 20210809_002336
#> 198 20210809_002354
#> 199 20210809_002427
#> 200 20210809_002455
#> 201 20210809_002458
#> 202 20210809_002535
#> 203 20210809_002622
#> 204 20210809_002731
#> 205 20210809_002753
#> 206 20210809_002824
#> 207 20210809_002854
#> 208 20210809_002923
#> 209 20210809_002947
#> 210 20210809_003020
#> 211 20210809_003032
#> 212 20210809_003115
#> 213 20210809_003129
#> 214 20210809_003151
#> 215 20210809_003158
#> 216 20210809_003425
#> 217 20210809_003432
#> 218 20210809_003446
#> 219 20210809_003501
#> 220 20210809_003517
#> 221 20210809_003535
#> 222 20210809_003550
#> 223 20210809_003554
#> 224 20210809_003616
#> 225 20210809_003625
#> 226 20210809_003643
#> 227 20210809_003649
#> 228 20210809_003723
#> 229 20210809_003729
#> 230 20210809_003742
#> 231 20210809_003801
#> 232 20210809_003807
#> 233 20210809_004051
#> 234 20210809_004052
#> 235 20210809_004106
#> 236 20210809_004122
#> 237 20210809_004136
#> 238 20210809_004142
#> 239 20210809_004211
#> 240 20210809_004217
#> 241 20210809_004232
#> 242 20210809_004300
#> 243 20210809_004307
#> 244 20210809_004334
#> 245 20210809_004340
#> 246 20210809_004356
#> 247 20210809_004419
#> 248 20210809_004437
#> 249 20210809_004509
#> 250 20210809_004540
#> 251 20210809_004545
#> 252 20210809_004645
#> 253 20210809_004646
#> 254 20210809_004833
#> 255 20210809_004842
#> 256 20210809_004913
#> 257 20210809_004918
#> 258 20210809_004940
#> 259 20210809_005005
#> 260 20210809_005026
#> 261 20210809_005030
#> 262 20210809_005045
#> 263 20210809_005100
#> 264 20210809_005130
#> 265 20210809_005154
#> 266 20210809_005212
#> 267 20210809_005217
#> 268 20210809_005243
#> 269 20210809_005245
#> 270 20210809_005307
#> 271 20210809_005351
#> 272 20210809_005408
#> 273 20210809_005429
#> 274 20210809_005449
#> 275 20210809_005510
#> 276 20210809_005535
#> 277 20210809_005544
#> 278 20210809_005610
#> 279 20210809_005612
#> 280 20210809_005631
#> 281 20210809_005656
#> 282 20210809_005723
#> 283 20210809_005746
#> 284 20210809_005808
#> 285 20210809_005815
#> 286 20210809_010052
#> 287 20210809_010053
#> 288 20210809_010107
#> 289 20210809_010123
#> 290 20210809_010127
#> 291 20210809_010155
#> 292 20210809_010159
#> 293 20210809_010223
#> 294 20210809_010245
#> 295 20210809_010305
#> 296 20210809_010310
#> 297 20210809_010335
#> 298 20210809_010343
#> 299 20210809_010402
#> 300 20210809_010428
#> 301 20210809_010454
#> 302 20210809_010514
#> 303 20210809_010534
#> 304 20210809_010538
#> 305 20210809_010815
#> 306 20210809_010817
#> 307 20210809_010824
#> 308 20210809_010851
#> 309 20210809_010852
#> 310 20210809_010902
#> 311 20210809_010917
#> 312 20210809_010933
#> 313 20210809_010957
#> 314 20210809_011024
#> 315 20210809_011028
#> 316 20210809_011057
#> 317 20210809_011100
#> 318 20210809_011119
#> 319 20210809_011139
#> 320 20210809_011207
#> 321 20210809_011232
#> 322 20210809_011253
#> 323 20210809_011257
#> 324 20210809_011532
#> 325 20210809_011556
#> 326 20210809_011611
#> 327 20210809_011638
#> 328 20210809_011725
#> 329 20210809_011730
#> 330 20210809_011803
#> 331 20210809_011818
#> 332 20210809_011839
#> 333 20210809_011856
#> 334 20210809_011917
#> 335 20210809_011939
#> 336 20210809_011947
#> 337 20210809_012020
#> 338 20210809_012022
#> 339 20210809_012045
#> 340 20210809_012101
#> 341 20210809_012118
#> 342 20210809_012126
#> 343 20210809_012144
#> 344 20210809_012212
#> 345 20210809_012238
#> 346 20210809_012302
#> 347 20210809_012328
#> 348 20210809_012339
#> 349 20210809_012401
#> 350 20210809_012426
#> 351 20210809_012430
#> 352 20210809_012525
#> 353 20210809_012710
#> 354 20210809_012713
#> 355 20210809_012740
#> 356 20210809_012757
#> 357 20210809_012821
#> 358 20210809_012833
#> 359 20210809_012857
#> 360 20210809_012906
#> 361 20210809_012935
#> 362 20210809_012943
#> 363 20210809_012950
#> 364 20210809_013025
#> 365 20210809_013046
#> 366 20210809_013106
#> 367 20210809_013128
#> 368 20210809_013154
#> 369 20210809_013222
#> 370 20210809_013228
#> 371 20210809_013302
#> 372 20210809_013311
#> 373 20210809_013333
#> 374 20210809_013355
#> 375 20210809_013403
#> 376 20210809_013431
#> 377 20210809_013436
#> 378 20210809_013500
#> 379 20210809_013507
#> 380 20210809_013802
#> 381 20210809_013813
#> 382 20210809_013828
#> 383 20210809_013845
#> 384 20210809_013904
#> 385 20210809_013942
#> 386 20210809_014004
#> 387 20210809_014013
#> 388 20210809_014050
#> 389 20210809_014103
#> 390 20210809_014108
#> 391 20210809_014115
#> 392 20210809_014135
#> 393 20210809_014150
#> 394 20210809_014205
#> 395 20210809_014223
#> 396 20210809_014243
#> 397 20210809_014308
#> 398 20210809_014328
#> 399 20210809_014354
#> 400 20210809_014358
#> 401 20210809_014500
#> 402 20210809_014651
#> 403 20210809_014658
#> 404 20210809_014713
#> 405 20210809_014714
#> 406 20210809_014757
#> 407 20210809_014813
#> 408 20210809_014840
#> 409 20210809_014922
#> 410 20210809_014942
#> 411 20210809_014947
#> 412 20210809_015013
#> 413 20210809_015024
#> 414 20210809_015046
#> 415 20210809_015105
#> 416 20210809_015127
#> 417 20210809_015139
#> 418 20210809_015219
#> 419 20210809_015225
#> 420 20210809_015253
#> 421 20210809_015318
#> 422 20210809_015343
#> 423 20210809_015407
#> 424 20210809_015429
#> 425 20210809_015433
#> 426 20210809_015606
#> 427 20210809_015744
#> 428 20210809_015802
#> 429 20210809_015820
#> 430 20210809_015842
#> 431 20210809_015903
#> 432 20210809_015913
#> 433 20210809_015946
#> 434 20210809_015949
#> 435 20210809_020003
#> 436 20210809_020022
#> 437 20210809_020028
#> 438 20210809_020056
#> 439 20210809_020105
#> 440 20210809_020114
#> 441 20210809_020328
#> 442 20210809_020350
#> 443 20210809_020410
#> 444 20210809_020418
#> 445 20210809_020441
#> 446 20210809_020452
#> 447 20210809_020513
#> 448 20210809_020535
#> 449 20210809_020556
#> 450 20210809_020619
#> 451 20210809_020626
#> 452 20210809_020653
#> 453 20210809_020706
#> 454 20210809_020726
#> 455 20210809_020732
#> 456 20210809_020900
#> 457 20210809_020901
#> 458 20210809_021034
#> 459 20210809_021051
#> 460 20210809_021111
#> 461 20210809_021138
#> 462 20210809_021143
#> 463 20210809_021202
#> 464 20210809_021213
#> 465 20210809_021232
#> 466 20210809_021255
#> 467 20210809_021323
#> 468 20210809_021343
#> 469 20210809_021414
#> 470 20210809_021418
#> 471 20210809_021441
#> 472 20210809_021454
#> 473 20210809_021512
#> 474 20210809_021531
#> 475 20210809_021534
# }