Acquire win probability for Major and Minor League games

Acquire win probability for Major and Minor League games

Usage

mlb_game_wp(game_pk, timecode = NULL)

Arguments

game_pk

The game_pk for the game requested

timecode

The time code for the MLB game (format: MMDDYYYY_HHMMSS)

Value

Returns a data frame that includes time codes from the game_pk requested

col_name	types

home_team_win_probability	numeric

away_team_win_probability	numeric

home_team_win_probability_added	numeric

at_bat_index	integer

leverage_index	numeric

Examples

# \donttest{
  try(mlb_game_wp(game_pk = 531060))
#>    home_team_win_probability away_team_win_probability
#> 1                       50.0                      50.0
#> 2                       53.7                      46.3
#> 3                       52.5                      47.5
#> 4                       54.7                      45.3
#> 5                       52.6                      47.4
#> 6                       51.0                      49.0
#> 7                       50.0                      50.0
#> 8                       46.2                      53.8
#> 9                       49.7                      50.3
#> 10                      52.6                      47.4
#> 11                      55.0                      45.0
#> 12                      66.0                      34.0
#> 13                      64.0                      36.0
#> 14                      66.2                      33.8
#> 15                      61.7                      38.3
#> 16                      64.3                      35.7
#> 17                      66.1                      33.9
#> 18                      67.3                      32.7
#> 19                      65.3                      34.7
#> 20                      67.5                      32.5
#> 21                      73.3                      26.7
#> 22                      75.4                      24.6
#> 23                      73.8                      26.2
#> 24                      76.4                      23.6
#> 25                      78.2                      21.8
#> 26                      67.7                      32.3
#> 27                      69.0                      31.0
#> 28                      67.0                      33.0
#> 29                      65.5                      34.5
#> 30                      66.6                      33.4
#> 31                      66.6                      33.4
#> 32                      67.6                      32.4
#> 33                      69.9                      30.1
#> 34                      68.0                      32.0
#> 35                      71.3                      28.7
#> 36                      69.2                      30.8
#> 37                      71.5                      28.5
#> 38                      77.6                      22.4
#> 39                      84.2                      15.8
#> 40                      84.2                      15.8
#> 41                      83.2                      16.8
#> 42                      78.9                      21.1
#> 43                      78.9                      21.1
#> 44                      84.0                      16.0
#> 45                      85.2                      14.8
#> 46                      85.2                      14.8
#> 47                      83.1                      16.9
#> 48                      83.1                      16.9
#> 49                      82.4                      17.6
#> 50                      88.0                      12.0
#> 51                      89.3                      10.7
#> 52                      88.3                      11.7
#> 53                      94.1                       5.9
#> 54                      93.7                       6.3
#> 55                      93.5                       6.5
#> 56                      93.5                       6.5
#> 57                      82.1                      17.9
#> 58                      82.1                      17.9
#> 59                      72.2                      27.8
#> 60                      87.3                      12.7
#> 61                      87.3                      12.7
#> 62                      81.3                      18.7
#> 63                      94.2                       5.8
#> 64                      97.2                       2.8
#> 65                      97.5                       2.5
#> 66                      98.3                       1.7
#> 67                      97.5                       2.5
#> 68                      96.9                       3.1
#> 69                      96.9                       3.1
#> 70                      85.8                      14.2
#> 71                      85.8                      14.2
#> 72                      67.3                      32.7
#> 73                      92.2                       7.8
#> 74                      92.2                       7.8
#>    home_team_win_probability_added at_bat_index leverage_index
#> 1                              0.0            0             NA
#> 2                              1.5            1           0.62
#> 3                             -1.2            2           0.40
#> 4                              2.2            3           0.79
#> 5                             -2.1            4           0.87
#> 6                             -1.6            5           0.62
#> 7                             -1.0            6           0.40
#> 8                             -3.8            7           0.93
#> 9                              3.5            8           1.53
#> 10                             2.9            9           1.23
#> 11                             2.4           10           0.84
#> 12                            11.0           11           0.92
#> 13                            -2.0           12           0.77
#> 14                             2.2           13           0.57
#> 15                            -4.5           14           1.04
#> 16                             2.6           15           1.03
#> 17                             1.8           16           0.73
#> 18                             1.2           17           0.46
#> 19                            -2.0           18           0.80
#> 20                             2.2           19           0.59
#> 21                             5.8           20           1.07
#> 22                             2.1           21           1.70
#> 23                            -1.6           22           0.58
#> 24                             2.6           23           1.04
#> 25                             1.8           24           0.72
#> 26                           -10.5           25           0.45
#> 27                             1.3           26           0.51
#> 28                            -2.0           27           0.83
#> 29                            -1.5           28           0.61
#> 30                             1.1           29           0.41
#> 31                             0.0           30           0.79
#> 32                             3.2           31           1.27
#> 33                             2.3           32           0.91
#> 34                            -1.9           33           0.57
#> 35                             3.3           34           1.16
#> 36                            -2.1           35           0.84
#> 37                             2.3           36           0.63
#> 38                             6.1           37           1.12
#> 39                             6.6           38           1.79
#> 40                             0.0           39           1.20
#> 41                             0.8           40           1.28
#> 42                            -4.3           41           1.77
#> 43                             0.0           42           1.23
#> 44                             2.1           43           0.85
#> 45                             1.2           44           0.49
#> 46                             0.0           45           0.50
#> 47                            -0.9           46           0.37
#> 48                             0.0           47           0.26
#> 49                             0.0           48           1.34
#> 50                             2.3           49           0.91
#> 51                             1.3           50           0.53
#> 52                            -1.0           51           0.40
#> 53                             5.8           52           0.31
#> 54                            -0.4           53           0.16
#> 55                            -0.2           54           0.11
#> 56                             0.0           55           0.88
#> 57                            -7.3           56           1.66
#> 58                             0.0           57           2.79
#> 59                             2.0           58           3.82
#> 60                            15.1           59           3.24
#> 61                             0.0           60           2.70
#> 62                             0.0           61           5.16
#> 63                             0.0           62           0.23
#> 64                             0.3           63           0.25
#> 65                             0.3           64           0.08
#> 66                             0.8           65           0.15
#> 67                            -0.8           66           0.24
#> 68                            -0.6           67           0.24
#> 69                             0.0           68           0.75
#> 70                            -7.5           69           1.53
#> 71                             0.0           70           2.89
#> 72                             2.7           71           4.18
#> 73                            24.9           72           4.60
#> 74                             0.0           73           2.80
# }