Acquire high/low stats for Major and Minor Leagues

Acquire high/low stats for Major and Minor Leagues

Usage

mlb_high_low_stats(
  org_type,
  season,
  sort_stat,
  team_ids = NULL,
  league_ids = NULL,
  sport_ids = NULL,
  game_type = NULL,
  stat_group = NULL,
  limit = NULL
)

Arguments

org_type

The organization type for return information (Required). Valid values include:

player

team

division

league

sport

season

The season for which you want to return information (Required).

sort_stat

The stat to sort the return (Required). Valid values can be found from 'stat_lookup_param' below

stat_name	stat_lookup_param	is_counting	stat_label	stat_groups	org_types	high_low_types

at_bats	atBats	TRUE	At bats	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM , GAME

total_plate_appearances	plateAppearances	TRUE	Total plate appearances	hitting	PLAYER, TEAM	PLAYER, TEAM , GAME

runs	runs	TRUE	Runs	hitting	PLAYER, TEAM	PLAYER, TEAM , GAME

runs_batted_in	rbi	TRUE	Runs batted in	hitting	PLAYER, TEAM	PLAYER, TEAM , GAME

home_team_runs	runs	TRUE	Home team runs	hitting	TEAM	GAME

away_team_runs	runs	TRUE	Away team runs	hitting	TEAM	GAME

hits	hits	TRUE	Hits	hitting	PLAYER, TEAM	PLAYER, TEAM , GAME

hits_risp	hitsRisp	TRUE	Hits risp	hitting	PLAYER, TEAM	PLAYER, TEAM

home_team_hits	hits	TRUE	Home team hits	hitting	TEAM	GAME

away_team_hits	hits	TRUE	Away team hits	hitting	TEAM	GAME

total_bases	totalBases	TRUE	Total bases	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM

doubles	doubles	TRUE	Doubles	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM , GAME

triples	triples	TRUE	Triples	hitting	PLAYER, TEAM	PLAYER, TEAM , GAME

home_runs	homeRuns	TRUE	Home runs	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM

extra_base_hits	extraBaseHits	TRUE	Extra base hits	hitting	PLAYER, TEAM	PLAYER, TEAM

walks	baseOnBalls	TRUE	Walks	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM

strikeouts	strikeouts	TRUE	Strikeouts	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM

stolen_bases	stolenBases	TRUE	Stolen bases	hitting	PLAYER, TEAM	PLAYER, TEAM

caught_stealing	caughtStealing	TRUE	Caught stealing	hitting , pitching, fielding	PLAYER, TEAM	PLAYER, TEAM

sacrifice_flies	sacFlies	TRUE	Sacrifice flies	hitting	PLAYER, TEAM	PLAYER, TEAM

sacrifice_bunts	sacBunts	TRUE	Sacrifice bunts	hitting	PLAYER, TEAM	PLAYER, TEAM

hit_by_pitches	hitByPitch	TRUE	Hit by pitches	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM

left_on_base	leftOnBase	TRUE	Left on base	hitting	PLAYER, TEAM	PLAYER, TEAM

ground_into_double_plays	groundIntoDoublePlay	TRUE	Ground into double plays	hitting , pitching	PLAYER, TEAM	PLAYER, TEAM

strikes	strikes	TRUE	Strikes	pitching	PLAYER, TEAM	PLAYER, TEAM , GAME

pitches	pitchesThrown	TRUE	Pitches	pitching	PLAYER, TEAM	PLAYER, TEAM , GAME

balks	balks	TRUE	Balks	pitching	PLAYER, TEAM	PLAYER, TEAM , GAME

innings_pitched	inningsPitched	TRUE	Innings pitched	pitching	PLAYER	PLAYER

errors	errors	TRUE	Errors	fielding	TEAM

home_team_errors	errors	TRUE	Home team errors	fielding	TEAM

away_team_errors	errors	TRUE	Away team errors	fielding	TEAM

chances	chances	TRUE	Chances	fielding	TEAM

put_outs	putOuts	TRUE	Put outs	fielding	TEAM

assists	assists	TRUE	Assists	fielding	TEAM

double_plays	doublePlays	TRUE	Double plays	fielding	TEAM

attendance	attendance	TRUE	Attendance	game	TEAM	GAME

game_time	gameDuration	TRUE	Game time	game	TEAM	GAME

delay_time	gameDuration	TRUE	Delay time	game	TEAM	GAME

longest	gameDuration	TRUE	Longest	game	TEAM

shortest	gameDuration	TRUE	Shortest	game	TEAM

inning	innings	TRUE	Inning	game	TEAM	GAME

win_streak	winStreak	TRUE	Win streak	streak	TEAM

loss_streak	lossStreak	TRUE	Loss streak	streak	TEAM

team_ids

The team_id(s) for which to return information.

league_ids

The league_id(s) for which to return information.

sport_ids

The sport_id(s) for which to return information.

game_type

The game_type for which to return information.

stat_group

Stat group for which to return information. Valid values include:

stat_group

hitting

pitching

fielding

catching

running

game

team

streak

limit

Number of records as the limit of the return.

Value

Returns a data frame with the following columns

col_name	types

total_splits	integer

season	integer

date	character

is_home	logical

rank	integer

game_innings	integer

stat_at_bats	integer

team_id	integer

team_name	character

team_link	character

opponent_id	integer

opponent_name	character

opponent_link	character

game_pk	integer

game_link	character

game_number	integer

game_content_link	character

home_team_id	integer

home_team_name	character

home_team_link	character

away_team_id	integer

away_team_name	character

away_team_link	character

combined_stats	logical

group_display_name	character

game_type_id	character

game_type_description	character

sort_stat_name	character

sort_stat_lookup_param	character

sort_stat_is_counting	logical

sort_stat_label	character

Examples

# \donttest{
  try(mlb_high_low_stats(org_type = 'Team', season = 2020, sort_stat = 'atBats'))
#> ── MLB High Low Stats data from MLB.com ───────────── baseballr 1.1.0 ──
#> ℹ Data updated: 2022-04-12 04:19:39 UTC
#> # A tibble: 20 × 32
#>    total_splits season date      is_home  rank game_innings stat_at_bats
#>           <int>  <int> <chr>     <lgl>   <int>        <int>        <int>
#>  1           31   2020 2020-09-… FALSE       1            9           52
#>  2           31   2020 2020-09-… FALSE       2           12           49
#>  3           31   2020 2020-08-… FALSE       2            9           49
#>  4           31   2020 2020-08-… FALSE       2           11           49
#>  5           31   2020 2020-09-… FALSE       2            9           49
#>  6           30   2020 2020-08-… FALSE       1           11           96
#>  7           30   2020 2020-09-… FALSE       2           12           94
#>  8           30   2020 2020-08-… FALSE       3           11           91
#>  9           30   2020 2020-08-… FALSE       4           13           90
#> 10           30   2020 2020-08-… FALSE       4            9           90
#> 11           31   2020 2020-09-… TRUE        1            9           52
#> 12           31   2020 2020-08-… TRUE        2            9           49
#> 13           31   2020 2020-08-… TRUE        3           11           49
#> 14           31   2020 2020-09-… TRUE        4           12           49
#> 15           31   2020 2020-09-… TRUE        5            9           49
#> 16           30   2020 2020-08-… FALSE       1           11           96
#> 17           30   2020 2020-09-… FALSE       2           12           94
#> 18           30   2020 2020-08-… FALSE       3           11           91
#> 19           30   2020 2020-08-… FALSE       4           13           90
#> 20           30   2020 2020-08-… FALSE       5            9           90
#> # … with 25 more variables: team_id <int>, team_name <chr>,
#> #   team_link <chr>, opponent_id <int>, opponent_name <chr>,
#> #   opponent_link <chr>, game_pk <int>, game_link <chr>,
#> #   game_number <int>, game_day_night <chr>, game_content_link <chr>,
#> #   home_team_id <int>, home_team_name <chr>, home_team_link <chr>,
#> #   away_team_id <int>, away_team_name <chr>, away_team_link <chr>,
#> #   combined_stats <lgl>, group_display_name <chr>, …
# }