Find MLB Rosters by Roster Type

Find MLB Rosters by Roster Type

Usage

mlb_rosters(team_id = NULL, season = NULL, date = NULL, roster_type = NULL)

Arguments

team_id

team_id to return team roster information for a particular club.

season

Year to return team roster information for a particular club in a specific season.

date

Date to return team roster and their coaching staff directorial information for a particular team.

roster_type

roster_type to return team directorial information for. See mlb_roster_types() for more options. Valid options include: '40Man', 'fullSeason', 'fullRoster', 'nonRosterInvitees', 'active', 'allTime', 'depthChart', 'gameday', 'coach'

Value

Returns a data frame with the following columns:

col_name	types

jersey_number	character

person_id	integer

person_full_name	character

person_link	character

position_code	character

position_name	character

position_type	character

position_abbreviation	character

status_code	character

status_description	character

link	character

team_id	integer

roster_type	character

season	numeric

date	character

Examples

# \donttest{
  try(mlb_rosters(team_id = 109, season = 2018, roster_type = 'active'))
#>    jersey_number person_id  person_full_name           person_link
#> 1             13    605113        Nick Ahmed /api/v1/people/605113
#> 2             35    542882     Matt Andriese /api/v1/people/542882
#> 3              5    488671        Alex Avila /api/v1/people/488671
#> 4             33    545332      Jake Barrett /api/v1/people/545332
#> 5             31    502202    Brad Boxberger /api/v1/people/502202
#> 6             61    611093    Silvino Bracho /api/v1/people/611093
#> 7             25    605151    Archie Bradley /api/v1/people/605151
#> 8             19    593647    Socrates Brito /api/v1/people/593647
#> 9             32    453329     Clay Buchholz /api/v1/people/453329
#> 10            40    605177     Andrew Chafin /api/v1/people/605177
#> 11            46    571578    Patrick Corbin /api/v1/people/571578
#> 12            29    407822  Jorge De La Rosa /api/v1/people/407822
#> 13            48    517414   Randall Delgado /api/v1/people/517414
#> 14             3    518614   Daniel Descalso /api/v1/people/518614
#> 15            41    518617      Jake Diekman /api/v1/people/518617
#> 16             1    502481      Jarrod Dyson /api/v1/people/502481
#> 17            14    500871   Eduardo Escobar /api/v1/people/500871
#> 18            52    643327       Zack Godley /api/v1/people/643327
#> 19            44    502671  Paul Goldschmidt /api/v1/people/502671
#> 20            21    425844      Zack Greinke /api/v1/people/425844
#> 21            66    673633  Yoshihisa Hirano /api/v1/people/673633
#> 22             9    445055           Jon Jay /api/v1/people/445055
#> 23            47    623182  Patrick Kivlehan /api/v1/people/623182
#> 24            55    571863         Matt Koch /api/v1/people/571863
#> 25            65    607216     Joey Krehbiel /api/v1/people/607216
#> 26            22    571875         Jake Lamb /api/v1/people/571875
#> 27            50    661255        Yoan Lopez /api/v1/people/661255
#> 28            10    571918     Deven Marrero /api/v1/people/571918
#> 29             4    606466       Ketel Marte /api/v1/people/606466
#> 30             2    425772       Jeff Mathis /api/v1/people/425772
#> 31            30    519008    T.J. McFarland /api/v1/people/519008
#> 32            47    450665       Kris Medlen /api/v1/people/450665
#> 33            26    571946     Shelby Miller /api/v1/people/571946
#> 34            36    571974  John Ryan Murphy /api/v1/people/571974
#> 35            45    502117 Kristopher Negron /api/v1/people/502117
#> 36            16    572008      Chris Owings /api/v1/people/572008
#> 37             6    444482     David Peralta /api/v1/people/444482
#> 38            11    572041        AJ Pollock /api/v1/people/572041
#> 39            38    592662        Robbie Ray /api/v1/people/592662
#> 40            57    477569    Fernando Salas /api/v1/people/477569
#> 41                  644428     Troy Scribner /api/v1/people/644428
#> 42            54    642073     Jimmie Sherfy /api/v1/people/642073
#> 43                  640463    Braden Shipley /api/v1/people/640463
#> 44            28    519306  Steven Souza Jr. /api/v1/people/519306
#> 45            37    455755     Chris Stewart /api/v1/people/455755
#> 46            15    545121   Ildemaro Vargas /api/v1/people/545121
#> 47            53    572233  Christian Walker /api/v1/people/572233
#> 48            99    592836    Taijuan Walker /api/v1/people/592836
#> 49            29    446899      Brad Ziegler /api/v1/people/446899
#>    position_code position_name position_type position_abbreviation
#> 1              6     Shortstop     Infielder                    SS
#> 2              1       Pitcher       Pitcher                     P
#> 3              2       Catcher       Catcher                     C
#> 4              1       Pitcher       Pitcher                     P
#> 5              1       Pitcher       Pitcher                     P
#> 6              1       Pitcher       Pitcher                     P
#> 7              1       Pitcher       Pitcher                     P
#> 8              9    Outfielder    Outfielder                    RF
#> 9              1       Pitcher       Pitcher                     P
#> 10             1       Pitcher       Pitcher                     P
#> 11             1       Pitcher       Pitcher                     P
#> 12             1       Pitcher       Pitcher                     P
#> 13             1       Pitcher       Pitcher                     P
#> 14             5    Third Base     Infielder                    3B
#> 15             1       Pitcher       Pitcher                     P
#> 16             8    Outfielder    Outfielder                    CF
#> 17             5    Third Base     Infielder                    3B
#> 18             1       Pitcher       Pitcher                     P
#> 19             3    First Base     Infielder                    1B
#> 20             1       Pitcher       Pitcher                     P
#> 21             1       Pitcher       Pitcher                     P
#> 22             9    Outfielder    Outfielder                    RF
#> 23             9    Outfielder    Outfielder                    RF
#> 24             1       Pitcher       Pitcher                     P
#> 25             1       Pitcher       Pitcher                     P
#> 26             5    Third Base     Infielder                    3B
#> 27             1       Pitcher       Pitcher                     P
#> 28             5    Third Base     Infielder                    3B
#> 29             4   Second Base     Infielder                    2B
#> 30             2       Catcher       Catcher                     C
#> 31             1       Pitcher       Pitcher                     P
#> 32             1       Pitcher       Pitcher                     P
#> 33             1       Pitcher       Pitcher                     P
#> 34             2       Catcher       Catcher                     C
#> 35             5    Third Base     Infielder                    3B
#> 36             4   Second Base     Infielder                    2B
#> 37             7    Outfielder    Outfielder                    LF
#> 38             8    Outfielder    Outfielder                    CF
#> 39             1       Pitcher       Pitcher                     P
#> 40             1       Pitcher       Pitcher                     P
#> 41             1       Pitcher       Pitcher                     P
#> 42             1       Pitcher       Pitcher                     P
#> 43             1       Pitcher       Pitcher                     P
#> 44             9    Outfielder    Outfielder                    RF
#> 45             2       Catcher       Catcher                     C
#> 46             6     Shortstop     Infielder                    SS
#> 47             3    First Base     Infielder                    1B
#> 48             1       Pitcher       Pitcher                     P
#> 49             1       Pitcher       Pitcher                     P
#>    status_code       status_description                     link
#> 1            A                   Active /api/v1/teams/109/roster
#> 2            A                   Active /api/v1/teams/109/roster
#> 3            A                   Active /api/v1/teams/109/roster
#> 4            A                   Active /api/v1/teams/109/roster
#> 5            A                   Active /api/v1/teams/109/roster
#> 6            A                   Active /api/v1/teams/109/roster
#> 7            A                   Active /api/v1/teams/109/roster
#> 8            A                   Active /api/v1/teams/109/roster
#> 9          D60           Injured 60-Day /api/v1/teams/109/roster
#> 10           A                   Active /api/v1/teams/109/roster
#> 11           A                   Active /api/v1/teams/109/roster
#> 12          RL                 Released /api/v1/teams/109/roster
#> 13           A                   Active /api/v1/teams/109/roster
#> 14           A                   Active /api/v1/teams/109/roster
#> 15           A                   Active /api/v1/teams/109/roster
#> 16         D60           Injured 60-Day /api/v1/teams/109/roster
#> 17           A                   Active /api/v1/teams/109/roster
#> 18           A                   Active /api/v1/teams/109/roster
#> 19           A                   Active /api/v1/teams/109/roster
#> 20           A                   Active /api/v1/teams/109/roster
#> 21           A                   Active /api/v1/teams/109/roster
#> 22           A                   Active /api/v1/teams/109/roster
#> 23         MIN    Minor League Contract /api/v1/teams/109/roster
#> 24           A                   Active /api/v1/teams/109/roster
#> 25           A                   Active /api/v1/teams/109/roster
#> 26         D60           Injured 60-Day /api/v1/teams/109/roster
#> 27           A                   Active /api/v1/teams/109/roster
#> 28         MIN    Minor League Contract /api/v1/teams/109/roster
#> 29           A                   Active /api/v1/teams/109/roster
#> 30           A                   Active /api/v1/teams/109/roster
#> 31           A                   Active /api/v1/teams/109/roster
#> 32         RET Voluntarily Retired List /api/v1/teams/109/roster
#> 33           A                   Active /api/v1/teams/109/roster
#> 34           A                   Active /api/v1/teams/109/roster
#> 35          TR                   Traded /api/v1/teams/109/roster
#> 36           A                   Active /api/v1/teams/109/roster
#> 37           A                   Active /api/v1/teams/109/roster
#> 38           A                   Active /api/v1/teams/109/roster
#> 39           A                   Active /api/v1/teams/109/roster
#> 40          RL                 Released /api/v1/teams/109/roster
#> 41         MIN    Minor League Contract /api/v1/teams/109/roster
#> 42           A                   Active /api/v1/teams/109/roster
#> 43           A                   Active /api/v1/teams/109/roster
#> 44           A                   Active /api/v1/teams/109/roster
#> 45           A                   Active /api/v1/teams/109/roster
#> 46           A                   Active /api/v1/teams/109/roster
#> 47         D60           Injured 60-Day /api/v1/teams/109/roster
#> 48         D60           Injured 60-Day /api/v1/teams/109/roster
#> 49         RET Voluntarily Retired List /api/v1/teams/109/roster
#>    team_id roster_type season date
#> 1      109  fullSeason   2018 <NA>
#> 2      109  fullSeason   2018 <NA>
#> 3      109  fullSeason   2018 <NA>
#> 4      109  fullSeason   2018 <NA>
#> 5      109  fullSeason   2018 <NA>
#> 6      109  fullSeason   2018 <NA>
#> 7      109  fullSeason   2018 <NA>
#> 8      109  fullSeason   2018 <NA>
#> 9      109  fullSeason   2018 <NA>
#> 10     109  fullSeason   2018 <NA>
#> 11     109  fullSeason   2018 <NA>
#> 12     109  fullSeason   2018 <NA>
#> 13     109  fullSeason   2018 <NA>
#> 14     109  fullSeason   2018 <NA>
#> 15     109  fullSeason   2018 <NA>
#> 16     109  fullSeason   2018 <NA>
#> 17     109  fullSeason   2018 <NA>
#> 18     109  fullSeason   2018 <NA>
#> 19     109  fullSeason   2018 <NA>
#> 20     109  fullSeason   2018 <NA>
#> 21     109  fullSeason   2018 <NA>
#> 22     109  fullSeason   2018 <NA>
#> 23     109  fullSeason   2018 <NA>
#> 24     109  fullSeason   2018 <NA>
#> 25     109  fullSeason   2018 <NA>
#> 26     109  fullSeason   2018 <NA>
#> 27     109  fullSeason   2018 <NA>
#> 28     109  fullSeason   2018 <NA>
#> 29     109  fullSeason   2018 <NA>
#> 30     109  fullSeason   2018 <NA>
#> 31     109  fullSeason   2018 <NA>
#> 32     109  fullSeason   2018 <NA>
#> 33     109  fullSeason   2018 <NA>
#> 34     109  fullSeason   2018 <NA>
#> 35     109  fullSeason   2018 <NA>
#> 36     109  fullSeason   2018 <NA>
#> 37     109  fullSeason   2018 <NA>
#> 38     109  fullSeason   2018 <NA>
#> 39     109  fullSeason   2018 <NA>
#> 40     109  fullSeason   2018 <NA>
#> 41     109  fullSeason   2018 <NA>
#> 42     109  fullSeason   2018 <NA>
#> 43     109  fullSeason   2018 <NA>
#> 44     109  fullSeason   2018 <NA>
#> 45     109  fullSeason   2018 <NA>
#> 46     109  fullSeason   2018 <NA>
#> 47     109  fullSeason   2018 <NA>
#> 48     109  fullSeason   2018 <NA>
#> 49     109  fullSeason   2018 <NA>
  try(mlb_rosters(team_id = 109, season = 2018, roster_type = 'coach'))
#>    jersey_number                            job job_id
#> 1             17                        Manager   MNGR
#> 2             12                    Bench Coach   COAB
#> 3             23                 Pitching Coach   COAP
#> 4             39               First Base Coach   COA1
#> 5              3               Third Base Coach   COA3
#> 6             71                  Bullpen Coach   COAU
#> 7             59        Assistant Hitting Coach   COAA
#> 8              7 Quality Control/Catching Coach   QCCC
#> 9             60                          Coach   COAC
#> 10            83                Bullpen Catcher   BCAT
#> 11            84                Bullpen Catcher   BCAT
#> 12                      Senior Baseball Advisor   SBAD
#>                                 title person_id  person_full_name
#> 1                             Manager    117950     Torey Lovullo
#> 2                         Bench Coach    119655      Jerry Narron
#> 3                      Pitching Coach    111807      Mike Butcher
#> 4                    First Base Coach    118763        Dave McKay
#> 5                    Third Base Coach    120418   Tony Perezchica
#> 6                       Bullpen Coach    114097      Mike Fetters
#> 7             Assistant Hitting Coach    117420         Tim Laker
#> 8      Quality Control/Catching Coach    425631     Robby Hammock
#> 9                  Major League Coach    468222       Luis Urueta
#> 10                    Bullpen Catcher    461861         Mark Reed
#> 11                    Bullpen Catcher    279827 Humberto Quintero
#> 12 Senior Manager, Player Development    678771     Shawn Marette
#>              person_link                      link team_id roster_type
#> 1  /api/v1/people/117950 /api/v1/teams/109/coaches     109       coach
#> 2  /api/v1/people/119655 /api/v1/teams/109/coaches     109       coach
#> 3  /api/v1/people/111807 /api/v1/teams/109/coaches     109       coach
#> 4  /api/v1/people/118763 /api/v1/teams/109/coaches     109       coach
#> 5  /api/v1/people/120418 /api/v1/teams/109/coaches     109       coach
#> 6  /api/v1/people/114097 /api/v1/teams/109/coaches     109       coach
#> 7  /api/v1/people/117420 /api/v1/teams/109/coaches     109       coach
#> 8  /api/v1/people/425631 /api/v1/teams/109/coaches     109       coach
#> 9  /api/v1/people/468222 /api/v1/teams/109/coaches     109       coach
#> 10 /api/v1/people/461861 /api/v1/teams/109/coaches     109       coach
#> 11 /api/v1/people/279827 /api/v1/teams/109/coaches     109       coach
#> 12 /api/v1/people/678771 /api/v1/teams/109/coaches     109       coach
#>    season date
#> 1    2018 <NA>
#> 2    2018 <NA>
#> 3    2018 <NA>
#> 4    2018 <NA>
#> 5    2018 <NA>
#> 6    2018 <NA>
#> 7    2018 <NA>
#> 8    2018 <NA>
#> 9    2018 <NA>
#> 10   2018 <NA>
#> 11   2018 <NA>
#> 12   2018 <NA>
# }