Find game_pk values for professional baseball games (major and minor leagues) that are tied

Find game_pk values for professional baseball games (major and minor leagues) that are tied

Usage

mlb_schedule_games_tied(season = 2021, game_type = "S")

Arguments

season

The season for which you want to find game_pk values for MLB games

game_type

game_type to return schedule information for all tied games in a particular game_type

game_type_id	game_type_description

S	Spring Training

R	Regular Season

F	Wild Card Game

D	Division Series

L	League Championship Series

W	World Series

C	Championship

N	Nineteenth Century Series

P	Playoffs

A	All-Star Game

I	Intrasquad

E	Exhibition

Value

Returns a tibble that includes game_pk values and additional information for games scheduled or played

col_name	types

date	character

total_items	integer

total_events	integer

total_games	integer

total_games_in_progress	integer

game_pk	integer

link	character

game_type	character

season	character

game_date	character

official_date	character

game_number	integer

public_facing	logical

double_header	character

gameday_type	character

tiebreaker	character

calendar_event_id	character

season_display	character

day_night	character

scheduled_innings	integer

reverse_home_away_status	logical

inning_break_length	integer

games_in_series	integer

series_game_number	integer

series_description	character

record_source	character

if_necessary	character

if_necessary_description	character

status_abstract_game_state	character

status_coded_game_state	character

status_detailed_state	character

status_status_code	character

status_start_time_tbd	logical

status_reason	character

status_abstract_game_code	character

teams_away_split_squad	logical

teams_away_series_number	integer

teams_away_league_record_wins	integer

teams_away_league_record_losses	integer

teams_away_league_record_pct	character

teams_away_team_id	integer

teams_away_team_name	character

teams_away_team_link	character

teams_home_split_squad	logical

teams_home_series_number	integer

teams_home_league_record_wins	integer

teams_home_league_record_losses	integer

teams_home_league_record_pct	character

teams_home_team_id	integer

teams_home_team_name	character

teams_home_team_link	character

venue_id	integer

venue_name	character

venue_link	character

content_link	character

is_tie	logical

description	character

teams_away_score	integer

teams_away_is_winner	logical

teams_home_score	integer

teams_home_is_winner	logical

reschedule_date	character

reschedule_game_date	character

rescheduled_from	character

rescheduled_from_date	character

resume_date	character

resume_game_date	character

resumed_from	character

resumed_from_date	character

events	list

Examples

# \donttest{
  try(mlb_schedule_games_tied(season = 2021))
#> ── MLB Schedule - Games Tied data from MLB.com ────── baseballr 1.4.0 ──
#> ℹ Data updated: 2022-12-29 00:08:29 UTC
#> # A tibble: 46 × 59
#>    date     total…¹ total…² total…³ total…⁴ game_pk link  game_…⁵ season
#>    <chr>      <int>   <int>   <int>   <int>   <int> <chr> <chr>   <chr> 
#>  1 2021-02…       1       0       1       0  641924 /api… S       2021  
#>  2 2021-03…       4       0       4       0  642190 /api… S       2021  
#>  3 2021-03…       4       0       4       0  641697 /api… S       2021  
#>  4 2021-03…       4       0       4       0  641634 /api… S       2021  
#>  5 2021-03…       4       0       4       0  641587 /api… S       2021  
#>  6 2021-03…       3       0       3       0  641540 /api… S       2021  
#>  7 2021-03…       3       0       3       0  641990 /api… S       2021  
#>  8 2021-03…       3       0       3       0  641627 /api… S       2021  
#>  9 2021-03…       2       0       2       0  642125 /api… S       2021  
#> 10 2021-03…       2       0       2       0  642129 /api… S       2021  
#> # … with 36 more rows, 50 more variables: game_date <chr>,
#> #   official_date <chr>, is_tie <lgl>, game_number <int>,
#> #   public_facing <lgl>, double_header <chr>, gameday_type <chr>,
#> #   tiebreaker <chr>, calendar_event_id <chr>, season_display <chr>,
#> #   day_night <chr>, scheduled_innings <int>,
#> #   reverse_home_away_status <lgl>, games_in_series <int>,
#> #   series_game_number <int>, series_description <chr>, …
# }