Find game_pk values for professional baseball postseason games (major and minor leagues)

Find game_pk values for professional baseball postseason games (major and minor leagues)

Usage

mlb_schedule_postseason(
  season = 2021,
  game_type = NULL,
  series_number = NULL,
  sport_id = 1,
  team_id = NULL
)

Arguments

season

The season for which you want to find game_pk values for MLB games

game_type

game_type to return schedule information for all tied games in a particular game_type

series_number

The Series number to return schedule information for all tied games in a particular series number

sport_id

The sport_id to return schedule information for.

team_id

The team_id to return schedule information for.

game_type_id	game_type_description

S	Spring Training

R	Regular Season

F	Wild Card Game

D	Division Series

L	League Championship Series

W	World Series

C	Championship

N	Nineteenth Century Series

P	Playoffs

A	All-Star Game

I	Intrasquad

E	Exhibition

Value

Returns a data frame that includes game_pk values and additional information for games scheduled or played

col_name	types

date	character

total_items	integer

total_events	integer

total_games	integer

total_games_in_progress	integer

game_pk	integer

link	character

game_type	character

season	character

game_date	character

official_date	character

is_tie	logical

is_featured_game	logical

game_number	integer

public_facing	logical

double_header	character

gameday_type	character

tiebreaker	character

calendar_event_id	character

season_display	character

day_night	character

description	character

scheduled_innings	integer

reverse_home_away_status	logical

games_in_series	integer

series_game_number	integer

series_description	character

record_source	character

if_necessary	character

if_necessary_description	character

status_abstract_game_state	character

status_coded_game_state	character

status_detailed_state	character

status_status_code	character

status_start_time_tbd	logical

status_abstract_game_code	character

teams_away_score	integer

teams_away_is_winner	logical

teams_away_split_squad	logical

teams_away_series_number	integer

teams_away_league_record_wins	integer

teams_away_league_record_losses	integer

teams_away_league_record_pct	character

teams_away_team_id	integer

teams_away_team_name	character

teams_away_team_link	character

teams_home_score	integer

teams_home_is_winner	logical

teams_home_split_squad	logical

teams_home_series_number	integer

teams_home_league_record_wins	integer

teams_home_league_record_losses	integer

teams_home_league_record_pct	character

teams_home_team_id	integer

teams_home_team_name	character

teams_home_team_link	character

venue_id	integer

venue_name	character

venue_link	character

content_link	character

inning_break_length	integer

reschedule_date	character

reschedule_game_date	character

status_reason	character

rescheduled_from	character

rescheduled_from_date	character

is_default_game	logical

events	list

Examples

# \donttest{
  try(mlb_schedule_postseason(season = 2021))
#>          date total_items total_events total_games
#> 1  2021-10-05           1            0           1
#> 2  2021-10-06           1            0           1
#> 3  2021-10-07           2            0           2
#> 4  2021-10-07           2            0           2
#> 5  2021-10-08           4            0           4
#> 6  2021-10-08           4            0           4
#> 7  2021-10-08           4            0           4
#> 8  2021-10-08           4            0           4
#> 9  2021-10-09           2            0           2
#> 10 2021-10-09           2            0           2
#> 11 2021-10-10           2            0           2
#> 12 2021-10-10           2            0           2
#> 13 2021-10-11           4            0           4
#> 14 2021-10-11           4            0           4
#> 15 2021-10-11           4            0           4
#> 16 2021-10-11           4            0           4
#> 17 2021-10-12           3            0           3
#> 18 2021-10-12           3            0           3
#> 19 2021-10-12           3            0           3
#> 20 2021-10-14           1            0           1
#> 21 2021-10-15           1            0           1
#> 22 2021-10-16           2            0           2
#> 23 2021-10-16           2            0           2
#> 24 2021-10-17           1            0           1
#> 25 2021-10-18           1            0           1
#> 26 2021-10-19           2            0           2
#> 27 2021-10-19           2            0           2
#> 28 2021-10-20           2            0           2
#> 29 2021-10-20           2            0           2
#> 30 2021-10-21           1            0           1
#> 31 2021-10-22           1            0           1
#> 32 2021-10-23           1            0           1
#> 33 2021-10-26           1            0           1
#> 34 2021-10-27           1            0           1
#> 35 2021-10-29           1            0           1
#> 36 2021-10-30           1            0           1
#> 37 2021-10-31           1            0           1
#> 38 2021-11-02           1            0           1
#>    total_games_in_progress game_pk                            link
#> 1                        0  660938 /api/v1.1/game/660938/feed/live
#> 2                        0  660937 /api/v1.1/game/660937/feed/live
#> 3                        0  660936 /api/v1.1/game/660936/feed/live
#> 4                        0  660929 /api/v1.1/game/660929/feed/live
#> 5                        0  660934 /api/v1.1/game/660934/feed/live
#> 6                        0  660917 /api/v1.1/game/660917/feed/live
#> 7                        0  660932 /api/v1.1/game/660932/feed/live
#> 8                        0  660933 /api/v1.1/game/660933/feed/live
#> 9                        0  660907 /api/v1.1/game/660907/feed/live
#> 10                       0  660925 /api/v1.1/game/660925/feed/live
#> 11                       0  660928 /api/v1.1/game/660928/feed/live
#> 12                       0  660931 /api/v1.1/game/660931/feed/live
#> 13                       0  660905 /api/v1.1/game/660905/feed/live
#> 14                       0  660926 /api/v1.1/game/660926/feed/live
#> 15                       0  660927 /api/v1.1/game/660927/feed/live
#> 16                       0  660919 /api/v1.1/game/660919/feed/live
#> 17                       0  660926 /api/v1.1/game/660926/feed/live
#> 18                       0  660914 /api/v1.1/game/660914/feed/live
#> 19                       0  660918 /api/v1.1/game/660918/feed/live
#> 20                       0  660920 /api/v1.1/game/660920/feed/live
#> 21                       0  660896 /api/v1.1/game/660896/feed/live
#> 22                       0  660913 /api/v1.1/game/660913/feed/live
#> 23                       0  660924 /api/v1.1/game/660924/feed/live
#> 24                       0  660922 /api/v1.1/game/660922/feed/live
#> 25                       0  660898 /api/v1.1/game/660898/feed/live
#> 26                       0  660916 /api/v1.1/game/660916/feed/live
#> 27                       0  660923 /api/v1.1/game/660923/feed/live
#> 28                       0  660900 /api/v1.1/game/660900/feed/live
#> 29                       0  660904 /api/v1.1/game/660904/feed/live
#> 30                       0  660903 /api/v1.1/game/660903/feed/live
#> 31                       0  660899 /api/v1.1/game/660899/feed/live
#> 32                       0  660921 /api/v1.1/game/660921/feed/live
#> 33                       0  660897 /api/v1.1/game/660897/feed/live
#> 34                       0  660908 /api/v1.1/game/660908/feed/live
#> 35                       0  660910 /api/v1.1/game/660910/feed/live
#> 36                       0  660902 /api/v1.1/game/660902/feed/live
#> 37                       0  660911 /api/v1.1/game/660911/feed/live
#> 38                       0  660906 /api/v1.1/game/660906/feed/live
#>    game_type season            game_date official_date is_tie
#> 1          F   2021 2021-10-06T00:08:00Z    2021-10-05  FALSE
#> 2          F   2021 2021-10-07T00:10:00Z    2021-10-06  FALSE
#> 3          D   2021 2021-10-07T20:07:00Z    2021-10-07  FALSE
#> 4          D   2021 2021-10-08T00:07:00Z    2021-10-07  FALSE
#> 5          D   2021 2021-10-08T18:07:00Z    2021-10-08  FALSE
#> 6          D   2021 2021-10-08T20:37:00Z    2021-10-08  FALSE
#> 7          D   2021 2021-10-08T23:02:00Z    2021-10-08  FALSE
#> 8          D   2021 2021-10-09T01:37:00Z    2021-10-08  FALSE
#> 9          D   2021 2021-10-09T21:07:00Z    2021-10-09  FALSE
#> 10         D   2021 2021-10-10T01:07:00Z    2021-10-09  FALSE
#> 11         D   2021 2021-10-10T20:07:00Z    2021-10-10  FALSE
#> 12         D   2021 2021-10-11T00:07:00Z    2021-10-10  FALSE
#> 13         D   2021 2021-10-11T17:07:00Z    2021-10-11  FALSE
#> 14         D   2021 2021-10-11T19:37:00Z    2021-10-12     NA
#> 15         D   2021 2021-10-11T23:07:00Z    2021-10-11  FALSE
#> 16         D   2021 2021-10-12T01:37:00Z    2021-10-11  FALSE
#> 17         D   2021 2021-10-12T18:07:00Z    2021-10-12  FALSE
#> 18         D   2021 2021-10-12T21:15:00Z    2021-10-12  FALSE
#> 19         D   2021 2021-10-13T01:07:00Z    2021-10-12  FALSE
#> 20         D   2021 2021-10-15T01:07:00Z    2021-10-14  FALSE
#> 21         L   2021 2021-10-16T00:07:00Z    2021-10-15  FALSE
#> 22         L   2021 2021-10-16T20:20:00Z    2021-10-16  FALSE
#> 23         L   2021 2021-10-17T00:08:00Z    2021-10-16  FALSE
#> 24         L   2021 2021-10-17T23:38:00Z    2021-10-17  FALSE
#> 25         L   2021 2021-10-19T00:08:00Z    2021-10-18  FALSE
#> 26         L   2021 2021-10-19T21:08:00Z    2021-10-19  FALSE
#> 27         L   2021 2021-10-20T00:08:00Z    2021-10-19  FALSE
#> 28         L   2021 2021-10-20T21:08:00Z    2021-10-20  FALSE
#> 29         L   2021 2021-10-21T00:08:00Z    2021-10-20  FALSE
#> 30         L   2021 2021-10-22T00:08:00Z    2021-10-21  FALSE
#> 31         L   2021 2021-10-23T00:08:00Z    2021-10-22  FALSE
#> 32         L   2021 2021-10-24T00:08:00Z    2021-10-23  FALSE
#> 33         W   2021 2021-10-27T00:09:00Z    2021-10-26  FALSE
#> 34         W   2021 2021-10-28T00:09:00Z    2021-10-27  FALSE
#> 35         W   2021 2021-10-30T00:09:00Z    2021-10-29  FALSE
#> 36         W   2021 2021-10-31T00:09:00Z    2021-10-30  FALSE
#> 37         W   2021 2021-11-01T00:15:00Z    2021-10-31  FALSE
#> 38         W   2021 2021-11-03T00:09:00Z    2021-11-02  FALSE
#>    is_featured_game game_number public_facing double_header
#> 1              TRUE           1          TRUE             N
#> 2              TRUE           1          TRUE             N
#> 3             FALSE           1          TRUE             N
#> 4             FALSE           1          TRUE             N
#> 5             FALSE           1          TRUE             N
#> 6             FALSE           1          TRUE             N
#> 7             FALSE           1          TRUE             N
#> 8             FALSE           1          TRUE             N
#> 9             FALSE           1          TRUE             N
#> 10            FALSE           1          TRUE             N
#> 11            FALSE           1          TRUE             N
#> 12            FALSE           1          TRUE             N
#> 13            FALSE           1          TRUE             N
#> 14            FALSE           1          TRUE             N
#> 15             TRUE           1          TRUE             N
#> 16            FALSE           1          TRUE             N
#> 17             TRUE           1          TRUE             N
#> 18             TRUE           1          TRUE             N
#> 19            FALSE           1          TRUE             N
#> 20             TRUE           1          TRUE             N
#> 21            FALSE           1          TRUE             N
#> 22            FALSE           1          TRUE             N
#> 23            FALSE           1          TRUE             N
#> 24            FALSE           1          TRUE             N
#> 25            FALSE           1          TRUE             N
#> 26            FALSE           1          TRUE             N
#> 27            FALSE           1          TRUE             N
#> 28            FALSE           1          TRUE             N
#> 29            FALSE           1          TRUE             N
#> 30            FALSE           1          TRUE             N
#> 31             TRUE           1          TRUE             N
#> 32             TRUE           1          TRUE             N
#> 33            FALSE           1          TRUE             N
#> 34            FALSE           1          TRUE             N
#> 35            FALSE           1          TRUE             N
#> 36            FALSE           1          TRUE             N
#> 37            FALSE           1          TRUE             N
#> 38             TRUE           1          TRUE             N
#>    gameday_type tiebreaker    calendar_event_id season_display
#> 1             P          N 14-660938-2021-10-05           2021
#> 2             P          N 14-660937-2021-10-06           2021
#> 3             P          N 14-660936-2021-10-07           2021
#> 4             P          N 14-660929-2021-10-07           2021
#> 5             P          N 14-660934-2021-10-08           2021
#> 6             P          N 14-660917-2021-10-08           2021
#> 7             P          N 14-660932-2021-10-08           2021
#> 8             P          N 14-660933-2021-10-08           2021
#> 9             P          N 14-660907-2021-10-09           2021
#> 10            P          N 14-660925-2021-10-09           2021
#> 11            P          N 14-660928-2021-10-10           2021
#> 12            P          N 14-660931-2021-10-10           2021
#> 13            P          N 14-660905-2021-10-11           2021
#> 14            P          N 14-660926-2021-10-11           2021
#> 15            P          N 14-660927-2021-10-11           2021
#> 16            P          N 14-660919-2021-10-11           2021
#> 17            P          N 14-660926-2021-10-12           2021
#> 18            P          N 14-660914-2021-10-12           2021
#> 19            P          N 14-660918-2021-10-12           2021
#> 20            P          N 14-660920-2021-10-14           2021
#> 21            P          N 14-660896-2021-10-15           2021
#> 22            P          N 14-660913-2021-10-16           2021
#> 23            P          N 14-660924-2021-10-16           2021
#> 24            P          N 14-660922-2021-10-17           2021
#> 25            P          N 14-660898-2021-10-18           2021
#> 26            P          N 14-660916-2021-10-19           2021
#> 27            P          N 14-660923-2021-10-19           2021
#> 28            P          N 14-660900-2021-10-20           2021
#> 29            P          N 14-660904-2021-10-20           2021
#> 30            P          N 14-660903-2021-10-21           2021
#> 31            P          N 14-660899-2021-10-22           2021
#> 32            P          N 14-660921-2021-10-23           2021
#> 33            P          N 14-660897-2021-10-26           2021
#> 34            P          N 14-660908-2021-10-27           2021
#> 35            P          N 14-660910-2021-10-29           2021
#> 36            P          N 14-660902-2021-10-30           2021
#> 37            P          N 14-660911-2021-10-31           2021
#> 38            P          N 14-660906-2021-11-02           2021
#>    day_night               description scheduled_innings
#> 1      night         AL Wild Card Game                 9
#> 2      night         NL Wild Card Game                 9
#> 3        day           ALDS 'B' Game 1                 9
#> 4      night           ALDS 'A' Game 1                 9
#> 5        day           ALDS 'B' Game 2                 9
#> 6        day           NLDS 'B' Game 1                 9
#> 7      night           ALDS 'A' Game 2                 9
#> 8      night           NLDS 'A' Game 1                 9
#> 9        day           NLDS 'B' Game 2                 9
#> 10     night           NLDS 'A' Game 2                 9
#> 11       day           ALDS 'A' Game 3                 9
#> 12     night           ALDS 'B' Game 3                 9
#> 13       day           NLDS 'B' Game 3                 9
#> 14       day           ALDS 'B' Game 4                 9
#> 15     night           ALDS 'A' Game 4                 9
#> 16     night           NLDS 'A' Game 3                 9
#> 17       day           ALDS 'B' Game 4                 9
#> 18     night           NLDS 'B' Game 4                 9
#> 19     night           NLDS 'A' Game 4                 9
#> 20     night           NLDS 'A' Game 5                 9
#> 21     night               ALCS Game 1                 9
#> 22       day ALCS Game 2 (FOX and FS1)                 9
#> 23     night               NLCS Game 1                 9
#> 24     night               NLCS Game 2                 9
#> 25     night               ALCS Game 3                 9
#> 26       day               NLCS Game 3                 9
#> 27     night               ALCS Game 4                 9
#> 28     night               ALCS Game 5                 9
#> 29     night               NLCS Game 4                 9
#> 30     night               NLCS Game 5                 9
#> 31     night               ALCS Game 6                 9
#> 32     night               NLCS Game 6                 9
#> 33     night       World Series Game 1                 9
#> 34     night       World Series Game 2                 9
#> 35     night       World Series Game 3                 9
#> 36     night       World Series Game 4                 9
#> 37     night       World Series Game 5                 9
#> 38     night       World Series Game 6                 9
#>    reverse_home_away_status games_in_series series_game_number
#> 1                     FALSE               1                  1
#> 2                     FALSE               1                  1
#> 3                     FALSE               5                  1
#> 4                     FALSE               5                  1
#> 5                     FALSE               5                  2
#> 6                     FALSE               5                  1
#> 7                     FALSE               5                  2
#> 8                     FALSE               5                  1
#> 9                     FALSE               5                  2
#> 10                    FALSE               5                  2
#> 11                    FALSE               5                  3
#> 12                    FALSE               5                  3
#> 13                    FALSE               5                  3
#> 14                    FALSE               5                  4
#> 15                    FALSE               5                  4
#> 16                    FALSE               5                  3
#> 17                    FALSE               5                  4
#> 18                    FALSE               5                  4
#> 19                    FALSE               5                  4
#> 20                    FALSE               5                  5
#> 21                    FALSE               7                  1
#> 22                    FALSE               7                  2
#> 23                    FALSE               7                  1
#> 24                    FALSE               7                  2
#> 25                    FALSE               7                  3
#> 26                    FALSE               7                  3
#> 27                    FALSE               7                  4
#> 28                    FALSE               7                  5
#> 29                    FALSE               7                  4
#> 30                    FALSE               7                  5
#> 31                    FALSE               7                  6
#> 32                    FALSE               7                  6
#> 33                    FALSE               7                  1
#> 34                    FALSE               7                  2
#> 35                    FALSE               7                  3
#> 36                    FALSE               7                  4
#> 37                    FALSE               7                  5
#> 38                    FALSE               7                  6
#>            series_description record_source if_necessary
#> 1              Wild Card Game             S            N
#> 2              Wild Card Game             S            N
#> 3             Division Series             S            N
#> 4             Division Series             S            N
#> 5             Division Series             S            N
#> 6             Division Series             S            N
#> 7             Division Series             S            N
#> 8             Division Series             S            N
#> 9             Division Series             S            N
#> 10            Division Series             S            N
#> 11            Division Series             S            N
#> 12            Division Series             S            N
#> 13            Division Series             S            N
#> 14            Division Series             S            N
#> 15            Division Series             S            N
#> 16            Division Series             S            N
#> 17            Division Series             S            N
#> 18            Division Series             S            N
#> 19            Division Series             S            N
#> 20            Division Series             S            N
#> 21 League Championship Series             S            N
#> 22 League Championship Series             S            N
#> 23 League Championship Series             S            N
#> 24 League Championship Series             S            N
#> 25 League Championship Series             S            N
#> 26 League Championship Series             S            N
#> 27 League Championship Series             S            N
#> 28 League Championship Series             S            N
#> 29 League Championship Series             S            N
#> 30 League Championship Series             S            N
#> 31 League Championship Series             S            N
#> 32 League Championship Series             S            N
#> 33               World Series             S            N
#> 34               World Series             S            N
#> 35               World Series             S            N
#> 36               World Series             S            N
#> 37               World Series             S            N
#> 38               World Series             S            N
#>    if_necessary_description status_abstract_game_state
#> 1               Normal Game                      Final
#> 2               Normal Game                      Final
#> 3               Normal Game                      Final
#> 4               Normal Game                      Final
#> 5               Normal Game                      Final
#> 6               Normal Game                      Final
#> 7               Normal Game                      Final
#> 8               Normal Game                      Final
#> 9               Normal Game                      Final
#> 10              Normal Game                      Final
#> 11              Normal Game                      Final
#> 12              Normal Game                      Final
#> 13              Normal Game                      Final
#> 14              Normal Game                      Final
#> 15              Normal Game                      Final
#> 16              Normal Game                      Final
#> 17              Normal Game                      Final
#> 18              Normal Game                      Final
#> 19              Normal Game                      Final
#> 20              Normal Game                      Final
#> 21              Normal Game                      Final
#> 22              Normal Game                      Final
#> 23              Normal Game                      Final
#> 24              Normal Game                      Final
#> 25              Normal Game                      Final
#> 26              Normal Game                      Final
#> 27              Normal Game                      Final
#> 28              Normal Game                      Final
#> 29              Normal Game                      Final
#> 30              Normal Game                      Final
#> 31              Normal Game                      Final
#> 32              Normal Game                      Final
#> 33              Normal Game                      Final
#> 34              Normal Game                      Final
#> 35              Normal Game                      Final
#> 36              Normal Game                      Final
#> 37              Normal Game                      Final
#> 38              Normal Game                      Final
#>    status_coded_game_state status_detailed_state status_status_code
#> 1                        F                 Final                  F
#> 2                        F                 Final                  F
#> 3                        F                 Final                  F
#> 4                        F                 Final                  F
#> 5                        F                 Final                  F
#> 6                        F                 Final                  F
#> 7                        F                 Final                  F
#> 8                        F                 Final                  F
#> 9                        F                 Final                  F
#> 10                       F                 Final                  F
#> 11                       F                 Final                  F
#> 12                       F                 Final                  F
#> 13                       F                 Final                  F
#> 14                       D             Postponed                 DR
#> 15                       F                 Final                  F
#> 16                       F                 Final                  F
#> 17                       F                 Final                  F
#> 18                       F                 Final                  F
#> 19                       F                 Final                  F
#> 20                       F                 Final                  F
#> 21                       F                 Final                  F
#> 22                       F                 Final                  F
#> 23                       F                 Final                  F
#> 24                       F                 Final                  F
#> 25                       F                 Final                  F
#> 26                       F                 Final                  F
#> 27                       F                 Final                  F
#> 28                       F                 Final                  F
#> 29                       F                 Final                  F
#> 30                       F                 Final                  F
#> 31                       F                 Final                  F
#> 32                       F                 Final                  F
#> 33                       F                 Final                  F
#> 34                       F                 Final                  F
#> 35                       F                 Final                  F
#> 36                       F                 Final                  F
#> 37                       F                 Final                  F
#> 38                       F                 Final                  F
#>    status_start_time_tbd status_abstract_game_code teams_away_score
#> 1                  FALSE                         F                2
#> 2                  FALSE                         F                1
#> 3                  FALSE                         F                1
#> 4                  FALSE                         F                0
#> 5                  FALSE                         F                4
#> 6                  FALSE                         F                1
#> 7                  FALSE                         F               14
#> 8                  FALSE                         F                0
#> 9                  FALSE                         F                3
#> 10                 FALSE                         F                9
#> 11                 FALSE                         F                4
#> 12                 FALSE                         F                6
#> 13                 FALSE                         F                0
#> 14                 FALSE                         F               NA
#> 15                 FALSE                         F                5
#> 16                 FALSE                         F                1
#> 17                 FALSE                         F               10
#> 18                 FALSE                         F                4
#> 19                 FALSE                         F                2
#> 20                 FALSE                         F                2
#> 21                 FALSE                         F                4
#> 22                 FALSE                         F                9
#> 23                 FALSE                         F                2
#> 24                 FALSE                         F                4
#> 25                 FALSE                         F                3
#> 26                 FALSE                         F                5
#> 27                 FALSE                         F                9
#> 28                 FALSE                         F                9
#> 29                 FALSE                         F                9
#> 30                 FALSE                         F                2
#> 31                 FALSE                         F                0
#> 32                 FALSE                         F                2
#> 33                 FALSE                         F                6
#> 34                 FALSE                         F                2
#> 35                 FALSE                         F                0
#> 36                 FALSE                         F                2
#> 37                 FALSE                         F                9
#> 38                 FALSE                         F                7
#>    teams_away_is_winner teams_away_split_squad teams_away_series_number
#> 1                 FALSE                  FALSE                        1
#> 2                 FALSE                  FALSE                        2
#> 3                 FALSE                  FALSE                        2
#> 4                 FALSE                  FALSE                        1
#> 5                 FALSE                  FALSE                        2
#> 6                 FALSE                  FALSE                        4
#> 7                  TRUE                  FALSE                        1
#> 8                 FALSE                  FALSE                        3
#> 9                  TRUE                  FALSE                        4
#> 10                 TRUE                  FALSE                        3
#> 11                FALSE                  FALSE                        1
#> 12                FALSE                  FALSE                        2
#> 13                FALSE                  FALSE                        4
#> 14                   NA                  FALSE                        2
#> 15                FALSE                  FALSE                        1
#> 16                 TRUE                  FALSE                        3
#> 17                 TRUE                  FALSE                        2
#> 18                FALSE                  FALSE                        4
#> 19                FALSE                  FALSE                        3
#> 20                 TRUE                  FALSE                        3
#> 21                FALSE                  FALSE                        1
#> 22                 TRUE                  FALSE                        1
#> 23                FALSE                  FALSE                        2
#> 24                FALSE                  FALSE                        2
#> 25                FALSE                  FALSE                        1
#> 26                FALSE                  FALSE                        2
#> 27                 TRUE                  FALSE                        1
#> 28                 TRUE                  FALSE                        1
#> 29                 TRUE                  FALSE                        2
#> 30                FALSE                  FALSE                        2
#> 31                FALSE                  FALSE                        1
#> 32                FALSE                  FALSE                        2
#> 33                 TRUE                  FALSE                        1
#> 34                FALSE                  FALSE                        1
#> 35                FALSE                  FALSE                        1
#> 36                FALSE                  FALSE                        1
#> 37                 TRUE                  FALSE                        1
#> 38                 TRUE                  FALSE                        1
#>    teams_away_league_record_wins teams_away_league_record_losses
#> 1                              0                               1
#> 2                              0                               1
#> 3                              0                               1
#> 4                              0                               1
#> 5                              0                               2
#> 6                              0                               1
#> 7                              1                               1
#> 8                              0                               1
#> 9                              1                               1
#> 10                             1                               1
#> 11                             1                               2
#> 12                             2                               1
#> 13                             1                               2
#> 14                             3                               1
#> 15                             1                               3
#> 16                             2                               1
#> 17                             3                               1
#> 18                             1                               3
#> 19                             2                               2
#> 20                             3                               2
#> 21                             0                               1
#> 22                             1                               1
#> 23                             0                               1
#> 24                             0                               2
#> 25                             1                               2
#> 26                             2                               1
#> 27                             2                               2
#> 28                             3                               2
#> 29                             3                               1
#> 30                             3                               2
#> 31                             2                               4
#> 32                             2                               4
#> 33                             1                               0
#> 34                             1                               1
#> 35                             1                               2
#> 36                             1                               3
#> 37                             2                               3
#> 38                             4                               2
#>    teams_away_league_record_pct teams_away_team_id teams_away_team_name
#> 1                          .000                147     New York Yankees
#> 2                          .000                138  St. Louis Cardinals
#> 3                          .000                145    Chicago White Sox
#> 4                          .000                111       Boston Red Sox
#> 5                          .000                145    Chicago White Sox
#> 6                          .000                144       Atlanta Braves
#> 7                          .500                111       Boston Red Sox
#> 8                          .000                119  Los Angeles Dodgers
#> 9                          .500                144       Atlanta Braves
#> 10                         .500                119  Los Angeles Dodgers
#> 11                         .333                139       Tampa Bay Rays
#> 12                         .667                117       Houston Astros
#> 13                         .333                158    Milwaukee Brewers
#> 14                         .750                117       Houston Astros
#> 15                         .250                139       Tampa Bay Rays
#> 16                         .667                137 San Francisco Giants
#> 17                         .750                117       Houston Astros
#> 18                         .250                158    Milwaukee Brewers
#> 19                         .500                137 San Francisco Giants
#> 20                         .600                119  Los Angeles Dodgers
#> 21                         .000                111       Boston Red Sox
#> 22                         .500                111       Boston Red Sox
#> 23                         .000                119  Los Angeles Dodgers
#> 24                         .000                119  Los Angeles Dodgers
#> 25                         .333                117       Houston Astros
#> 26                         .667                144       Atlanta Braves
#> 27                         .500                117       Houston Astros
#> 28                         .600                117       Houston Astros
#> 29                         .750                144       Atlanta Braves
#> 30                         .600                144       Atlanta Braves
#> 31                         .333                111       Boston Red Sox
#> 32                         .333                119  Los Angeles Dodgers
#> 33                        1.000                144       Atlanta Braves
#> 34                         .500                144       Atlanta Braves
#> 35                         .333                117       Houston Astros
#> 36                         .250                117       Houston Astros
#> 37                         .400                117       Houston Astros
#> 38                         .667                144       Atlanta Braves
#>    teams_away_team_link teams_home_score teams_home_is_winner
#> 1     /api/v1/teams/147                6                 TRUE
#> 2     /api/v1/teams/138                3                 TRUE
#> 3     /api/v1/teams/145                6                 TRUE
#> 4     /api/v1/teams/111                5                 TRUE
#> 5     /api/v1/teams/145                9                 TRUE
#> 6     /api/v1/teams/144                2                 TRUE
#> 7     /api/v1/teams/111                6                FALSE
#> 8     /api/v1/teams/119                4                 TRUE
#> 9     /api/v1/teams/144                0                FALSE
#> 10    /api/v1/teams/119                2                FALSE
#> 11    /api/v1/teams/139                6                 TRUE
#> 12    /api/v1/teams/117               12                 TRUE
#> 13    /api/v1/teams/158                3                 TRUE
#> 14    /api/v1/teams/117               NA                   NA
#> 15    /api/v1/teams/139                6                 TRUE
#> 16    /api/v1/teams/137                0                FALSE
#> 17    /api/v1/teams/117                1                FALSE
#> 18    /api/v1/teams/158                5                 TRUE
#> 19    /api/v1/teams/137                7                 TRUE
#> 20    /api/v1/teams/119                1                FALSE
#> 21    /api/v1/teams/111                5                 TRUE
#> 22    /api/v1/teams/111                5                FALSE
#> 23    /api/v1/teams/119                3                 TRUE
#> 24    /api/v1/teams/119                5                 TRUE
#> 25    /api/v1/teams/117               12                 TRUE
#> 26    /api/v1/teams/144                6                 TRUE
#> 27    /api/v1/teams/117                2                FALSE
#> 28    /api/v1/teams/117                1                FALSE
#> 29    /api/v1/teams/144                2                FALSE
#> 30    /api/v1/teams/144               11                 TRUE
#> 31    /api/v1/teams/111                5                 TRUE
#> 32    /api/v1/teams/119                4                 TRUE
#> 33    /api/v1/teams/144                2                FALSE
#> 34    /api/v1/teams/144                7                 TRUE
#> 35    /api/v1/teams/117                2                 TRUE
#> 36    /api/v1/teams/117                3                 TRUE
#> 37    /api/v1/teams/117                5                FALSE
#> 38    /api/v1/teams/144                0                FALSE
#>    teams_home_split_squad teams_home_series_number
#> 1                   FALSE                        1
#> 2                   FALSE                        2
#> 3                   FALSE                        2
#> 4                   FALSE                        1
#> 5                   FALSE                        2
#> 6                   FALSE                        4
#> 7                   FALSE                        1
#> 8                   FALSE                        3
#> 9                   FALSE                        4
#> 10                  FALSE                        3
#> 11                  FALSE                        1
#> 12                  FALSE                        2
#> 13                  FALSE                        4
#> 14                  FALSE                        2
#> 15                  FALSE                        1
#> 16                  FALSE                        3
#> 17                  FALSE                        2
#> 18                  FALSE                        4
#> 19                  FALSE                        3
#> 20                  FALSE                        3
#> 21                  FALSE                        1
#> 22                  FALSE                        1
#> 23                  FALSE                        2
#> 24                  FALSE                        2
#> 25                  FALSE                        1
#> 26                  FALSE                        2
#> 27                  FALSE                        1
#> 28                  FALSE                        1
#> 29                  FALSE                        2
#> 30                  FALSE                        2
#> 31                  FALSE                        1
#> 32                  FALSE                        2
#> 33                  FALSE                        1
#> 34                  FALSE                        1
#> 35                  FALSE                        1
#> 36                  FALSE                        1
#> 37                  FALSE                        1
#> 38                  FALSE                        1
#>    teams_home_league_record_wins teams_home_league_record_losses
#> 1                              1                               0
#> 2                              1                               0
#> 3                              1                               0
#> 4                              1                               0
#> 5                              2                               0
#> 6                              1                               0
#> 7                              1                               1
#> 8                              1                               0
#> 9                              1                               1
#> 10                             1                               1
#> 11                             2                               1
#> 12                             1                               2
#> 13                             2                               1
#> 14                             1                               3
#> 15                             3                               1
#> 16                             1                               2
#> 17                             1                               3
#> 18                             3                               1
#> 19                             2                               2
#> 20                             2                               3
#> 21                             1                               0
#> 22                             1                               1
#> 23                             1                               0
#> 24                             2                               0
#> 25                             2                               1
#> 26                             1                               2
#> 27                             2                               2
#> 28                             2                               3
#> 29                             1                               3
#> 30                             2                               3
#> 31                             4                               2
#> 32                             4                               2
#> 33                             0                               1
#> 34                             1                               1
#> 35                             2                               1
#> 36                             3                               1
#> 37                             3                               2
#> 38                             2                               4
#>    teams_home_league_record_pct teams_home_team_id teams_home_team_name
#> 1                         1.000                111       Boston Red Sox
#> 2                         1.000                119  Los Angeles Dodgers
#> 3                         1.000                117       Houston Astros
#> 4                         1.000                139       Tampa Bay Rays
#> 5                         1.000                117       Houston Astros
#> 6                         1.000                158    Milwaukee Brewers
#> 7                          .500                139       Tampa Bay Rays
#> 8                         1.000                137 San Francisco Giants
#> 9                          .500                158    Milwaukee Brewers
#> 10                         .500                137 San Francisco Giants
#> 11                         .667                111       Boston Red Sox
#> 12                         .333                145    Chicago White Sox
#> 13                         .667                144       Atlanta Braves
#> 14                         .250                145    Chicago White Sox
#> 15                         .750                111       Boston Red Sox
#> 16                         .333                119  Los Angeles Dodgers
#> 17                         .250                145    Chicago White Sox
#> 18                         .750                144       Atlanta Braves
#> 19                         .500                119  Los Angeles Dodgers
#> 20                         .400                137 San Francisco Giants
#> 21                        1.000                117       Houston Astros
#> 22                         .500                117       Houston Astros
#> 23                        1.000                144       Atlanta Braves
#> 24                        1.000                144       Atlanta Braves
#> 25                         .667                111       Boston Red Sox
#> 26                         .333                119  Los Angeles Dodgers
#> 27                         .500                111       Boston Red Sox
#> 28                         .400                111       Boston Red Sox
#> 29                         .250                119  Los Angeles Dodgers
#> 30                         .400                119  Los Angeles Dodgers
#> 31                         .667                117       Houston Astros
#> 32                         .667                144       Atlanta Braves
#> 33                         .000                117       Houston Astros
#> 34                         .500                117       Houston Astros
#> 35                         .667                144       Atlanta Braves
#> 36                         .750                144       Atlanta Braves
#> 37                         .600                144       Atlanta Braves
#> 38                         .333                117       Houston Astros
#>    teams_home_team_link venue_id            venue_name
#> 1     /api/v1/teams/111        3           Fenway Park
#> 2     /api/v1/teams/119       22        Dodger Stadium
#> 3     /api/v1/teams/117     2392      Minute Maid Park
#> 4     /api/v1/teams/139       12       Tropicana Field
#> 5     /api/v1/teams/117     2392      Minute Maid Park
#> 6     /api/v1/teams/158       32 American Family Field
#> 7     /api/v1/teams/139       12       Tropicana Field
#> 8     /api/v1/teams/137     2395           Oracle Park
#> 9     /api/v1/teams/158       32 American Family Field
#> 10    /api/v1/teams/137     2395           Oracle Park
#> 11    /api/v1/teams/111        3           Fenway Park
#> 12    /api/v1/teams/145        4 Guaranteed Rate Field
#> 13    /api/v1/teams/144     4705           Truist Park
#> 14    /api/v1/teams/145        4 Guaranteed Rate Field
#> 15    /api/v1/teams/111        3           Fenway Park
#> 16    /api/v1/teams/119       22        Dodger Stadium
#> 17    /api/v1/teams/145        4 Guaranteed Rate Field
#> 18    /api/v1/teams/144     4705           Truist Park
#> 19    /api/v1/teams/119       22        Dodger Stadium
#> 20    /api/v1/teams/137     2395           Oracle Park
#> 21    /api/v1/teams/117     2392      Minute Maid Park
#> 22    /api/v1/teams/117     2392      Minute Maid Park
#> 23    /api/v1/teams/144     4705           Truist Park
#> 24    /api/v1/teams/144     4705           Truist Park
#> 25    /api/v1/teams/111        3           Fenway Park
#> 26    /api/v1/teams/119       22        Dodger Stadium
#> 27    /api/v1/teams/111        3           Fenway Park
#> 28    /api/v1/teams/111        3           Fenway Park
#> 29    /api/v1/teams/119       22        Dodger Stadium
#> 30    /api/v1/teams/119       22        Dodger Stadium
#> 31    /api/v1/teams/117     2392      Minute Maid Park
#> 32    /api/v1/teams/144     4705           Truist Park
#> 33    /api/v1/teams/117     2392      Minute Maid Park
#> 34    /api/v1/teams/117     2392      Minute Maid Park
#> 35    /api/v1/teams/144     4705           Truist Park
#> 36    /api/v1/teams/144     4705           Truist Park
#> 37    /api/v1/teams/144     4705           Truist Park
#> 38    /api/v1/teams/117     2392      Minute Maid Park
#>             venue_link                content_link inning_break_length
#> 1     /api/v1/venues/3 /api/v1/game/660938/content                  NA
#> 2    /api/v1/venues/22 /api/v1/game/660937/content                  NA
#> 3  /api/v1/venues/2392 /api/v1/game/660936/content                 120
#> 4    /api/v1/venues/12 /api/v1/game/660929/content                 120
#> 5  /api/v1/venues/2392 /api/v1/game/660934/content                 120
#> 6    /api/v1/venues/32 /api/v1/game/660917/content                 120
#> 7    /api/v1/venues/12 /api/v1/game/660932/content                 120
#> 8  /api/v1/venues/2395 /api/v1/game/660933/content                  NA
#> 9    /api/v1/venues/32 /api/v1/game/660907/content                 120
#> 10 /api/v1/venues/2395 /api/v1/game/660925/content                  NA
#> 11    /api/v1/venues/3 /api/v1/game/660928/content                 120
#> 12    /api/v1/venues/4 /api/v1/game/660931/content                 120
#> 13 /api/v1/venues/4705 /api/v1/game/660905/content                  NA
#> 14    /api/v1/venues/4 /api/v1/game/660926/content                  NA
#> 15    /api/v1/venues/3 /api/v1/game/660927/content                  NA
#> 16   /api/v1/venues/22 /api/v1/game/660919/content                  NA
#> 17    /api/v1/venues/4 /api/v1/game/660926/content                  NA
#> 18 /api/v1/venues/4705 /api/v1/game/660914/content                  NA
#> 19   /api/v1/venues/22 /api/v1/game/660918/content                  NA
#> 20 /api/v1/venues/2395 /api/v1/game/660920/content                  NA
#> 21 /api/v1/venues/2392 /api/v1/game/660896/content                  NA
#> 22 /api/v1/venues/2392 /api/v1/game/660913/content                  NA
#> 23 /api/v1/venues/4705 /api/v1/game/660924/content                  NA
#> 24 /api/v1/venues/4705 /api/v1/game/660922/content                 120
#> 25    /api/v1/venues/3 /api/v1/game/660898/content                  NA
#> 26   /api/v1/venues/22 /api/v1/game/660916/content                 120
#> 27    /api/v1/venues/3 /api/v1/game/660923/content                  NA
#> 28    /api/v1/venues/3 /api/v1/game/660900/content                  NA
#> 29   /api/v1/venues/22 /api/v1/game/660904/content                  NA
#> 30   /api/v1/venues/22 /api/v1/game/660903/content                 120
#> 31 /api/v1/venues/2392 /api/v1/game/660899/content                  NA
#> 32 /api/v1/venues/4705 /api/v1/game/660921/content                 120
#> 33 /api/v1/venues/2392 /api/v1/game/660897/content                 120
#> 34 /api/v1/venues/2392 /api/v1/game/660908/content                 120
#> 35 /api/v1/venues/4705 /api/v1/game/660910/content                 120
#> 36 /api/v1/venues/4705 /api/v1/game/660902/content                 120
#> 37 /api/v1/venues/4705 /api/v1/game/660911/content                 120
#> 38 /api/v1/venues/2392 /api/v1/game/660906/content                 120
#>         reschedule_date reschedule_game_date status_reason
#> 1                  <NA>                 <NA>          <NA>
#> 2                  <NA>                 <NA>          <NA>
#> 3                  <NA>                 <NA>          <NA>
#> 4                  <NA>                 <NA>          <NA>
#> 5                  <NA>                 <NA>          <NA>
#> 6                  <NA>                 <NA>          <NA>
#> 7                  <NA>                 <NA>          <NA>
#> 8                  <NA>                 <NA>          <NA>
#> 9                  <NA>                 <NA>          <NA>
#> 10                 <NA>                 <NA>          <NA>
#> 11                 <NA>                 <NA>          <NA>
#> 12                 <NA>                 <NA>          <NA>
#> 13                 <NA>                 <NA>          <NA>
#> 14 2021-10-12T18:07:00Z           2021-10-12          Rain
#> 15                 <NA>                 <NA>          <NA>
#> 16                 <NA>                 <NA>          <NA>
#> 17                 <NA>                 <NA>          <NA>
#> 18                 <NA>                 <NA>          <NA>
#> 19                 <NA>                 <NA>          <NA>
#> 20                 <NA>                 <NA>          <NA>
#> 21                 <NA>                 <NA>          <NA>
#> 22                 <NA>                 <NA>          <NA>
#> 23                 <NA>                 <NA>          <NA>
#> 24                 <NA>                 <NA>          <NA>
#> 25                 <NA>                 <NA>          <NA>
#> 26                 <NA>                 <NA>          <NA>
#> 27                 <NA>                 <NA>          <NA>
#> 28                 <NA>                 <NA>          <NA>
#> 29                 <NA>                 <NA>          <NA>
#> 30                 <NA>                 <NA>          <NA>
#> 31                 <NA>                 <NA>          <NA>
#> 32                 <NA>                 <NA>          <NA>
#> 33                 <NA>                 <NA>          <NA>
#> 34                 <NA>                 <NA>          <NA>
#> 35                 <NA>                 <NA>          <NA>
#> 36                 <NA>                 <NA>          <NA>
#> 37                 <NA>                 <NA>          <NA>
#> 38                 <NA>                 <NA>          <NA>
#>        rescheduled_from rescheduled_from_date is_default_game events
#> 1                  <NA>                  <NA>              NA   NULL
#> 2                  <NA>                  <NA>              NA   NULL
#> 3                  <NA>                  <NA>              NA   NULL
#> 4                  <NA>                  <NA>              NA   NULL
#> 5                  <NA>                  <NA>              NA   NULL
#> 6                  <NA>                  <NA>              NA   NULL
#> 7                  <NA>                  <NA>              NA   NULL
#> 8                  <NA>                  <NA>              NA   NULL
#> 9                  <NA>                  <NA>              NA   NULL
#> 10                 <NA>                  <NA>              NA   NULL
#> 11                 <NA>                  <NA>              NA   NULL
#> 12                 <NA>                  <NA>              NA   NULL
#> 13                 <NA>                  <NA>              NA   NULL
#> 14                 <NA>                  <NA>              NA   NULL
#> 15                 <NA>                  <NA>              NA   NULL
#> 16                 <NA>                  <NA>              NA   NULL
#> 17 2021-10-11T19:37:00Z            2021-10-11              NA   NULL
#> 18                 <NA>                  <NA>              NA   NULL
#> 19                 <NA>                  <NA>              NA   NULL
#> 20                 <NA>                  <NA>              NA   NULL
#> 21                 <NA>                  <NA>              NA   NULL
#> 22                 <NA>                  <NA>              NA   NULL
#> 23                 <NA>                  <NA>              NA   NULL
#> 24                 <NA>                  <NA>              NA   NULL
#> 25                 <NA>                  <NA>              NA   NULL
#> 26                 <NA>                  <NA>              NA   NULL
#> 27                 <NA>                  <NA>              NA   NULL
#> 28                 <NA>                  <NA>              NA   NULL
#> 29                 <NA>                  <NA>              NA   NULL
#> 30                 <NA>                  <NA>              NA   NULL
#> 31                 <NA>                  <NA>              NA   NULL
#> 32                 <NA>                  <NA>              NA   NULL
#> 33                 <NA>                  <NA>              NA   NULL
#> 34                 <NA>                  <NA>              NA   NULL
#> 35                 <NA>                  <NA>              NA   NULL
#> 36                 <NA>                  <NA>              NA   NULL
#> 37                 <NA>                  <NA>              NA   NULL
#> 38                 <NA>                  <NA>            TRUE   NULL
# }