MLB Standings

MLB Standings

Usage

mlb_standings(
  season = NULL,
  date = NULL,
  standings_type = NULL,
  league_id = NULL
)

Arguments

season

Year(s) to return to return standings information for.

date

Date to return to return standings information for.

standings_type

The standings_type(s) to return standings information for. Description of all standings_types

regularSeason - Regular Season Standings

wildCard - Wild card standings

divisionLeaders - Division Leader standings

wildCardWithLeaders - Wild card standings with Division Leaders

firstHalf - First half standings. Only valid for leagues with a split season (Mexican League).

secondHalf - Second half standings. Only valid for leagues with a split season (Mexican League).

springTraining - Spring Training Standings

postseason - Postseason Standings

byDivision - Standings by Division

byConference - Standings by Conference

byLeague - Standings by League

league_id

The league_id(s) to return standings information for.

Value

Returns a tibble with the following columns

col_name	types

standings_type	character

last_updated	character

team_records_season	character

team_records_clinch_indicator	character

team_records_division_rank	character

team_records_league_rank	character

team_records_sport_rank	character

team_records_games_played	integer

team_records_games_back	character

team_records_wild_card_games_back	character

team_records_league_games_back	character

team_records_spring_league_games_back	character

team_records_sport_games_back	character

team_records_division_games_back	character

team_records_conference_games_back	character

team_records_last_updated	character

team_records_runs_allowed	integer

team_records_runs_scored	integer

team_records_division_champ	logical

team_records_division_leader	logical

team_records_has_wildcard	logical

team_records_clinched	logical

team_records_elimination_number	character

team_records_wild_card_elimination_number	character

team_records_magic_number	character

team_records_wins	integer

team_records_losses	integer

team_records_run_differential	integer

team_records_winning_percentage	character

team_records_wild_card_rank	character

team_records_wild_card_leader	logical

team_records_team_id	integer

team_records_team_name	character

team_records_team_link	character

team_records_streak_streak_type	character

team_records_streak_streak_number	integer

team_records_streak_streak_code	character

team_records_league_record_wins	integer

team_records_league_record_losses	integer

team_records_league_record_ties	integer

team_records_league_record_pct	character

team_records_records_split_records	list

team_records_records_division_records	list

team_records_records_overall_records	list

team_records_records_league_records	list

team_records_records_expected_records	list

league_id	integer

league_link	character

division_id	integer

division_link	character

sport_id	integer

sport_link	character

Examples

# \donttest{
  try(mlb_standings(season = 2021, league_id = 103))
#>    standings_type             last_updated team_records_season
#> 1   regularSeason 2022-03-14T05:15:01.572Z                2021
#> 2   regularSeason 2022-03-14T05:15:01.572Z                2021
#> 3   regularSeason 2022-03-14T05:15:01.572Z                2021
#> 4   regularSeason 2022-03-14T05:15:01.572Z                2021
#> 5   regularSeason 2022-03-14T05:15:01.572Z                2021
#> 6   regularSeason 2022-03-31T19:50:45.503Z                2021
#> 7   regularSeason 2022-03-31T19:50:45.503Z                2021
#> 8   regularSeason 2022-03-31T19:50:45.503Z                2021
#> 9   regularSeason 2022-03-31T19:50:45.503Z                2021
#> 10  regularSeason 2022-03-31T19:50:45.503Z                2021
#> 11  regularSeason 2022-03-14T05:14:55.108Z                2021
#> 12  regularSeason 2022-03-14T05:14:55.108Z                2021
#> 13  regularSeason 2022-03-14T05:14:55.108Z                2021
#> 14  regularSeason 2022-03-14T05:14:55.108Z                2021
#> 15  regularSeason 2022-03-14T05:14:55.108Z                2021
#>    team_records_clinch_indicator team_records_division_rank
#> 1                              z                          1
#> 2                              w                          2
#> 3                              w                          3
#> 4                           <NA>                          4
#> 5                           <NA>                          5
#> 6                              y                          1
#> 7                           <NA>                          2
#> 8                           <NA>                          3
#> 9                           <NA>                          4
#> 10                          <NA>                          5
#> 11                             y                          1
#> 12                          <NA>                          2
#> 13                          <NA>                          3
#> 14                          <NA>                          4
#> 15                          <NA>                          5
#>    team_records_league_rank team_records_sport_rank
#> 1                         1                       1
#> 2                         4                       4
#> 3                         5                       5
#> 4                         6                       6
#> 5                        15                      15
#> 6                         3                       3
#> 7                         9                       9
#> 8                        10                      10
#> 9                        12                      12
#> 10                       13                      13
#> 11                        2                       2
#> 12                        7                       7
#> 13                        8                       8
#> 14                       11                      11
#> 15                       14                      14
#>    team_records_games_played team_records_games_back
#> 1                        162                       -
#> 2                        162                     8.0
#> 3                        162                     8.0
#> 4                        162                     9.0
#> 5                        162                    48.0
#> 6                        162                       -
#> 7                        162                    13.0
#> 8                        162                    16.0
#> 9                        162                    19.0
#> 10                       162                    20.0
#> 11                       162                       -
#> 12                       162                     5.0
#> 13                       162                     9.0
#> 14                       162                    18.0
#> 15                       162                    35.0
#>    team_records_wild_card_games_back team_records_league_games_back
#> 1                                  -                              -
#> 2                                  -                            8.0
#> 3                                  -                            8.0
#> 4                                1.0                            9.0
#> 5                               40.0                           48.0
#> 6                                  -                            7.0
#> 7                               12.0                           20.0
#> 8                               15.0                           23.0
#> 9                               18.0                           26.0
#> 10                              19.0                           27.0
#> 11                                 -                            5.0
#> 12                               2.0                           10.0
#> 13                               6.0                           14.0
#> 14                              15.0                           23.0
#> 15                              32.0                           40.0
#>    team_records_spring_league_games_back team_records_sport_games_back
#> 1                                      -                             -
#> 2                                      -                           8.0
#> 3                                      -                           8.0
#> 4                                      -                           9.0
#> 5                                      -                          48.0
#> 6                                      -                           7.0
#> 7                                      -                          20.0
#> 8                                      -                          23.0
#> 9                                      -                          26.0
#> 10                                     -                          27.0
#> 11                                     -                           5.0
#> 12                                     -                          10.0
#> 13                                     -                          14.0
#> 14                                     -                          23.0
#> 15                                     -                          40.0
#>    team_records_division_games_back team_records_conference_games_back
#> 1                                 -                                  -
#> 2                               8.0                                  -
#> 3                               8.0                                  -
#> 4                               9.0                                  -
#> 5                              48.0                                  -
#> 6                                 -                                  -
#> 7                              13.0                                  -
#> 8                              16.0                                  -
#> 9                              19.0                                  -
#> 10                             20.0                                  -
#> 11                                -                                  -
#> 12                              5.0                                  -
#> 13                              9.0                                  -
#> 14                             18.0                                  -
#> 15                             35.0                                  -
#>    team_records_last_updated team_records_runs_allowed
#> 1       2022-02-15T20:15:03Z                       651
#> 2       2022-02-16T17:53:30Z                       749
#> 3       2022-02-15T19:55:25Z                       669
#> 4       2022-02-15T20:09:55Z                       663
#> 5       2022-03-14T05:15:01Z                       956
#> 6       2022-03-31T19:50:45Z                       636
#> 7       2022-02-16T17:10:25Z                       727
#> 8       2021-11-22T16:34:42Z                       756
#> 9       2021-11-23T17:57:50Z                       788
#> 10      2021-11-23T17:57:51Z                       834
#> 11      2022-03-14T05:14:55Z                       658
#> 12      2021-11-22T16:33:55Z                       748
#> 13      2021-11-22T16:34:26Z                       687
#> 14      2022-01-11T19:27:15Z                       804
#> 15      2021-11-23T00:06:06Z                       815
#>    team_records_runs_scored team_records_division_champ
#> 1                       857                        TRUE
#> 2                       829                       FALSE
#> 3                       711                       FALSE
#> 4                       846                       FALSE
#> 5                       659                       FALSE
#> 6                       796                        TRUE
#> 7                       717                       FALSE
#> 8                       697                       FALSE
#> 9                       686                       FALSE
#> 10                      729                       FALSE
#> 11                      863                        TRUE
#> 12                      697                       FALSE
#> 13                      743                       FALSE
#> 14                      723                       FALSE
#> 15                      625                       FALSE
#>    team_records_division_leader team_records_has_wildcard
#> 1                          TRUE                      TRUE
#> 2                         FALSE                      TRUE
#> 3                         FALSE                      TRUE
#> 4                         FALSE                      TRUE
#> 5                         FALSE                      TRUE
#> 6                          TRUE                      TRUE
#> 7                         FALSE                      TRUE
#> 8                         FALSE                      TRUE
#> 9                         FALSE                      TRUE
#> 10                        FALSE                      TRUE
#> 11                         TRUE                      TRUE
#> 12                        FALSE                      TRUE
#> 13                        FALSE                      TRUE
#> 14                        FALSE                      TRUE
#> 15                        FALSE                      TRUE
#>    team_records_clinched team_records_elimination_number
#> 1                   TRUE                               -
#> 2                   TRUE                               E
#> 3                   TRUE                               E
#> 4                  FALSE                               E
#> 5                  FALSE                               E
#> 6                   TRUE                               -
#> 7                  FALSE                               E
#> 8                  FALSE                               E
#> 9                  FALSE                               E
#> 10                 FALSE                               E
#> 11                  TRUE                               -
#> 12                 FALSE                               E
#> 13                 FALSE                               E
#> 14                 FALSE                               E
#> 15                 FALSE                               E
#>    team_records_wild_card_elimination_number team_records_magic_number
#> 1                                          -                         -
#> 2                                          -                      <NA>
#> 3                                          -                      <NA>
#> 4                                          E                      <NA>
#> 5                                          E                      <NA>
#> 6                                          -                         -
#> 7                                          E                      <NA>
#> 8                                          E                      <NA>
#> 9                                          E                      <NA>
#> 10                                         E                      <NA>
#> 11                                         -                         -
#> 12                                         E                      <NA>
#> 13                                         E                      <NA>
#> 14                                         E                      <NA>
#> 15                                         E                      <NA>
#>    team_records_wins team_records_losses team_records_run_differential
#> 1                100                  62                           206
#> 2                 92                  70                            80
#> 3                 92                  70                            42
#> 4                 91                  71                           183
#> 5                 52                 110                          -297
#> 6                 93                  69                           160
#> 7                 80                  82                           -10
#> 8                 77                  85                           -59
#> 9                 74                  88                          -102
#> 10                73                  89                          -105
#> 11                95                  67                           205
#> 12                90                  72                           -51
#> 13                86                  76                            56
#> 14                77                  85                           -81
#> 15                60                 102                          -190
#>    team_records_winning_percentage team_records_wild_card_rank
#> 1                             .617                        <NA>
#> 2                             .568                           1
#> 3                             .568                           2
#> 4                             .562                           3
#> 5                             .321                          12
#> 6                             .574                        <NA>
#> 7                             .494                           6
#> 8                             .475                           7
#> 9                             .457                           9
#> 10                            .451                          10
#> 11                            .586                        <NA>
#> 12                            .556                           4
#> 13                            .531                           5
#> 14                            .475                           8
#> 15                            .370                          11
#>    team_records_wild_card_leader team_records_team_id
#> 1                             NA                  139
#> 2                           TRUE                  111
#> 3                           TRUE                  147
#> 4                             NA                  141
#> 5                             NA                  110
#> 6                             NA                  145
#> 7                             NA                  114
#> 8                             NA                  116
#> 9                             NA                  118
#> 10                            NA                  142
#> 11                            NA                  117
#> 12                            NA                  136
#> 13                            NA                  133
#> 14                            NA                  108
#> 15                            NA                  140
#>    team_records_team_name team_records_team_link
#> 1          Tampa Bay Rays      /api/v1/teams/139
#> 2          Boston Red Sox      /api/v1/teams/111
#> 3        New York Yankees      /api/v1/teams/147
#> 4       Toronto Blue Jays      /api/v1/teams/141
#> 5       Baltimore Orioles      /api/v1/teams/110
#> 6       Chicago White Sox      /api/v1/teams/145
#> 7       Cleveland Indians      /api/v1/teams/114
#> 8          Detroit Tigers      /api/v1/teams/116
#> 9      Kansas City Royals      /api/v1/teams/118
#> 10        Minnesota Twins      /api/v1/teams/142
#> 11         Houston Astros      /api/v1/teams/117
#> 12       Seattle Mariners      /api/v1/teams/136
#> 13      Oakland Athletics      /api/v1/teams/133
#> 14     Los Angeles Angels      /api/v1/teams/108
#> 15          Texas Rangers      /api/v1/teams/140
#>    team_records_streak_streak_type team_records_streak_streak_number
#> 1                           losses                                 1
#> 2                             wins                                 3
#> 3                             wins                                 1
#> 4                             wins                                 3
#> 5                           losses                                 3
#> 6                           losses                                 1
#> 7                             wins                                 1
#> 8                             wins                                 1
#> 9                           losses                                 2
#> 10                            wins                                 2
#> 11                            wins                                 2
#> 12                          losses                                 1
#> 13                          losses                                 2
#> 14                            wins                                 1
#> 15                          losses                                 1
#>    team_records_streak_streak_code team_records_league_record_wins
#> 1                               L1                             100
#> 2                               W3                              92
#> 3                               W1                              92
#> 4                               W3                              91
#> 5                               L3                              52
#> 6                               L1                              93
#> 7                               W1                              80
#> 8                               W1                              77
#> 9                               L2                              74
#> 10                              W2                              73
#> 11                              W2                              95
#> 12                              L1                              90
#> 13                              L2                              86
#> 14                              W1                              77
#> 15                              L1                              60
#>    team_records_league_record_losses team_records_league_record_ties
#> 1                                 62                               0
#> 2                                 70                               0
#> 3                                 70                               0
#> 4                                 71                               0
#> 5                                110                               0
#> 6                                 69                               0
#> 7                                 82                               0
#> 8                                 85                               0
#> 9                                 88                               0
#> 10                                89                               0
#> 11                                67                               0
#> 12                                72                               0
#> 13                                76                               0
#> 14                                85                               0
#> 15                               102                               0
#>    team_records_league_record_pct
#> 1                            .617
#> 2                            .568
#> 3                            .568
#> 4                            .562
#> 5                            .321
#> 6                            .574
#> 7                            .494
#> 8                            .475
#> 9                            .457
#> 10                           .451
#> 11                           .586
#> 12                           .556
#> 13                           .531
#> 14                           .475
#> 15                           .370
#>                                                                                                                                                                                                                                                                                                                                team_records_records_split_records
#> 1  52, 48, 38, 18, 20, 34, 28, 62, 7, 7, 20, 48, 41, 59, 45, 55, 29, 33, 27, 15, 12, 14, 21, 35, 3, 12, 25, 42, 20, 42, 30, 32, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .642, .593, .585, .545, .625, .708, .571, .639, .700, .368, .444, .533, .672, .584, .600, .632
#> 2   49, 43, 35, 20, 15, 29, 28, 57, 5, 11, 26, 46, 31, 61, 88, 4, 32, 38, 31, 13, 18, 19, 20, 39, 5, 5, 18, 48, 31, 39, 57, 13, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .605, .531, .530, .606, .455, .604, .583, .594, .500, .688, .591, .489, .500, .610, .607, .235
#> 3   46, 46, 32, 16, 16, 30, 30, 60, 7, 10, 28, 48, 25, 67, 82, 10, 35, 35, 24, 13, 11, 22, 24, 46, 3, 8, 20, 43, 34, 36, 64, 6, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .568, .568, .571, .552, .593, .577, .556, .566, .700, .556, .583, .527, .424, .650, .562, .625
#> 4   47, 44, 23, 12, 11, 35, 33, 68, 6, 3, 15, 48, 41, 50, 61, 30, 33, 38, 20, 10, 10, 23, 28, 51, 4, 9, 15, 44, 21, 50, 54, 17, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .588, .537, .535, .545, .524, .603, .541, .571, .600, .250, .500, .522, .661, .500, .530, .638
#> 5     27, 25, 22, 14, 8, 13, 17, 30, 4, 4, 13, 30, 15, 37, 48, 4, 54, 56, 43, 20, 23, 34, 33, 67, 6, 7, 24, 78, 44, 66, 95, 15, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .333, .309, .338, .412, .258, .277, .340, .309, .400, .364, .351, .278, .254, .359, .336, .211
#> 6       53, 40, 25, 14, 11, 39, 29, 68, 7, 6, 18, 27, 35, 58, 88, 5, 28, 41, 17, 8, 9, 20, 32, 52, 3, 7, 24, 29, 29, 40, 64, 5, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .654, .494, .595, .636, .550, .661, .475, .567, .700, .462, .429, .482, .547, .592, .579, .500
#> 7     40, 40, 29, 14, 15, 26, 25, 51, 5, 7, 17, 28, 34, 46, 77, 3, 41, 41, 30, 15, 15, 26, 26, 52, 5, 9, 22, 51, 26, 56, 75, 7, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .494, .494, .492, .483, .500, .500, .490, .495, .500, .438, .436, .354, .567, .451, .507, .300
#> 8    42, 35, 29, 15, 14, 27, 21, 48, 3, 12, 23, 39, 35, 42, 71, 6, 39, 46, 25, 11, 14, 28, 32, 60, 7, 8, 23, 36, 37, 48, 81, 4, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .519, .432, .537, .577, .500, .491, .396, .444, .300, .600, .500, .520, .486, .467, .467, .600
#> 9     39, 35, 26, 14, 12, 25, 23, 48, 5, 4, 21, 35, 30, 44, 73, 1, 42, 46, 24, 13, 11, 29, 35, 64, 5, 5, 19, 44, 36, 52, 80, 8, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .481, .432, .520, .519, .522, .463, .397, .429, .500, .444, .525, .443, .455, .458, .477, .111
#> 10  38, 35, 20, 10, 10, 28, 25, 53, 6, 11, 25, 29, 29, 44, 68, 5, 43, 46, 36, 21, 15, 22, 31, 53, 4, 12, 19, 49, 42, 47, 85, 4, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .469, .432, .357, .323, .400, .560, .446, .500, .600, .478, .568, .372, .408, .484, .444, .556
#> 11    51, 44, 34, 18, 16, 33, 28, 61, 4, 9, 21, 45, 26, 69, 88, 7, 30, 37, 28, 12, 16, 18, 21, 39, 6, 8, 19, 32, 26, 41, 61, 6, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .630, .543, .548, .600, .500, .647, .571, .610, .400, .529, .525, .584, .500, .627, .591, .538
#> 12  46, 44, 36, 18, 18, 28, 26, 54, 7, 14, 33, 44, 33, 57, 80, 10, 35, 37, 28, 14, 14, 21, 23, 44, 3, 7, 19, 34, 23, 49, 67, 5, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .568, .543, .563, .563, .563, .571, .531, .551, .700, .667, .635, .564, .589, .538, .544, .667
#> 13   43, 43, 37, 16, 21, 27, 22, 49, 4, 8, 23, 30, 39, 47, 77, 9, 38, 38, 26, 16, 10, 22, 28, 50, 6, 3, 27, 51, 31, 45, 66, 10, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .531, .531, .587, .500, .677, .551, .440, .495, .400, .727, .460, .370, .557, .511, .538, .474
#> 14   40, 37, 30, 17, 13, 23, 24, 47, 5, 3, 25, 39, 22, 55, 67, 10, 42, 43, 25, 10, 15, 32, 28, 60, 5, 5, 14, 62, 27, 58, 79, 6, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .488, .463, .545, .630, .464, .418, .462, .439, .500, .375, .641, .386, .449, .487, .459, .625
#> 15  36, 24, 27, 15, 12, 21, 12, 33, 5, 9, 18, 34, 22, 38, 19, 41, 45, 57, 34, 11, 23, 34, 34, 68, 5, 9, 21, 63, 34, 68, 56, 46, home, away, left, leftHome, leftAway, rightHome, rightAway, right, lastTen, extraInning, oneRun, winners, day, night, grass, turf, .444, .296, .443, .577, .343, .382, .261, .327, .500, .500, .462, .351, .393, .358, .253, .471
#>                                                                                                                                                                team_records_records_division_records
#> 1  15, 51, 19, 19, 25, 13, .441, .671, .594, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 2  15, 41, 20, 18, 35, 13, .455, .539, .606, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 3  22, 36, 22, 12, 40, 10, .647, .474, .688, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 4  16, 42, 19, 16, 34, 15, .500, .553, .559, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 5  15, 20, 10, 17, 56, 24, .469, .263, .294, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 6  16, 19, 44, 17, 14, 32, .485, .576, .579, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 7  15, 13, 43, 17, 21, 33, .469, .382, .566, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 8  18, 18, 30, 16, 14, 46, .529, .563, .395, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 9  14, 12, 36, 19, 21, 40, .424, .364, .474, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 10 13, 13, 37, 20, 20, 39, .394, .394, .487, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 11 49, 18, 19, 27, 13, 16, .645, .581, .543, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 12 47, 19, 15, 29, 15, 17, .618, .559, .469, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 13 37, 15, 23, 39, 18, 10, .487, .455, .697, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 14 29, 16, 21, 47, 17, 12, .382, .485, .636, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#> 15 28, 14, 11, 48, 20, 21, .368, .412, .344, 200, 201, 202, American League West, American League East, American League Central, /api/v1/divisions/200, /api/v1/divisions/201, /api/v1/divisions/202
#>      team_records_records_overall_records
#> 1  52, 48, 29, 33, home, away, .642, .593
#> 2  49, 43, 32, 38, home, away, .605, .531
#> 3  46, 46, 35, 35, home, away, .568, .568
#> 4  47, 44, 33, 38, home, away, .588, .537
#> 5  27, 25, 54, 56, home, away, .333, .309
#> 6  53, 40, 28, 41, home, away, .654, .494
#> 7  40, 40, 41, 41, home, away, .494, .494
#> 8  42, 35, 39, 46, home, away, .519, .432
#> 9  39, 35, 42, 46, home, away, .481, .432
#> 10 38, 35, 43, 46, home, away, .469, .432
#> 11 51, 44, 30, 37, home, away, .630, .543
#> 12 46, 44, 35, 37, home, away, .568, .543
#> 13 43, 43, 38, 38, home, away, .531, .531
#> 14 40, 37, 42, 43, home, away, .488, .463
#> 15 36, 24, 45, 57, home, away, .444, .296
#>                                                                               team_records_records_league_records
#> 1   85, 15, 57, 5, .599, .750, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 2   76, 16, 66, 4, .535, .800, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 3   80, 12, 62, 8, .563, .600, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 4   77, 14, 65, 6, .542, .700, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 5   45, 7, 97, 13, .317, .350, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 6   79, 14, 63, 6, .556, .700, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 7   71, 9, 71, 11, .500, .450, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 8   66, 11, 76, 9, .465, .550, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 9   62, 12, 80, 8, .437, .600, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 10 63, 10, 79, 10, .444, .500, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 11  86, 9, 56, 11, .606, .450, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 12  81, 9, 61, 11, .570, .450, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 13  75, 11, 67, 9, .528, .550, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 14  66, 11, 76, 9, .465, .550, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#> 15  53, 7, 89, 13, .373, .350, 103, 104, American League, National League, /api/v1/league/103, /api/v1/league/104
#>                     team_records_records_expected_records league_id
#> 1  101, 101, 61, 61, xWinLoss, xWinLossSeason, .623, .623       103
#> 2    88, 88, 74, 74, xWinLoss, xWinLossSeason, .543, .543       103
#> 3    86, 86, 76, 76, xWinLoss, xWinLossSeason, .531, .531       103
#> 4    99, 99, 63, 63, xWinLoss, xWinLossSeason, .611, .611       103
#> 5  54, 54, 108, 108, xWinLoss, xWinLossSeason, .333, .333       103
#> 6    97, 97, 65, 65, xWinLoss, xWinLossSeason, .599, .599       103
#> 7    80, 80, 82, 82, xWinLoss, xWinLossSeason, .494, .494       103
#> 8    75, 75, 87, 87, xWinLoss, xWinLossSeason, .463, .463       103
#> 9    71, 71, 91, 91, xWinLoss, xWinLossSeason, .438, .438       103
#> 10   71, 71, 91, 91, xWinLoss, xWinLossSeason, .438, .438       103
#> 11 101, 101, 61, 61, xWinLoss, xWinLossSeason, .623, .623       103
#> 12   76, 76, 86, 86, xWinLoss, xWinLossSeason, .469, .469       103
#> 13   87, 87, 75, 75, xWinLoss, xWinLossSeason, .537, .537       103
#> 14   73, 73, 89, 89, xWinLoss, xWinLossSeason, .451, .451       103
#> 15 62, 62, 100, 100, xWinLoss, xWinLossSeason, .383, .383       103
#>           league_link division_id         division_link sport_id
#> 1  /api/v1/league/103         201 /api/v1/divisions/201        1
#> 2  /api/v1/league/103         201 /api/v1/divisions/201        1
#> 3  /api/v1/league/103         201 /api/v1/divisions/201        1
#> 4  /api/v1/league/103         201 /api/v1/divisions/201        1
#> 5  /api/v1/league/103         201 /api/v1/divisions/201        1
#> 6  /api/v1/league/103         202 /api/v1/divisions/202        1
#> 7  /api/v1/league/103         202 /api/v1/divisions/202        1
#> 8  /api/v1/league/103         202 /api/v1/divisions/202        1
#> 9  /api/v1/league/103         202 /api/v1/divisions/202        1
#> 10 /api/v1/league/103         202 /api/v1/divisions/202        1
#> 11 /api/v1/league/103         200 /api/v1/divisions/200        1
#> 12 /api/v1/league/103         200 /api/v1/divisions/200        1
#> 13 /api/v1/league/103         200 /api/v1/divisions/200        1
#> 14 /api/v1/league/103         200 /api/v1/divisions/200        1
#> 15 /api/v1/league/103         200 /api/v1/divisions/200        1
#>          sport_link
#> 1  /api/v1/sports/1
#> 2  /api/v1/sports/1
#> 3  /api/v1/sports/1
#> 4  /api/v1/sports/1
#> 5  /api/v1/sports/1
#> 6  /api/v1/sports/1
#> 7  /api/v1/sports/1
#> 8  /api/v1/sports/1
#> 9  /api/v1/sports/1
#> 10 /api/v1/sports/1
#> 11 /api/v1/sports/1
#> 12 /api/v1/sports/1
#> 13 /api/v1/sports/1
#> 14 /api/v1/sports/1
#> 15 /api/v1/sports/1
# }