MLB Standings

MLB Standings

Usage

mlb_standings(
  season = NULL,
  date = NULL,
  standings_type = NULL,
  league_id = NULL
)

Arguments

season

Year(s) to return to return standings information for.

date

Date to return to return standings information for.

standings_type

The standings_type(s) to return standings information for. Description of all standings_types

regularSeason - Regular Season Standings

wildCard - Wild card standings

divisionLeaders - Division Leader standings

wildCardWithLeaders - Wild card standings with Division Leaders

firstHalf - First half standings. Only valid for leagues with a split season (Mexican League).

secondHalf - Second half standings. Only valid for leagues with a split season (Mexican League).

springTraining - Spring Training Standings

postseason - Postseason Standings

byDivision - Standings by Division

byConference - Standings by Conference

byLeague - Standings by League

league_id

The league_id(s) to return standings information for.

Value

Returns a tibble with the following columns

col_name	types

standings_type	character

last_updated	character

team_records_season	character

team_records_clinch_indicator	character

team_records_division_rank	character

team_records_league_rank	character

team_records_sport_rank	character

team_records_games_played	integer

team_records_games_back	character

team_records_wild_card_games_back	character

team_records_league_games_back	character

team_records_spring_league_games_back	character

team_records_sport_games_back	character

team_records_division_games_back	character

team_records_conference_games_back	character

team_records_last_updated	character

team_records_runs_allowed	integer

team_records_runs_scored	integer

team_records_division_champ	logical

team_records_division_leader	logical

team_records_has_wildcard	logical

team_records_clinched	logical

team_records_elimination_number	character

team_records_wild_card_elimination_number	character

team_records_magic_number	character

team_records_wins	integer

team_records_losses	integer

team_records_run_differential	integer

team_records_winning_percentage	character

team_records_wild_card_rank	character

team_records_wild_card_leader	logical

team_records_team_id	integer

team_records_team_name	character

team_records_team_link	character

team_records_streak_streak_type	character

team_records_streak_streak_number	integer

team_records_streak_streak_code	character

team_records_league_record_wins	integer

team_records_league_record_losses	integer

team_records_league_record_ties	integer

team_records_league_record_pct	character

team_records_records_split_records	list

team_records_records_division_records	list

team_records_records_overall_records	list

team_records_records_league_records	list

team_records_records_expected_records	list

league_id	integer

league_link	character

division_id	integer

division_link	character

sport_id	integer

sport_link	character

Examples

# \donttest{
  try(mlb_standings(season = 2021, league_id = 103))
#> ── MLB Standings data from MLB.com ────────────────── baseballr 1.1.0 ──
#> ℹ Data updated: 2022-04-11 14:34:51 UTC
#> # A tibble: 15 × 52
#>    standings_type last_updated         team_records_se… team_records_cl…
#>    <chr>          <chr>                <chr>            <chr>           
#>  1 regularSeason  2022-03-14T05:15:01… 2021             z               
#>  2 regularSeason  2022-03-14T05:15:01… 2021             w               
#>  3 regularSeason  2022-03-14T05:15:01… 2021             w               
#>  4 regularSeason  2022-03-14T05:15:01… 2021             NA              
#>  5 regularSeason  2022-03-14T05:15:01… 2021             NA              
#>  6 regularSeason  2022-03-31T19:50:45… 2021             y               
#>  7 regularSeason  2022-03-31T19:50:45… 2021             NA              
#>  8 regularSeason  2022-03-31T19:50:45… 2021             NA              
#>  9 regularSeason  2022-03-31T19:50:45… 2021             NA              
#> 10 regularSeason  2022-03-31T19:50:45… 2021             NA              
#> 11 regularSeason  2022-03-14T05:14:55… 2021             y               
#> 12 regularSeason  2022-03-14T05:14:55… 2021             NA              
#> 13 regularSeason  2022-03-14T05:14:55… 2021             NA              
#> 14 regularSeason  2022-03-14T05:14:55… 2021             NA              
#> 15 regularSeason  2022-03-14T05:14:55… 2021             NA              
#> # … with 48 more variables: team_records_division_rank <chr>,
#> #   team_records_league_rank <chr>, team_records_sport_rank <chr>,
#> #   team_records_games_played <int>, team_records_games_back <chr>,
#> #   team_records_wild_card_games_back <chr>,
#> #   team_records_league_games_back <chr>,
#> #   team_records_spring_league_games_back <chr>,
#> #   team_records_sport_games_back <chr>, …
# }