MLB Stats

MLB Stats

Usage

mlb_stats(
  stat_type = NULL,
  player_pool = NULL,
  game_type = NULL,
  team_id = NULL,
  position = NULL,
  stat_group = NULL,
  season = NULL,
  league_id = NULL,
  sport_ids = NULL,
  sort_stat = NULL,
  order = NULL,
  limit = 1000,
  offset = NULL
)

Arguments

stat_type

Stat type to return statistics for.

player_pool

There are 4 different types of player pools to return statistics for a particular player pool across a sport. Acceptable values include: All, Qualified, Rookies, or Qualified_rookies

game_type

Game type to return information for a particular statistic in a particular game type.

team_id

Team ID to return information and ranking for a particular statistic for a particular team.

position

Position to return statistics for a given position. Default to "Qualified" player pool Acceptable values include:

P

C

1B

2B

3B

SS

LF

CF

RF

DH

PH

PR

BR

OF

IF

SP

RP

CP

UT

UI

UO

RHP

LHP

RHS

LHS

LHR

RHR

B

X

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

league_id

League ID to return statistics for a given league. Default to "Qualified" player pool.

sport_ids

The sport_id(s) to return information and ranking information for.

sort_stat

Sort return based on stat.

order

Order return based on either desc or asc.

limit

A limit to limit return to a particular number of records.

offset

An offset to returns i+1 as the first record in the set of players.

Value

Returns a tibble with the following columns

col_name	types

total_splits	integer

season	character

num_teams	integer

rank	integer

games_played	integer

ground_outs	integer

air_outs	integer

runs	integer

doubles	integer

triples	integer

home_runs	integer

strike_outs	integer

base_on_balls	integer

intentional_walks	integer

hits	integer

hit_by_pitch	integer

avg	character

at_bats	integer

obp	character

slg	character

ops	character

caught_stealing	integer

stolen_bases	integer

stolen_base_percentage	character

ground_into_double_play	integer

number_of_pitches	integer

plate_appearances	integer

total_bases	integer

rbi	integer

left_on_base	integer

sac_bunts	integer

sac_flies	integer

babip	character

ground_outs_to_airouts	character

catchers_interference	integer

at_bats_per_home_run	character

team_id	integer

team_name	character

team_link	character

player_id	integer

player_full_name	character

player_link	character

player_first_name	character

player_last_name	character

league_id	integer

league_name	character

league_link	character

sport_id	integer

sport_link	character

sport_abbreviation	character

position_code	character

position_name	character

position_type	character

position_abbreviation	character

splits_tied_with_offset	list

splits_tied_with_limit	list

player_pool	character

type_display_name	character

group_display_name	character

Examples

# \donttest{
  try(mlb_stats(stat_type = 'season', stat_group = 'hitting', season = 2021))
#> ── MLB Stats data from MLB.com ────────────────────── baseballr 1.2.0 ──
#> ℹ Data updated: 2022-06-17 03:02:20 UTC
#> # A tibble: 132 × 59
#>    total_splits season num_teams  rank games_played ground_outs air_outs
#>           <int> <chr>      <int> <int>        <int>       <int>    <int>
#>  1          132 2021           2     1          148         141      153
#>  2          132 2021           1     2          143         131      174
#>  3          132 2021           1     3          151         160       97
#>  4          132 2021           1     4          121         141      130
#>  5          132 2021           1     5          161         161      147
#>  6          132 2021           2     6          120         140       86
#>  7          132 2021           1     7          141         112       97
#>  8          132 2021           1     8          123         163       83
#>  9          132 2021           1     9          138         117      135
#> 10          132 2021           2    10          155         156      180
#> # … with 122 more rows, and 52 more variables: runs <int>,
#> #   doubles <int>, triples <int>, home_runs <int>, strike_outs <int>,
#> #   base_on_balls <int>, intentional_walks <int>, hits <int>,
#> #   hit_by_pitch <int>, avg <chr>, at_bats <int>, obp <chr>, slg <chr>,
#> #   ops <chr>, caught_stealing <int>, stolen_bases <int>,
#> #   stolen_base_percentage <chr>, ground_into_double_play <int>,
#> #   number_of_pitches <int>, plate_appearances <int>, …
# }