MLB Stats Leaders

MLB Stats Leaders

Usage

mlb_stats_leaders(
  leader_categories = NULL,
  player_pool = NULL,
  leader_game_types = NULL,
  sit_codes = NULL,
  position = NULL,
  stat_group = NULL,
  season = NULL,
  league_id = NULL,
  sport_id = NULL,
  start_date = NULL,
  end_date = NULL,
  stat_type = NULL,
  limit = 1000
)

Arguments

leader_categories

League leader category to return information and ranking for a particular statistic.

player_pool

There are 4 different types of player pools to return statistics for a particular player pool across a sport. Acceptable values include: All, Qualified, Rookies, or Qualified_rookies

leader_game_types

Game type to return information and ranking for a particular statistic in a particular game type.

sit_codes

Situation code to return information and ranking for a particular statistic in a particular game type.

position

Position to return statistics for a given position. Default to "Qualified" player pool Acceptable values include:

P

C

1B

2B

3B

SS

LF

CF

RF

DH

PH

PR

BR

OF

IF

SP

RP

CP

UT

UI

UO

RHP

LHP

RHS

LHS

LHR

RHR

B

X

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

league_id

League ID to return statistics for a given league. Default to "Qualified" player pool.

sport_id

The sport_id to return information and ranking information for.

start_date

Start date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY start_date must be coupled with end_date and byDateRange stat_type

end_date

End date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY end_date must be coupled with start_date and byDateRange stat_type

stat_type

The stat_type to return information and ranking for a particular statistic for a particular stat type.

limit

A limit to limit return to a particular number of records.

Value

Returns a tibble with the following columns

col_name	types

leader_category	character

rank	integer

value	character

season	character

num_teams	integer

team_id	integer

team_name	character

team_link	character

league_id	integer

league_name	character

league_link	character

person_id	integer

person_full_name	character

person_link	character

person_first_name	character

person_last_name	character

sport_id	integer

sport_link	character

sport_abbreviation	character

stat_group	character

total_splits	integer

game_type_id	character

game_type_description	character

Examples

# \donttest{
 try(mlb_stats_leaders(leader_categories='homeRuns',sport_id=1, season = 2021))
#> ── MLB Stats Leaders data from MLB.com ────────────── baseballr 1.5.0 ──
#> ℹ Data updated: 2023-03-15 18:36:05 UTC
#> # A tibble: 306 × 23
#>    leader_c…¹  rank value season num_t…² team_id team_…³ team_…⁴ leagu…⁵
#>    <chr>      <int> <chr> <chr>    <int>   <int> <chr>   <chr>     <int>
#>  1 homeRuns       1 48    2021         1     141 Toront… /api/v…     103
#>  2 homeRuns       1 48    2021         1     118 Kansas… /api/v…     103
#>  3 homeRuns       3 46    2021         1     108 Los An… /api/v…     103
#>  4 homeRuns       4 45    2021         1     141 Toront… /api/v…     103
#>  5 homeRuns       5 42    2021         1     135 San Di… /api/v…     104
#>  6 homeRuns       6 39    2021         1     136 Seattl… /api/v…     103
#>  7 homeRuns       6 39    2021         1     147 New Yo… /api/v…     103
#>  8 homeRuns       6 39    2021         1     139 Tampa … /api/v…     103
#>  9 homeRuns       6 39    2021         1     133 Oaklan… /api/v…     103
#> 10 homeRuns      10 38    2021         1     111 Boston… /api/v…     103
#> # … with 296 more rows, 14 more variables: league_name <chr>,
#> #   league_link <chr>, person_id <int>, person_full_name <chr>,
#> #   person_link <chr>, person_first_name <chr>, person_last_name <chr>,
#> #   sport_id <int>, sport_link <chr>, sport_abbreviation <chr>,
#> #   stat_group <chr>, total_splits <int>, game_type_id <chr>,
#> #   game_type_description <chr>, and abbreviated variable names
#> #   ¹​leader_category, ²​num_teams, ³​team_name, ⁴​team_link, ⁵​league_id
# }