MLB Team Alumni

MLB Team Alumni

Usage

mlb_team_alumni(team_id = NULL, stat_group = NULL, season = NULL)

Arguments

team_id

Team ID to return information and ranking for a particular statistic for a particular team.

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

Value

Returns a tibble with the following columns

col_name	types

player_id	integer

player_full_name	character

link	character

first_name	character

last_name	character

primary_number	character

birth_date	character

current_age	integer

birth_city	character

birth_country	character

height	character

weight	integer

active	logical

use_name	character

middle_name	character

boxscore_name	character

nick_name	character

gender	character

name_matrilineal	character

is_player	logical

is_verified	logical

pronunciation	character

mlb_debut_date	character

name_first_last	character

name_slug	character

first_last_name	character

last_first_name	character

last_init_name	character

init_last_name	character

full_fml_name	character

full_lfm_name	character

strike_zone_top	numeric

strike_zone_bottom	numeric

alumni_last_season	character

birth_state_province	character

draft_year	integer

primary_position_code	character

primary_position_name	character

primary_position_type	character

primary_position_abbreviation	character

bat_side_code	character

bat_side_description	character

pitch_hand_code	character

pitch_hand_description	character

Examples

# \donttest{
  try(mlb_team_alumni(team_id = 137, stat_group = 'hitting', season = 2021))
#>    player_id      player_full_name                  link  first_name
#> 1     501303        Ehire Adrianza /api/v1/people/501303       Ehire
#> 2     542881        Tyler Anderson /api/v1/people/542881       Tyler
#> 3     641312        Shaun Anderson /api/v1/people/641312       Shaun
#> 4     624414      Christian Arroyo /api/v1/people/624414   Christian
#> 5     623214       Abiatal Avelino /api/v1/people/623214     Abiatal
#> 6     642772 Luis Alexander Basabe /api/v1/people/642772        Luis
#> 7     542963           Rob Brantly /api/v1/people/542963      Robert
#> 8     518516     Madison Bumgarner /api/v1/people/518516     Madison
#> 9     502239         Trevor Cahill /api/v1/people/502239      Trevor
#> 10    593525       Orlando Calixte /api/v1/people/593525     Orlando
#> 11    656322           Sam Coonrod /api/v1/people/656322      Samuel
#> 12    518586     Charlie Culberson /api/v1/people/518586     Charles
#> 13    622110            Matt Duffy /api/v1/people/622110     Matthew
#> 14    594807           Adam Duvall /api/v1/people/594807        Adam
#> 15    543148        Tim Federowicz /api/v1/people/543148     Timothy
#> 16    605244         Aramis Garcia /api/v1/people/605244      Aramis
#> 17    608339            Zach Green /api/v1/people/608339        Zach
#> 18    594311        Jandel Gustave /api/v1/people/594311      Jandel
#> 19    593700           Alen Hanson /api/v1/people/593700        Alen
#> 20    623168        Tyler Heineman /api/v1/people/623168       Tyler
#> 21    592390         Heath Hembree /api/v1/people/592390     Richard
#> 22    656582            Connor Joe /api/v1/people/656582      Connor
#> 23    624507           Ryder Jones /api/v1/people/624507       Ryder
#> 24    543377         Corban Joseph /api/v1/people/543377      Corban
#> 25    592464          Mike Kickham /api/v1/people/592464     Michael
#> 26    663411          Joe McCarthy /api/v1/people/663411      Joseph
#> 27    457705      Andrew McCutchen /api/v1/people/457705      Andrew
#> 28    453343         Mark Melancon /api/v1/people/453343        Mark
#> 29    519043            Matt Moore /api/v1/people/519043     Matthew
#> 30    595345          Steven Okert /api/v1/people/595345      Steven
#> 31    605412             Joe Panik /api/v1/people/605412      Joseph
#> 32    467827         Gerardo Parra /api/v1/people/467827     Gerardo
#> 33    607680          Kevin Pillar /api/v1/people/607680       Kevin
#> 34    621002      Daniel Robertson /api/v1/people/621002      Daniel
#> 35    605446      Dereck Rodriguez /api/v1/people/605446        Ivan
#> 36    467055        Pablo Sandoval /api/v1/people/467055       Pablo
#> 37    622046            Chris Shaw /api/v1/people/622046 Christopher
#> 38    592767            Drew Smyly /api/v1/people/592767        Todd
#> 39    608717        Chris Stratton /api/v1/people/608717 Christopher
#> 40    519326     Hunter Strickland /api/v1/people/519326      Hunter
#> 41    572180          Andrew Susac /api/v1/people/572180      Andrew
#> 42    514913          Ruben Tejada /api/v1/people/514913       Ruben
#> 43    605509       Kelby Tomlinson /api/v1/people/605509       Kelby
#> 44    519390          Stephen Vogt /api/v1/people/519390     Stephen
#>     last_name primary_number birth_date current_age
#> 1    Adrianza              5 1989-08-21          32
#> 2    Anderson             31 1989-12-30          32
#> 3    Anderson             64 1994-10-29          27
#> 4      Arroyo             39 1995-05-30          26
#> 5     Avelino             82 1995-02-14          27
#> 6      Basabe             75 1996-08-26          25
#> 7     Brantly             62 1989-07-14          32
#> 8   Bumgarner             40 1989-08-01          32
#> 9      Cahill             35 1988-03-01          34
#> 10    Calixte             16 1992-02-03          30
#> 11    Coonrod             54 1992-09-22          29
#> 12  Culberson             11 1989-04-10          33
#> 13      Duffy              5 1991-01-15          31
#> 14     Duvall             14 1988-09-04          33
#> 15 Federowicz             18 1987-08-05          34
#> 16     Garcia             33 1993-01-12          29
#> 17      Green             96 1994-03-07          28
#> 18    Gustave             31 1992-10-12          29
#> 19     Hanson           <NA> 1992-10-22          29
#> 20   Heineman             50 1991-06-19          30
#> 21    Hembree             53 1989-01-13          33
#> 22        Joe              9 1992-08-16          29
#> 23      Jones             86 1994-06-07          27
#> 24     Joseph             77 1988-10-28          33
#> 25    Kickham             59 1988-12-12          33
#> 26   McCarthy             71 1994-02-23          28
#> 27  McCutchen             24 1986-10-10          35
#> 28   Melancon             34 1985-03-28          37
#> 29      Moore             45 1989-06-18          32
#> 30      Okert             48 1991-07-09          30
#> 31      Panik             12 1990-10-30          31
#> 32      Parra             88 1987-05-06          34
#> 33     Pillar             37 1989-01-04          33
#> 34  Robertson             39 1994-03-22          28
#> 35  Rodriguez             33 1992-06-05          29
#> 36   Sandoval           <NA> 1986-08-11          35
#> 37       Shaw             26 1993-10-20          28
#> 38      Smyly             11 1989-06-13          32
#> 39   Stratton             46 1990-08-22          31
#> 40 Strickland             54 1988-09-24          33
#> 41      Susac             63 1990-03-22          32
#> 42     Tejada           <NA> 1989-10-27          32
#> 43  Tomlinson           <NA> 1990-06-16          31
#> 44       Vogt             21 1984-11-01          37
#>                 birth_city      birth_country height weight active
#> 1                 Guarenas                VEN  6' 1"    198   TRUE
#> 2                Las Vegas                USA  6' 2"    220   TRUE
#> 3            Coral Springs                USA  6' 4"    228   TRUE
#> 4                    Tampa                USA  6' 1"    210   TRUE
#> 5     San Pedro De Macoris Dominican Republic  6' 0"    210   TRUE
#> 6                 El Vigia          Venezuela  6' 0"    180   TRUE
#> 7                San Diego                USA  6' 0"    191   TRUE
#> 8                  Hickory                USA  6' 4"    257   TRUE
#> 9                Oceanside                USA  6' 4"    223   TRUE
#> 10           Santo Domingo Dominican Republic  6' 0"    183   TRUE
#> 11               St. Louis                USA  6' 1"    225   TRUE
#> 12                    Rome                USA  6' 1"    200   TRUE
#> 13              Long Beach                USA  6' 2"    190   TRUE
#> 14              Louisville                USA  6' 1"    215   TRUE
#> 15                    Erie                USA 5' 10"    215   TRUE
#> 16                 Hialeah                USA  6' 1"    228   TRUE
#> 17              Carmichael                USA  6' 2"    224   TRUE
#> 18                Pimentel Dominican Republic  6' 3"    220   TRUE
#> 19               La Romana Dominican Republic  6' 0"    170   TRUE
#> 20       Pacific Palisades                USA 5' 10"    199   TRUE
#> 21             Spartanburg                USA  6' 4"    220   TRUE
#> 22               San Diego                USA  6' 0"    205   TRUE
#> 23                 Seattle                USA  6' 2"    221   TRUE
#> 24               Nashville                USA  6' 0"    185   TRUE
#> 25               St. Louis                USA  6' 4"    220   TRUE
#> 26                Scranton                USA  6' 3"    220   TRUE
#> 27              Fort Meade                USA 5' 11"    195   TRUE
#> 28             Wheat Ridge                USA  6' 1"    215   TRUE
#> 29       Fort Walton Beach                USA  6' 3"    210   TRUE
#> 30               Riverside                USA  6' 2"    202   TRUE
#> 31                 Yonkers                USA  6' 1"    205   TRUE
#> 32 Santa Barbara del Zulia                VEN 5' 11"    209   TRUE
#> 33              West Hills                USA  6' 0"    200   TRUE
#> 34                  Upland                USA 5' 11"    210   TRUE
#> 35               Arlington                USA  6' 0"    208   TRUE
#> 36          Puerto Cabello          Venezuela 5' 10"    268   TRUE
#> 37                Stoneham                USA  6' 4"    220   TRUE
#> 38             Little Rock                USA  6' 2"    188   TRUE
#> 39                  Tupelo                USA  6' 2"    205   TRUE
#> 40               Thomaston                USA  6' 3"    225   TRUE
#> 41               Roseville                USA  6' 1"    220   TRUE
#> 42                Veraguas             Panama 5' 11"    200   TRUE
#> 43               Chickasha                USA  6' 2"    175   TRUE
#> 44                 Visalia                USA  6' 0"    216   TRUE
#>          use_name middle_name     boxscore_name        nick_name gender
#> 1           Ehire     Enrique          Adrianza        Guarenero      M
#> 2           Tyler        John       Anderson, T         Mr. Duck      M
#> 3           Shaun        <NA>       Anderson, S       Bigg Shaun      M
#> 4       Christian      Israel            Arroyo             <NA>      M
#> 5         Abiatal        <NA>           Avelino             <NA>      M
#> 6  Luis Alexander   Alexander Basabe, Luis Alex             <NA>      M
#> 7             Rob       Jacob           Brantly             <NA>      M
#> 8         Madison        Kyle         Bumgarner          Mad-Bum      M
#> 9          Trevor        John            Cahill             T.C.      M
#> 10        Orlando        <NA>           Calixte             <NA>      M
#> 11            Sam     Timothy           Coonrod            Coach      M
#> 12        Charlie      Edward         Culberson         A.C.E.S.      M
#> 13           Matt     Michael          Duffy, M          Duffman      M
#> 14           Adam        Lynn            Duvall             Duvy      M
#> 15            Tim      Joseph        Federowicz             <NA>      M
#> 16         Aramis     Michael        Garcia, Ar             <NA>      M
#> 17           Zach          C.             Green             <NA>      M
#> 18         Jandel   Francisco           Gustave              Gus      M
#> 19           Alen        Rery            Hanson    El Chamaquito      M
#> 20          Tyler      Andrew          Heineman             <NA>      M
#> 21          Heath       Heath           Hembree      Heater Here      M
#> 22         Connor      Kok-Wy               Joe             <NA>      M
#> 23          Ryder    McKinley          Jones, R           Jonesy      M
#> 24         Corban       Reece            Joseph             <NA>      M
#> 25           Mike      Joseph           Kickham             <NA>      M
#> 26            Joe      Edward          McCarthy             <NA>      M
#> 27         Andrew      Stefan         McCutchen        Drusneeze      M
#> 28           Mark       David          Melancon      Muh Lan Son      M
#> 29           Matt        Cody             Moore        Matty Moe      M
#> 30         Steven    Chandler             Okert                O      M
#> 31            Joe     Matthew             Panik             J.P.      M
#> 32        Gerardo     Enrique             Parra          El Yolo      M
#> 33          Kevin      Andrew            Pillar               KP      M
#> 34         Daniel         Ray      Robertson, D            D-Rob      M
#> 35         Dereck      Dereck      Rodríguez, D            D-Rod      M
#> 36          Pablo      Emilio          Sandoval            Panda      M
#> 37          Chris       James           Shaw, C             <NA>      M
#> 38           Drew      Andrew             Smyly           Smiles      M
#> 39          Chris         Lee          Stratton            Strat      M
#> 40         Hunter        Drew        Strickland Southern Thunder      M
#> 41         Andrew        John             Susac             Suze      M
#> 42          Ruben       Dario            Tejada             <NA>      M
#> 43          Kelby       Aaron         Tomlinson            Tommy      M
#> 44        Stephen         Guy              Vogt        I Believe      M
#>    name_matrilineal is_player is_verified            pronunciation
#> 1             Palma      TRUE        TRUE  EH-dey aye-dree-AHN-zah
#> 2              <NA>      TRUE        TRUE                     <NA>
#> 3              <NA>      TRUE       FALSE                     <NA>
#> 4              <NA>      TRUE        TRUE                uh-ROY-oh
#> 5              <NA>      TRUE       FALSE                     <NA>
#> 6              <NA>      TRUE       FALSE              buh-SAH-bay
#> 7              <NA>      TRUE        TRUE                     <NA>
#> 8              <NA>      TRUE        TRUE MAD-i-suhn BUHM-gahr-ner
#> 9              <NA>      TRUE        TRUE                     <NA>
#> 10             <NA>      TRUE        TRUE                     <NA>
#> 11             <NA>      TRUE        TRUE                KOON-rahd
#> 12             <NA>      TRUE        TRUE                     <NA>
#> 13             <NA>      TRUE        TRUE              mat DUH-fee
#> 14             <NA>      TRUE        TRUE                     <NA>
#> 15             <NA>      TRUE        TRUE          fehd-er-oh-vich
#> 16             <NA>      TRUE        TRUE               AY-ruh-mis
#> 17             <NA>      TRUE        TRUE                     <NA>
#> 18             <NA>      TRUE        TRUE  yahn-DELL goo-STAH-vayh
#> 19           Michel      TRUE        TRUE                     <NA>
#> 20             <NA>      TRUE        TRUE                     <NA>
#> 21             <NA>      TRUE        TRUE                     <NA>
#> 22             <NA>      TRUE        TRUE                     <NA>
#> 23             <NA>      TRUE        TRUE                     <NA>
#> 24             <NA>      TRUE        TRUE                     <NA>
#> 25             <NA>      TRUE        TRUE                  kick-um
#> 26             <NA>      TRUE        TRUE                     <NA>
#> 27             <NA>      TRUE        TRUE              mik-KUCH-in
#> 28             <NA>      TRUE        TRUE              muh-LAN-son
#> 29             <NA>      TRUE        TRUE                     moor
#> 30             <NA>      TRUE        TRUE                     <NA>
#> 31             <NA>      TRUE        TRUE                    panic
#> 32             <NA>      TRUE        TRUE       her-ARR-doe PAR-uh
#> 33             <NA>      TRUE        TRUE                     <NA>
#> 34             <NA>      TRUE        TRUE                     <NA>
#> 35             <NA>      TRUE        TRUE                     <NA>
#> 36             <NA>      TRUE        TRUE             san-doh-VAHL
#> 37             <NA>      TRUE        TRUE                     <NA>
#> 38             <NA>      TRUE        TRUE                 SMIY-lee
#> 39             <NA>      TRUE        TRUE                     <NA>
#> 40             <NA>      TRUE        TRUE                     <NA>
#> 41             <NA>      TRUE        TRUE                 SUE-sack
#> 42             <NA>      TRUE        TRUE                     <NA>
#> 43             <NA>      TRUE        TRUE                     <NA>
#> 44             <NA>      TRUE        TRUE                     VOTE
#>    mlb_debut_date       name_first_last                    name_slug
#> 1      2013-09-08        Ehire Adrianza        ehire-adrianza-501303
#> 2      2016-06-12        Tyler Anderson        tyler-anderson-542881
#> 3      2019-05-15        Shaun Anderson        shaun-anderson-641312
#> 4      2017-04-24      Christian Arroyo      christian-arroyo-624414
#> 5      2018-09-08       Abiatal Avelino       abiatal-avelino-623214
#> 6      2020-08-27 Luis Alexander Basabe luis-alexander-basabe-642772
#> 7      2012-08-14           Rob Brantly           rob-brantly-542963
#> 8      2009-09-08     Madison Bumgarner     madison-bumgarner-518516
#> 9      2009-04-07         Trevor Cahill         trevor-cahill-502239
#> 10     2015-04-19       Orlando Calixte       orlando-calixte-593525
#> 11     2019-05-26           Sam Coonrod           sam-coonrod-656322
#> 12     2012-05-13     Charlie Culberson     charlie-culberson-518586
#> 13     2014-05-22            Matt Duffy            matt-duffy-622110
#> 14     2014-06-26           Adam Duvall           adam-duvall-594807
#> 15     2011-09-11        Tim Federowicz        tim-federowicz-543148
#> 16     2018-08-31         Aramis Garcia         aramis-garcia-605244
#> 17     2019-07-21            Zach Green            zach-green-608339
#> 18     2016-08-11        Jandel Gustave        jandel-gustave-594311
#> 19     2016-05-16           Alen Hanson           alen-hanson-593700
#> 20     2019-09-04        Tyler Heineman        tyler-heineman-623168
#> 21     2013-09-03         Heath Hembree         heath-hembree-592390
#> 22     2019-03-28            Connor Joe            connor-joe-656582
#> 23     2017-06-24           Ryder Jones           ryder-jones-624507
#> 24     2013-05-13         Corban Joseph         corban-joseph-543377
#> 25     2013-05-28          Mike Kickham          mike-kickham-592464
#> 26     2020-07-23          Joe McCarthy          joe-mccarthy-663411
#> 27     2009-06-04      Andrew McCutchen      andrew-mccutchen-457705
#> 28     2009-04-26         Mark Melancon         mark-melancon-453343
#> 29     2011-09-14            Matt Moore            matt-moore-519043
#> 30     2016-04-19          Steven Okert          steven-okert-595345
#> 31     2014-05-22             Joe Panik             joe-panik-605412
#> 32     2009-05-13         Gerardo Parra         gerardo-parra-467827
#> 33     2013-08-14          Kevin Pillar          kevin-pillar-607680
#> 34     2017-04-04      Daniel Robertson      daniel-robertson-621002
#> 35     2018-05-29      Dereck Rodriguez      dereck-rodriguez-605446
#> 36     2008-08-14        Pablo Sandoval        pablo-sandoval-467055
#> 37     2018-08-31            Chris Shaw            chris-shaw-622046
#> 38     2012-04-12            Drew Smyly            drew-smyly-592767
#> 39     2016-05-30        Chris Stratton        chris-stratton-608717
#> 40     2014-09-01     Hunter Strickland     hunter-strickland-519326
#> 41     2014-05-22          Andrew Susac          andrew-susac-572180
#> 42     2010-04-07          Ruben Tejada          ruben-tejada-514913
#> 43     2015-08-03       Kelby Tomlinson       kelby-tomlinson-605509
#> 44     2012-04-06          Stephen Vogt          stephen-vogt-519390
#>          first_last_name        last_first_name last_init_name
#> 1         Ehire Adrianza        Adrianza, Ehire    Adrianza, E
#> 2         Tyler Anderson        Anderson, Tyler    Anderson, T
#> 3         Shaun Anderson        Anderson, Shaun    Anderson, S
#> 4       Christian Arroyo      Arroyo, Christian      Arroyo, C
#> 5        Abiatal Avelino       Avelino, Abiatal     Avelino, A
#> 6  Luis Alexander Basabe Basabe, Luis Alexander      Basabe, L
#> 7            Rob Brantly           Brantly, Rob     Brantly, R
#> 8      Madison Bumgarner     Bumgarner, Madison   Bumgarner, M
#> 9          Trevor Cahill         Cahill, Trevor      Cahill, T
#> 10       Orlando Calixte       Calixte, Orlando     Calixte, O
#> 11           Sam Coonrod           Coonrod, Sam     Coonrod, S
#> 12     Charlie Culberson     Culberson, Charlie   Culberson, C
#> 13            Matt Duffy            Duffy, Matt       Duffy, M
#> 14           Adam Duvall           Duvall, Adam      Duvall, A
#> 15        Tim Federowicz        Federowicz, Tim  Federowicz, T
#> 16         Aramis Garcia         Garcia, Aramis      Garcia, A
#> 17            Zach Green            Green, Zach       Green, Z
#> 18        Jandel Gustave        Gustave, Jandel     Gustave, J
#> 19           Alen Hanson           Hanson, Alen      Hanson, A
#> 20        Tyler Heineman        Heineman, Tyler    Heineman, T
#> 21         Heath Hembree         Hembree, Heath     Hembree, H
#> 22            Connor Joe            Joe, Connor         Joe, C
#> 23           Ryder Jones           Jones, Ryder       Jones, R
#> 24         Corban Joseph         Joseph, Corban      Joseph, C
#> 25          Mike Kickham          Kickham, Mike     Kickham, M
#> 26          Joe McCarthy          McCarthy, Joe    McCarthy, J
#> 27      Andrew McCutchen      McCutchen, Andrew   McCutchen, A
#> 28         Mark Melancon         Melancon, Mark    Melancon, M
#> 29            Matt Moore            Moore, Matt       Moore, M
#> 30          Steven Okert          Okert, Steven       Okert, S
#> 31             Joe Panik             Panik, Joe       Panik, J
#> 32         Gerardo Parra         Parra, Gerardo       Parra, G
#> 33          Kevin Pillar          Pillar, Kevin      Pillar, K
#> 34      Daniel Robertson      Robertson, Daniel   Robertson, D
#> 35      Dereck Rodríguez      Rodríguez, Dereck   Rodríguez, D
#> 36        Pablo Sandoval        Sandoval, Pablo    Sandoval, P
#> 37            Chris Shaw            Shaw, Chris        Shaw, C
#> 38            Drew Smyly            Smyly, Drew       Smyly, D
#> 39        Chris Stratton        Stratton, Chris    Stratton, C
#> 40     Hunter Strickland     Strickland, Hunter  Strickland, H
#> 41          Andrew Susac          Susac, Andrew       Susac, A
#> 42          Rubén Tejada          Tejada, Rubén      Tejada, R
#> 43       Kelby Tomlinson       Tomlinson, Kelby   Tomlinson, K
#> 44          Stephen Vogt          Vogt, Stephen        Vogt, S
#>    init_last_name             full_fml_name              full_lfm_name
#> 1      E Adrianza    Ehire Enrique Adrianza    Adrianza, Ehire Enrique
#> 2      T Anderson       Tyler John Anderson       Anderson, Tyler John
#> 3      S Anderson            Shaun Anderson            Anderson, Shaun
#> 4        C Arroyo   Christian Israel Arroyo   Arroyo, Christian Israel
#> 5       A Avelino           Abiatal Avelino           Avelino, Abiatal
#> 6       LA Basabe     Luis Alexander Basabe     Basabe, Luis Alexander
#> 7       R Brantly      Robert Jacob Brantly      Brantly, Robert Jacob
#> 8     M Bumgarner    Madison Kyle Bumgarner    Bumgarner, Madison Kyle
#> 9        T Cahill        Trevor John Cahill        Cahill, Trevor John
#> 10      O Calixte           Orlando Calixte           Calixte, Orlando
#> 11      S Coonrod    Samuel Timothy Coonrod    Coonrod, Samuel Timothy
#> 12    C Culberson  Charles Edward Culberson  Culberson, Charles Edward
#> 13        M Duffy     Matthew Michael Duffy     Duffy, Matthew Michael
#> 14       A Duvall          Adam Lynn Duvall          Duvall, Adam Lynn
#> 15   T Federowicz Timothy Joseph Federowicz Federowicz, Timothy Joseph
#> 16       A Garcia     Aramis Michael Garcia     Garcia, Aramis Michael
#> 17        Z Green             Zach C. Green             Green, Zach C.
#> 18      J Gustave  Jandel Francisco Gustave  Gustave, Jandel Francisco
#> 19       A Hanson          Alen Rery Hanson          Hanson, Alen Rery
#> 20     T Heineman     Tyler Andrew Heineman     Heineman, Tyler Andrew
#> 21      H Hembree     Richard Heath Hembree     Hembree, Richard Heath
#> 22          C Joe         Connor Kok-Wy Joe         Joe, Connor Kok-Wy
#> 23        R Jones      Ryder McKinley Jones      Jones, Ryder McKinley
#> 24       C Joseph       Corban Reece Joseph       Joseph, Corban Reece
#> 25      M Kickham    Michael Joseph Kickham    Kickham, Michael Joseph
#> 26     J McCarthy    Joseph Edward McCarthy    McCarthy, Joseph Edward
#> 27    A McCutchen   Andrew Stefan McCutchen   McCutchen, Andrew Stefan
#> 28     M Melancon       Mark David Melancon       Melancon, Mark David
#> 29        M Moore        Matthew Cody Moore        Moore, Matthew Cody
#> 30        S Okert     Steven Chandler Okert     Okert, Steven Chandler
#> 31        J Panik      Joseph Matthew Panik      Panik, Joseph Matthew
#> 32        G Parra     Gerardo Enrique Parra     Parra, Gerardo Enrique
#> 33       K Pillar       Kevin Andrew Pillar       Pillar, Kevin Andrew
#> 34    D Robertson      Daniel Ray Robertson      Robertson, Daniel Ray
#> 35    D Rodríguez     Ivan Dereck Rodríguez     Rodríguez, Ivan Dereck
#> 36     P Sandoval     Pablo Emilio Sandoval     Sandoval, Pablo Emilio
#> 37         C Shaw    Christopher James Shaw    Shaw, Christopher James
#> 38        D Smyly         Todd Andrew Smyly         Smyly, Todd Andrew
#> 39     C Stratton  Christopher Lee Stratton  Stratton, Christopher Lee
#> 40   H Strickland    Hunter Drew Strickland    Strickland, Hunter Drew
#> 41        A Susac         Andrew John Susac         Susac, Andrew John
#> 42       R Tejada        Rubén Dario Tejada        Tejada, Rubén Dario
#> 43    K Tomlinson     Kelby Aaron Tomlinson     Tomlinson, Kelby Aaron
#> 44         S Vogt          Stephen Guy Vogt          Vogt, Stephen Guy
#>    strike_zone_top strike_zone_bottom alumni_last_season
#> 1            3.411              1.565               2016
#> 2            3.467              1.589               2020
#> 3            3.549              1.627               2020
#> 4            3.411              1.565               2017
#> 5            3.371              1.535               2020
#> 6            3.371              1.535               2020
#> 7            3.371              1.535               2020
#> 8            3.549              1.627               2019
#> 9            3.549              1.627               2020
#> 10           3.371              1.535               2017
#> 11           3.411              1.565               2020
#> 12           3.411              1.565               2012
#> 13           3.467              1.589               2016
#> 14           3.411              1.565               2015
#> 15           3.301              1.504               2017
#> 16           3.411              1.565               2020
#> 17           3.467              1.589               2020
#> 18           3.490              1.601               2020
#> 19           3.371              1.535               2019
#> 20           3.301              1.504               2020
#> 21           3.549              1.627               2014
#> 22           3.371              1.535               2019
#> 23           3.467              1.589               2019
#> 24           3.371              1.535               2019
#> 25           3.549              1.627               2016
#> 26           3.490              1.601               2020
#> 27           3.319              1.513               2018
#> 28           3.411              1.565               2019
#> 29           3.490              1.601               2017
#> 30           3.467              1.589               2019
#> 31           3.411              1.565               2019
#> 32           3.319              1.513               2019
#> 33           3.371              1.535               2019
#> 34           3.319              1.513               2020
#> 35           3.371              1.535               2020
#> 36           3.301              1.504               2020
#> 37           3.549              1.627               2020
#> 38           3.467              1.589               2020
#> 39           3.467              1.589               2018
#> 40           3.490              1.601               2018
#> 41           3.411              1.565               2016
#> 42           3.319              1.513               2016
#> 43           3.467              1.589               2018
#> 44           3.371              1.535               2019
#>    birth_state_province draft_year primary_position_code
#> 1                  <NA>         NA                     5
#> 2                    NV       2011                     1
#> 3                    FL       2016                     1
#> 4                    FL       2013                     4
#> 5                  <NA>         NA                     6
#> 6                  <NA>         NA                     8
#> 7                    CA       2010                     2
#> 8                    NC       2007                     1
#> 9                    CA       2006                     1
#> 10                 <NA>         NA                     6
#> 11                   MO       2014                     1
#> 12                   GA       2007                     5
#> 13                   CA       2012                     5
#> 14                   KY       2010                     9
#> 15                   PA       2008                     2
#> 16                   FL       2014                     2
#> 17                   CA       2012                     I
#> 18                 <NA>         NA                     1
#> 19                 <NA>         NA                     4
#> 20                   CA       2012                     2
#> 21                   SC       2010                     1
#> 22                   CA       2014                     3
#> 23                   WA       2013                     5
#> 24                   TN       2008                     4
#> 25                   MO       2010                     1
#> 26                   PA       2015                     7
#> 27                   FL       2005                     7
#> 28                   CO       2006                     1
#> 29                   FL       2007                     1
#> 30                   CA       2012                     1
#> 31                   NY       2011                     5
#> 32                 <NA>         NA                     O
#> 33                   CA       2011                     8
#> 34                   CA       2012                     6
#> 35                   TX       2011                     1
#> 36                 <NA>         NA                     3
#> 37                   MA       2015                     O
#> 38                   AR       2010                     1
#> 39                   MS       2012                     1
#> 40                   GA       2007                     1
#> 41                   CA       2011                     2
#> 42                 <NA>         NA                     4
#> 43                   OK       2011                     5
#> 44                   CA       2007                     2
#>    primary_position_name primary_position_type
#> 1             Third Base             Infielder
#> 2                Pitcher               Pitcher
#> 3                Pitcher               Pitcher
#> 4            Second Base             Infielder
#> 5              Shortstop             Infielder
#> 6             Outfielder            Outfielder
#> 7                Catcher               Catcher
#> 8                Pitcher               Pitcher
#> 9                Pitcher               Pitcher
#> 10             Shortstop             Infielder
#> 11               Pitcher               Pitcher
#> 12            Third Base             Infielder
#> 13            Third Base             Infielder
#> 14            Outfielder            Outfielder
#> 15               Catcher               Catcher
#> 16               Catcher               Catcher
#> 17               Infield             Infielder
#> 18               Pitcher               Pitcher
#> 19           Second Base             Infielder
#> 20               Catcher               Catcher
#> 21               Pitcher               Pitcher
#> 22            First Base             Infielder
#> 23            Third Base             Infielder
#> 24           Second Base             Infielder
#> 25               Pitcher               Pitcher
#> 26            Outfielder            Outfielder
#> 27            Outfielder            Outfielder
#> 28               Pitcher               Pitcher
#> 29               Pitcher               Pitcher
#> 30               Pitcher               Pitcher
#> 31            Third Base             Infielder
#> 32              Outfield            Outfielder
#> 33            Outfielder            Outfielder
#> 34             Shortstop             Infielder
#> 35               Pitcher               Pitcher
#> 36            First Base             Infielder
#> 37              Outfield            Outfielder
#> 38               Pitcher               Pitcher
#> 39               Pitcher               Pitcher
#> 40               Pitcher               Pitcher
#> 41               Catcher               Catcher
#> 42           Second Base             Infielder
#> 43            Third Base             Infielder
#> 44               Catcher               Catcher
#>    primary_position_abbreviation bat_side_code bat_side_description
#> 1                             3B             S               Switch
#> 2                              P             L                 Left
#> 3                              P             R                Right
#> 4                             2B             R                Right
#> 5                             SS             R                Right
#> 6                             CF             S               Switch
#> 7                              C             L                 Left
#> 8                              P             R                Right
#> 9                              P             R                Right
#> 10                            SS             R                Right
#> 11                             P             R                Right
#> 12                            3B             R                Right
#> 13                            3B             R                Right
#> 14                            RF             R                Right
#> 15                             C             R                Right
#> 16                             C             R                Right
#> 17                            IF             R                Right
#> 18                             P             R                Right
#> 19                            2B             S               Switch
#> 20                             C             S               Switch
#> 21                             P             R                Right
#> 22                            1B             R                Right
#> 23                            3B             L                 Left
#> 24                            2B             L                 Left
#> 25                             P             L                 Left
#> 26                            LF             L                 Left
#> 27                            LF             R                Right
#> 28                             P             R                Right
#> 29                             P             L                 Left
#> 30                             P             L                 Left
#> 31                            3B             L                 Left
#> 32                            OF             L                 Left
#> 33                            CF             R                Right
#> 34                            SS             R                Right
#> 35                             P             R                Right
#> 36                            1B             S               Switch
#> 37                            OF             L                 Left
#> 38                             P             L                 Left
#> 39                             P             R                Right
#> 40                             P             R                Right
#> 41                             C             R                Right
#> 42                            2B             R                Right
#> 43                            3B             R                Right
#> 44                             C             L                 Left
#>    pitch_hand_code pitch_hand_description
#> 1                R                  Right
#> 2                L                   Left
#> 3                R                  Right
#> 4                R                  Right
#> 5                R                  Right
#> 6                R                  Right
#> 7                R                  Right
#> 8                L                   Left
#> 9                R                  Right
#> 10               R                  Right
#> 11               R                  Right
#> 12               R                  Right
#> 13               R                  Right
#> 14               R                  Right
#> 15               R                  Right
#> 16               R                  Right
#> 17               R                  Right
#> 18               R                  Right
#> 19               R                  Right
#> 20               R                  Right
#> 21               R                  Right
#> 22               R                  Right
#> 23               R                  Right
#> 24               R                  Right
#> 25               L                   Left
#> 26               L                   Left
#> 27               R                  Right
#> 28               R                  Right
#> 29               L                   Left
#> 30               L                   Left
#> 31               R                  Right
#> 32               L                   Left
#> 33               R                  Right
#> 34               R                  Right
#> 35               R                  Right
#> 36               R                  Right
#> 37               R                  Right
#> 38               L                   Left
#> 39               R                  Right
#> 40               R                  Right
#> 41               R                  Right
#> 42               R                  Right
#> 43               R                  Right
#> 44               R                  Right
# }