MLB Team Leaders

MLB Team Leaders

Usage

mlb_team_leaders(
  team_id = NULL,
  leader_categories = NULL,
  leader_game_types = NULL,
  season = NULL,
  limit = 1000
)

Arguments

team_id

Team ID to return team leader information for.

leader_categories

Team leader category to return information and ranking for a particular statistic.

leader_game_types

Game type to return information and ranking for a particular statistic in a particular game type.

season

Season to return team leader information for.

limit

A limit to limit return to a particular number of records.

Value

Returns a tibble with the following columns

col_name	types

leader_category	character

rank	integer

value	character

season	character

team_id	integer

team_name	character

team_link	character

league_id	integer

league_name	character

league_link	character

person_id	integer

person_full_name	character

person_link	character

person_first_name	character

person_last_name	character

sport_id	integer

sport_link	character

sport_abbreviation	character

stat_group	character

total_splits	integer

game_type_id	character

game_type_description	character

Examples

# \donttest{
  try(mlb_team_leaders(team_id = 137, leader_categories = "homeRuns", season = 2021))
#>    leader_category rank value season team_id            team_name
#> 1         homeRuns    1    29   2021     137 San Francisco Giants
#> 2         homeRuns    2    25   2021     137 San Francisco Giants
#> 3         homeRuns    3    24   2021     137 San Francisco Giants
#> 4         homeRuns    4    18   2021     137 San Francisco Giants
#> 5         homeRuns    4    18   2021     137 San Francisco Giants
#> 6         homeRuns    4    18   2021     137 San Francisco Giants
#> 7         homeRuns    7    16   2021     137 San Francisco Giants
#> 8         homeRuns    8    13   2021     137 San Francisco Giants
#> 9         homeRuns    8    13   2021     137 San Francisco Giants
#> 10        homeRuns   10    12   2021     137 San Francisco Giants
#> 11        homeRuns   11     8   2021     137 San Francisco Giants
#> 12        homeRuns   12     7   2021     137 San Francisco Giants
#> 13        homeRuns   12     7   2021     137 San Francisco Giants
#> 14        homeRuns   12     7   2021     137 San Francisco Giants
#> 15        homeRuns   12     7   2021     137 San Francisco Giants
#> 16        homeRuns   16     5   2021     137 San Francisco Giants
#> 17        homeRuns   16     5   2021     137 San Francisco Giants
#> 18        homeRuns   18     4   2021     137 San Francisco Giants
#> 19        homeRuns   19     3   2021     137 San Francisco Giants
#> 20        homeRuns   20     1   2021     137 San Francisco Giants
#> 21        homeRuns   20     1   2021     137 San Francisco Giants
#> 22        homeRuns    1    20   2021     137 San Francisco Giants
#> 23        homeRuns    2    19   2021     137 San Francisco Giants
#> 24        homeRuns    3    15   2021     137 San Francisco Giants
#> 25        homeRuns    4    14   2021     137 San Francisco Giants
#> 26        homeRuns    5     9   2021     137 San Francisco Giants
#> 27        homeRuns    5     9   2021     137 San Francisco Giants
#> 28        homeRuns    7     7   2021     137 San Francisco Giants
#> 29        homeRuns    7     7   2021     137 San Francisco Giants
#> 30        homeRuns    9     5   2021     137 San Francisco Giants
#> 31        homeRuns    9     5   2021     137 San Francisco Giants
#> 32        homeRuns   11     4   2021     137 San Francisco Giants
#> 33        homeRuns   11     4   2021     137 San Francisco Giants
#> 34        homeRuns   11     4   2021     137 San Francisco Giants
#> 35        homeRuns   14     3   2021     137 San Francisco Giants
#> 36        homeRuns   14     3   2021     137 San Francisco Giants
#> 37        homeRuns   14     3   2021     137 San Francisco Giants
#> 38        homeRuns   14     3   2021     137 San Francisco Giants
#> 39        homeRuns   18     2   2021     137 San Francisco Giants
#> 40        homeRuns   18     2   2021     137 San Francisco Giants
#> 41        homeRuns   18     2   2021     137 San Francisco Giants
#> 42        homeRuns   18     2   2021     137 San Francisco Giants
#> 43        homeRuns   18     2   2021     137 San Francisco Giants
#> 44        homeRuns   18     2   2021     137 San Francisco Giants
#> 45        homeRuns   24     1   2021     137 San Francisco Giants
#> 46        homeRuns   24     1   2021     137 San Francisco Giants
#> 47        homeRuns   24     1   2021     137 San Francisco Giants
#> 48        homeRuns   24     1   2021     137 San Francisco Giants
#> 49        homeRuns   24     1   2021     137 San Francisco Giants
#> 50        homeRuns    1   105   2021     137 San Francisco Giants
#> 51        homeRuns    2    42   2021     137 San Francisco Giants
#> 52        homeRuns    3     4   2021     137 San Francisco Giants
#>            team_link league_id league_name        league_link person_id
#> 1  /api/v1/teams/137       104          NL /api/v1/league/104    474832
#> 2  /api/v1/teams/137       104          NL /api/v1/league/104    573262
#> 3  /api/v1/teams/137       104          NL /api/v1/league/104    543063
#> 4  /api/v1/teams/137       104          NL /api/v1/league/104    527038
#> 5  /api/v1/teams/137       104          NL /api/v1/league/104    457763
#> 6  /api/v1/teams/137       104          NL /api/v1/league/104    664774
#> 7  /api/v1/teams/137       104          NL /api/v1/league/104    573131
#> 8  /api/v1/teams/137       104          NL /api/v1/league/104    543105
#> 9  /api/v1/teams/137       104          NL /api/v1/league/104    446334
#> 10 /api/v1/teams/137       104          NL /api/v1/league/104    596103
#> 11 /api/v1/teams/137       104          NL /api/v1/league/104    621453
#> 12 /api/v1/teams/137       104          NL /api/v1/league/104    592178
#> 13 /api/v1/teams/137       104          NL /api/v1/league/104    642731
#> 14 /api/v1/teams/137       104          NL /api/v1/league/104    600303
#> 15 /api/v1/teams/137       104          NL /api/v1/league/104    456781
#> 16 /api/v1/teams/137       104          NL /api/v1/league/104    592200
#> 17 /api/v1/teams/137       104          NL /api/v1/league/104    643289
#> 18 /api/v1/teams/137       104          NL /api/v1/league/104    643565
#> 19 /api/v1/teams/137       104          NL /api/v1/league/104    613564
#> 20 /api/v1/teams/137       104          NL /api/v1/league/104    644433
#> 21 /api/v1/teams/137       104          NL /api/v1/league/104    657277
#> 22 /api/v1/teams/137       104          NL /api/v1/league/104    592332
#> 23 /api/v1/teams/137       104          NL /api/v1/league/104    543101
#> 24 /api/v1/teams/137       104          NL /api/v1/league/104    456501
#> 25 /api/v1/teams/137       104          NL /api/v1/league/104    622072
#> 26 /api/v1/teams/137       104          NL /api/v1/league/104    606424
#> 27 /api/v1/teams/137       104          NL /api/v1/league/104    657277
#> 28 /api/v1/teams/137       104          NL /api/v1/league/104    641793
#> 29 /api/v1/teams/137       104          NL /api/v1/league/104    459429
#> 30 /api/v1/teams/137       104          NL /api/v1/league/104    669674
#> 31 /api/v1/teams/137       104          NL /api/v1/league/104    643511
#> 32 /api/v1/teams/137       104          NL /api/v1/league/104    605154
#> 33 /api/v1/teams/137       104          NL /api/v1/league/104    666808
#> 34 /api/v1/teams/137       104          NL /api/v1/league/104    605538
#> 35 /api/v1/teams/137       104          NL /api/v1/league/104    543351
#> 36 /api/v1/teams/137       104          NL /api/v1/league/104    431148
#> 37 /api/v1/teams/137       104          NL /api/v1/league/104    500779
#> 38 /api/v1/teams/137       104          NL /api/v1/league/104    666619
#> 39 /api/v1/teams/137       104          NL /api/v1/league/104    501625
#> 40 /api/v1/teams/137       104          NL /api/v1/league/104    608678
#> 41 /api/v1/teams/137       104          NL /api/v1/league/104    669214
#> 42 /api/v1/teams/137       104          NL /api/v1/league/104    592717
#> 43 /api/v1/teams/137       104          NL /api/v1/league/104    572125
#> 44 /api/v1/teams/137       104          NL /api/v1/league/104    642073
#> 45 /api/v1/teams/137       104          NL /api/v1/league/104    670056
#> 46 /api/v1/teams/137       104          NL /api/v1/league/104    543022
#> 47 /api/v1/teams/137       104          NL /api/v1/league/104    606625
#> 48 /api/v1/teams/137       104          NL /api/v1/league/104    593974
#> 49 /api/v1/teams/137       104          NL /api/v1/league/104    453265
#> 50 /api/v1/teams/137       104          NL /api/v1/league/104    457763
#> 51 /api/v1/teams/137       104          NL /api/v1/league/104    592200
#> 52 /api/v1/teams/137       104          NL /api/v1/league/104    644433
#>      person_full_name           person_link person_first_name
#> 1        Brandon Belt /api/v1/people/474832           Brandon
#> 2    Mike Yastrzemski /api/v1/people/573262              Mike
#> 3    Brandon Crawford /api/v1/people/543063           Brandon
#> 4       Wilmer Flores /api/v1/people/527038            Wilmer
#> 5        Buster Posey /api/v1/people/457763            Buster
#> 6    LaMonte Wade Jr. /api/v1/people/664774           LaMonte
#> 7           Darin Ruf /api/v1/people/573131             Darin
#> 8      Alex Dickerson /api/v1/people/543105              Alex
#> 9       Evan Longoria /api/v1/people/446334              Evan
#> 10      Austin Slater /api/v1/people/596103            Austin
#> 11      Steven Duggar /api/v1/people/621453            Steven
#> 12        Kris Bryant /api/v1/people/592178              Kris
#> 13     Thairo Estrada /api/v1/people/642731            Thairo
#> 14    Tommy La Stella /api/v1/people/600303             Tommy
#> 15     Donovan Solano /api/v1/people/456781           Donovan
#> 16        Curt Casali /api/v1/people/592200              Curt
#> 17     Mauricio Dubon /api/v1/people/643289          Mauricio
#> 18      Mike Tauchman /api/v1/people/643565              Mike
#> 19       Jason Vosler /api/v1/people/613564             Jason
#> 20     Chadwick Tromp /api/v1/people/644433          Chadwick
#> 21         Logan Webb /api/v1/people/657277             Logan
#> 22      Kevin Gausman /api/v1/people/592332             Kevin
#> 23 Anthony DeSclafani /api/v1/people/543101           Anthony
#> 24       Johnny Cueto /api/v1/people/456501            Johnny
#> 25          Alex Wood /api/v1/people/622072              Alex
#> 26      Jarlin Garcia /api/v1/people/606424            Jarlin
#> 27         Logan Webb /api/v1/people/657277             Logan
#> 28       Zack Littell /api/v1/people/641793              Zack
#> 29         Jake McGee /api/v1/people/459429              Jake
#> 30           Sam Long /api/v1/people/669674               Sam
#> 31       Tyler Rogers /api/v1/people/643511             Tyler
#> 32       John Brebbia /api/v1/people/605154              John
#> 33       Camilo Doval /api/v1/people/666808            Camilo
#> 34        Matt Wisler /api/v1/people/605538              Matt
#> 35        Jay Jackson /api/v1/people/543351               Jay
#> 36       Scott Kazmir /api/v1/people/431148             Scott
#> 37      Jose Quintana /api/v1/people/500779              Jose
#> 38     Gregory Santos /api/v1/people/666619           Gregory
#> 39       Jose Alvarez /api/v1/people/501625              Jose
#> 40      Dominic Leone /api/v1/people/608678           Dominic
#> 41       Conner Menez /api/v1/people/669214            Conner
#> 42      Aaron Sanchez /api/v1/people/592717             Aaron
#> 43         Sam Selman /api/v1/people/572125               Sam
#> 44      Jimmie Sherfy /api/v1/people/642073             James
#> 45      Caleb Baragar /api/v1/people/670056             Caleb
#> 46     Tyler Chatwood /api/v1/people/543022             Tyler
#> 47      Reyes Moronta /api/v1/people/606625             Reyes
#> 48      Wandy Peralta /api/v1/people/593974             Wandy
#> 49        Tony Watson /api/v1/people/453265              Tony
#> 50       Buster Posey /api/v1/people/457763            Buster
#> 51        Curt Casali /api/v1/people/592200              Curt
#> 52     Chadwick Tromp /api/v1/people/644433          Chadwick
#>    person_last_name sport_id       sport_link sport_abbreviation
#> 1              Belt        1 /api/v1/sports/1                MLB
#> 2       Yastrzemski        1 /api/v1/sports/1                MLB
#> 3          Crawford        1 /api/v1/sports/1                MLB
#> 4            Flores        1 /api/v1/sports/1                MLB
#> 5             Posey        1 /api/v1/sports/1                MLB
#> 6              Wade        1 /api/v1/sports/1                MLB
#> 7               Ruf        1 /api/v1/sports/1                MLB
#> 8         Dickerson        1 /api/v1/sports/1                MLB
#> 9          Longoria        1 /api/v1/sports/1                MLB
#> 10           Slater        1 /api/v1/sports/1                MLB
#> 11           Duggar        1 /api/v1/sports/1                MLB
#> 12           Bryant        1 /api/v1/sports/1                MLB
#> 13          Estrada        1 /api/v1/sports/1                MLB
#> 14        La Stella        1 /api/v1/sports/1                MLB
#> 15           Solano        1 /api/v1/sports/1                MLB
#> 16           Casali        1 /api/v1/sports/1                MLB
#> 17            Dubon        1 /api/v1/sports/1                MLB
#> 18         Tauchman        1 /api/v1/sports/1                MLB
#> 19           Vosler        1 /api/v1/sports/1                MLB
#> 20            Tromp        1 /api/v1/sports/1                MLB
#> 21             Webb        1 /api/v1/sports/1                MLB
#> 22          Gausman        1 /api/v1/sports/1                MLB
#> 23       DeSclafani        1 /api/v1/sports/1                MLB
#> 24            Cueto        1 /api/v1/sports/1                MLB
#> 25             Wood        1 /api/v1/sports/1                MLB
#> 26           Garcia        1 /api/v1/sports/1                MLB
#> 27             Webb        1 /api/v1/sports/1                MLB
#> 28          Littell        1 /api/v1/sports/1                MLB
#> 29            McGee        1 /api/v1/sports/1                MLB
#> 30             Long        1 /api/v1/sports/1                MLB
#> 31           Rogers        1 /api/v1/sports/1                MLB
#> 32          Brebbia        1 /api/v1/sports/1                MLB
#> 33            Doval        1 /api/v1/sports/1                MLB
#> 34           Wisler        1 /api/v1/sports/1                MLB
#> 35          Jackson        1 /api/v1/sports/1                MLB
#> 36           Kazmir        1 /api/v1/sports/1                MLB
#> 37         Quintana        1 /api/v1/sports/1                MLB
#> 38           Santos        1 /api/v1/sports/1                MLB
#> 39          Alvarez        1 /api/v1/sports/1                MLB
#> 40            Leone        1 /api/v1/sports/1                MLB
#> 41            Menez        1 /api/v1/sports/1                MLB
#> 42          Sanchez        1 /api/v1/sports/1                MLB
#> 43           Selman        1 /api/v1/sports/1                MLB
#> 44           Sherfy        1 /api/v1/sports/1                MLB
#> 45          Baragar        1 /api/v1/sports/1                MLB
#> 46         Chatwood        1 /api/v1/sports/1                MLB
#> 47          Moronta        1 /api/v1/sports/1                MLB
#> 48          Peralta        1 /api/v1/sports/1                MLB
#> 49           Watson        1 /api/v1/sports/1                MLB
#> 50            Posey        1 /api/v1/sports/1                MLB
#> 51           Casali        1 /api/v1/sports/1                MLB
#> 52            Tromp        1 /api/v1/sports/1                MLB
#>    stat_group total_splits game_type_id game_type_description
#> 1     hitting           21            R        Regular Season
#> 2     hitting           21            R        Regular Season
#> 3     hitting           21            R        Regular Season
#> 4     hitting           21            R        Regular Season
#> 5     hitting           21            R        Regular Season
#> 6     hitting           21            R        Regular Season
#> 7     hitting           21            R        Regular Season
#> 8     hitting           21            R        Regular Season
#> 9     hitting           21            R        Regular Season
#> 10    hitting           21            R        Regular Season
#> 11    hitting           21            R        Regular Season
#> 12    hitting           21            R        Regular Season
#> 13    hitting           21            R        Regular Season
#> 14    hitting           21            R        Regular Season
#> 15    hitting           21            R        Regular Season
#> 16    hitting           21            R        Regular Season
#> 17    hitting           21            R        Regular Season
#> 18    hitting           21            R        Regular Season
#> 19    hitting           21            R        Regular Season
#> 20    hitting           21            R        Regular Season
#> 21    hitting           21            R        Regular Season
#> 22   pitching           28            R        Regular Season
#> 23   pitching           28            R        Regular Season
#> 24   pitching           28            R        Regular Season
#> 25   pitching           28            R        Regular Season
#> 26   pitching           28            R        Regular Season
#> 27   pitching           28            R        Regular Season
#> 28   pitching           28            R        Regular Season
#> 29   pitching           28            R        Regular Season
#> 30   pitching           28            R        Regular Season
#> 31   pitching           28            R        Regular Season
#> 32   pitching           28            R        Regular Season
#> 33   pitching           28            R        Regular Season
#> 34   pitching           28            R        Regular Season
#> 35   pitching           28            R        Regular Season
#> 36   pitching           28            R        Regular Season
#> 37   pitching           28            R        Regular Season
#> 38   pitching           28            R        Regular Season
#> 39   pitching           28            R        Regular Season
#> 40   pitching           28            R        Regular Season
#> 41   pitching           28            R        Regular Season
#> 42   pitching           28            R        Regular Season
#> 43   pitching           28            R        Regular Season
#> 44   pitching           28            R        Regular Season
#> 45   pitching           28            R        Regular Season
#> 46   pitching           28            R        Regular Season
#> 47   pitching           28            R        Regular Season
#> 48   pitching           28            R        Regular Season
#> 49   pitching           28            R        Regular Season
#> 50   catching            3            R        Regular Season
#> 51   catching            3            R        Regular Season
#> 52   catching            3            R        Regular Season
# }