MLB Team Individual Stats

MLB Team Individual Stats

Usage

mlb_team_stats(
  team_id = NULL,
  stat_type = NULL,
  game_type = NULL,
  stat_group = NULL,
  season = NULL,
  sport_ids = NULL
)

Arguments

team_id

Team ID to return information and ranking for a particular statistic for a particular team.

stat_type

Stat type to return statistics for.

game_type

Game type to return information for a particular statistic in a particular game type.

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

sport_ids

The sport_id(s) to return information and ranking information for.

Value

Returns a tibble with the following columns

col_name	types

season	character

games_played	integer

ground_outs	integer

air_outs	integer

runs	integer

doubles	integer

triples	integer

home_runs	integer

strike_outs	integer

base_on_balls	integer

intentional_walks	integer

hits	integer

hit_by_pitch	integer

avg	character

at_bats	integer

obp	character

slg	character

ops	character

caught_stealing	integer

stolen_bases	integer

stolen_base_percentage	character

ground_into_double_play	integer

number_of_pitches	integer

plate_appearances	integer

total_bases	integer

rbi	integer

left_on_base	integer

sac_bunts	integer

sac_flies	integer

babip	character

ground_outs_to_airouts	character

catchers_interference	integer

at_bats_per_home_run	character

team_id	integer

team_name	character

team_link	character

type_display_name	character

group_display_name	character

Examples

# \donttest{
  try(mlb_team_stats(team_id = 137, stat_type = 'season', stat_group = 'hitting', season = 2021))
#> ── MLB Team Stats data from MLB.com ───────────────── baseballr 1.3.0 ──
#> ℹ Data updated: 2022-09-09 07:20:26 UTC
#> # A tibble: 1 × 38
#>   season games_p…¹ groun…² air_o…³  runs doubles triples home_…⁴ strik…⁵
#>   <chr>      <int>   <int>   <int> <int>   <int>   <int>   <int>   <int>
#> 1 2021         162    1304    1403   804     271      25     241    1461
#> # … with 29 more variables: base_on_balls <int>,
#> #   intentional_walks <int>, hits <int>, hit_by_pitch <int>, avg <chr>,
#> #   at_bats <int>, obp <chr>, slg <chr>, ops <chr>,
#> #   caught_stealing <int>, stolen_bases <int>,
#> #   stolen_base_percentage <chr>, ground_into_double_play <int>,
#> #   number_of_pitches <int>, plate_appearances <int>,
#> #   total_bases <int>, rbi <int>, left_on_base <int>, …
# }