MLB Teams

MLB Teams

Usage

mlb_teams(
  season = NULL,
  active_status = NULL,
  all_star_statuses = NULL,
  league_ids = NULL,
  sport_ids = NULL,
  game_type = NULL
)

Arguments

season

Year to return to return team information for.

active_status

The active statuses to populate teams for a given season.

all_star_statuses

The all-star statuses to populate teams for a given season.

league_ids

The league_id(s) to return team information for.

sport_ids

The sport_id(s) to return team information for.

game_type

The game_type to return team information for.

Value

Returns a data frame with the following columns

col_name	types

team_id	integer

team_full_name	character

link	character

season	integer

team_code	character

file_code	character

team_abbreviation	character

team_name	character

location_name	character

first_year_of_play	character

short_name	character

franchise_name	character

club_name	character

all_star_status	character

active	logical

venue_id	integer

venue_name	character

venue_link	character

spring_venue_id	integer

spring_venue_link	character

league_id	integer

league_name	character

league_link	character

division_id	integer

division_name	character

division_link	character

sport_id	integer

sport_link	character

sport_name	character

spring_league_id	integer

spring_league_name	character

spring_league_link	character

spring_league_abbreviation	character

Examples

# \donttest{
  try(mlb_teams(season = 2021, sport_ids = c(1)))
#>    all_star_status team_id        team_full_name              link
#> 1                N     133     Oakland Athletics /api/v1/teams/133
#> 2                N     134    Pittsburgh Pirates /api/v1/teams/134
#> 3                N     135      San Diego Padres /api/v1/teams/135
#> 4                N     136      Seattle Mariners /api/v1/teams/136
#> 5                N     137  San Francisco Giants /api/v1/teams/137
#> 6                N     138   St. Louis Cardinals /api/v1/teams/138
#> 7                N     139        Tampa Bay Rays /api/v1/teams/139
#> 8                N     140         Texas Rangers /api/v1/teams/140
#> 9                N     141     Toronto Blue Jays /api/v1/teams/141
#> 10               N     142       Minnesota Twins /api/v1/teams/142
#> 11               N     143 Philadelphia Phillies /api/v1/teams/143
#> 12               N     144        Atlanta Braves /api/v1/teams/144
#> 13               N     145     Chicago White Sox /api/v1/teams/145
#> 14               N     146         Miami Marlins /api/v1/teams/146
#> 15               N     147      New York Yankees /api/v1/teams/147
#> 16               N     158     Milwaukee Brewers /api/v1/teams/158
#> 17               N     108    Los Angeles Angels /api/v1/teams/108
#> 18               N     109  Arizona Diamondbacks /api/v1/teams/109
#> 19               N     110     Baltimore Orioles /api/v1/teams/110
#> 20               N     111        Boston Red Sox /api/v1/teams/111
#> 21               N     112          Chicago Cubs /api/v1/teams/112
#> 22               N     113       Cincinnati Reds /api/v1/teams/113
#> 23               N     114     Cleveland Indians /api/v1/teams/114
#> 24               N     115      Colorado Rockies /api/v1/teams/115
#> 25               N     116        Detroit Tigers /api/v1/teams/116
#> 26               N     117        Houston Astros /api/v1/teams/117
#> 27               N     118    Kansas City Royals /api/v1/teams/118
#> 28               N     119   Los Angeles Dodgers /api/v1/teams/119
#> 29               N     120  Washington Nationals /api/v1/teams/120
#> 30               N     121         New York Mets /api/v1/teams/121
#>    season team_code file_code team_abbreviation team_name
#> 1    2021       oak       oak               OAK Athletics
#> 2    2021       pit       pit               PIT   Pirates
#> 3    2021       sdn        sd                SD    Padres
#> 4    2021       sea       sea               SEA  Mariners
#> 5    2021       sfn        sf                SF    Giants
#> 6    2021       sln       stl               STL Cardinals
#> 7    2021       tba        tb                TB      Rays
#> 8    2021       tex       tex               TEX   Rangers
#> 9    2021       tor       tor               TOR Blue Jays
#> 10   2021       min       min               MIN     Twins
#> 11   2021       phi       phi               PHI  Phillies
#> 12   2021       atl       atl               ATL    Braves
#> 13   2021       cha       cws               CWS White Sox
#> 14   2021       mia       mia               MIA   Marlins
#> 15   2021       nya       nyy               NYY   Yankees
#> 16   2021       mil       mil               MIL   Brewers
#> 17   2021       ana       ana               LAA    Angels
#> 18   2021       ari       ari               ARI   D-backs
#> 19   2021       bal       bal               BAL   Orioles
#> 20   2021       bos       bos               BOS   Red Sox
#> 21   2021       chn       chc               CHC      Cubs
#> 22   2021       cin       cin               CIN      Reds
#> 23   2021       cle       cle               CLE   Indians
#> 24   2021       col       col               COL   Rockies
#> 25   2021       det       det               DET    Tigers
#> 26   2021       hou       hou               HOU    Astros
#> 27   2021       kca        kc                KC    Royals
#> 28   2021       lan        la               LAD   Dodgers
#> 29   2021       was       was               WSH Nationals
#> 30   2021       nyn       nym               NYM      Mets
#>     location_name first_year_of_play    short_name franchise_name
#> 1         Oakland               1901       Oakland        Oakland
#> 2      Pittsburgh               1882    Pittsburgh     Pittsburgh
#> 3       San Diego               1968     San Diego      San Diego
#> 4         Seattle               1977       Seattle        Seattle
#> 5   San Francisco               1883 San Francisco  San Francisco
#> 6       St. Louis               1892     St. Louis      St. Louis
#> 7  St. Petersburg               1996     Tampa Bay      Tampa Bay
#> 8       Arlington               1961         Texas          Texas
#> 9         Toronto               1977       Toronto        Toronto
#> 10    Minneapolis               1901     Minnesota      Minnesota
#> 11   Philadelphia               1883  Philadelphia   Philadelphia
#> 12        Atlanta               1871       Atlanta        Atlanta
#> 13        Chicago               1901 Chi White Sox        Chicago
#> 14          Miami               1991         Miami          Miami
#> 15          Bronx               1903    NY Yankees       New York
#> 16      Milwaukee               1968     Milwaukee      Milwaukee
#> 17        Anaheim               1961     LA Angels    Los Angeles
#> 18        Phoenix               1996       Arizona        Arizona
#> 19      Baltimore               1901     Baltimore      Baltimore
#> 20         Boston               1901        Boston         Boston
#> 21        Chicago               1874      Chi Cubs        Chicago
#> 22     Cincinnati               1882    Cincinnati     Cincinnati
#> 23      Cleveland               1901     Cleveland      Cleveland
#> 24         Denver               1992      Colorado       Colorado
#> 25        Detroit               1901       Detroit        Detroit
#> 26        Houston               1962       Houston        Houston
#> 27    Kansas City               1968   Kansas City    Kansas City
#> 28    Los Angeles               1884    LA Dodgers    Los Angeles
#> 29     Washington               1968    Washington     Washington
#> 30       Flushing               1962       NY Mets       New York
#>       club_name active spring_league_id spring_league_name
#> 1     Athletics   TRUE              114      Cactus League
#> 2       Pirates   TRUE              115  Grapefruit League
#> 3        Padres   TRUE              114      Cactus League
#> 4      Mariners   TRUE              114      Cactus League
#> 5        Giants   TRUE              114      Cactus League
#> 6     Cardinals   TRUE              115  Grapefruit League
#> 7          Rays   TRUE              115  Grapefruit League
#> 8       Rangers   TRUE              114      Cactus League
#> 9     Blue Jays   TRUE              115  Grapefruit League
#> 10        Twins   TRUE              115  Grapefruit League
#> 11     Phillies   TRUE              115  Grapefruit League
#> 12       Braves   TRUE              115  Grapefruit League
#> 13    White Sox   TRUE              114      Cactus League
#> 14      Marlins   TRUE              115  Grapefruit League
#> 15      Yankees   TRUE              115  Grapefruit League
#> 16      Brewers   TRUE              114      Cactus League
#> 17       Angels   TRUE              114      Cactus League
#> 18 Diamondbacks   TRUE              114      Cactus League
#> 19      Orioles   TRUE              115  Grapefruit League
#> 20      Red Sox   TRUE              115  Grapefruit League
#> 21         Cubs   TRUE              114      Cactus League
#> 22         Reds   TRUE              114      Cactus League
#> 23      Indians   TRUE              114      Cactus League
#> 24      Rockies   TRUE              114      Cactus League
#> 25       Tigers   TRUE              115  Grapefruit League
#> 26       Astros   TRUE              115  Grapefruit League
#> 27       Royals   TRUE              114      Cactus League
#> 28      Dodgers   TRUE              114      Cactus League
#> 29    Nationals   TRUE              115  Grapefruit League
#> 30         Mets   TRUE              115  Grapefruit League
#>    spring_league_link spring_league_abbreviation venue_id
#> 1  /api/v1/league/114                         CL       10
#> 2  /api/v1/league/115                         GL       31
#> 3  /api/v1/league/114                         CL     2680
#> 4  /api/v1/league/114                         CL      680
#> 5  /api/v1/league/114                         CL     2395
#> 6  /api/v1/league/115                         GL     2889
#> 7  /api/v1/league/115                         GL       12
#> 8  /api/v1/league/114                         CL     5325
#> 9  /api/v1/league/115                         GL       14
#> 10 /api/v1/league/115                         GL     3312
#> 11 /api/v1/league/115                         GL     2681
#> 12 /api/v1/league/115                         GL     4705
#> 13 /api/v1/league/114                         CL        4
#> 14 /api/v1/league/115                         GL     4169
#> 15 /api/v1/league/115                         GL     3313
#> 16 /api/v1/league/114                         CL       32
#> 17 /api/v1/league/114                         CL        1
#> 18 /api/v1/league/114                         CL       15
#> 19 /api/v1/league/115                         GL        2
#> 20 /api/v1/league/115                         GL        3
#> 21 /api/v1/league/114                         CL       17
#> 22 /api/v1/league/114                         CL     2602
#> 23 /api/v1/league/114                         CL        5
#> 24 /api/v1/league/114                         CL       19
#> 25 /api/v1/league/115                         GL     2394
#> 26 /api/v1/league/115                         GL     2392
#> 27 /api/v1/league/114                         CL        7
#> 28 /api/v1/league/114                         CL       22
#> 29 /api/v1/league/115                         GL     3309
#> 30 /api/v1/league/115                         GL     3289
#>                     venue_name          venue_link spring_venue_id
#> 1             Oakland Coliseum   /api/v1/venues/10            2507
#> 2                     PNC Park   /api/v1/venues/31            2526
#> 3                   Petco Park /api/v1/venues/2680            2530
#> 4                T-Mobile Park  /api/v1/venues/680            2530
#> 5                  Oracle Park /api/v1/venues/2395            2532
#> 6                Busch Stadium /api/v1/venues/2889            2520
#> 7              Tropicana Field   /api/v1/venues/12            2534
#> 8             Globe Life Field /api/v1/venues/5325            2603
#> 9                Rogers Centre   /api/v1/venues/14            2536
#> 10                Target Field /api/v1/venues/3312            2862
#> 11          Citizens Bank Park /api/v1/venues/2681            2700
#> 12                 Truist Park /api/v1/venues/4705            5380
#> 13       Guaranteed Rate Field    /api/v1/venues/4            3809
#> 14              loanDepot park /api/v1/venues/4169            2520
#> 15              Yankee Stadium /api/v1/venues/3313            2523
#> 16       American Family Field   /api/v1/venues/32            2518
#> 17               Angel Stadium    /api/v1/venues/1            2500
#> 18                 Chase Field   /api/v1/venues/15            4249
#> 19 Oriole Park at Camden Yards    /api/v1/venues/2            2508
#> 20                 Fenway Park    /api/v1/venues/3            4309
#> 21               Wrigley Field   /api/v1/venues/17            4629
#> 22    Great American Ball Park /api/v1/venues/2602            3834
#> 23           Progressive Field    /api/v1/venues/5            3834
#> 24                 Coors Field   /api/v1/venues/19            4249
#> 25               Comerica Park /api/v1/venues/2394            2511
#> 26            Minute Maid Park /api/v1/venues/2392            5000
#> 27            Kauffman Stadium    /api/v1/venues/7            2603
#> 28              Dodger Stadium   /api/v1/venues/22            3809
#> 29              Nationals Park /api/v1/venues/3309            5000
#> 30                  Citi Field /api/v1/venues/3289            2856
#>      spring_venue_link league_id     league_name        league_link
#> 1  /api/v1/venues/2507       103 American League /api/v1/league/103
#> 2  /api/v1/venues/2526       104 National League /api/v1/league/104
#> 3  /api/v1/venues/2530       104 National League /api/v1/league/104
#> 4  /api/v1/venues/2530       103 American League /api/v1/league/103
#> 5  /api/v1/venues/2532       104 National League /api/v1/league/104
#> 6  /api/v1/venues/2520       104 National League /api/v1/league/104
#> 7  /api/v1/venues/2534       103 American League /api/v1/league/103
#> 8  /api/v1/venues/2603       103 American League /api/v1/league/103
#> 9  /api/v1/venues/2536       103 American League /api/v1/league/103
#> 10 /api/v1/venues/2862       103 American League /api/v1/league/103
#> 11 /api/v1/venues/2700       104 National League /api/v1/league/104
#> 12 /api/v1/venues/5380       104 National League /api/v1/league/104
#> 13 /api/v1/venues/3809       103 American League /api/v1/league/103
#> 14 /api/v1/venues/2520       104 National League /api/v1/league/104
#> 15 /api/v1/venues/2523       103 American League /api/v1/league/103
#> 16 /api/v1/venues/2518       104 National League /api/v1/league/104
#> 17 /api/v1/venues/2500       103 American League /api/v1/league/103
#> 18 /api/v1/venues/4249       104 National League /api/v1/league/104
#> 19 /api/v1/venues/2508       103 American League /api/v1/league/103
#> 20 /api/v1/venues/4309       103 American League /api/v1/league/103
#> 21 /api/v1/venues/4629       104 National League /api/v1/league/104
#> 22 /api/v1/venues/3834       104 National League /api/v1/league/104
#> 23 /api/v1/venues/3834       103 American League /api/v1/league/103
#> 24 /api/v1/venues/4249       104 National League /api/v1/league/104
#> 25 /api/v1/venues/2511       103 American League /api/v1/league/103
#> 26 /api/v1/venues/5000       103 American League /api/v1/league/103
#> 27 /api/v1/venues/2603       103 American League /api/v1/league/103
#> 28 /api/v1/venues/3809       104 National League /api/v1/league/104
#> 29 /api/v1/venues/5000       104 National League /api/v1/league/104
#> 30 /api/v1/venues/2856       104 National League /api/v1/league/104
#>    division_id           division_name         division_link sport_id
#> 1          200    American League West /api/v1/divisions/200        1
#> 2          205 National League Central /api/v1/divisions/205        1
#> 3          203    National League West /api/v1/divisions/203        1
#> 4          200    American League West /api/v1/divisions/200        1
#> 5          203    National League West /api/v1/divisions/203        1
#> 6          205 National League Central /api/v1/divisions/205        1
#> 7          201    American League East /api/v1/divisions/201        1
#> 8          200    American League West /api/v1/divisions/200        1
#> 9          201    American League East /api/v1/divisions/201        1
#> 10         202 American League Central /api/v1/divisions/202        1
#> 11         204    National League East /api/v1/divisions/204        1
#> 12         204    National League East /api/v1/divisions/204        1
#> 13         202 American League Central /api/v1/divisions/202        1
#> 14         204    National League East /api/v1/divisions/204        1
#> 15         201    American League East /api/v1/divisions/201        1
#> 16         205 National League Central /api/v1/divisions/205        1
#> 17         200    American League West /api/v1/divisions/200        1
#> 18         203    National League West /api/v1/divisions/203        1
#> 19         201    American League East /api/v1/divisions/201        1
#> 20         201    American League East /api/v1/divisions/201        1
#> 21         205 National League Central /api/v1/divisions/205        1
#> 22         205 National League Central /api/v1/divisions/205        1
#> 23         202 American League Central /api/v1/divisions/202        1
#> 24         203    National League West /api/v1/divisions/203        1
#> 25         202 American League Central /api/v1/divisions/202        1
#> 26         200    American League West /api/v1/divisions/200        1
#> 27         202 American League Central /api/v1/divisions/202        1
#> 28         203    National League West /api/v1/divisions/203        1
#> 29         204    National League East /api/v1/divisions/204        1
#> 30         204    National League East /api/v1/divisions/204        1
#>          sport_link            sport_name
#> 1  /api/v1/sports/1 Major League Baseball
#> 2  /api/v1/sports/1 Major League Baseball
#> 3  /api/v1/sports/1 Major League Baseball
#> 4  /api/v1/sports/1 Major League Baseball
#> 5  /api/v1/sports/1 Major League Baseball
#> 6  /api/v1/sports/1 Major League Baseball
#> 7  /api/v1/sports/1 Major League Baseball
#> 8  /api/v1/sports/1 Major League Baseball
#> 9  /api/v1/sports/1 Major League Baseball
#> 10 /api/v1/sports/1 Major League Baseball
#> 11 /api/v1/sports/1 Major League Baseball
#> 12 /api/v1/sports/1 Major League Baseball
#> 13 /api/v1/sports/1 Major League Baseball
#> 14 /api/v1/sports/1 Major League Baseball
#> 15 /api/v1/sports/1 Major League Baseball
#> 16 /api/v1/sports/1 Major League Baseball
#> 17 /api/v1/sports/1 Major League Baseball
#> 18 /api/v1/sports/1 Major League Baseball
#> 19 /api/v1/sports/1 Major League Baseball
#> 20 /api/v1/sports/1 Major League Baseball
#> 21 /api/v1/sports/1 Major League Baseball
#> 22 /api/v1/sports/1 Major League Baseball
#> 23 /api/v1/sports/1 Major League Baseball
#> 24 /api/v1/sports/1 Major League Baseball
#> 25 /api/v1/sports/1 Major League Baseball
#> 26 /api/v1/sports/1 Major League Baseball
#> 27 /api/v1/sports/1 Major League Baseball
#> 28 /api/v1/sports/1 Major League Baseball
#> 29 /api/v1/sports/1 Major League Baseball
#> 30 /api/v1/sports/1 Major League Baseball
# }