MLB Teams Stats

MLB Teams Stats

Usage

mlb_teams_stats(
  stat_type = NULL,
  game_type = NULL,
  stat_group = NULL,
  season = NULL,
  sport_ids = NULL,
  sort_stat = NULL,
  order = NULL
)

Arguments

stat_type

Stat type to return statistics for.

game_type

Game type to return information for a particular statistic in a particular game type.

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

sport_ids

The sport_id(s) to return information and ranking information for.

sort_stat

Sort return based on stat.

order

Order return based on either desc or asc.

Value

Returns a tibble with the following columns

col_name	types

total_splits	integer

season	character

rank	integer

games_played	integer

ground_outs	integer

air_outs	integer

runs	integer

doubles	integer

triples	integer

home_runs	integer

strike_outs	integer

base_on_balls	integer

intentional_walks	integer

hits	integer

hit_by_pitch	integer

avg	character

at_bats	integer

obp	character

slg	character

ops	character

caught_stealing	integer

stolen_bases	integer

stolen_base_percentage	character

ground_into_double_play	integer

number_of_pitches	integer

plate_appearances	integer

total_bases	integer

rbi	integer

left_on_base	integer

sac_bunts	integer

sac_flies	integer

babip	character

ground_outs_to_airouts	character

catchers_interference	integer

at_bats_per_home_run	character

team_id	integer

team_name	character

team_link	character

splits_tied_with_offset	list

splits_tied_with_limit	list

type_display_name	character

group_display_name	character

Examples

# \donttest{
  try(mlb_teams_stats(stat_type = 'season', stat_group = 'hitting', season = 2021))
#>    total_splits season rank games_played ground_outs air_outs runs
#> 1           371   2021    1            2           8        6   10
#> 2           371   2021    2            2          14       19   18
#> 3           371   2021    3            2          14        8    9
#> 4           371   2021    4            1           4       17    7
#> 5           371   2021    5            2          18       15   19
#> 6           371   2021    6            1           9        7   13
#> 7           371   2021    7           63         546      495  500
#> 8           371   2021    8            1           8        4    9
#> 9           371   2021    9            3          24       26   23
#> 10          371   2021   10           66         593      594  438
#> 11          371   2021   11           64         575      608  361
#> 12          371   2021   12           62         590      479  376
#> 13          371   2021   13            4          30       36   24
#> 14          371   2021   14            2          16       12   14
#> 15          371   2021   15           49         476      473  313
#> 16          371   2021   16            5          37       46   32
#> 17          371   2021   17           49         464      446  288
#> 18          371   2021   18           61         559      502  322
#> 19          371   2021   19           59         448      377  376
#> 20          371   2021   20            1           6       12    8
#> 21          371   2021   21          124        1014     1047  872
#> 22          371   2021   22           65         555      605  351
#> 23          371   2021   23          120         933     1146  884
#> 24          371   2021   24            3          22       30   19
#> 25          371   2021   25           66         662      589  369
#> 26          371   2021   26           66         641      576  350
#> 27          371   2021   27           49         437      457  298
#> 28          371   2021   28           66         663      565  309
#> 29          371   2021   29           49         442      481  287
#> 30          371   2021   30           50         503      510  278
#> 31          371   2021   31           66         633      595  358
#> 32          371   2021   32           65         627      550  374
#> 33          371   2021   33            5          35       47   42
#> 34          371   2021   34           60         423      375  384
#> 35          371   2021   35           66         574      506  327
#> 36          371   2021   36            5          29       34   33
#> 37          371   2021   37          120         875      981  753
#> 38          371   2021   38           64         572      512  338
#> 39          371   2021   39          120        1017     1119  826
#> 40          371   2021   40            1          15        8    5
#> 41          371   2021   41            3          31       22   21
#> 42          371   2021   42           68         707      646  309
#> 43          371   2021   43           58         444      441  381
#> 44          371   2021   44           66         653      570  384
#> 45          371   2021   45          120        1029     1284  760
#> 46          371   2021   46          127        1015     1050  753
#> 47          371   2021   47           31         251      279  170
#> 48          371   2021   48          120         955     1136  647
#> 49          371   2021   49          120        1026     1023  686
#> 50          371   2021   50          126        1087     1096  794
#>    doubles triples home_runs strike_outs base_on_balls
#> 1        2       2         0          20             6
#> 2        4       1         6          12            13
#> 3        4       3         0          14             0
#> 4        0       0         1           6             7
#> 5        6       0         4          20            18
#> 6        3       1         3           9             5
#> 7      130      15        86         431           254
#> 8        2       0         1          13             7
#> 9        4       0         2          23            11
#> 10     143      11       102         469           226
#> 11     119       7        67         404           249
#> 12     119       7        81         412           187
#> 13       2       1         3          27             4
#> 14       4       1         1          19            12
#> 15      99       8        50         269           188
#> 16       4       1         7          29            24
#> 17      92       4        31         326           216
#> 18     115       7        55         391           187
#> 19      96      16        46         399           248
#> 20       2       0         1           8             5
#> 21     284      50       198        1086           502
#> 22     111      10        73         444           216
#> 23     215      28       205         876           574
#> 24       6       0         0          16            15
#> 25     151       4        72         417           202
#> 26      98       9        55         435           216
#> 27     103      13        56         336           199
#> 28     112       7        37         387           207
#> 29      95       9        50         287           185
#> 30      96       4        44         280           163
#> 31     111      11        58         455           220
#> 32     112       9        65         432           262
#> 33       9       0        11          43            21
#> 34     100      29        57         564           252
#> 35     120      17        51         514           214
#> 36      11       2         3          52            17
#> 37     208      24       157        1098           585
#> 38     112      12        55         533           252
#> 39     227      23       180         850           591
#> 40       2       0         0           6             0
#> 41       3       1         4          26            13
#> 42     102       5        38         363           237
#> 43     113      14        52         498           267
#> 44     103       6        74         493           240
#> 45     213      13       124         621           563
#> 46     236      34       186        1202           477
#> 47      51       6        28         219           137
#> 48     211      10       118         846           499
#> 49     189      24       130         860           457
#> 50     248      37       193        1070           494
#>    intentional_walks hits hit_by_pitch  avg at_bats  obp  slg   ops
#> 1                  0   20            0 .377      53 .433 .491  .924
#> 2                  0   26            1 .371      70 .476 .714 1.190
#> 3                  0   21            0 .362      58 .362 .534  .896
#> 4                  1   14            0 .359      39 .447 .436  .883
#> 5                  0   28            1 .350      80 .470 .575 1.045
#> 6                  0   12            1 .343      35 .429 .743 1.172
#> 7                  9  735           32 .338    2173 .411 .531  .942
#> 8                  0   11            0 .324      34 .419 .471  .890
#> 9                  0   33            7 .320     103 .415 .417  .832
#> 10                 6  719           25 .308    2335 .372 .510  .882
#> 11                 9  691           23 .308    2245 .380 .457  .837
#> 12                10  638           28 .306    2085 .368 .486  .854
#> 13                 0   38            9 .304     125 .362 .408  .770
#> 14                 0   20            1 .303      66 .413 .439  .852
#> 15                14  499           36 .299    1670 .376 .457  .833
#> 16                 0   46            1 .299     154 .394 .474  .868
#> 17                10  511           19 .296    1729 .378 .407  .785
#> 18                 5  594           18 .295    2011 .358 .442  .800
#> 19                 3  495           53 .295    1680 .397 .453  .850
#> 20                 1   10            0 .294      34 .375 .441  .816
#> 21                 6 1284           84 .294    4370 .373 .518  .891
#> 22                 5  654           33 .293    2229 .361 .450  .811
#> 23                14 1203           96 .293    4107 .389 .509  .898
#> 24                 0   28            3 .292      96 .404 .354  .758
#> 25                11  671           39 .291    2303 .356 .454  .810
#> 26                 6  661           36 .291    2271 .359 .415  .774
#> 27                 6  496           27 .291    1705 .371 .465  .836
#> 28                 8  640           31 .291    2202 .356 .398  .754
#> 29                13  484           15 .289    1672 .362 .447  .809
#> 30                12  509           43 .288    1768 .358 .421  .779
#> 31                10  659           42 .287    2298 .357 .420  .777
#> 32                 9  632           21 .287    2204 .365 .434  .799
#> 33                 0   49            3 .287     171 .374 .532  .906
#> 34                 1  536           30 .286    1871 .376 .462  .838
#> 35                 6  625           33 .286    2189 .355 .426  .781
#> 36                 0   45            4 .285     158 .369 .437  .806
#> 37                 9 1140           63 .283    4024 .378 .464  .842
#> 38                 4  624           29 .283    2206 .362 .419  .781
#> 39                14 1147           54 .282    4065 .376 .482  .858
#> 40                 0   11            0 .282      39 .275 .333  .608
#> 41                 0   31            0 .282     110 .358 .436  .794
#> 42                18  651           28 .281    2315 .352 .379  .731
#> 43                 4  526           49 .280    1878 .379 .438  .817
#> 44                 4  651           36 .280    2325 .354 .425  .779
#> 45                10 1110           79 .279    3977 .375 .433  .808
#> 46                 6 1241           67 .279    4455 .354 .472  .826
#> 47                 9  283           26 .279    1016 .375 .423  .798
#> 48                11 1108           60 .278    3983 .363 .425  .788
#> 49                 7 1094          117 .278    3939 .366 .437  .803
#> 50                10 1227           62 .277    4424 .355 .481  .836
#>    caught_stealing stolen_bases stolen_base_percentage
#> 1                3            8                   .727
#> 2                0            1                  1.000
#> 3                1            2                   .667
#> 4                0            1                  1.000
#> 5                0            0                   .---
#> 6                1            3                   .750
#> 7               35           69                   .663
#> 8                0            1                  1.000
#> 9                3            4                   .571
#> 10              14           30                   .682
#> 11              32           31                   .492
#> 12              20           21                   .512
#> 13               2            7                   .778
#> 14               4            9                   .692
#> 15               9           13                   .591
#> 16               0            3                  1.000
#> 17               9           17                   .654
#> 18              22           33                   .600
#> 19              42          115                   .732
#> 20               0            0                   .---
#> 21              33           94                   .740
#> 22              34           49                   .590
#> 23              18          111                   .860
#> 24               2            1                   .333
#> 25              23           34                   .596
#> 26              17           35                   .673
#> 27              17           27                   .614
#> 28              22           46                   .676
#> 29              12           21                   .636
#> 30              10           15                   .600
#> 31              17           23                   .575
#> 32              24           45                   .652
#> 33               0            3                  1.000
#> 34              21           67                   .761
#> 35              13           27                   .675
#> 36               2            3                   .600
#> 37              29          123                   .809
#> 38              16           27                   .628
#> 39              34          159                   .824
#> 40               0            0                   .---
#> 41               1            1                   .500
#> 42              60           74                   .552
#> 43              31           72                   .699
#> 44              12           41                   .774
#> 45              25          137                   .846
#> 46              25           89                   .781
#> 47              10           27                   .730
#> 48              27           84                   .757
#> 49              26           72                   .735
#> 50              11           53                   .828
#>    ground_into_double_play number_of_pitches plate_appearances
#> 1                        1               281                60
#> 2                        1               323                85
#> 3                        0               209                58
#> 4                        0               183                48
#> 5                        1               421               100
#> 6                        1               169                43
#> 7                       55              9691              2494
#> 8                        0               189                43
#> 9                        1               510               124
#> 10                      48              9743              2626
#> 11                      54              9646              2550
#> 12                      51              8510              2334
#> 13                       0               497               144
#> 14                       1               311                80
#> 15                      57              6993              1941
#> 16                       5               713               183
#> 17                      55              7561              1982
#> 18                      59              8346              2251
#> 19                      36              3820              2020
#> 20                       0               166                42
#> 21                      71             19239              5018
#> 22                      57              9290              2507
#> 23                      80             18302              4828
#> 24                       3               443               114
#> 25                      56              9079              2580
#> 26                      67              9542              2565
#> 27                      40              7521              1953
#> 28                      79              9219              2493
#> 29                      42              6963              1894
#> 30                      59              7263              2009
#> 31                      55              9690              2606
#> 32                      49              9483              2525
#> 33                       3               761               198
#> 34                      27              4300              2180
#> 35                      69              9273              2466
#> 36                       0               702               181
#> 37                      95             18994              4745
#> 38                      55              9582              2522
#> 39                      73             18410              4778
#> 40                       1               138                40
#> 41                       4               463               123
#> 42                      69              9682              2633
#> 43                      20              4347              2225
#> 44                      63             10047              2643
#> 45                      93             17447              4686
#> 46                      84             19887              5053
#> 47                      19              2063              1195
#> 48                     101             17854              4604
#> 49                      69             16735              4577
#> 50                      98             19874              5037
#>    total_bases rbi left_on_base sac_bunts sac_flies babip
#> 1           26   8           18         0         1  .588
#> 2           50  18           21         1         0  .385
#> 3           31   9           10         0         0  .477
#> 4           17   7           14         1         1  .394
#> 5           46  18           27         0         1  .421
#> 6           26  12            3         1         1  .375
#> 7         1153 460          445        10        24  .386
#> 8           16   8           10         0         2  .455
#> 9           43  21           23         1         2  .388
#> 10        1190 413          481        16        24  .345
#> 11        1025 340          521        19        14  .349
#> 12        1014 353          430        17        17  .346
#> 13          51  21           21         3         3  .357
#> 14          29  12           15         0         1  .404
#> 15         764 295          364        16        31  .325
#> 16          73  28           37         3         1  .328
#> 17         704 254          407         9         9  .348
#> 18         888 297          398        21        14  .341
#> 19         761 316          377        16        23  .357
#> 20          15   8            7         1         1  .346
#> 21        2262 816          910        10        52  .346
#> 22        1004 334          458         4        25  .334
#> 23        2089 823          937        11        40  .326
#> 24          34  17           23         0         0  .350
#> 25        1046 343          490        17        19  .327
#> 26         942 320          495        22        20  .336
#> 27         793 280          386         6        15  .331
#> 28         877 285          467        26        27  .334
#> 29         747 268          360         7        15  .321
#> 30         745 257          397        13        21  .317
#> 31         966 343          513        23        21  .333
#> 32         957 352          479        19        18  .329
#> 33          91  40           30         3         0  .325
#> 34         865 334          403         4        23  .376
#> 35         932 293          465         9        21  .349
#> 36          69  32           29         2         0  .408
#> 37        1867 700          946        17        53  .348
#> 38         925 318          504        20        15  .348
#> 39        1960 760          923        17        51  .313
#> 40          13   3            8         0         1  .324
#> 41          48  18           21         0         0  .338
#> 42         877 285          494        27        25  .316
#> 43         823 321          436         3        28  .350
#> 44         988 353          489        23        19  .325
#> 45        1721 689          922        15        52  .300
#> 46        2103 713          960        13        40  .340
#> 47         430 150          251         6        10  .327
#> 48        1693 602          932        16        46  .323
#> 49        1721 638          913        25        39  .323
#> 50        2128 758          922         7        49  .322
#>    ground_outs_to_airouts catchers_interference at_bats_per_home_run
#> 1                    1.33                     0                 -.--
#> 2                    0.74                     0                11.67
#> 3                    1.75                     0                 -.--
#> 4                    0.24                     0                39.00
#> 5                    1.20                     0                20.00
#> 6                    1.29                     0                11.67
#> 7                    1.10                     1                25.27
#> 8                    2.00                     0                34.00
#> 9                    0.92                     0                51.50
#> 10                   1.00                     0                22.89
#> 11                   0.95                     0                33.51
#> 12                   1.23                     0                25.74
#> 13                   0.83                     0                41.67
#> 14                   1.33                     0                66.00
#> 15                   1.01                     0                33.40
#> 16                   0.80                     0                22.00
#> 17                   1.04                     0                55.77
#> 18                   1.11                     0                36.56
#> 19                   1.19                     0                36.52
#> 20                   0.50                     1                34.00
#> 21                   0.97                     0                22.07
#> 22                   0.92                     0                30.53
#> 23                   0.81                     0                20.03
#> 24                   0.73                     0                 -.--
#> 25                   1.12                     0                31.99
#> 26                   1.11                     0                41.29
#> 27                   0.96                     1                30.45
#> 28                   1.17                     0                59.51
#> 29                   0.92                     0                33.44
#> 30                   0.99                     1                40.18
#> 31                   1.06                     2                39.62
#> 32                   1.14                     1                33.91
#> 33                   0.74                     0                15.54
#> 34                   1.13                     0                32.82
#> 35                   1.13                     0                42.92
#> 36                   0.85                     0                52.67
#> 37                   0.89                     3                25.63
#> 38                   1.12                     0                40.11
#> 39                   0.91                     0                22.58
#> 40                   1.88                     0                 -.--
#> 41                   1.41                     0                27.50
#> 42                   1.09                     1                60.92
#> 43                   1.01                     0                36.12
#> 44                   1.15                     0                31.42
#> 45                   0.80                     0                32.07
#> 46                   0.97                     1                23.95
#> 47                   0.90                     0                36.29
#> 48                   0.84                     0                33.75
#> 49                   1.00                     0                30.30
#> 50                   0.99                     1                22.92
#>    team_id                    team_name          team_link
#> 1     5600        Brewers (Mid-Atl '22) /api/v1/teams/5600
#> 2      805           Dominican Republic  /api/v1/teams/805
#> 3     5595                Mets (SC '22) /api/v1/teams/5595
#> 4      940                United States  /api/v1/teams/940
#> 5     5465          Cincinnati Bearcats /api/v1/teams/5465
#> 6     5556    Louisiana-Monroe Warhawks /api/v1/teams/5556
#> 7     5566     Mariachis de Guadalajara /api/v1/teams/5566
#> 8     5055                      PG East /api/v1/teams/5055
#> 9     5067          St. Mary's Rattlers /api/v1/teams/5067
#> 10     502        Saraperos de Saltillo  /api/v1/teams/502
#> 11     562        Sultanes de Monterrey  /api/v1/teams/562
#> 12     520            Pericos de Puebla  /api/v1/teams/520
#> 13    5070       Central Missouri Mules /api/v1/teams/5070
#> 14    5058                     National /api/v1/teams/5058
#> 15     694        Caribes de Anzoategui  /api/v1/teams/694
#> 16    5572        East Carolina Pirates /api/v1/teams/5572
#> 17     695           Leones del Caracas  /api/v1/teams/695
#> 18     528   Rieleros de Aguascalientes  /api/v1/teams/528
#> 19    1994                  ACL Rockies /api/v1/teams/1994
#> 20    3032             Texas A&M Aggies /api/v1/teams/3032
#> 21    2310                    Reno Aces /api/v1/teams/2310
#> 22    5010             Toros de Tijuana /api/v1/teams/5010
#> 23     495            Lexington Legends  /api/v1/teams/495
#> 24    5324    Texas-Arlington Mavericks /api/v1/teams/5324
#> 25    5567        El Aguila de Veracruz /api/v1/teams/5567
#> 26     536 Tecolotes de los Dos Laredos  /api/v1/teams/536
#> 27     696    Navegantes del Magallanes  /api/v1/teams/696
#> 28     442           Olmecas de Tabasco  /api/v1/teams/442
#> 29     693           Cardenales de Lara  /api/v1/teams/693
#> 30     699             Tigres de Aragua  /api/v1/teams/699
#> 31     560           Acereros del Norte  /api/v1/teams/560
#> 32     532     Diablos Rojos del Mexico  /api/v1/teams/532
#> 33    5575      Central Florida Knights /api/v1/teams/5575
#> 34     407                     ACL Cubs  /api/v1/teams/407
#> 35    4444         Generales de Durango /api/v1/teams/4444
#> 36    5573            Tulane Green Wave /api/v1/teams/5573
#> 37    5435              York Revolution /api/v1/teams/5435
#> 38     447     Algodoneros Union Laguna  /api/v1/teams/447
#> 39    5508       Gastonia Honey Hunters /api/v1/teams/5508
#> 40     798                         Cuba  /api/v1/teams/798
#> 41    5080   Mississippi State Bulldogs /api/v1/teams/5080
#> 42     678        Tomateros de Culiacan  /api/v1/teams/678
#> 43    5374                  ACL Indians /api/v1/teams/5374
#> 44     434               Bravos de Leon  /api/v1/teams/434
#> 45    1896            Long Island Ducks /api/v1/teams/1896
#> 46     105        Sacramento River Cats  /api/v1/teams/105
#> 47     688           Indios de Mayaguez  /api/v1/teams/688
#> 48     525       Charleston Dirty Birds  /api/v1/teams/525
#> 49    5433 Southern Maryland Blue Crabs /api/v1/teams/5433
#> 50     400           Las Vegas Aviators  /api/v1/teams/400
#>    splits_tied_with_offset splits_tied_with_limit type_display_name
#> 1                     NULL                   NULL            season
#> 2                     NULL                   NULL            season
#> 3                     NULL                   NULL            season
#> 4                     NULL                   NULL            season
#> 5                     NULL                   NULL            season
#> 6                     NULL                   NULL            season
#> 7                     NULL                   NULL            season
#> 8                     NULL                   NULL            season
#> 9                     NULL                   NULL            season
#> 10                    NULL                   NULL            season
#> 11                    NULL                   NULL            season
#> 12                    NULL                   NULL            season
#> 13                    NULL                   NULL            season
#> 14                    NULL                   NULL            season
#> 15                    NULL                   NULL            season
#> 16                    NULL                   NULL            season
#> 17                    NULL                   NULL            season
#> 18                    NULL                   NULL            season
#> 19                    NULL                   NULL            season
#> 20                    NULL                   NULL            season
#> 21                    NULL                   NULL            season
#> 22                    NULL                   NULL            season
#> 23                    NULL                   NULL            season
#> 24                    NULL                   NULL            season
#> 25                    NULL                   NULL            season
#> 26                    NULL                   NULL            season
#> 27                    NULL                   NULL            season
#> 28                    NULL                   NULL            season
#> 29                    NULL                   NULL            season
#> 30                    NULL                   NULL            season
#> 31                    NULL                   NULL            season
#> 32                    NULL                   NULL            season
#> 33                    NULL                   NULL            season
#> 34                    NULL                   NULL            season
#> 35                    NULL                   NULL            season
#> 36                    NULL                   NULL            season
#> 37                    NULL                   NULL            season
#> 38                    NULL                   NULL            season
#> 39                    NULL                   NULL            season
#> 40                    NULL                   NULL            season
#> 41                    NULL                   NULL            season
#> 42                    NULL                   NULL            season
#> 43                    NULL                   NULL            season
#> 44                    NULL                   NULL            season
#> 45                    NULL                   NULL            season
#> 46                    NULL                   NULL            season
#> 47                    NULL                   NULL            season
#> 48                    NULL                   NULL            season
#> 49                    NULL                   NULL            season
#> 50                    NULL                   NULL            season
#>    group_display_name
#> 1             hitting
#> 2             hitting
#> 3             hitting
#> 4             hitting
#> 5             hitting
#> 6             hitting
#> 7             hitting
#> 8             hitting
#> 9             hitting
#> 10            hitting
#> 11            hitting
#> 12            hitting
#> 13            hitting
#> 14            hitting
#> 15            hitting
#> 16            hitting
#> 17            hitting
#> 18            hitting
#> 19            hitting
#> 20            hitting
#> 21            hitting
#> 22            hitting
#> 23            hitting
#> 24            hitting
#> 25            hitting
#> 26            hitting
#> 27            hitting
#> 28            hitting
#> 29            hitting
#> 30            hitting
#> 31            hitting
#> 32            hitting
#> 33            hitting
#> 34            hitting
#> 35            hitting
#> 36            hitting
#> 37            hitting
#> 38            hitting
#> 39            hitting
#> 40            hitting
#> 41            hitting
#> 42            hitting
#> 43            hitting
#> 44            hitting
#> 45            hitting
#> 46            hitting
#> 47            hitting
#> 48            hitting
#> 49            hitting
#> 50            hitting
# }