MLB Teams Stats Leaders

MLB Teams Stats Leaders

Usage

mlb_teams_stats_leaders(
  leader_categories = NULL,
  leader_game_types = NULL,
  sit_codes = NULL,
  stat_group = NULL,
  season = NULL,
  league_id = NULL,
  sport_id = NULL,
  start_date = NULL,
  end_date = NULL,
  stat_type = NULL,
  limit = 1000
)

Arguments

leader_categories

League leader category to return information and ranking for a particular statistic.

leader_game_types

Game type to return information and ranking for a particular statistic in a particular game type.

sit_codes

Situation code to return information and ranking for a particular statistic in a particular game type.

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

league_id

League ID to return statistics for a given league. Default to "Qualified" player pool.

sport_id

The sport_id to return information and ranking information for.

start_date

Start date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY start_date must be coupled with end_date and byDateRange stat_type

end_date

End date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY end_date must be coupled with start_date and byDateRange stat_type

stat_type

The stat_type to return information and ranking for a particular statistic for a particular stat type.

limit

A limit to limit return to a particular number of records.

Value

Returns a tibble with the following columns

col_name	types

leader_category	character

rank	integer

value	character

season	character

num_teams	integer

team_id	integer

team_name	character

team_link	character

league_id	integer

league_name	character

league_link	character

person_id	integer

person_full_name	character

person_link	character

person_first_name	character

person_last_name	character

sport_id	integer

sport_link	character

sport_abbreviation	character

stat_group	character

total_splits	integer

game_type_id	character

game_type_description	character

Examples

# \donttest{
 try(mlb_teams_stats_leaders(leader_categories='homeRuns',sport_id=1, season = 2021))
#>    leader_category rank value season team_id             team_name
#> 1         homeRuns    1   262   2021     141     Toronto Blue Jays
#> 2         homeRuns    2   241   2021     137  San Francisco Giants
#> 3         homeRuns    3   239   2021     144        Atlanta Braves
#> 4         homeRuns    4   237   2021     119   Los Angeles Dodgers
#> 5         homeRuns    5   228   2021     142       Minnesota Twins
#> 6         homeRuns    6   222   2021     113       Cincinnati Reds
#> 7         homeRuns    6   222   2021     147      New York Yankees
#> 8         homeRuns    6   222   2021     139        Tampa Bay Rays
#> 9         homeRuns    9   221   2021     117        Houston Astros
#> 10        homeRuns   10   219   2021     111        Boston Red Sox
#> 11        homeRuns   11   210   2021     112          Chicago Cubs
#> 12        homeRuns   12   203   2021     114     Cleveland Indians
#> 13        homeRuns   13   199   2021     133     Oakland Athletics
#> 14        homeRuns   13   199   2021     136      Seattle Mariners
#> 15        homeRuns   15   198   2021     143 Philadelphia Phillies
#> 16        homeRuns   15   198   2021     138   St. Louis Cardinals
#> 17        homeRuns   17   195   2021     110     Baltimore Orioles
#> 18        homeRuns   18   194   2021     158     Milwaukee Brewers
#> 19        homeRuns   19   190   2021     145     Chicago White Sox
#> 20        homeRuns   19   190   2021     108    Los Angeles Angels
#> 21        homeRuns   21   182   2021     115      Colorado Rockies
#> 22        homeRuns   21   182   2021     120  Washington Nationals
#> 23        homeRuns   23   180   2021     135      San Diego Padres
#> 24        homeRuns   24   179   2021     116        Detroit Tigers
#> 25        homeRuns   25   176   2021     121         New York Mets
#> 26        homeRuns   26   167   2021     140         Texas Rangers
#> 27        homeRuns   27   163   2021     118    Kansas City Royals
#> 28        homeRuns   28   158   2021     146         Miami Marlins
#> 29        homeRuns   29   144   2021     109  Arizona Diamondbacks
#> 30        homeRuns   30   124   2021     134    Pittsburgh Pirates
#> 31        homeRuns    1   258   2021     110     Baltimore Orioles
#> 32        homeRuns    2   247   2021     120  Washington Nationals
#> 33        homeRuns    3   239   2021     142       Minnesota Twins
#> 34        homeRuns    4   235   2021     112          Chicago Cubs
#> 35        homeRuns    5   232   2021     109  Arizona Diamondbacks
#> 36        homeRuns    5   232   2021     140         Texas Rangers
#> 37        homeRuns    7   216   2021     114     Cleveland Indians
#> 38        homeRuns    8   213   2021     134    Pittsburgh Pirates
#> 39        homeRuns    9   209   2021     141     Toronto Blue Jays
#> 40        homeRuns   10   206   2021     113       Cincinnati Reds
#> 41        homeRuns   11   205   2021     135      San Diego Padres
#> 42        homeRuns   12   200   2021     143 Philadelphia Phillies
#> 43        homeRuns   13   199   2021     116        Detroit Tigers
#> 44        homeRuns   14   197   2021     136      Seattle Mariners
#> 45        homeRuns   15   196   2021     115      Colorado Rockies
#> 46        homeRuns   15   196   2021     147      New York Yankees
#> 47        homeRuns   17   191   2021     133     Oakland Athletics
#> 48        homeRuns   18   190   2021     121         New York Mets
#> 49        homeRuns   19   189   2021     118    Kansas City Royals
#> 50        homeRuns   20   188   2021     108    Los Angeles Angels
#> 51        homeRuns   21   187   2021     117        Houston Astros
#> 52        homeRuns   22   184   2021     139        Tampa Bay Rays
#> 53        homeRuns   23   183   2021     144        Atlanta Braves
#> 54        homeRuns   24   182   2021     145     Chicago White Sox
#> 55        homeRuns   25   176   2021     111        Boston Red Sox
#> 56        homeRuns   26   168   2021     158     Milwaukee Brewers
#> 57        homeRuns   27   162   2021     146         Miami Marlins
#> 58        homeRuns   28   161   2021     119   Los Angeles Dodgers
#> 59        homeRuns   29   152   2021     138   St. Louis Cardinals
#> 60        homeRuns   30   151   2021     137  San Francisco Giants
#> 61        homeRuns    1  2115   2021     112          Chicago Cubs
#> 62        homeRuns    2  1976   2021     120  Washington Nationals
#> 63        homeRuns    3  1281   2021     144        Atlanta Braves
#> 64        homeRuns    4  1128   2021     108    Los Angeles Angels
#> 65        homeRuns    5  1080   2021     114     Cleveland Indians
#> 66        homeRuns    6  1032   2021     110     Baltimore Orioles
#> 67        homeRuns    7   995   2021     116        Detroit Tigers
#> 68        homeRuns    8   972   2021     146         Miami Marlins
#> 69        homeRuns    9   956   2021     142       Minnesota Twins
#> 70        homeRuns   10   928   2021     109  Arizona Diamondbacks
#> 71        homeRuns   10   928   2021     140         Texas Rangers
#> 72        homeRuns   12   836   2021     141     Toronto Blue Jays
#> 73        homeRuns   13   820   2021     135      San Diego Padres
#> 74        homeRuns   14   788   2021     136      Seattle Mariners
#> 75        homeRuns   15   764   2021     133     Oakland Athletics
#> 76        homeRuns   16   760   2021     121         New York Mets
#> 77        homeRuns   17   736   2021     139        Tampa Bay Rays
#> 78        homeRuns   18   728   2021     145     Chicago White Sox
#> 79        homeRuns   19   672   2021     158     Milwaukee Brewers
#> 80        homeRuns   20   639   2021     134    Pittsburgh Pirates
#> 81        homeRuns   21   604   2021     137  San Francisco Giants
#> 82        homeRuns   22   600   2021     143 Philadelphia Phillies
#> 83        homeRuns   23   588   2021     147      New York Yankees
#> 84        homeRuns   24   567   2021     118    Kansas City Royals
#> 85        homeRuns   25   561   2021     117        Houston Astros
#> 86        homeRuns   26   528   2021     111        Boston Red Sox
#> 87        homeRuns   27   483   2021     119   Los Angeles Dodgers
#> 88        homeRuns   28   456   2021     138   St. Louis Cardinals
#> 89        homeRuns   29   412   2021     113       Cincinnati Reds
#> 90        homeRuns   30   392   2021     115      Colorado Rockies
#>            team_link stat_group total_splits game_type_id
#> 1  /api/v1/teams/141    hitting           30            R
#> 2  /api/v1/teams/137    hitting           30            R
#> 3  /api/v1/teams/144    hitting           30            R
#> 4  /api/v1/teams/119    hitting           30            R
#> 5  /api/v1/teams/142    hitting           30            R
#> 6  /api/v1/teams/113    hitting           30            R
#> 7  /api/v1/teams/147    hitting           30            R
#> 8  /api/v1/teams/139    hitting           30            R
#> 9  /api/v1/teams/117    hitting           30            R
#> 10 /api/v1/teams/111    hitting           30            R
#> 11 /api/v1/teams/112    hitting           30            R
#> 12 /api/v1/teams/114    hitting           30            R
#> 13 /api/v1/teams/133    hitting           30            R
#> 14 /api/v1/teams/136    hitting           30            R
#> 15 /api/v1/teams/143    hitting           30            R
#> 16 /api/v1/teams/138    hitting           30            R
#> 17 /api/v1/teams/110    hitting           30            R
#> 18 /api/v1/teams/158    hitting           30            R
#> 19 /api/v1/teams/145    hitting           30            R
#> 20 /api/v1/teams/108    hitting           30            R
#> 21 /api/v1/teams/115    hitting           30            R
#> 22 /api/v1/teams/120    hitting           30            R
#> 23 /api/v1/teams/135    hitting           30            R
#> 24 /api/v1/teams/116    hitting           30            R
#> 25 /api/v1/teams/121    hitting           30            R
#> 26 /api/v1/teams/140    hitting           30            R
#> 27 /api/v1/teams/118    hitting           30            R
#> 28 /api/v1/teams/146    hitting           30            R
#> 29 /api/v1/teams/109    hitting           30            R
#> 30 /api/v1/teams/134    hitting           30            R
#> 31 /api/v1/teams/110   pitching           30            R
#> 32 /api/v1/teams/120   pitching           30            R
#> 33 /api/v1/teams/142   pitching           30            R
#> 34 /api/v1/teams/112   pitching           30            R
#> 35 /api/v1/teams/109   pitching           30            R
#> 36 /api/v1/teams/140   pitching           30            R
#> 37 /api/v1/teams/114   pitching           30            R
#> 38 /api/v1/teams/134   pitching           30            R
#> 39 /api/v1/teams/141   pitching           30            R
#> 40 /api/v1/teams/113   pitching           30            R
#> 41 /api/v1/teams/135   pitching           30            R
#> 42 /api/v1/teams/143   pitching           30            R
#> 43 /api/v1/teams/116   pitching           30            R
#> 44 /api/v1/teams/136   pitching           30            R
#> 45 /api/v1/teams/115   pitching           30            R
#> 46 /api/v1/teams/147   pitching           30            R
#> 47 /api/v1/teams/133   pitching           30            R
#> 48 /api/v1/teams/121   pitching           30            R
#> 49 /api/v1/teams/118   pitching           30            R
#> 50 /api/v1/teams/108   pitching           30            R
#> 51 /api/v1/teams/117   pitching           30            R
#> 52 /api/v1/teams/139   pitching           30            R
#> 53 /api/v1/teams/144   pitching           30            R
#> 54 /api/v1/teams/145   pitching           30            R
#> 55 /api/v1/teams/111   pitching           30            R
#> 56 /api/v1/teams/158   pitching           30            R
#> 57 /api/v1/teams/146   pitching           30            R
#> 58 /api/v1/teams/119   pitching           30            R
#> 59 /api/v1/teams/138   pitching           30            R
#> 60 /api/v1/teams/137   pitching           30            R
#> 61 /api/v1/teams/112   catching           30            R
#> 62 /api/v1/teams/120   catching           30            R
#> 63 /api/v1/teams/144   catching           30            R
#> 64 /api/v1/teams/108   catching           30            R
#> 65 /api/v1/teams/114   catching           30            R
#> 66 /api/v1/teams/110   catching           30            R
#> 67 /api/v1/teams/116   catching           30            R
#> 68 /api/v1/teams/146   catching           30            R
#> 69 /api/v1/teams/142   catching           30            R
#> 70 /api/v1/teams/109   catching           30            R
#> 71 /api/v1/teams/140   catching           30            R
#> 72 /api/v1/teams/141   catching           30            R
#> 73 /api/v1/teams/135   catching           30            R
#> 74 /api/v1/teams/136   catching           30            R
#> 75 /api/v1/teams/133   catching           30            R
#> 76 /api/v1/teams/121   catching           30            R
#> 77 /api/v1/teams/139   catching           30            R
#> 78 /api/v1/teams/145   catching           30            R
#> 79 /api/v1/teams/158   catching           30            R
#> 80 /api/v1/teams/134   catching           30            R
#> 81 /api/v1/teams/137   catching           30            R
#> 82 /api/v1/teams/143   catching           30            R
#> 83 /api/v1/teams/147   catching           30            R
#> 84 /api/v1/teams/118   catching           30            R
#> 85 /api/v1/teams/117   catching           30            R
#> 86 /api/v1/teams/111   catching           30            R
#> 87 /api/v1/teams/119   catching           30            R
#> 88 /api/v1/teams/138   catching           30            R
#> 89 /api/v1/teams/113   catching           30            R
#> 90 /api/v1/teams/115   catching           30            R
#>    game_type_description
#> 1         Regular Season
#> 2         Regular Season
#> 3         Regular Season
#> 4         Regular Season
#> 5         Regular Season
#> 6         Regular Season
#> 7         Regular Season
#> 8         Regular Season
#> 9         Regular Season
#> 10        Regular Season
#> 11        Regular Season
#> 12        Regular Season
#> 13        Regular Season
#> 14        Regular Season
#> 15        Regular Season
#> 16        Regular Season
#> 17        Regular Season
#> 18        Regular Season
#> 19        Regular Season
#> 20        Regular Season
#> 21        Regular Season
#> 22        Regular Season
#> 23        Regular Season
#> 24        Regular Season
#> 25        Regular Season
#> 26        Regular Season
#> 27        Regular Season
#> 28        Regular Season
#> 29        Regular Season
#> 30        Regular Season
#> 31        Regular Season
#> 32        Regular Season
#> 33        Regular Season
#> 34        Regular Season
#> 35        Regular Season
#> 36        Regular Season
#> 37        Regular Season
#> 38        Regular Season
#> 39        Regular Season
#> 40        Regular Season
#> 41        Regular Season
#> 42        Regular Season
#> 43        Regular Season
#> 44        Regular Season
#> 45        Regular Season
#> 46        Regular Season
#> 47        Regular Season
#> 48        Regular Season
#> 49        Regular Season
#> 50        Regular Season
#> 51        Regular Season
#> 52        Regular Season
#> 53        Regular Season
#> 54        Regular Season
#> 55        Regular Season
#> 56        Regular Season
#> 57        Regular Season
#> 58        Regular Season
#> 59        Regular Season
#> 60        Regular Season
#> 61        Regular Season
#> 62        Regular Season
#> 63        Regular Season
#> 64        Regular Season
#> 65        Regular Season
#> 66        Regular Season
#> 67        Regular Season
#> 68        Regular Season
#> 69        Regular Season
#> 70        Regular Season
#> 71        Regular Season
#> 72        Regular Season
#> 73        Regular Season
#> 74        Regular Season
#> 75        Regular Season
#> 76        Regular Season
#> 77        Regular Season
#> 78        Regular Season
#> 79        Regular Season
#> 80        Regular Season
#> 81        Regular Season
#> 82        Regular Season
#> 83        Regular Season
#> 84        Regular Season
#> 85        Regular Season
#> 86        Regular Season
#> 87        Regular Season
#> 88        Regular Season
#> 89        Regular Season
#> 90        Regular Season
# }