MLB Teams Stats Leaders

MLB Teams Stats Leaders

Usage

mlb_teams_stats_leaders(
  leader_categories = NULL,
  leader_game_types = NULL,
  sit_codes = NULL,
  stat_group = NULL,
  season = NULL,
  league_id = NULL,
  sport_id = NULL,
  start_date = NULL,
  end_date = NULL,
  stat_type = NULL,
  limit = 1000
)

Arguments

leader_categories

League leader category to return information and ranking for a particular statistic.

leader_game_types

Game type to return information and ranking for a particular statistic in a particular game type.

sit_codes

Situation code to return information and ranking for a particular statistic in a particular game type.

stat_group

Stat group to return information and ranking for a particular statistic in a particular group.

season

Year to return information and ranking for a particular statistic in a given year.

league_id

League ID to return statistics for a given league. Default to "Qualified" player pool.

sport_id

The sport_id to return information and ranking information for.

start_date

Start date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY start_date must be coupled with end_date and byDateRange stat_type

end_date

End date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY end_date must be coupled with start_date and byDateRange stat_type

stat_type

The stat_type to return information and ranking for a particular statistic for a particular stat type.

limit

A limit to limit return to a particular number of records.

Value

Returns a tibble with the following columns

col_name	types

leader_category	character

rank	integer

value	character

season	character

num_teams	integer

team_id	integer

team_name	character

team_link	character

league_id	integer

league_name	character

league_link	character

person_id	integer

person_full_name	character

person_link	character

person_first_name	character

person_last_name	character

sport_id	integer

sport_link	character

sport_abbreviation	character

stat_group	character

total_splits	integer

game_type_id	character

game_type_description	character

Examples

# \donttest{
 try(mlb_teams_stats_leaders(leader_categories='homeRuns',sport_id=1, season = 2021))
#> ── MLB Teams Stats Leaders data from MLB.com ──────── baseballr 1.5.0 ──
#> ℹ Data updated: 2023-03-21 18:46:49 UTC
#> # A tibble: 90 × 11
#>    leader_c…¹  rank value season team_id team_…² team_…³ stat_…⁴ total…⁵
#>    <chr>      <int> <chr> <chr>    <int> <chr>   <chr>   <chr>     <int>
#>  1 homeRuns       1 262   2021       141 Toront… /api/v… hitting      30
#>  2 homeRuns       2 241   2021       137 San Fr… /api/v… hitting      30
#>  3 homeRuns       3 239   2021       144 Atlant… /api/v… hitting      30
#>  4 homeRuns       4 237   2021       119 Los An… /api/v… hitting      30
#>  5 homeRuns       5 228   2021       142 Minnes… /api/v… hitting      30
#>  6 homeRuns       6 222   2021       113 Cincin… /api/v… hitting      30
#>  7 homeRuns       6 222   2021       147 New Yo… /api/v… hitting      30
#>  8 homeRuns       6 222   2021       139 Tampa … /api/v… hitting      30
#>  9 homeRuns       9 221   2021       117 Housto… /api/v… hitting      30
#> 10 homeRuns      10 219   2021       111 Boston… /api/v… hitting      30
#> # … with 80 more rows, 2 more variables: game_type_id <chr>,
#> #   game_type_description <chr>, and abbreviated variable names
#> #   ¹​leader_category, ²​team_name, ³​team_link, ⁴​stat_group,
#> #   ⁵​total_splits
# }