Get Play-By-Play Data for NCAA Baseball Games

Get Play-By-Play Data for NCAA Baseball Games

Usage

ncaa_baseball_pbp(game_info_url)

Arguments

game_info_url

The url for the game's play-by-play data. This can be found using the get_ncaa_schedule_info function.

Value

A data frame with play-by-play data for an individual game. |col_name |types | |:--------------|:---------| |date |character | |location |character | |attendance |logical | |inning |character | |inning_top_bot |character | |score |character | |batting |character | |fielding |character | |description |character |

Examples

# \donttest{
  x <- ncaa_schedule_info(736, 2021)$game_info_url[2]
  ncaa_baseball_pbp(game_info_url = x)
#>           date         location attendance inning inning_top_bot score
#> 1   02/22/2021 Nashville, Tenn.         NA      1            top   0-0
#> 2   02/22/2021 Nashville, Tenn.         NA      1            top   0-0
#> 3   02/22/2021 Nashville, Tenn.         NA      1            top   0-0
#> 4   02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 5   02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 6   02/22/2021 Nashville, Tenn.         NA      1            bot   0-0
#> 7   02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 8   02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 9   02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 10  02/22/2021 Nashville, Tenn.         NA      2            top   0-0
#> 11  02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 12  02/22/2021 Nashville, Tenn.         NA      2            bot   0-0
#> 13  02/22/2021 Nashville, Tenn.         NA      2            bot   0-1
#> 14  02/22/2021 Nashville, Tenn.         NA      2            bot   0-2
#> 15  02/22/2021 Nashville, Tenn.         NA      2            bot   0-2
#> 16  02/22/2021 Nashville, Tenn.         NA      2            bot   0-2
#> 17  02/22/2021 Nashville, Tenn.         NA      2            bot   0-2
#> 18  02/22/2021 Nashville, Tenn.         NA      3            top   0-2
#> 19  02/22/2021 Nashville, Tenn.         NA      3            top   0-2
#> 20  02/22/2021 Nashville, Tenn.         NA      3            top   0-2
#> 21  02/22/2021 Nashville, Tenn.         NA      3            top   0-2
#> 22  02/22/2021 Nashville, Tenn.         NA      3            top   0-2
#> 23  02/22/2021 Nashville, Tenn.         NA      3            bot   0-2
#> 24  02/22/2021 Nashville, Tenn.         NA      3            bot   0-2
#> 25  02/22/2021 Nashville, Tenn.         NA      3            bot   0-2
#> 26  02/22/2021 Nashville, Tenn.         NA      4            top   0-2
#> 27  02/22/2021 Nashville, Tenn.         NA      4            top   0-2
#> 28  02/22/2021 Nashville, Tenn.         NA      4            top   0-2
#> 29  02/22/2021 Nashville, Tenn.         NA      4            top   0-2
#> 30  02/22/2021 Nashville, Tenn.         NA      4            top   0-3
#> 31  02/22/2021 Nashville, Tenn.         NA      4            bot   0-3
#> 32  02/22/2021 Nashville, Tenn.         NA      4            bot   0-3
#> 33  02/22/2021 Nashville, Tenn.         NA      4            bot   0-3
#> 34  02/22/2021 Nashville, Tenn.         NA      4            bot   0-3
#> 35  02/22/2021 Nashville, Tenn.         NA      4            top   0-3
#> 36  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 37  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 38  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 39  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 40  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 41  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 42  02/22/2021 Nashville, Tenn.         NA      5            top   0-3
#> 43  02/22/2021 Nashville, Tenn.         NA      5            top   0-5
#> 44  02/22/2021 Nashville, Tenn.         NA      5            bot   0-3
#> 45  02/22/2021 Nashville, Tenn.         NA      5            bot   0-3
#> 46  02/22/2021 Nashville, Tenn.         NA      5            bot   0-5
#> 47  02/22/2021 Nashville, Tenn.         NA      5            bot   0-5
#> 48  02/22/2021 Nashville, Tenn.         NA      5            bot   0-5
#> 49  02/22/2021 Nashville, Tenn.         NA      5            bot   0-5
#> 50  02/22/2021 Nashville, Tenn.         NA      5            top   0-5
#> 51  02/22/2021 Nashville, Tenn.         NA      6            top   0-5
#> 52  02/22/2021 Nashville, Tenn.         NA      6            top   0-5
#> 53  02/22/2021 Nashville, Tenn.         NA      6            top   0-5
#> 54  02/22/2021 Nashville, Tenn.         NA      6            top   0-5
#> 55  02/22/2021 Nashville, Tenn.         NA      6            top   0-5
#> 56  02/22/2021 Nashville, Tenn.         NA      6            top   0-5
#> 57  02/22/2021 Nashville, Tenn.         NA      6            top   0-7
#> 58  02/22/2021 Nashville, Tenn.         NA      6            bot   0-5
#> 59  02/22/2021 Nashville, Tenn.         NA      6            bot   0-5
#> 60  02/22/2021 Nashville, Tenn.         NA      6            bot   0-5
#> 61  02/22/2021 Nashville, Tenn.         NA      6            bot   0-5
#> 62  02/22/2021 Nashville, Tenn.         NA      6            bot   0-6
#> 63  02/22/2021 Nashville, Tenn.         NA      6            bot   0-6
#> 64  02/22/2021 Nashville, Tenn.         NA      6            bot   0-7
#> 65  02/22/2021 Nashville, Tenn.         NA      6            bot   0-7
#> 66  02/22/2021 Nashville, Tenn.         NA      6            bot   0-7
#> 67  02/22/2021 Nashville, Tenn.         NA      6            top   0-7
#> 68  02/22/2021 Nashville, Tenn.         NA      7            top   0-7
#> 69  02/22/2021 Nashville, Tenn.         NA      7            top   0-7
#> 70  02/22/2021 Nashville, Tenn.         NA      7            top   0-7
#> 71  02/22/2021 Nashville, Tenn.         NA      7            top   0-7
#> 72  02/22/2021 Nashville, Tenn.         NA      7            top   0-7
#> 73  02/22/2021 Nashville, Tenn.         NA      7            top   0-8
#> 74  02/22/2021 Nashville, Tenn.         NA      7            bot   0-7
#> 75  02/22/2021 Nashville, Tenn.         NA      7            bot   0-7
#> 76  02/22/2021 Nashville, Tenn.         NA      7            bot   0-7
#> 77  02/22/2021 Nashville, Tenn.         NA      7            bot   0-7
#> 78  02/22/2021 Nashville, Tenn.         NA      7            bot   0-8
#> 79  02/22/2021 Nashville, Tenn.         NA      7            bot   0-8
#> 80  02/22/2021 Nashville, Tenn.         NA      7            top   0-8
#> 81  02/22/2021 Nashville, Tenn.         NA      8            top   0-8
#> 82  02/22/2021 Nashville, Tenn.         NA      8            top   0-8
#> 83  02/22/2021 Nashville, Tenn.         NA      8            top   0-8
#> 84  02/22/2021 Nashville, Tenn.         NA      8            top   0-8
#> 85  02/22/2021 Nashville, Tenn.         NA      8            top   0-8
#> 86  02/22/2021 Nashville, Tenn.         NA      8            top   0-8
#> 87  02/22/2021 Nashville, Tenn.         NA      8            top   1-8
#> 88  02/22/2021 Nashville, Tenn.         NA      8            top   1-8
#> 89  02/22/2021 Nashville, Tenn.         NA      8            top   1-8
#> 90  02/22/2021 Nashville, Tenn.         NA      8            bot   1-8
#> 91  02/22/2021 Nashville, Tenn.         NA      8            bot   1-8
#> 92  02/22/2021 Nashville, Tenn.         NA      8            bot   1-8
#> 93  02/22/2021 Nashville, Tenn.         NA      8            bot   1-8
#> 94  02/22/2021 Nashville, Tenn.         NA      8            bot   1-8
#> 95  02/22/2021 Nashville, Tenn.         NA      8            bot   1-9
#> 96  02/22/2021 Nashville, Tenn.         NA      8            bot   1-9
#> 97  02/22/2021 Nashville, Tenn.         NA      8            bot   1-9
#> 98  02/22/2021 Nashville, Tenn.         NA      8            bot   1-9
#> 99  02/22/2021 Nashville, Tenn.         NA      8            bot  1-10
#> 100 02/22/2021 Nashville, Tenn.         NA      8            bot  1-12
#> 101 02/22/2021 Nashville, Tenn.         NA      8            bot  1-12
#> 102 02/22/2021 Nashville, Tenn.         NA      8            bot  1-12
#> 103 02/22/2021 Nashville, Tenn.         NA      8            bot  1-14
#> 104 02/22/2021 Nashville, Tenn.         NA      8            bot  1-14
#> 105 02/22/2021 Nashville, Tenn.         NA      8            bot  1-14
#> 106 02/22/2021 Nashville, Tenn.         NA      8            bot  1-14
#> 107 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 108 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 109 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 110 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 111 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 112 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 113 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 114 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 115 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 116 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 117 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#> 118 02/22/2021 Nashville, Tenn.         NA      9            top  1-14
#>        batting   fielding
#> 1   Wright St. Vanderbilt
#> 2   Wright St. Vanderbilt
#> 3   Wright St. Vanderbilt
#> 4   Vanderbilt Wright St.
#> 5   Vanderbilt Wright St.
#> 6   Vanderbilt Wright St.
#> 7   Wright St. Vanderbilt
#> 8   Wright St. Vanderbilt
#> 9   Wright St. Vanderbilt
#> 10  Wright St. Vanderbilt
#> 11  Vanderbilt Wright St.
#> 12  Vanderbilt Wright St.
#> 13  Vanderbilt Wright St.
#> 14  Vanderbilt Wright St.
#> 15  Vanderbilt Wright St.
#> 16  Vanderbilt Wright St.
#> 17  Vanderbilt Wright St.
#> 18  Wright St. Vanderbilt
#> 19  Wright St. Vanderbilt
#> 20  Wright St. Vanderbilt
#> 21  Wright St. Vanderbilt
#> 22  Wright St. Vanderbilt
#> 23  Vanderbilt Wright St.
#> 24  Vanderbilt Wright St.
#> 25  Vanderbilt Wright St.
#> 26  Wright St. Vanderbilt
#> 27  Wright St. Vanderbilt
#> 28  Wright St. Vanderbilt
#> 29  Wright St. Vanderbilt
#> 30  Wright St. Vanderbilt
#> 31  Vanderbilt Wright St.
#> 32  Vanderbilt Wright St.
#> 33  Vanderbilt Wright St.
#> 34  Vanderbilt Wright St.
#> 35  Wright St. Vanderbilt
#> 36  Wright St. Vanderbilt
#> 37  Wright St. Vanderbilt
#> 38  Wright St. Vanderbilt
#> 39  Wright St. Vanderbilt
#> 40  Wright St. Vanderbilt
#> 41  Wright St. Vanderbilt
#> 42  Wright St. Vanderbilt
#> 43  Wright St. Vanderbilt
#> 44  Vanderbilt Wright St.
#> 45  Vanderbilt Wright St.
#> 46  Vanderbilt Wright St.
#> 47  Vanderbilt Wright St.
#> 48  Vanderbilt Wright St.
#> 49  Vanderbilt Wright St.
#> 50  Wright St. Vanderbilt
#> 51  Wright St. Vanderbilt
#> 52  Wright St. Vanderbilt
#> 53  Wright St. Vanderbilt
#> 54  Wright St. Vanderbilt
#> 55  Wright St. Vanderbilt
#> 56  Wright St. Vanderbilt
#> 57  Wright St. Vanderbilt
#> 58  Vanderbilt Wright St.
#> 59  Vanderbilt Wright St.
#> 60  Vanderbilt Wright St.
#> 61  Vanderbilt Wright St.
#> 62  Vanderbilt Wright St.
#> 63  Vanderbilt Wright St.
#> 64  Vanderbilt Wright St.
#> 65  Vanderbilt Wright St.
#> 66  Vanderbilt Wright St.
#> 67  Wright St. Vanderbilt
#> 68  Wright St. Vanderbilt
#> 69  Wright St. Vanderbilt
#> 70  Wright St. Vanderbilt
#> 71  Wright St. Vanderbilt
#> 72  Wright St. Vanderbilt
#> 73  Wright St. Vanderbilt
#> 74  Vanderbilt Wright St.
#> 75  Vanderbilt Wright St.
#> 76  Vanderbilt Wright St.
#> 77  Vanderbilt Wright St.
#> 78  Vanderbilt Wright St.
#> 79  Vanderbilt Wright St.
#> 80  Wright St. Vanderbilt
#> 81  Wright St. Vanderbilt
#> 82  Wright St. Vanderbilt
#> 83  Wright St. Vanderbilt
#> 84  Wright St. Vanderbilt
#> 85  Wright St. Vanderbilt
#> 86  Wright St. Vanderbilt
#> 87  Wright St. Vanderbilt
#> 88  Wright St. Vanderbilt
#> 89  Wright St. Vanderbilt
#> 90  Vanderbilt Wright St.
#> 91  Vanderbilt Wright St.
#> 92  Vanderbilt Wright St.
#> 93  Vanderbilt Wright St.
#> 94  Vanderbilt Wright St.
#> 95  Vanderbilt Wright St.
#> 96  Vanderbilt Wright St.
#> 97  Vanderbilt Wright St.
#> 98  Vanderbilt Wright St.
#> 99  Vanderbilt Wright St.
#> 100 Vanderbilt Wright St.
#> 101 Vanderbilt Wright St.
#> 102 Vanderbilt Wright St.
#> 103 Vanderbilt Wright St.
#> 104 Vanderbilt Wright St.
#> 105 Vanderbilt Wright St.
#> 106 Vanderbilt Wright St.
#> 107 Wright St. Vanderbilt
#> 108 Wright St. Vanderbilt
#> 109 Wright St. Vanderbilt
#> 110 Wright St. Vanderbilt
#> 111 Wright St. Vanderbilt
#> 112 Wright St. Vanderbilt
#> 113 Wright St. Vanderbilt
#> 114 Wright St. Vanderbilt
#> 115 Wright St. Vanderbilt
#> 116 Wright St. Vanderbilt
#> 117 Wright St. Vanderbilt
#> 118 Wright St. Vanderbilt
#>                                                                                                         description
#> 1                                                                           Dues, D. struck out swinging (0-2 KKS).
#> 2                                                                      Hamilton, Q. struck out swinging (1-2 BKKS).
#> 3                                                                     Black, T. struck out swinging (3-2 KBFBFFBS).
#> 4                                                                            Davis struck out looking (1-2 BFKFFK).
#> 5                                                                                     Young lined out to p (0-1 K).
#> 6                                                                         Rodriguez struck out looking (2-2 BBFFK).
#> 7                                                                                                     Bulger to lf.
#> 8                                                                             Harris, Z. flied out to rf (2-1 BKB).
#> 9                                                                   Anglin, G. grounded out to 1b unassisted (0-0).
#> 10                                                                       Alders, A. struck out swinging (1-2 KBSS).
#> 11                                                                           Thomas struck out swinging (1-2 SBKS).
#> 12                                                                           Noland singled to left center (1-0 B).
#> 13                                                      Kolwyck doubled to right center, RBI (0-1 K) Noland scored.
#> 14                                                    Keegan doubled down the rf line, RBI (2-0 BB) Kolwyck scored.
#> 15                                           Bradfield Jr grounded out to ss (3-2 BFKBFB) Keegan advanced to third.
#> 16                                                                        Gonzalez struck out swinging (1-2 SSBFS).
#> 17                                                                                                 R: 2 H: 3 LOB: 1
#> 18                                                                     Piotto, K. struck out swinging (2-2 BSSFBS).
#> 19                                                                                 Greenwell, J walked (3-1 BBSBB).
#> 20                                                                      McConnell, J struck out swinging (0-2 KFS).
#> 21                                                                          Dues, D. grounded out to 2b (2-2 BKBF).
#> 22                                                                                                           LOB: 1
#> 23                                                                                    Bulger lined out to lf (0-0).
#> 24                                                                                  Young popped up to cf (0-2 FK).
#> 25                                                                              Rodriguez popped up to 1b (1-1 BK).
#> 26                                                                    Hamilton, Q. struck out swinging (2-2 BKBFS).
#> 27                                                                     Black, T. doubled to right center (1-2 FBF).
#> 28                                                                       Harris, Z. struck out swinging (1-2 SSBS).
#> 29                                                                             Anglin, G. flied out to cf (1-1 BK).
#> 30                                                                                                      H: 1 LOB: 1
#> 31                                                                     Thomas homered to left field, RBI (2-1 BBK).
#> 32                                                                                      Noland singled, bunt (0-0).
#> 33                                 Kolwyck grounded into double play ss to 2b to 1b (1-0 B) Noland out on the play.
#> 34                                                                                 Keegan popped up to 2b (1-1 FB).
#> 35                                                                                                        R: 1 H: 2
#> 36                                                                                    Alders, A. walked (3-0 BBBB).
#> 37                                                      Piotto, K. walked (3-0 BBBB) Alders, A. advanced to second.
#> 38                                                                                         Hliboki to p for Rocker.
#> 39  Greenwell, J singled through the right side (0-1 K) Piotto, K. advanced to second Alders, A. advanced to third.
#> 40                                                                          McConnell, J infield fly to c (0-2 SS).
#> 41                                                                         Dues, D. struck out looking (1-2 BKSFK).
#> 42                                                                            Hamilton, Q. fouled out to c (0-1 K).
#> 43                                                                                                      H: 1 LOB: 3
#> 44                                                              Bradfield Jr singled to third base, bunt (2-1 BBK).
#> 45                                                                                       Bradfield Jr stole second.
#> 46                                           Gonzalez homered to center field, 2 RBI (1-2 FBF) Bradfield Jr scored.
#> 47                                                                          Bulger struck out swinging (2-2 FKBBS).
#> 48                                                                           Young struck out swinging (2-2 BKBKS).
#> 49                                                                       Rodriguez struck out looking (3-2 KBFBBK).
#> 50                                                                                                        R: 2 H: 2
#> 51                                                                                    Black, T. walked (3-1 SBBBB).
#> 52                                                                             Harris, Z. flied out to cf (0-2 SS).
#> 53                                                                       Anglin, G. struck out swinging (1-2 SSBS).
#> 54                       Alders, A. reached on an error by 1b, assist by ss (2-1 KBB) Black, T. advanced to second.
#> 55                    Piotto, K. hit by pitch (2-2 BBFF) Alders, A. advanced to second Black, T. advanced to third.
#> 56                                                                         Greenwell, J lined out to 2b (2-2 KBFB).
#> 57                                                                                                      E: 1 LOB: 3
#> 58                                                                                 Hornsmeier to p for Schrand, J..
#> 59                                                                         Thomas tripled to right field (1-2 FBS).
#> 60                                                                                     Noland walked (3-2 KBBBKFB).
#> 61                                                Kolwyck grounded out to 3b (3-2 KBBBK) Noland advanced to second.
#> 62                                                  Keegan reached on a fielders choice, RBI (1-0 B) Thomas scored.
#> 63                             Bradfield Jr walked (3-2 BKBFBB) Keegan advanced to second Noland advanced to third.
#> 64     Gonzalez hit by pitch, RBI (1-2 BFF) Bradfield Jr advanced to second Keegan advanced to third Noland scored.
#> 65                                                                                   Shirk, J. to p for Hornsmeier.
#> 66                                  Bulger grounded into double play 2b to 1b (3-2 KBBKB) Gonzalez out on the play.
#> 67                                                                                                 R: 2 H: 1 LOB: 2
#> 68                                                                                       McElvain to p for Hliboki.
#> 69                                                                   McConnell, J struck out swinging (3-2 BBBKFS).
#> 70                                                                                Dues, D. flied out to 2b (0-1 S).
#> 71                                                             Hamilton, Q. reached on a fielding error by p (0-0).
#> 72                                                                             Black, T. flied out to 3b (2-1 BBK).
#> 73                                                                                                      E: 1 LOB: 1
#> 74                                                                         Young singled to center field (2-1 KBB).
#> 75                                                                        Young advanced to second on a wild pitch.
#> 76                                                                       Rodriguez struck out swinging (2-2 BBKFS).
#> 77                                                                           Thomas struck out swinging (1-2 BFFS).
#> 78                                                       Noland doubled to left field, RBI (1-2 FFFB) Young scored.
#> 79                                                                           Kolwyck grounded out to 2b (2-2 KBBF).
#> 80                                                                                                 R: 1 H: 2 LOB: 1
#> 81                                                                                         Hewett to lf for Bulger.
#> 82                                                                           Harris, Z. grounded out to ss (0-1 K).
#> 83                                                                            Luikart, J. pinch hit for Anglin, G..
#> 84                                                                         Luikart, J. singled to left field (0-0).
#> 85                                                                    Alders, A. struck out swinging (3-2 BKFBBFS).
#> 86                                                                               Sass, S. pinch hit for Piotto, K..
#> 87                                                Sass, S. doubled to left center, RBI (1-1 KB) Luikart, J. scored.
#> 88                                                                         Castleman, L pinch hit for Greenwell, J.
#> 89                                                                   Castleman, L struck out swinging (3-2 BBBKFS).
#> 90                                                                                               Luikart, J. to 1b.
#> 91                                                                                                   Sass, S. to c.
#> 92                                                                                Sayre, A. to rf for Castleman, L.
#> 93                                                                                 Tabeling, L. to p for Shirk, J..
#> 94                                                                   Keegan tripled to right field (3-2 KSFBFBFFB).
#> 95                                          Bradfield Jr singled through the right side, RBI (1-0 B) Keegan scored.
#> 96                                          Gonzalez grounded out to 2b (2-2 FKBB) Bradfield Jr advanced to second.
#> 97                                                                                   Hewett hit by pitch (2-1 BFB).
#> 98                               Young walked (3-1 BBFBB) Hewett advanced to second Bradfield Jr advanced to third.
#> 99       Rodriguez hit by pitch, RBI (0-1 K) Young advanced to second Hewett advanced to third Bradfield Jr scored.
#> 100            Thomas singled to center field, 2 RBI (0-0) Rodriguez advanced to second Young scored Hewett scored.
#> 101                                                                                     ODowd pinch hit for Noland.
#> 102                                                                            ODowd struck out looking (1-2 BFFK).
#> 103                                  Kolwyck doubled to right center, 2 RBI (1-0 B) Thomas scored Rodriguez scored.
#> 104                                                                               Haught, T. to p for Tabeling, L..
#> 105                                                                                     Casas pinch hit for Keegan.
#> 106                                                                                  Casas grounded out to p (0-0).
#> 107                                                                                                   Hewett to cf.
#> 108                                                                                                    Casas to 1b.
#> 109                                                                                  LaNeve to lf for Bradfield Jr.
#> 110                                                                        Rountree, C. pinch hit for McConnell, J.
#> 111                                                                     Rountree, C. struck out looking (0-2 FKFK).
#> 112                                                                                Neff, A. pinch hit for Dues, D..
#> 113                                                                           Neff, A. singled to left field (0-0).
#> 114                                                                         Merlano, D. pinch hit for Hamilton, Q..
#> 115                                                                    Merlano, D. struck out looking (2-2 KKBBFK).
#> 116                                                   Black, T. walked (3-2 BKBFFFBFB) Neff, A. advanced to second.
#> 117                                                                            Harris, Z. flied out to cf (0-2 FF).
#> 118                                                                                                     H: 1 LOB: 2
# }