Scrape NCAA baseball data (Division I, II, and III)

This function allows the user to obtain batting or pitching statistics for any school affiliated with the NCAA at the division I, II, or III levels. The function acquires data from the NCAA's website (stats.ncaa.org) and returns a data frame.

Usage

ncaa_scrape(teamid, year, type = "batting")

Arguments

teamid

The numerical ID that the NCAA website uses to identify a team

year

The season for which data should be returned, in the form of "YYYY". Years currently available: 2013-2017.

type

A string indicating whether to return "batting" or "pitching" statistics

Value

A data frame with the following variables |col_name |types | |:-------------|:---------| |year |integer | |school |character | |conference |character | |division |numeric | |Jersey |character | |Player |character | |Yr |character | |Pos |character | |GP |numeric | |GS |numeric | |BA |numeric | |OBPct |numeric | |SlgPct |numeric | |R |numeric | |AB |numeric | |H |numeric | |2B |numeric | |3B |numeric | |TB |numeric | |HR |numeric | |RBI |numeric | |BB |numeric | |HBP |numeric | |SF |numeric | |SH |numeric | |K |numeric | |DP |numeric | |CS |numeric | |Picked |numeric | |SB |numeric | |RBI2out |numeric | |teamid |numeric | |conference_id |integer | |player_id |integer | |player_url |character |

Examples

# \donttest{
  ncaa_scrape(teamid=255, year=2013, type = "batting")
#>    year       school conference division Jersey              Player Yr
#> 1  2013 Georgia Tech        ACC        1      9        Murray, A.J. So
#> 2  2013 Georgia Tech        ACC        1     10         Evans, Zane Jr
#> 3  2013 Georgia Tech        ACC        1      8          Wren, Kyle Jr
#> 4  2013 Georgia Tech        ACC        1      3          Hyde, Mott Jr
#> 5  2013 Georgia Tech        ACC        1     32       Palka, Daniel Jr
#> 6  2013 Georgia Tech        ACC        1     31           Dove, Sam Sr
#> 7  2013 Georgia Tech        ACC        1     14      Gonzalez, Matt Fr
#> 8  2013 Georgia Tech        ACC        1      6     Thomas, Brandon Sr
#> 9  2013 Georgia Tech        ACC        1     35    Spingola, Daniel So
#> 10 2013 Georgia Tech        ACC        1     13       Smith, Thomas So
#> 11 2013 Georgia Tech        ACC        1     29      Earnest, Mitch Jr
#> 12 2013 Georgia Tech        ACC        1      1         Dore, Dylan Fr
#> 13 2013 Georgia Tech        ACC        1     20        Miller, Cole Fr
#> 14 2013 Georgia Tech        ACC        1      2       Wruble, Grant Fr
#> 15 2013 Georgia Tech        ACC        1      4 Hodorowski, Michael So
#> 16 2013 Georgia Tech        ACC        1     16       Lynch, Connor So
#> 17 2013 Georgia Tech        ACC        1      5  Sheffield, Charles So
#> 18 2013 Georgia Tech        ACC        1     23         Young, Will Fr
#> 19 2013 Georgia Tech        ACC        1     12   Fineberg, Gregory Fr
#> 20 2013 Georgia Tech        ACC        1      -              Totals  -
#> 21 2013 Georgia Tech        ACC        1      -     Opponent Totals  -
#>    Pos GP GS    BA OBPct SlgPct   R  AB   H  2B 3B TB HR RBI  BB HBP SF
#> 1    C 64 64 0.271 0.366  0.394  41 236  64  10  2 NA  5  35  30   6  1
#> 2    C 64 64 0.361 0.430  0.590  47 244  88  12  1 NA 14  66  33   1  6
#> 3   OF 64 64 0.360 0.423  0.467  50 272  98  11  6 NA  2  28  29   2  2
#> 4  INF 64 64 0.260 0.357  0.430  32 223  58  18  1 NA  6  32  26   8  1
#> 5   1B 62 62 0.342 0.436  0.637  55 237  81  13  3 NA 17  66  31  10  2
#> 6  INF 61 61 0.288 0.371  0.412  47 233  67  14  3 NA  3  34  30   4  5
#> 7  INF 61 58 0.291 0.324  0.388  37 227  66   9  2 NA  3  37  11   1  2
#> 8   OF 53 53 0.321 0.406  0.445  49 209  67  15  4 NA  1  30  30   2  3
#> 9   OF 50 21 0.256 0.375  0.321  12  78  20   2 NA NA  1  13  14   2  2
#> 10 INF 45 30 0.286 0.362  0.446  16 112  32   5  2 NA  3  13  13   1  1
#> 11   C 32 17 0.306 0.368  0.516  13  62  19   4 NA NA  3  15   3   3 NA
#> 12  OF 27  8 0.289 0.386  0.368   8  38  11   3 NA NA NA   4   3   3 NA
#> 13  1B 18  1 0.053 0.095  0.053   2  19   1  NA NA NA NA   3   1  NA  1
#> 14   C 11  0 0.250 0.333  0.250   1   4   1  NA NA NA NA   2   1  NA  1
#> 15  OF 11  2 0.278 0.300  0.278  NA  18   5  NA NA NA NA   3   1  NA  1
#> 16   C 10  7 0.192 0.179  0.231   3  26   5   1 NA NA NA   3  NA  NA  2
#> 17  OF  3  0 0.000 0.500  0.000   1   2  NA  NA NA NA NA  NA   1   1 NA
#> 18 INF  2  0 0.000 0.000  0.000  NA   2  NA  NA NA NA NA   1  NA  NA NA
#> 19   C  2  0 0.000 0.000  0.000  NA   2  NA  NA NA NA NA  NA  NA  NA NA
#> 20   - NA NA 0.304 0.382  0.455 414  NA 683 117 24 NA 58 385 257  44 30
#> 21   - NA NA 0.264 0.353  0.368 317  NA 561  81 11 NA 40 296 258  51 30
#>    SH   K DP CS Picked SB RBI2out teamid conference_id player_id
#> 1   2  56  8 NA     NA  4      NA    255           821   1323370
#> 2  NA  40  8  1     NA NA      NA    255           821   1211076
#> 3   2  36  2 13     NA 28      NA    255           821   1211075
#> 4  11  46  1  5     NA  6      NA    255           821   1211073
#> 5  NA  60  8 NA     NA  6      NA    255           821   1211087
#> 6   5  44  5  3     NA  7      NA    255           821   1106044
#> 7   1  47  8  4     NA 10      NA    255           821   1423138
#> 8   3  40  3  2     NA  9      NA    255           821   1106039
#> 9  NA  13  1  2     NA  3      NA    255           821   1323375
#> 10  5  16  4  3     NA  3      NA    255           821   1323374
#> 11  2  19  3 NA     NA  1      NA    255           821   1211086
#> 12  1  10 NA NA     NA NA      NA    255           821   1423134
#> 13 NA  11  2 NA     NA NA      NA    255           821   1423140
#> 14 NA   1 NA NA     NA NA      NA    255           821   1423135
#> 15 NA   2  1 NA     NA NA      NA    255           821   1323368
#> 16 NA   4 NA NA     NA NA      NA    255           821   1323369
#> 17 NA   2 NA NA     NA NA      NA    255           821   1323373
#> 18 NA  NA NA NA     NA NA      NA    255           821   1423141
#> 19 NA   2 NA NA     NA NA      NA    255           821   1423137
#> 20 32 449 54 33     NA 77      NA    255           821        NA
#> 21 51 447 59 29     NA 79      NA    255           821        NA
#>                                                                                             player_url
#> 1  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1323370
#> 2  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1211076
#> 3  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1211075
#> 4  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1211073
#> 5  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1211087
#> 6  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1106044
#> 7  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1423138
#> 8  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1106039
#> 9  http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1323375
#> 10 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1323374
#> 11 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1211086
#> 12 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1423134
#> 13 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1423140
#> 14 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1423135
#> 15 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1323368
#> 16 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1323369
#> 17 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1323373
#> 18 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1423141
#> 19 http://stats.ncaa.org/player/index?game_sport_year_ctl_id=11320&org_id=255&stats_player_seq=1423137
#> 20                                                                                                <NA>
#> 21                                                                                                <NA>
# }