Query Baseball Savant Leaderboards

This function allows you to read leaderboard data from BaseballSavant directly into R as data frame.

Usage

statcast_leaderboards(
  leaderboard = "exit_velocity_barrels",
  year = 2018,
  abs = 50,
  min_pa = "q",
  min_pitches = 100,
  min_field = "q",
  min_run = 0,
  player_type = "batter",
  fielding_type = "player",
  oaa_position = "",
  oaa_roles = "",
  team = "",
  arsenal_type = "n_",
  run_type = "raw",
  min2b = 5,
  min3b = 0,
  position = "",
  bats = "",
  hand = ""
)

Arguments

leaderboard

The type of leaderboard to retrieve, input as a string. Current options include exit_velocity_barrels, expected_statistics, pitch_arsenal, outs_above_average, directional_oaa, catch_probability, pop_time, sprint_speed, and running_splits_90_ft.

year

The season for which you want data.

abs

The minimum number of batted balls. Applies only to exit_velocity_barrels leaderboards.

min_pa

Minimum number of plate appearances. Can be a number or 'q' for qualified batters.

min_pitches

Minimum number of pitches thrown.

min_field

Minimum number of fieding opportunities.

min_run

Minimum number of running opportunities.

player_type

One of either 'batter' or pitcher. For the expected_statistics leaderboard, 'batter-team' and 'pitcher-team' are also available.

fielding_type

One of either 'player' or 'team'.

oaa_position

Can be either the number position of a player or 'if' or 'of' for position categories.

oaa_roles

Can be either the number position of a player or 'if' or 'of' for position categories.

team

An abbreviation for a team. Can be left blank.

arsenal_type

One of either 'n_', 'avg_spin', or 'avg_speed'.

run_type

One of either 'percent' or 'raw'.

min2b

The minimum number of throwing attempts to second base.

min3b

The minimum number of throwing attempts to third base.

position

The numeric position of the player. For DH use 10. Can be left blank.

bats

The handedness of the batter. One of 'R' or 'L'. Can be left blank.

hand

The handedness of the pitcher. One of 'R' or 'L'. Can be left blank.

Value

Returns a tibble of Statcast leaderboard data with the following columns (for leaderboard: 'exit_velocity_barrels'):

col_name	types

year	numeric

last_name	character

first_name	character

player_id	integer

attempts	integer

avg_hit_angle	numeric

anglesweetspotpercent	numeric

max_hit_speed	numeric

avg_hit_speed	numeric

fbld	numeric

gb	numeric

max_distance	integer

avg_distance	integer

avg_hr_distance	integer

ev95plus	integer

ev95per-swing	numeric

ev95percent	numeric

barrels	integer

brl_percent	numeric

brl_pa	numeric

Details

oaa_roles argument: 30 = 1B - Straight Up 31 = 1B - Towards 1B/2B Hole 32 = 1B - Close to Line 40 = 2B - Straight Up 41 = 2B - Shaded Towards 2B Bag 42 = 2B - Towards 1B/2B Hole 43 = 2B - Behind First Basemen 46 = 2B - Up the Middle 60 = SS - Straight Up 61 = SS - Towards 3B/SS Hole 62 = SS - Shaded Towards 2B Bag 64 = SS - Up the Middle 50 = 3B - Straight Up 51 = 3B - Close to Line 52 = 3B - Towards 3B/SS Hole 77 = LF - Close to Line 71 = LF - Leaning Left 70 = LF - Straight Up 72 = LF - Leaning Right 78 = LF - LF Gap 87 = CF - LF Gap 81 = CF - Leaning Left 82 = CF - Leaning Right 89 = CF - RF Gap 98 = RF - RF Gap 91 = RF - Leaning Left 90 = RF - Straight Up 92 = RF - Leaning Right 99 = RF - Close to Line

Examples

# \donttest{
  try(statcast_leaderboards(leaderboard = "exit_velocity_barrels", year = 2018))
#> ── MLB Baseball Savant Statcast Leaderboards data from baseballsavant.ml
#> ℹ Data updated: 2022-10-05 17:37:51 UTC
#> # A tibble: 249 × 19
#>     year last_…¹ first…² playe…³ attem…⁴ avg_h…⁵ angle…⁶ max_h…⁷ avg_h…⁸
#>    <dbl> <chr>   <chr>     <int>   <int>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1  2018 Peraza  José     606299     570    15      37.7    109.    85.8
#>  2  2018 Lindor  Franci…  596019     560    14.7    34.3    114.    90.6
#>  3  2018 Markak… Nick     455976     552    10.1    38.4    109.    90.6
#>  4  2018 Turner  Trea     607208     534     9.3    28.8    112.    89.1
#>  5  2018 Machado Manny    592518     533    14.4    33.8    115.    91.6
#>  6  2018 Segura  Jean     516416     527     7.8    29.8    112.    87.7
#>  7  2018 Albies  Ozzie    645277     527    16.1    34      110.    86.8
#>  8  2018 Merrif… Whit     593160     526    16.9    40.7    106.    87  
#>  9  2018 Inciar… Ender    542255     519    10.9    32.6    105     84.9
#> 10  2018 Brantl… Michael  488726     517     9.5    38.1    109.    90.2
#> # … with 239 more rows, 10 more variables: fbld <dbl>, gb <dbl>,
#> #   max_distance <int>, avg_distance <int>, avg_hr_distance <int>,
#> #   ev95plus <int>, ev95percent <dbl>, barrels <int>,
#> #   brl_percent <dbl>, brl_pa <dbl>, and abbreviated variable names
#> #   ¹​last_name, ²​first_name, ³​player_id, ⁴​attempts, ⁵​avg_hit_angle,
#> #   ⁶​anglesweetspotpercent, ⁷​max_hit_speed, ⁸​avg_hit_speed
# }